The AddExpression class.

This class represents the sum of any number of expressions.

# S4 method for Expression,missing
+(e1, e2)

# S4 method for Expression,Expression
+(e1, e2)

# S4 method for Expression,ConstVal
+(e1, e2)

# S4 method for ConstVal,Expression
+(e1, e2)

# S4 method for AddExpression
dim_from_args(object)

# S4 method for AddExpression
name(x)

# S4 method for AddExpression
to_numeric(object, values)

# S4 method for AddExpression
is_atom_log_log_convex(object)

# S4 method for AddExpression
is_atom_log_log_concave(object)

# S4 method for AddExpression
is_symmetric(object)

# S4 method for AddExpression
is_hermitian(object)

# S4 method for AddExpression
copy(object, args = NULL, id_objects = list())

# S4 method for AddExpression
graph_implementation(object, arg_objs, dim, data = NA_real_)

Arguments

e1, e2

The Expression objects or numeric constants to add.

x, object

An AddExpression object.

values

A list of arguments to the atom.

args

An optional list of arguments to reconstruct the atom. Default is to use current args of the atom.

id_objects

Currently unused.

arg_objs

A list of linear expressions for each argument.

dim

A vector representing the dimensions of the resulting expression.

data

A list of additional data required by the atom.

Methods (by generic)

dim_from_args(AddExpression): The dimensions of the expression.

name(AddExpression): The string form of the expression.

to_numeric(AddExpression): Sum all the values.

is_atom_log_log_convex(AddExpression): Is the atom log-log convex?

is_atom_log_log_concave(AddExpression): Is the atom log-log convex?

is_symmetric(AddExpression): Is the atom symmetric?

is_hermitian(AddExpression): Is the atom hermitian?

copy(AddExpression): Returns a shallow copy of the AddExpression atom

graph_implementation(AddExpression): The graph implementation of the expression.

Slots

arg_groups

A list of Expressions and numeric data.frame, matrix, or vector objects.