An interface for the CPLEX solver.

An interface for the CPLEX solver.

CPLEX_QP()

# S4 method for CPLEX_QP
mip_capable(solver)

# S4 method for CPLEX_QP
status_map(solver, status)

# S4 method for CPLEX_QP
name(x)

# S4 method for CPLEX_QP
import_solver(solver)

# S4 method for CPLEX_QP,list,InverseData
invert(object, solution, inverse_data)

# S4 method for CPLEX_QP
solve_via_data(
  object,
  data,
  warm_start,
  verbose,
  feastol,
  reltol,
  abstol,
  num_iter,
  solver_opts,
  solver_cache
)

Arguments

status
A status code returned by the solver.

x, object, solver
A CPLEX_QP object.

solution
The raw solution returned by the solver.

inverse_data
A InverseData object containing data necessary for the inversion.

data
Data generated via an apply call.

warm_start
A boolean of whether to warm start the solver.

verbose
A boolean of whether to enable solver verbosity.

feastol
The feasible tolerance on the primal and dual residual.

reltol
The relative tolerance on the duality gap.

abstol
The absolute tolerance on the duality gap.

num_iter
The maximum number of iterations.

solver_opts
A list of Solver specific options

solver_cache
Cache for the solver.

Methods (by generic)

mip_capable: Can the solver handle mixed-integer programs?

status_map: Converts status returned by the CPLEX solver to its respective CVXPY status.

name: Returns the name of the solver.

import_solver: Imports the solver.

invert: Returns the solution to the original problem given the inverse_data.

solve_via_data: Solve a problem represented by data returned from apply.