The ConicSolver class.

Conic solver class with reduction semantics.

# S4 method for ConicSolver,Problem
accepts(object, problem)

# S4 method for ConicSolver
reduction_format_constr(object, problem, constr, exp_cone_order)

# S4 method for ConicSolver
group_coeff_offset(object, problem, constraints, exp_cone_order)

# S4 method for ConicSolver,Solution,InverseData
invert(object, solution, inverse_data)

Arguments

object
A ConicSolver object.

problem
A Problem object.

constr
A Constraint to format.

exp_cone_order
A list indicating how the exponential cone arguments are ordered.

constraints
A list of Constraint objects.

solution
A Solution object to invert.

inverse_data
A InverseData object containing data necessary for the inversion.

Methods (by generic)

accepts: Can the problem be solved with a conic solver?

reduction_format_constr: Return a list representing a cone program whose problem data tensors will yield the coefficient "A" and offset "b" for the respective constraints: Linear Equations: A Linear inequalities: A Second order cone: A Exponential cone: A Semidefinite cone: A

group_coeff_offset: Combine the constraints into a single matrix, offset.

invert: Returns the solution to the original problem given the inverse_data.