The DivExpression class.

This class represents one expression divided by another expression.

# S4 method for class 'Expression,Expression'
e1/e2

# S4 method for class 'Expression,ConstVal'
e1/e2

# S4 method for class 'ConstVal,Expression'
e1/e2

# S4 method for class 'DivExpression'
to_numeric(object, values)

# S4 method for class 'DivExpression'
is_quadratic(object)

# S4 method for class 'DivExpression'
is_qpwa(object)

# S4 method for class 'DivExpression'
dim_from_args(object)

# S4 method for class 'DivExpression'
is_atom_convex(object)

# S4 method for class 'DivExpression'
is_atom_concave(object)

# S4 method for class 'DivExpression'
is_atom_log_log_convex(object)

# S4 method for class 'DivExpression'
is_atom_log_log_concave(object)

# S4 method for class 'DivExpression'
is_incr(object, idx)

# S4 method for class 'DivExpression'
is_decr(object, idx)

# S4 method for class 'DivExpression'
graph_implementation(object, arg_objs, dim, data = NA_real_)

Arguments

e1, e2

The Expression objects or numeric constants to divide. The denominator, e2, must be a scalar constant.

object

A DivExpression object.

values

A list of arguments to the atom.

idx

An index into the atom.

arg_objs

A list of linear expressions for each argument.

dim

A vector representing the dimensions of the resulting expression.

data

A list of additional data required by the atom.

Methods (by generic)

to_numeric(DivExpression): Matrix division by a scalar.

is_quadratic(DivExpression): Is the left-hand expression quadratic and the right-hand expression constant?

is_qpwa(DivExpression): Is the expression quadratic of piecewise affine?

dim_from_args(DivExpression): The (row, col) dimensions of the left-hand expression.

is_atom_convex(DivExpression): Division is convex (affine) in its arguments only if the denominator is constant.

is_atom_concave(DivExpression): Division is concave (affine) in its arguments only if the denominator is constant.

is_atom_log_log_convex(DivExpression): Is the atom log-log convex?

is_atom_log_log_concave(DivExpression): Is the atom log-log concave?

is_incr(DivExpression): Is the right-hand expression positive?

is_decr(DivExpression): Is the right-hand expression negative?

graph_implementation(DivExpression): The graph implementation of the expression.