The ECOS class.

This class is an interface for the ECOS solver.

ECOS()

# S4 method for ECOS
lp_capable(solver)

# S4 method for ECOS
socp_capable(solver)

# S4 method for ECOS
sdp_capable(solver)

# S4 method for ECOS
exp_capable(solver)

# S4 method for ECOS
mip_capable(solver)

# S4 method for ECOS
name(object)

# S4 method for ECOS
import_solver(solver)

# S4 method for ECOS
Solver.solve(
  solver,
  objective,
  constraints,
  cached_data,
  warm_start,
  verbose,
  ...
)

# S4 method for ECOS
format_results(solver, results_dict, data, cached_data)

Arguments

object, solver
An ECOS object.

objective
A list representing the canonicalized objective.

constraints
A list of canonicalized constraints.

cached_data
A list mapping solver name to cached problem data.

warm_start
A logical value indicating whether the previous solver result should be used to warm start.

verbose
A logical value indicating whether to print solver output.

...
Additional arguments to the solver.

results_dict
A list containing the solver output.

data
A list containing information about the problem.

Methods (by generic)

lp_capable: ECOS can handle linear programs.

socp_capable: ECOS can handle second-order cone programs.

sdp_capable: ECOS cannot handle semidefinite programs.

exp_capable: ECOS can handle exponential cone programs.

mip_capable: ECOS cannot handle mixed-integer programs.

name: The name of the solver.

import_solver: Imports the ECOSolveR library.

Solver.solve: Call the solver on the canonicalized problem.

format_results: Convert raw solver output into standard list of results.

References

A. Domahidi, E. Chu, and S. Boyd. "ECOS: An SOCP solver for Embedded Systems." Proceedings of the European Control Conference, pp. 3071-3076, 2013.

See also

Examples

ecos <- ECOS()
lp_capable(ecos)

#> [1] TRUE

sdp_capable(ecos)

#> [1] FALSE

socp_capable(ecos)

#> [1] TRUE

exp_capable(ecos)

#> [1] TRUE

mip_capable(ecos)

#> [1] FALSE