An interface for the GLPK MI solver.

An interface for the GLPK MI solver.

GLPK_MI()

# S4 method for class 'GLPK_MI'
mip_capable(solver)

# S4 method for class 'GLPK_MI'
status_map(solver, status)

# S4 method for class 'GLPK_MI'
name(x)

# S4 method for class 'GLPK_MI'
solve_via_data(
  object,
  data,
  warm_start,
  verbose,
  feastol,
  reltol,
  abstol,
  num_iter,
  solver_opts,
  solver_cache
)

Arguments

solver, object, x

A GLPK_MI object.

status

A status code returned by the solver.

data

Data generated via an apply call.

warm_start

A boolean of whether to warm start the solver.

verbose

A boolean of whether to enable solver verbosity.

feastol

The feasible tolerance.

reltol

The relative tolerance.

abstol

The absolute tolerance.

num_iter

The maximum number of iterations.

solver_opts

A list of Solver specific options

solver_cache

Cache for the solver.

Methods (by generic)

mip_capable(GLPK_MI): Can the solver handle mixed-integer programs?

status_map(GLPK_MI): Converts status returned by the GLPK_MI solver to its respective CVXPY status.

name(GLPK_MI): Returns the name of the solver.

solve_via_data(GLPK_MI): Solve a problem represented by data returned from apply.