The Index class.

This class represents indexing or slicing into a matrix.

# S4 method for Expression,missing,missing,ANY
[(x, i, j, ..., drop = TRUE)

# S4 method for Expression,numeric,missing,ANY
[(x, i, j, ..., drop = TRUE)

# S4 method for Expression,missing,numeric,ANY
[(x, i, j, ..., drop = TRUE)

# S4 method for Expression,numeric,numeric,ANY
[(x, i, j, ..., drop = TRUE)

Index(expr, key)

# S4 method for Index
to_numeric(object, values)

# S4 method for Index
dim_from_args(object)

# S4 method for Index
is_atom_log_log_convex(object)

# S4 method for Index
is_atom_log_log_concave(object)

# S4 method for Index
get_data(object)

# S4 method for Index
graph_implementation(object, arg_objs, dim, data = NA_real_)

# S4 method for SpecialIndex
to_numeric(object, values)

# S4 method for SpecialIndex
dim_from_args(object)

Arguments

x

A Expression object.

i, j

The row and column indices of the slice.

...

(Unimplemented) Optional arguments.

drop

(Unimplemented) A logical value indicating whether the result should be coerced to the lowest possible dimension.

expr

An Expression representing a vector or matrix.

key

A list containing the start index, end index, and step size of the slice.

object

An Index object.

values

A list of arguments to the atom.

arg_objs

A list of linear expressions for each argument.

dim

A vector representing the dimensions of the resulting expression.

data

A list of additional data required by the atom.

Methods (by generic)

to_numeric(Index): The index/slice into the given value.

dim_from_args(Index): The dimensions of the atom.

is_atom_log_log_convex(Index): Is the atom log-log convex?

is_atom_log_log_concave(Index): Is the atom log-log concave?

get_data(Index): A list containing key.

graph_implementation(Index): The graph implementation of the atom.

to_numeric(SpecialIndex): The index/slice into the given value.

dim_from_args(SpecialIndex): The dimensions of the atom.

Slots

expr

An Expression representing a vector or matrix.

key

A list containing the start index, end index, and step size of the slice.