The Int class.

This class represents an integer variable.

Int(rows = 1, cols = 1, name = NA_character_)

# S4 method for Int
as.character(x)

# S4 method for Int
canonicalize(object)

Arguments

rows
The number of rows in the variable.

cols
The number of columns in the variable.

name
(Optional) A character string representing the name of the variable.

x, object
An Int object.

Methods (by generic)

canonicalize: Enforce that the variable be an integer.

Slots

id

(Internal) A unique identification number used internally.

rows

The number of rows in the variable.

cols

The number of columns in the variable.

name

(Optional) A character string representing the name of the variable.

primal_value

(Internal) The primal value of the variable stored internally.

Examples

x <- Int(3, name = "i") ## 3-int variable
y <- Int(3, 3, name = "j") # Matrix variable
as.character(y)

#> [1] "Int(3, 3)"

id(y)

#> [1] 9

is_positive(x)

#> [1] FALSE

is_negative(x)

#> [1] FALSE

size(y)

#> [1] 3 3

name(y)

#> [1] "j"

value(y) <- matrix(1:9, nrow = 3)
value(y)

#>      [,1] [,2] [,3]
#> [1,]    1    4    7
#> [2,]    2    5    8
#> [3,]    3    6    9

grad(y)

#> $`9`
#> 9 x 9 sparse Matrix of class "dgCMatrix"
#>                        
#>  [1,] 1 . . . . . . . .
#>  [2,] . 1 . . . . . . .
#>  [3,] . . 1 . . . . . .
#>  [4,] . . . 1 . . . . .
#>  [5,] . . . . 1 . . . .
#>  [6,] . . . . . 1 . . .
#>  [7,] . . . . . . 1 . .
#>  [8,] . . . . . . . 1 .
#>  [9,] . . . . . . . . 1
#> 

variables(y)

#> [[1]]
#> Int(3, 3)

canonicalize(y)

#> [[1]]
#> [[1]]$type
#> [1] "variable"
#> 
#> [[1]]$size
#> [1] 3 3
#> 
#> [[1]]$args
#> list()
#> 
#> [[1]]$data
#> [1] 9
#> 
#> [[1]]$class
#> [1] "LinOp"
#> 
#> 
#> [[2]]
#> [[2]][[1]]
#> An object of class "IntConstr"
#> Slot "lin_op":
#> $type
#> [1] "variable"
#> 
#> $size
#> [1] 3 3
#> 
#> $args
#> list()
#> 
#> $data
#> [1] 9
#> 
#> $class
#> [1] "LinOp"
#> 
#> 
#> Slot ".noncvx_var":
#> $type
#> [1] "variable"
#> 
#> $size
#> [1] 3 3
#> 
#> $args
#> list()
#> 
#> $data
#> [1] 9
#> 
#> $class
#> [1] "LinOp"
#> 
#> 
#> Slot "constr_id":
#> [1] 10
#> 
#> 
#>