An interface for the MOSEK solver.

An interface for the MOSEK solver.

MOSEK()

# S4 method for MOSEK
mip_capable(solver)

# S4 method for MOSEK
import_solver(solver)

# S4 method for MOSEK
name(x)

# S4 method for MOSEK,Problem
accepts(object, problem)

# S4 method for MOSEK
block_format(object, problem, constraints, exp_cone_order = NA)

# S4 method for MOSEK,Problem
perform(object, problem)

# S4 method for MOSEK
solve_via_data(
  object,
  data,
  warm_start,
  verbose,
  feastol,
  reltol,
  abstol,
  num_iter,
  solver_opts,
  solver_cache
)

# S4 method for MOSEK,ANY,ANY
invert(object, solution, inverse_data)

Arguments

solver, object, x

A MOSEK object.

problem

A Problem object.

constraints

A list of Constraint objects for which coefficient andd offset data ("G", "h" respectively) is needed

exp_cone_order

A parameter that is only used when a Constraint object describes membership in the exponential cone.

data

Data generated via an apply call.

warm_start

A boolean of whether to warm start the solver.

verbose

A boolean of whether to enable solver verbosity.

feastol

The feasible tolerance.

reltol

The relative tolerance.

abstol

The absolute tolerance.

num_iter

The maximum number of iterations.

solver_opts

A list of Solver specific options

solver_cache

Cache for the solver.

solution

The raw solution returned by the solver.

inverse_data

A list containing data necessary for the inversion.

Methods (by generic)

mip_capable(MOSEK): Can the solver handle mixed-integer programs?

import_solver(MOSEK): Imports the solver.

name(MOSEK): Returns the name of the solver.

accepts(object = MOSEK, problem = Problem): Can MOSEK solve the problem?

block_format(MOSEK): Returns a large matrix "coeff" and a vector of constants "offset" such that every Constraint in "constraints" holds at z in R^n iff "coeff" * z <=_K offset", where K is a product of cones supported by MOSEK and CVXR (zero cone, nonnegative orthant, second order cone, exponential cone). The nature of K is inferred later by accessing the data in "lengths" and "ids".

perform(object = MOSEK, problem = Problem): Returns a new problem and data for inverting the new solution.

solve_via_data(MOSEK): Solve a problem represented by data returned from apply.

invert(object = MOSEK, solution = ANY, inverse_data = ANY): Returns the solution to the original problem given the inverse_data.