An interface for the OSQP solver.

An interface for the OSQP solver.

OSQP()

# S4 method for OSQP
status_map(solver, status)

# S4 method for OSQP
name(x)

# S4 method for OSQP
import_solver(solver)

# S4 method for OSQP,list,InverseData
invert(object, solution, inverse_data)

# S4 method for OSQP
solve_via_data(
  object,
  data,
  warm_start,
  verbose,
  solver_opts,
  solver_cache = list()
)

Arguments

solver, object, x
A OSQP object.

status
A status code returned by the solver.

solution
The raw solution returned by the solver.

inverse_data
A InverseData object containing data necessary for the inversion.

data
Data generated via an apply call.

warm_start
A boolean of whether to warm start the solver.

verbose
A boolean of whether to enable solver verbosity.

solver_opts
A list of Solver specific options

solver_cache
Cache for the solver.

Methods (by generic)

status_map: Converts status returned by the OSQP solver to its respective CVXPY status.

name: Returns the name of the solver.

import_solver: Imports the solver.

invert: Returns the solution to the original problem given the inverse_data.

solve_via_data: Solve a problem represented by data returned from apply.