The Parameter class.

This class represents a parameter, either scalar or a matrix.

Parameter(
  rows = NULL,
  cols = NULL,
  name = NA_character_,
  value = NA_real_,
  ...
)

# S4 method for Parameter
get_data(object)

# S4 method for Parameter
name(x)

# S4 method for Parameter
value(object)

# S4 method for Parameter
value(object) <- value

# S4 method for Parameter
grad(object)

# S4 method for Parameter
parameters(object)

# S4 method for Parameter
canonicalize(object)

Arguments

rows

The number of rows in the parameter.

cols

The number of columns in the parameter.

name

(Optional) A character string representing the name of the parameter.

value

(Optional) A numeric element, vector, matrix, or data.frame. Defaults to NA and may be changed with value<- later.

...

Additional attribute arguments. See Leaf for details.

object, x

A Parameter object.

Methods (by generic)

get_data(Parameter): Returns list(dim, name, value, attributes).

name(Parameter): The name of the parameter.

value(Parameter): The value of the parameter.

value(Parameter) <- value: Set the value of the parameter.

grad(Parameter): An empty list since the gradient of a parameter is zero.

parameters(Parameter): Returns itself as a parameter.

canonicalize(Parameter): The canonical form of the parameter.

Slots

rows

The number of rows in the parameter.

cols

The number of columns in the parameter.

name

(Optional) A character string representing the name of the parameter.

value

(Optional) A numeric element, vector, matrix, or data.frame. Defaults to NA and may be changed with value<- later.

Examples

x <- Parameter(3, name = "x0", nonpos = TRUE) ## 3-vec negative
is_nonneg(x)
#> [1] FALSE
is_nonpos(x)
#> [1] TRUE
size(x)
#> [1] 3