The QuadOverLin class.

This class represents the sum of squared entries in X divided by a scalar y, \(\sum_{i,j} X_{i,j}^2/y\).

QuadOverLin(x, y)

# S4 method for class 'QuadOverLin'
allow_complex(object)

# S4 method for class 'QuadOverLin'
to_numeric(object, values)

# S4 method for class 'QuadOverLin'
validate_args(object)

# S4 method for class 'QuadOverLin'
dim_from_args(object)

# S4 method for class 'QuadOverLin'
sign_from_args(object)

# S4 method for class 'QuadOverLin'
is_atom_convex(object)

# S4 method for class 'QuadOverLin'
is_atom_concave(object)

# S4 method for class 'QuadOverLin'
is_atom_log_log_convex(object)

# S4 method for class 'QuadOverLin'
is_atom_log_log_concave(object)

# S4 method for class 'QuadOverLin'
is_incr(object, idx)

# S4 method for class 'QuadOverLin'
is_decr(object, idx)

# S4 method for class 'QuadOverLin'
is_quadratic(object)

# S4 method for class 'QuadOverLin'
is_qpwa(object)

# S4 method for class 'QuadOverLin'
.domain(object)

# S4 method for class 'QuadOverLin'
.grad(object, values)

Arguments

x

An Expression or numeric matrix.

y

A scalar Expression or numeric constant.

object

A QuadOverLin object.

values

A list of numeric values for the arguments

idx

An index into the atom.

Methods (by generic)

allow_complex(QuadOverLin): Does the atom handle complex numbers?

to_numeric(QuadOverLin): The sum of the entries of x squared over y.

validate_args(QuadOverLin): Check the dimensions of the arguments.

dim_from_args(QuadOverLin): The atom is a scalar.

sign_from_args(QuadOverLin): The atom is positive.

is_atom_convex(QuadOverLin): The atom is convex.

is_atom_concave(QuadOverLin): The atom is not concave.

is_atom_log_log_convex(QuadOverLin): Is the atom log-log convex?

is_atom_log_log_concave(QuadOverLin): Is the atom log-log concave?

is_incr(QuadOverLin): A logical value indicating whether the atom is weakly increasing in argument idx.

is_decr(QuadOverLin): A logical value indicating whether the atom is weakly decreasing in argument idx.

is_quadratic(QuadOverLin): Quadratic if x is affine and y is constant.

is_qpwa(QuadOverLin): Quadratic of piecewise affine if x is piecewise linear and y is constant.

.domain(QuadOverLin): Returns constraints describing the domain of the node

.grad(QuadOverLin): Gives the (sub/super)gradient of the atom w.r.t. each variable

Slots

x

An Expression or numeric matrix.

y

A scalar Expression or numeric constant.