Call to Solver

Returns the result of the call to the solver.

Solver.solve(solver, objective, constraints, cached_data, warm_start, verbose,
  ...)

Arguments

solver
A Solver object.

objective
A list representing the canonicalized objective.

constraints
A list of canonicalized constraints.

cached_data
A list mapping solver name to cached problem data.

warm_start
A logical value indicating whether the previous solver result should be used to warm start.

verbose
A logical value indicating whether to print solver output.

...
Additional arguments to the solver.

Value

A list containing the status, optimal value, primal variable, and dual variables for the equality and inequality constraints.