The SolvingChain class.

This class represents a reduction chain that ends with a solver.

# S4 method for class 'SolvingChain,Chain'
prepend(object, chain)

# S4 method for class 'SolvingChain,Problem'
reduction_solve(
  object,
  problem,
  warm_start,
  verbose,
  feastol,
  reltol,
  abstol,
  num_iter,
  solver_opts
)

# S4 method for class 'SolvingChain'
reduction_solve_via_data(
  object,
  problem,
  data,
  warm_start,
  verbose,
  feastol,
  reltol,
  abstol,
  num_iter,
  solver_opts
)

Arguments

object

A SolvingChain object.

chain

A Chain to prepend.

problem

The problem to solve.

warm_start

A boolean of whether to warm start the solver.

verbose

A boolean of whether to enable solver verbosity.

feastol

The feasible tolerance.

reltol

The relative tolerance.

abstol

The absolute tolerance.

num_iter

The maximum number of iterations.

solver_opts

A list of Solver specific options

data

Data for the solver.

Methods (by generic)

prepend(object = SolvingChain, chain = Chain): Create and return a new SolvingChain by concatenating chain with this instance.

reduction_solve(object = SolvingChain, problem = Problem): Applies each reduction in the chain to the problem, solves it, and then inverts the chain to return a solution of the supplied problem.

reduction_solve_via_data(SolvingChain): Solves the problem using the data output by the an apply invocation.