Matrix Diagonal

Extracts the diagonal from a matrix or makes a vector into a diagonal matrix.

# S4 method for Expression
diag(x = 1, nrow, ncol)

Arguments

x
An Expression, vector, or square matrix.

nrow, ncol
(Optional) Dimensions for the result when x is not a matrix.

Value

An Expression representing the diagonal vector or matrix.

Examples

C <- Variable(3,3)
obj <- Maximize(C[1,3])
constraints <- list(diag(C) == 1, C[1,2] == 0.6, C[2,3] == -0.3, C == Variable(3,3, PSD = TRUE))
prob <- Problem(obj, constraints)
result <- solve(prob)

#> Acquiring MOSEK environment
#> Problem
#>   Name                   :                 
#>   Objective sense        : min             
#>   Type                   : CONIC (conic optimization problem)
#>   Constraints            : 23              
#>   Cones                  : 0               
#>   Scalar variables       : 15              
#>   Matrix variables       : 1               
#>   Integer variables      : 0               
#> 
#> Optimizer started.
#> Presolve started.
#> Linear dependency checker started.
#> Linear dependency checker terminated.
#> Eliminator started.
#> Freed constraints in eliminator : 0
#> Eliminator terminated.
#> Eliminator - tries                  : 1                 time                   : 0.00            
#> Lin. dep.  - tries                  : 1                 time                   : 0.00            
#> Lin. dep.  - number                 : 0               
#> Presolve terminated. Time: 0.00    
#> Problem
#>   Name                   :                 
#>   Objective sense        : min             
#>   Type                   : CONIC (conic optimization problem)
#>   Constraints            : 23              
#>   Cones                  : 0               
#>   Scalar variables       : 15              
#>   Matrix variables       : 1               
#>   Integer variables      : 0               
#> 
#> Optimizer  - threads                : 8               
#> Optimizer  - solved problem         : the primal      
#> Optimizer  - Constraints            : 9
#> Optimizer  - Cones                  : 1
#> Optimizer  - Scalar variables       : 2                 conic                  : 2               
#> Optimizer  - Semi-definite variables: 1                 scalarized             : 6               
#> Factor     - setup time             : 0.00              dense det. time        : 0.00            
#> Factor     - ML order time          : 0.00              GP order time          : 0.00            
#> Factor     - nonzeros before factor : 45                after factor           : 45              
#> Factor     - dense dim.             : 0                 flops                  : 5.88e+02        
#> ITE PFEAS    DFEAS    GFEAS    PRSTATUS   POBJ              DOBJ              MU       TIME  
#> 0   1.2e+00  1.0e+00  1.0e+00  0.00e+00   0.000000000e+00   0.000000000e+00   1.0e+00  0.00  
#> 1   1.7e-01  1.4e-01  6.7e-02  5.29e-01   -3.870442272e-01  -3.406220964e-01  1.4e-01  0.00  
#> 2   8.7e-03  7.3e-03  9.2e-04  7.46e-01   -5.823257506e-01  -5.765797356e-01  7.3e-03  0.00  
#> 3   3.3e-05  2.7e-05  2.2e-07  9.90e-01   -5.831664073e-01  -5.831415151e-01  2.7e-05  0.00  
#> 4   6.8e-07  5.7e-07  6.5e-10  1.00e+00   -5.831518639e-01  -5.831513437e-01  5.7e-07  0.00  
#> 5   1.3e-08  1.1e-08  1.7e-12  1.00e+00   -5.831513724e-01  -5.831513626e-01  1.1e-08  0.00  
#> Optimizer terminated. Time: 0.00    
#> 
#> 
#> Interior-point solution summary
#>   Problem status  : PRIMAL_AND_DUAL_FEASIBLE
#>   Solution status : OPTIMAL
#>   Primal.  obj: -5.8315137237e-01   nrm: 1e+00    Viol.  con: 9e-09    var: 0e+00    barvar: 0e+00  
#>   Dual.    obj: -5.8315136261e-01   nrm: 1e+00    Viol.  con: 0e+00    var: 0e+00    barvar: 2e-08  
#> Optimizer summary
#>   Optimizer                 -                        time: 0.00    
#>     Interior-point          - iterations : 5         time: 0.00    
#>       Basis identification  -                        time: 0.00    
#>         Primal              - iterations : 0         time: 0.00    
#>         Dual                - iterations : 0         time: 0.00    
#>         Clean primal        - iterations : 0         time: 0.00    
#>         Clean dual          - iterations : 0         time: 0.00    
#>     Simplex                 -                        time: 0.00    
#>       Primal simplex        - iterations : 0         time: 0.00    
#>       Dual simplex          - iterations : 0         time: 0.00    
#>     Mixed integer           - relaxations: 0         time: 0.00    
#> 

result$value

#> [1] 0.5831514

result$getValue(C)

#>           [,1] [,2]       [,3]
#> [1,] 1.0000000  0.6  0.5831514
#> [2,] 0.6000000  1.0 -0.3000000
#> [3,] 0.5831514 -0.3  1.0000000