Geometric Mean

The (weighted) geometric mean of vector \(x\) with optional powers given by \(p\).

geo_mean(x, p = NA_real_, max_denom = 1024)

Arguments

x
An Expression or vector.

p
(Optional) A vector of weights for the weighted geometric mean. Defaults to a vector of ones, giving the unweighted geometric mean \(x_1^{1/n} \cdots x_n^{1/n}\).

max_denom
(Optional) The maximum denominator to use in approximating p/sum(p) with w. If w is not an exact representation, increasing max_denom may offer a more accurate representation, at the cost of requiring more convex inequalities to represent the geometric mean. Defaults to 1024.

Value

An Expression representing the geometric mean of the input.

Details

$$\left(x_1^{p_1} \cdots x_n^{p_n} \right)^{\frac{1}{\mathbf{1}^Tp}}$$

The geometric mean includes an implicit constraint that \(x_i \geq 0\) whenever \(p_i > 0\). If \(p_i = 0, x_i\) will be unconstrained. The only exception to this rule occurs when \(p\) has exactly one nonzero element, say \(p_i\), in which case geo_mean(x,p) is equivalent to \(x_i\) (without the nonnegativity constraint). A specific case of this is when \(x \in \mathbf{R}^1\).

Examples

x <- Variable(2)
cost <- geo_mean(x)
prob <- Problem(Maximize(cost), list(sum(x) <= 1))
result <- solve(prob)

#> Problem
#>   Name                   :                 
#>   Objective sense        : min             
#>   Type                   : CONIC (conic optimization problem)
#>   Constraints            : 4               
#>   Cones                  : 1               
#>   Scalar variables       : 6               
#>   Matrix variables       : 0               
#>   Integer variables      : 0               
#> 
#> Optimizer started.
#> Presolve started.
#> Linear dependency checker started.
#> Linear dependency checker terminated.
#> Eliminator started.
#> Freed constraints in eliminator : 3
#> Eliminator terminated.
#> Eliminator started.
#> Freed constraints in eliminator : 0
#> Eliminator terminated.
#> Eliminator - tries                  : 2                 time                   : 0.00            
#> Lin. dep.  - tries                  : 1                 time                   : 0.00            
#> Lin. dep.  - number                 : 0               
#> Presolve terminated. Time: 0.00    
#> Optimizer terminated. Time: 0.00    
#> 
#> 
#> Interior-point solution summary
#>   Problem status  : PRIMAL_AND_DUAL_FEASIBLE
#>   Solution status : OPTIMAL
#>   Primal.  obj: -5.0000000000e-01   nrm: 1e+00    Viol.  con: 0e+00    var: 0e+00    cones: 0e+00  
#>   Dual.    obj: -5.0000000000e-01   nrm: 5e-01    Viol.  con: 0e+00    var: 0e+00    cones: 0e+00  
#> Optimizer summary
#>   Optimizer                 -                        time: 0.00    
#>     Interior-point          - iterations : 0         time: 0.00    
#>       Basis identification  -                        time: 0.00    
#>         Primal              - iterations : 0         time: 0.00    
#>         Dual                - iterations : 0         time: 0.00    
#>         Clean primal        - iterations : 0         time: 0.00    
#>         Clean dual          - iterations : 0         time: 0.00    
#>     Simplex                 -                        time: 0.00    
#>       Primal simplex        - iterations : 0         time: 0.00    
#>       Dual simplex          - iterations : 0         time: 0.00    
#>     Mixed integer           - relaxations: 0         time: 0.00    
#> 

result$value

#> [1] 0.5

result$getValue(x)

#>      [,1]
#> [1,]  0.5
#> [2,]  0.5

x <- Variable(5)
p <- c(0.07, 0.12, 0.23, 0.19, 0.39)
prob <- Problem(Maximize(geo_mean(x,p)), list(p_norm(x) <= 1))
result <- solve(prob)

#> Problem
#>   Name                   :                 
#>   Objective sense        : min             
#>   Type                   : CONIC (conic optimization problem)
#>   Constraints            : 220             
#>   Cones                  : 72              
#>   Scalar variables       : 297             
#>   Matrix variables       : 0               
#>   Integer variables      : 0               
#> 
#> Optimizer started.
#> Presolve started.
#> Linear dependency checker started.
#> Linear dependency checker terminated.
#> Eliminator started.
#> Freed constraints in eliminator : 0
#> Eliminator terminated.
#> Eliminator - tries                  : 1                 time                   : 0.00            
#> Lin. dep.  - tries                  : 1                 time                   : 0.00            
#> Lin. dep.  - number                 : 0               
#> Presolve terminated. Time: 0.00    
#> Problem
#>   Name                   :                 
#>   Objective sense        : min             
#>   Type                   : CONIC (conic optimization problem)
#>   Constraints            : 220             
#>   Cones                  : 72              
#>   Scalar variables       : 297             
#>   Matrix variables       : 0               
#>   Integer variables      : 0               
#> 
#> Optimizer  - threads                : 8               
#> Optimizer  - solved problem         : the dual        
#> Optimizer  - Constraints            : 76
#> Optimizer  - Cones                  : 72
#> Optimizer  - Scalar variables       : 219               conic                  : 219             
#> Optimizer  - Semi-definite variables: 0                 scalarized             : 0               
#> Factor     - setup time             : 0.00              dense det. time        : 0.00            
#> Factor     - ML order time          : 0.00              GP order time          : 0.00            
#> Factor     - nonzeros before factor : 285               after factor           : 596             
#> Factor     - dense dim.             : 0                 flops                  : 8.38e+03        
#> ITE PFEAS    DFEAS    GFEAS    PRSTATUS   POBJ              DOBJ              MU       TIME  
#> 0   1.0e+00  9.0e+00  2.0e+00  0.00e+00   0.000000000e+00   -1.000000000e+00  1.0e+00  0.00  
#> 1   4.0e-01  3.6e+00  2.7e-01  7.78e-01   -1.253801695e-01  -6.850556377e-01  4.0e-01  0.00  
#> 2   1.4e-01  1.2e+00  5.1e-02  1.46e+00   -2.893451134e-01  -4.443705350e-01  1.4e-01  0.00  
#> 3   3.4e-02  3.0e-01  6.4e-03  1.07e+00   -4.198346388e-01  -4.553619963e-01  3.4e-02  0.00  
#> 4   6.3e-03  5.7e-02  5.4e-04  1.02e+00   -4.612837333e-01  -4.676890522e-01  6.3e-03  0.00  
#> 5   1.8e-03  1.7e-02  9.3e-05  9.56e-01   -4.748979372e-01  -4.767657304e-01  1.8e-03  0.00  
#> 6   4.6e-04  4.2e-03  1.2e-05  9.61e-01   -4.793165857e-01  -4.797734508e-01  4.6e-04  0.00  
#> 7   7.9e-05  7.1e-04  9.2e-07  9.55e-01   -4.810567078e-01  -4.811344667e-01  7.9e-05  0.00  
#> 8   1.9e-05  1.7e-04  1.1e-07  9.54e-01   -4.814402871e-01  -4.814589433e-01  1.9e-05  0.00  
#> 9   4.7e-06  4.2e-05  1.4e-08  9.65e-01   -4.815394288e-01  -4.815439842e-01  4.7e-06  0.00  
#> 10  7.8e-07  7.1e-06  1.0e-09  9.73e-01   -4.815705991e-01  -4.815713620e-01  7.8e-07  0.00  
#> 11  1.4e-07  1.3e-06  8.0e-11  9.70e-01   -4.815770388e-01  -4.815771745e-01  1.4e-07  0.00  
#> 12  2.9e-08  2.6e-07  7.6e-12  9.77e-01   -4.815782518e-01  -4.815782797e-01  2.9e-08  0.00  
#> 13  5.4e-09  4.8e-08  6.1e-13  1.01e+00   -4.815785468e-01  -4.815785519e-01  5.4e-09  0.00  
#> 14  2.8e-09  8.0e-09  4.1e-14  1.00e+00   -4.815786064e-01  -4.815786072e-01  8.8e-10  0.00  
#> Optimizer terminated. Time: 0.00    
#> 
#> 
#> Interior-point solution summary
#>   Problem status  : PRIMAL_AND_DUAL_FEASIBLE
#>   Solution status : OPTIMAL
#>   Primal.  obj: -4.8157860636e-01   nrm: 1e+00    Viol.  con: 0e+00    var: 0e+00    cones: 0e+00  
#>   Dual.    obj: -4.8157860720e-01   nrm: 5e-01    Viol.  con: 0e+00    var: 0e+00    cones: 4e-11  
#> Optimizer summary
#>   Optimizer                 -                        time: 0.00    
#>     Interior-point          - iterations : 14        time: 0.00    
#>       Basis identification  -                        time: 0.00    
#>         Primal              - iterations : 0         time: 0.00    
#>         Dual                - iterations : 0         time: 0.00    
#>         Clean primal        - iterations : 0         time: 0.00    
#>         Clean dual          - iterations : 0         time: 0.00    
#>     Simplex                 -                        time: 0.00    
#>       Primal simplex        - iterations : 0         time: 0.00    
#>       Dual simplex          - iterations : 0         time: 0.00    
#>     Mixed integer           - relaxations: 0         time: 0.00    
#> 

result$value

#> [1] 0.4815786

result$getValue(x)

#>           [,1]
#> [1,] 0.2645762
#> [2,] 0.3464108
#> [3,] 0.4795829
#> [4,] 0.4358901
#> [5,] 0.6244990