Get Problem Data

Get the problem data used in the call to the solver.

get_problem_data(object, solver)

Arguments

object
A Problem object.

solver
A string indicating the solver that the problem data is for. Call installed_solvers() to see all available.

Value

A list of arguments for the solver.

Examples

a <- Variable(name = "a")
data <- get_problem_data(Problem(Maximize(exp(a) + 2)), "SCS")
data[["dims"]]

#> $f
#> [1] 0
#> 
#> $l
#> [1] 0
#> 
#> $ep
#> [1] 1
#> 

data[["c"]]

#> [1]  0 -1

data[["A"]]

#> 3 x 2 sparse Matrix of class "dgCMatrix"
#>           
#> [1,] -1  .
#> [2,]  .  .
#> [3,]  . -1

x <- Variable(2, name = "x")
data <- get_problem_data(Problem(Minimize(p_norm(x) + 3)), "ECOS")
data[["dims"]]

#> $f
#> [1] 0
#> 
#> $l
#> [1] 0
#> 
#> $ep
#> [1] 0
#> 
#> $q
#> [1] 3
#> 

data[["c"]]

#> [1] 0 0 1

data[["A"]]

#> 0 x 3 sparse Matrix of class "ngCMatrix"
#>        

data[["G"]]

#> 3 x 3 sparse Matrix of class "dgCMatrix"
#>              
#> [1,]  .  . -1
#> [2,] -1  .  .
#> [3,]  . -1  .