Harmonic Mean

The harmonic mean, \(\left(\frac{1}{n} \sum_{i=1}^n x_i^{-1}\right)^{-1}\). For a matrix, the function is applied over all entries.

harmonic_mean(x)

Arguments

x
An Expression, vector, or matrix.

Value

An Expression representing the harmonic mean of the input.

Examples

x <- Variable()
prob <- Problem(Maximize(harmonic_mean(x)), list(x >= 0, x <= 5))
result <- solve(prob)

#> Problem
#>   Name                   :                 
#>   Objective sense        : min             
#>   Type                   : CONIC (conic optimization problem)
#>   Constraints            : 6               
#>   Cones                  : 1               
#>   Scalar variables       : 6               
#>   Matrix variables       : 0               
#>   Integer variables      : 0               
#> 
#> Optimizer started.
#> Presolve started.
#> Linear dependency checker started.
#> Linear dependency checker terminated.
#> Eliminator started.
#> Freed constraints in eliminator : 1
#> Eliminator terminated.
#> Eliminator started.
#> Freed constraints in eliminator : 0
#> Eliminator terminated.
#> Eliminator - tries                  : 2                 time                   : 0.00            
#> Lin. dep.  - tries                  : 1                 time                   : 0.00            
#> Lin. dep.  - number                 : 0               
#> Presolve terminated. Time: 0.00    
#> Problem
#>   Name                   :                 
#>   Objective sense        : min             
#>   Type                   : CONIC (conic optimization problem)
#>   Constraints            : 6               
#>   Cones                  : 1               
#>   Scalar variables       : 6               
#>   Matrix variables       : 0               
#>   Integer variables      : 0               
#> 
#> Optimizer  - threads                : 8               
#> Optimizer  - solved problem         : the primal      
#> Optimizer  - Constraints            : 2
#> Optimizer  - Cones                  : 1
#> Optimizer  - Scalar variables       : 4                 conic                  : 3               
#> Optimizer  - Semi-definite variables: 0                 scalarized             : 0               
#> Factor     - setup time             : 0.00              dense det. time        : 0.00            
#> Factor     - ML order time          : 0.00              GP order time          : 0.00            
#> Factor     - nonzeros before factor : 3                 after factor           : 3               
#> Factor     - dense dim.             : 0                 flops                  : 2.90e+01        
#> ITE PFEAS    DFEAS    GFEAS    PRSTATUS   POBJ              DOBJ              MU       TIME  
#> 0   4.0e+00  1.0e+00  1.0e+00  0.00e+00   0.000000000e+00   0.000000000e+00   1.0e+00  0.00  
#> 1   7.5e-01  1.9e-01  1.8e-01  -2.89e-01  -5.757073360e-01  1.895599516e-02   1.9e-01  0.00  
#> 2   2.7e-01  6.8e-02  7.2e-02  1.31e-01   -3.520530680e+00  -2.650758113e+00  6.8e-02  0.00  
#> 3   2.6e-02  6.5e-03  2.7e-03  6.42e-01   -4.863143061e+00  -4.723267253e+00  6.5e-03  0.00  
#> 4   2.6e-05  6.6e-06  9.2e-08  9.68e-01   -4.999880971e+00  -4.999714697e+00  6.6e-06  0.00  
#> 5   8.1e-10  2.0e-10  2.8e-12  1.00e+00   -4.999999997e+00  -4.999999990e+00  2.0e-10  0.00  
#> Optimizer terminated. Time: 0.00    
#> 
#> 
#> Interior-point solution summary
#>   Problem status  : PRIMAL_AND_DUAL_FEASIBLE
#>   Solution status : OPTIMAL
#>   Primal.  obj: -4.9999999972e+00   nrm: 1e+01    Viol.  con: 9e-10    var: 0e+00    cones: 0e+00  
#>   Dual.    obj: -4.9999999899e+00   nrm: 1e+00    Viol.  con: 5e-10    var: 1e-15    cones: 0e+00  
#> Optimizer summary
#>   Optimizer                 -                        time: 0.00    
#>     Interior-point          - iterations : 5         time: 0.00    
#>       Basis identification  -                        time: 0.00    
#>         Primal              - iterations : 0         time: 0.00    
#>         Dual                - iterations : 0         time: 0.00    
#>         Clean primal        - iterations : 0         time: 0.00    
#>         Clean dual          - iterations : 0         time: 0.00    
#>     Simplex                 -                        time: 0.00    
#>       Primal simplex        - iterations : 0         time: 0.00    
#>       Dual simplex          - iterations : 0         time: 0.00    
#>     Mixed integer           - relaxations: 0         time: 0.00    
#> 

result$value

#> [1] 5

result$getValue(x)

#> [1] 5