Logistic Function

The elementwise logistic function, \(\log(1 + e^x)\). This is a special case of log(sum(exp)) that evaluates to a vector rather than to a scalar, which is useful for logistic regression.

logistic(x)

Arguments

x
An Expression, vector, or matrix.

Value

An Expression representing the logistic function evaluated at the input.

Examples

set.seed(92)
n <- 20
m <- 1000
sigma <- 45

beta_true <- stats::rnorm(n)
idxs <- sample(n, size = 0.8*n, replace = FALSE)
beta_true[idxs] <- 0
X <- matrix(stats::rnorm(m*n, 0, 5), nrow = m, ncol = n)
y <- sign(X %*% beta_true + stats::rnorm(m, 0, sigma))

beta <- Variable(n)
X_sign <- apply(X, 2, function(x) { ifelse(y <= 0, -1, 1) * x })
obj <- -sum(logistic(-X[y <= 0,] %*% beta)) - sum(logistic(X[y == 1,] %*% beta))
prob <- Problem(Maximize(obj))
result <- solve(prob)

log_odds <- result$getValue(X %*% beta)
beta_res <- result$getValue(beta)
y_probs <- 1/(1 + exp(-X %*% beta_res))
log(y_probs/(1 - y_probs))

#>                  [,1]
#>    [1,] -0.6704630943
#>    [2,] -0.5925209891
#>    [3,] -0.2870325126
#>    [4,] -0.3903460838
#>    [5,]  0.1512545808
#>    [6,] -0.3212345731
#>    [7,]  0.0710814624
#>    [8,] -0.0005971982
#>    [9,]  0.6231597991
#>   [10,] -0.3461583046
#>   [11,]  0.6165958988
#>   [12,]  0.2493824618
#>   [13,] -0.4229509296
#>   [14,]  1.3550725182
#>   [15,] -0.3265632379
#>   [16,] -0.5974123435
#>   [17,]  1.4251078729
#>   [18,] -0.8064208571
#>   [19,]  0.6354287657
#>   [20,] -0.5313350436
#>   [21,] -0.5836882522
#>   [22,] -0.2018627665
#>   [23,] -1.5741226918
#>   [24,] -0.2996852805
#>   [25,] -0.2075347411
#>   [26,]  0.4437004152
#>   [27,] -0.5653485055
#>   [28,] -1.2550708184
#>   [29,] -0.5346294301
#>   [30,]  0.4834093125
#>   [31,]  0.2311038484
#>   [32,]  0.2894497519
#>   [33,] -0.0125280479
#>   [34,]  0.7295815742
#>   [35,] -0.8747179060
#>   [36,] -0.3928946793
#>   [37,] -0.1155685247
#>   [38,]  1.1672549202
#>   [39,] -0.7217500006
#>   [40,]  0.0611253461
#>   [41,] -0.6167710917
#>   [42,] -0.0245109387
#>   [43,]  0.6908279842
#>   [44,] -0.8034528496
#>   [45,]  0.8837170463
#>   [46,] -1.0940707098
#>   [47,]  0.0151100359
#>   [48,] -0.0365093011
#>   [49,]  0.1208107821
#>   [50,]  0.0228331233
#>   [51,]  0.0203853239
#>   [52,] -0.4370914681
#>   [53,]  0.0381301779
#>   [54,] -0.0670267935
#>   [55,] -0.5450527619
#>   [56,] -0.1197502391
#>   [57,] -0.1350437527
#>   [58,]  0.1775482590
#>   [59,]  1.1398039059
#>   [60,] -0.4142606158
#>   [61,]  0.3993020597
#>   [62,] -0.4202167020
#>   [63,]  0.0406903893
#>   [64,] -0.6489144130
#>   [65,] -0.9702993001
#>   [66,]  0.1117457778
#>   [67,] -0.1453580608
#>   [68,]  0.0616531834
#>   [69,]  0.0779095880
#>   [70,]  0.4237450465
#>   [71,] -0.4578364043
#>   [72,] -0.2049726780
#>   [73,] -0.2303909471
#>   [74,]  0.1166966835
#>   [75,] -0.3651982685
#>   [76,] -0.2369515665
#>   [77,]  0.1457245187
#>   [78,] -1.0896509795
#>   [79,] -0.1308720502
#>   [80,]  0.2291696041
#>   [81,]  0.1373257407
#>   [82,] -0.0015646314
#>   [83,]  0.3467600623
#>   [84,] -0.1543902000
#>   [85,] -0.2309035762
#>   [86,] -0.5247155374
#>   [87,]  0.4603781400
#>   [88,] -0.3126503176
#>   [89,]  0.2120065820
#>   [90,]  0.3146989486
#>   [91,]  0.4193055548
#>   [92,]  0.6556830968
#>   [93,] -0.5111163077
#>   [94,]  1.3013374346
#>   [95,]  0.8030203710
#>   [96,]  0.2830379912
#>   [97,] -0.6704019054
#>   [98,]  0.6392110092
#>   [99,]  0.0848846862
#>  [100,] -0.0064341268
#>  [101,]  0.1647477723
#>  [102,]  0.2973145486
#>  [103,]  0.3564216441
#>  [104,] -0.8130515865
#>  [105,] -0.5596621702
#>  [106,]  0.2440641301
#>  [107,] -0.5330565274
#>  [108,]  0.4528455885
#>  [109,] -0.5708986004
#>  [110,] -0.4624013945
#>  [111,] -0.2336904024
#>  [112,]  0.1511217068
#>  [113,] -0.0619004030
#>  [114,]  0.5838999680
#>  [115,] -0.0005893495
#>  [116,] -0.7089783791
#>  [117,] -0.3172727555
#>  [118,]  0.3970610532
#>  [119,] -1.5820026886
#>  [120,]  0.5686307363
#>  [121,] -0.3273449985
#>  [122,] -0.8146018356
#>  [123,] -0.0947998006
#>  [124,]  0.3203461145
#>  [125,] -0.4429181712
#>  [126,]  0.1170340592
#>  [127,] -0.7652298928
#>  [128,] -0.0573324384
#>  [129,]  0.0095935053
#>  [130,] -0.1932161145
#>  [131,]  0.1964240374
#>  [132,] -0.3175629926
#>  [133,]  0.4246242947
#>  [134,]  0.1266898630
#>  [135,]  0.3305635541
#>  [136,] -0.0061835713
#>  [137,] -0.5537671920
#>  [138,] -0.1133696638
#>  [139,] -0.7994656146
#>  [140,]  0.8081654134
#>  [141,]  0.4153117334
#>  [142,]  0.8942311554
#>  [143,] -0.3139475941
#>  [144,]  0.1458384812
#>  [145,] -0.2050280908
#>  [146,] -0.0601898146
#>  [147,] -1.1748166854
#>  [148,]  0.5454748011
#>  [149,]  0.2075685441
#>  [150,] -0.2618777792
#>  [151,]  0.8038945945
#>  [152,] -0.0455080952
#>  [153,]  0.5741366552
#>  [154,]  0.6919909214
#>  [155,]  0.0716659997
#>  [156,] -0.4965780438
#>  [157,]  0.4423937338
#>  [158,] -0.3816873436
#>  [159,]  0.5206047494
#>  [160,] -0.1021938450
#>  [161,] -0.2226422504
#>  [162,]  0.7098894947
#>  [163,]  0.5765288888
#>  [164,]  0.2450316307
#>  [165,] -0.1737092700
#>  [166,]  0.3589674175
#>  [167,] -0.3290063128
#>  [168,] -0.8993424573
#>  [169,] -0.8218023129
#>  [170,] -0.7797818752
#>  [171,] -0.2137195442
#>  [172,]  0.5276324395
#>  [173,] -0.8487864481
#>  [174,]  0.1034555714
#>  [175,] -0.4103645849
#>  [176,] -0.7528826142
#>  [177,] -0.4402419824
#>  [178,] -1.0685712121
#>  [179,]  0.4078557225
#>  [180,]  0.6313630446
#>  [181,]  0.3750910809
#>  [182,] -0.0151764805
#>  [183,] -0.5310858859
#>  [184,]  0.4077537150
#>  [185,]  0.1571557429
#>  [186,] -0.6328638623
#>  [187,] -0.3822564787
#>  [188,]  0.5205178556
#>  [189,]  0.2562708438
#>  [190,]  0.8532652213
#>  [191,] -0.5649183765
#>  [192,]  0.0370592396
#>  [193,] -0.3390815660
#>  [194,] -0.0375175498
#>  [195,]  0.0474711987
#>  [196,] -0.0326520004
#>  [197,]  0.2159994217
#>  [198,] -1.0662561903
#>  [199,] -0.9549963582
#>  [200,] -0.6453859870
#>  [201,]  0.5183360237
#>  [202,] -0.6995044833
#>  [203,]  0.2180985389
#>  [204,]  0.1893089323
#>  [205,]  0.4929912721
#>  [206,]  0.1844672354
#>  [207,]  0.3573924748
#>  [208,] -0.2500562062
#>  [209,]  1.3917221599
#>  [210,] -0.7617842542
#>  [211,]  0.5279267179
#>  [212,] -1.0117761552
#>  [213,] -0.2423738896
#>  [214,] -0.6133964016
#>  [215,] -1.1656276815
#>  [216,]  0.3821355129
#>  [217,] -0.4501219477
#>  [218,] -1.0973375643
#>  [219,]  0.8937381489
#>  [220,]  0.3140684791
#>  [221,]  0.1340518593
#>  [222,] -1.2230935005
#>  [223,]  0.5678232270
#>  [224,]  0.8463636554
#>  [225,]  0.6349199386
#>  [226,]  0.7005939499
#>  [227,] -0.7992179575
#>  [228,]  0.5683432554
#>  [229,]  0.4113019419
#>  [230,]  0.0902421750
#>  [231,] -0.3665193801
#>  [232,]  0.0127787978
#>  [233,]  0.1629168771
#>  [234,] -0.3527184433
#>  [235,] -0.0758209905
#>  [236,]  0.0091707366
#>  [237,] -0.4611199859
#>  [238,]  0.0657048908
#>  [239,] -0.5298589974
#>  [240,] -0.5962216513
#>  [241,]  0.0637662108
#>  [242,] -0.3510143309
#>  [243,]  0.5476080702
#>  [244,]  0.5265622528
#>  [245,] -0.8598192311
#>  [246,]  0.6395315450
#>  [247,]  0.3215956905
#>  [248,] -0.4795930641
#>  [249,]  0.3975225600
#>  [250,]  0.1215639790
#>  [251,] -0.3140290293
#>  [252,] -0.3046505326
#>  [253,]  0.3923203321
#>  [254,] -0.1856449828
#>  [255,] -0.0851611816
#>  [256,] -0.0029478025
#>  [257,]  0.5734787937
#>  [258,] -0.0359979676
#>  [259,] -0.9568001289
#>  [260,]  0.6395681968
#>  [261,]  0.4795746769
#>  [262,]  0.1872611791
#>  [263,] -0.2498902398
#>  [264,]  0.3260975147
#>  [265,]  0.1986606024
#>  [266,] -0.6790657152
#>  [267,]  0.6809566567
#>  [268,] -0.2952565107
#>  [269,]  0.5265816021
#>  [270,]  0.6812931870
#>  [271,] -0.4787236950
#>  [272,]  0.6944713248
#>  [273,]  0.3116193282
#>  [274,]  0.6770417954
#>  [275,]  0.0024187407
#>  [276,]  0.1779772250
#>  [277,]  0.1332649675
#>  [278,]  0.2881618668
#>  [279,] -0.8398287038
#>  [280,]  0.8293094758
#>  [281,]  0.3237942274
#>  [282,]  0.3688730275
#>  [283,] -0.2838432992
#>  [284,]  0.1698083723
#>  [285,] -0.0972796484
#>  [286,]  0.1692317039
#>  [287,]  0.6139347302
#>  [288,]  0.1736842273
#>  [289,] -0.0806357938
#>  [290,]  0.0115361111
#>  [291,] -0.3183890650
#>  [292,]  0.0866711389
#>  [293,] -0.7586370144
#>  [294,]  0.0584521343
#>  [295,] -0.1537379881
#>  [296,]  0.8435353662
#>  [297,]  0.8815472075
#>  [298,]  0.1534473131
#>  [299,]  1.2271321376
#>  [300,]  0.3098944194
#>  [301,] -0.2555286308
#>  [302,] -0.2073298362
#>  [303,] -0.2298583788
#>  [304,] -0.5772172872
#>  [305,] -0.4885411882
#>  [306,]  0.5318317570
#>  [307,] -0.1440925919
#>  [308,]  0.6375914730
#>  [309,]  0.2563585650
#>  [310,] -0.0603845653
#>  [311,] -0.2223213659
#>  [312,] -0.4951118581
#>  [313,] -0.2533166605
#>  [314,]  0.3508833108
#>  [315,]  0.0868145897
#>  [316,] -0.3995795729
#>  [317,] -0.2455257719
#>  [318,] -0.0988029681
#>  [319,]  0.6998293778
#>  [320,] -0.4848228366
#>  [321,] -0.3735409115
#>  [322,] -0.0385927913
#>  [323,]  0.0463456198
#>  [324,] -0.8583735096
#>  [325,] -0.4583224458
#>  [326,]  0.6281461322
#>  [327,] -0.6350990763
#>  [328,] -0.9581905886
#>  [329,]  0.9368536508
#>  [330,]  0.1483863835
#>  [331,] -0.0538268231
#>  [332,] -1.2323118921
#>  [333,] -0.1706905972
#>  [334,]  0.1205512480
#>  [335,] -0.4870888559
#>  [336,]  0.0540017947
#>  [337,] -0.1344850823
#>  [338,]  1.1543677874
#>  [339,]  0.8619039797
#>  [340,]  0.0228922609
#>  [341,] -0.5239051305
#>  [342,]  0.3351250325
#>  [343,] -0.7517216179
#>  [344,]  0.6050344871
#>  [345,] -0.0211553811
#>  [346,]  0.4027210964
#>  [347,]  0.8252914038
#>  [348,] -0.3307107790
#>  [349,]  0.3177290923
#>  [350,] -0.0224525234
#>  [351,] -0.5546794847
#>  [352,] -0.2831599314
#>  [353,] -0.1794669095
#>  [354,] -0.5933559358
#>  [355,]  0.5243847617
#>  [356,]  0.2004131891
#>  [357,]  0.2904696844
#>  [358,] -0.4817135716
#>  [359,] -0.1010059101
#>  [360,]  0.3605564874
#>  [361,]  0.3838614231
#>  [362,] -0.6909305597
#>  [363,] -0.3578699591
#>  [364,]  0.7571715975
#>  [365,]  0.4263380266
#>  [366,] -0.7292050856
#>  [367,]  1.3051143798
#>  [368,]  0.3406709893
#>  [369,]  0.1624632189
#>  [370,] -0.4134291670
#>  [371,]  0.8469834519
#>  [372,]  0.1445704831
#>  [373,] -1.0051192671
#>  [374,]  0.1813266335
#>  [375,]  0.3127989349
#>  [376,] -0.2938476248
#>  [377,] -1.1220807211
#>  [378,]  0.0153643351
#>  [379,]  0.6470340989
#>  [380,] -1.1136346557
#>  [381,] -0.6871787650
#>  [382,]  0.5673112934
#>  [383,] -0.2756630634
#>  [384,] -0.8618140217
#>  [385,] -0.1516840589
#>  [386,] -0.2098221117
#>  [387,] -1.2010421602
#>  [388,]  1.2558559000
#>  [389,] -0.5078151112
#>  [390,] -0.8335911163
#>  [391,]  0.8759323090
#>  [392,]  0.5860483916
#>  [393,] -1.2938098973
#>  [394,] -0.7939923322
#>  [395,] -0.5236388459
#>  [396,]  0.1775993935
#>  [397,] -1.4013526309
#>  [398,]  0.7131577448
#>  [399,]  0.0354501129
#>  [400,]  0.1344397221
#>  [401,]  0.4879786668
#>  [402,]  0.3638103580
#>  [403,] -0.1982433027
#>  [404,]  0.1427058083
#>  [405,]  0.0306158980
#>  [406,]  0.6818210203
#>  [407,]  0.0828449771
#>  [408,]  0.1285041169
#>  [409,]  0.7544537358
#>  [410,]  0.2249915170
#>  [411,]  0.0320380242
#>  [412,]  0.4893762923
#>  [413,]  0.3546944908
#>  [414,] -0.3652958428
#>  [415,]  0.8342700434
#>  [416,] -0.2503075068
#>  [417,]  0.5247274379
#>  [418,]  0.9033112073
#>  [419,]  0.6160263267
#>  [420,]  0.1392540046
#>  [421,]  0.7268299327
#>  [422,] -0.5110412634
#>  [423,] -0.6768779887
#>  [424,]  0.8344327890
#>  [425,]  0.3100223884
#>  [426,]  0.4325917152
#>  [427,]  0.1333015809
#>  [428,]  0.5181987565
#>  [429,] -0.3190033028
#>  [430,]  0.4315650637
#>  [431,]  0.3930292982
#>  [432,] -0.4950452640
#>  [433,] -0.7391312311
#>  [434,] -0.3018866894
#>  [435,] -0.2371898569
#>  [436,]  0.9648006435
#>  [437,] -0.5663648856
#>  [438,] -0.2494181011
#>  [439,] -0.8171903198
#>  [440,] -0.7060629710
#>  [441,] -0.5057634920
#>  [442,]  0.3789786440
#>  [443,] -0.9727690416
#>  [444,]  0.4803748875
#>  [445,]  0.6154997032
#>  [446,] -0.2193776609
#>  [447,]  0.2770470116
#>  [448,]  0.1446139602
#>  [449,]  0.7294718634
#>  [450,]  0.1318391605
#>  [451,] -0.9285978375
#>  [452,]  0.5649934188
#>  [453,] -0.0734980908
#>  [454,]  0.6055783505
#>  [455,]  0.5110950362
#>  [456,]  0.3522425614
#>  [457,] -0.7569996894
#>  [458,] -0.1832534183
#>  [459,] -0.7545347483
#>  [460,] -0.0289834913
#>  [461,]  0.7738117560
#>  [462,]  0.4916010153
#>  [463,] -0.3123306984
#>  [464,] -1.0281761315
#>  [465,]  0.2271018663
#>  [466,]  0.7107839854
#>  [467,] -0.3046738854
#>  [468,] -0.5839132909
#>  [469,]  1.4979160378
#>  [470,] -0.2798554854
#>  [471,]  0.0598992311
#>  [472,] -0.1166457859
#>  [473,]  0.5759246347
#>  [474,]  0.0739417146
#>  [475,]  0.5817835136
#>  [476,]  0.2869988865
#>  [477,] -0.1296581482
#>  [478,]  0.4669543137
#>  [479,] -0.0833644020
#>  [480,]  0.0444098596
#>  [481,] -0.2424492241
#>  [482,]  0.4828024234
#>  [483,]  0.3103943940
#>  [484,]  0.0108842979
#>  [485,]  0.7864454839
#>  [486,] -0.9775950596
#>  [487,] -0.3104788274
#>  [488,] -0.3305451043
#>  [489,] -0.2215545770
#>  [490,]  0.0721147916
#>  [491,]  0.2177866891
#>  [492,]  0.1394661416
#>  [493,]  0.2520113003
#>  [494,]  0.5967330686
#>  [495,]  0.3058412189
#>  [496,]  0.5917396952
#>  [497,]  0.3640687346
#>  [498,]  0.1549105577
#>  [499,] -0.4251377243
#>  [500,] -0.5954678992
#>  [501,]  0.1971918644
#>  [502,]  0.3262388709
#>  [503,]  0.2089614475
#>  [504,] -0.3834710821
#>  [505,]  0.4277982292
#>  [506,]  0.0551673688
#>  [507,] -0.0386644264
#>  [508,]  0.2733804115
#>  [509,] -0.5726106305
#>  [510,] -0.2647311980
#>  [511,] -0.1636706499
#>  [512,]  0.2262565214
#>  [513,] -0.0280558932
#>  [514,] -0.2338558530
#>  [515,]  0.6861946942
#>  [516,] -0.5119172017
#>  [517,] -0.0523612319
#>  [518,]  0.1044179985
#>  [519,] -0.1237408193
#>  [520,]  0.0757047706
#>  [521,] -0.5462386337
#>  [522,]  0.0475511330
#>  [523,]  1.1280829868
#>  [524,]  2.1538725956
#>  [525,] -0.3064813486
#>  [526,] -0.0496100706
#>  [527,] -0.8556069840
#>  [528,]  0.0941741890
#>  [529,] -0.2142734187
#>  [530,]  0.4862952814
#>  [531,] -0.6682489502
#>  [532,] -0.1670082401
#>  [533,]  0.5639744511
#>  [534,]  0.0179409626
#>  [535,] -0.3645823165
#>  [536,]  0.0408228142
#>  [537,]  0.9396835211
#>  [538,] -0.1184808161
#>  [539,]  0.4494359154
#>  [540,] -0.2161924581
#>  [541,] -0.0674631000
#>  [542,]  0.2276774635
#>  [543,]  0.1874767676
#>  [544,] -0.2695367287
#>  [545,] -0.1355872264
#>  [546,]  0.8939486084
#>  [547,] -0.3868984184
#>  [548,]  0.0633872214
#>  [549,]  0.4120493351
#>  [550,]  0.2301518473
#>  [551,]  0.9149526279
#>  [552,]  1.1397675052
#>  [553,] -0.4759072001
#>  [554,] -0.2316817941
#>  [555,]  0.5579745089
#>  [556,] -0.4722423575
#>  [557,] -0.1854953888
#>  [558,] -1.1758166154
#>  [559,]  0.4799535299
#>  [560,]  0.2889641123
#>  [561,] -1.2229510114
#>  [562,] -0.6987285132
#>  [563,] -0.7161550448
#>  [564,] -0.6979248374
#>  [565,] -0.3020491063
#>  [566,] -0.4899998114
#>  [567,] -0.8276038607
#>  [568,] -0.4674991911
#>  [569,]  0.1342953798
#>  [570,]  0.4109908809
#>  [571,]  0.6809737866
#>  [572,]  0.0412524890
#>  [573,] -0.1366373253
#>  [574,] -1.1207434628
#>  [575,]  0.3470099606
#>  [576,] -0.0923424897
#>  [577,] -0.3819376280
#>  [578,]  0.8333122290
#>  [579,] -0.8232316111
#>  [580,]  0.4841933217
#>  [581,]  0.3710602331
#>  [582,] -0.1472165107
#>  [583,] -0.6323964292
#>  [584,] -0.0760928271
#>  [585,]  0.3433281094
#>  [586,]  0.2907458131
#>  [587,]  0.1804670090
#>  [588,]  0.0960758384
#>  [589,]  0.6891833191
#>  [590,] -0.5049619061
#>  [591,]  1.2848962949
#>  [592,] -0.3799319584
#>  [593,]  0.7477771022
#>  [594,] -0.1956214798
#>  [595,] -0.4801365588
#>  [596,] -1.1238793288
#>  [597,] -0.1742748683
#>  [598,]  0.7379016658
#>  [599,] -0.3434222068
#>  [600,] -0.1916686113
#>  [601,]  0.5453954063
#>  [602,]  0.5152117861
#>  [603,] -0.0849468117
#>  [604,] -0.9375782982
#>  [605,]  0.0342859006
#>  [606,] -0.6655001969
#>  [607,]  0.1873358796
#>  [608,] -0.0769445688
#>  [609,]  0.0459696854
#>  [610,] -1.1788788675
#>  [611,]  0.3286439489
#>  [612,] -0.0070406452
#>  [613,]  0.0901116562
#>  [614,] -0.6918573163
#>  [615,] -0.4205835982
#>  [616,]  0.4543308578
#>  [617,] -1.0384659050
#>  [618,]  0.1549885632
#>  [619,]  0.0778147845
#>  [620,]  0.7126003751
#>  [621,] -0.4782872030
#>  [622,]  0.0135671196
#>  [623,]  0.4106845711
#>  [624,] -1.0563351529
#>  [625,] -0.9110845650
#>  [626,]  0.4652633278
#>  [627,]  0.0051111338
#>  [628,] -0.7323712310
#>  [629,]  0.1945694518
#>  [630,]  0.1503684341
#>  [631,] -0.4189372942
#>  [632,] -0.1985270571
#>  [633,]  0.3485485299
#>  [634,] -0.2546342682
#>  [635,]  0.1227112444
#>  [636,] -0.1856134324
#>  [637,]  0.2738344981
#>  [638,]  1.0011676193
#>  [639,]  0.4347957205
#>  [640,] -1.6299772488
#>  [641,]  0.2096881632
#>  [642,]  0.0085564740
#>  [643,]  0.8349114198
#>  [644,] -0.3045662395
#>  [645,]  0.6219764269
#>  [646,] -0.2574940186
#>  [647,]  0.0345350859
#>  [648,]  0.0117655821
#>  [649,] -1.4764558527
#>  [650,] -0.1628251015
#>  [651,] -1.2739512449
#>  [652,]  0.8570896057
#>  [653,]  0.1669312245
#>  [654,] -0.5317906525
#>  [655,] -1.0839297506
#>  [656,] -0.3629756170
#>  [657,]  0.2942429359
#>  [658,] -0.7617615215
#>  [659,] -0.1690555849
#>  [660,]  0.4039995092
#>  [661,] -0.3634629728
#>  [662,] -0.8665707704
#>  [663,] -0.0941831195
#>  [664,]  0.5124761774
#>  [665,]  0.3797346859
#>  [666,]  0.4798142041
#>  [667,] -0.5581945658
#>  [668,]  0.3612952156
#>  [669,]  0.5623029607
#>  [670,]  0.1053972536
#>  [671,]  0.8528508307
#>  [672,]  0.3209360902
#>  [673,] -0.2066732733
#>  [674,] -0.4629214635
#>  [675,]  0.0776331754
#>  [676,]  0.0504122360
#>  [677,]  0.0518264135
#>  [678,]  0.2186389681
#>  [679,]  0.9749586949
#>  [680,] -1.1149582238
#>  [681,]  0.3516395313
#>  [682,]  0.3667504414
#>  [683,]  0.8789325100
#>  [684,] -0.2670387681
#>  [685,] -0.1829113050
#>  [686,]  0.1890775345
#>  [687,]  1.0046734881
#>  [688,]  0.3745007246
#>  [689,] -0.0766472504
#>  [690,]  0.6810611880
#>  [691,]  0.4759962036
#>  [692,]  0.6531523419
#>  [693,]  1.4329123786
#>  [694,] -0.7168583666
#>  [695,] -0.9394586746
#>  [696,]  0.2211785165
#>  [697,] -0.6708400660
#>  [698,] -0.6266761197
#>  [699,] -0.2085815674
#>  [700,] -0.5009836953
#>  [701,] -0.0736223995
#>  [702,]  0.0828177416
#>  [703,]  0.3307963818
#>  [704,] -0.7440159095
#>  [705,] -0.3728372470
#>  [706,]  0.2977476771
#>  [707,]  0.7548277324
#>  [708,]  0.3886601273
#>  [709,]  0.4132897191
#>  [710,]  0.1895265842
#>  [711,]  0.3994029084
#>  [712,]  0.9802412222
#>  [713,]  0.7551543250
#>  [714,] -0.2736306146
#>  [715,]  0.1938535597
#>  [716,]  0.2522116117
#>  [717,]  0.4456374814
#>  [718,] -0.1725815278
#>  [719,]  1.0797141577
#>  [720,]  0.3771454548
#>  [721,]  1.0647239704
#>  [722,]  0.5670855333
#>  [723,] -0.3559525192
#>  [724,] -0.7792760136
#>  [725,]  0.3513232697
#>  [726,] -0.2204913107
#>  [727,] -0.7116093102
#>  [728,] -0.1274127442
#>  [729,]  0.1966028391
#>  [730,] -0.3877922299
#>  [731,] -0.0199311849
#>  [732,]  0.1184648051
#>  [733,]  0.2743294548
#>  [734,]  0.5775163651
#>  [735,] -0.5056837523
#>  [736,] -0.2552236081
#>  [737,] -0.6560235148
#>  [738,]  0.0348107266
#>  [739,]  0.0902877569
#>  [740,] -0.6225091461
#>  [741,]  0.0797127152
#>  [742,] -0.1238019421
#>  [743,] -0.2795899558
#>  [744,]  0.4582662524
#>  [745,] -0.2354744170
#>  [746,]  0.2209555148
#>  [747,] -0.5618237903
#>  [748,]  0.2231037185
#>  [749,]  0.4956425742
#>  [750,] -0.5832015255
#>  [751,]  0.2152637809
#>  [752,]  0.0542756606
#>  [753,] -0.2871976117
#>  [754,]  0.8426849160
#>  [755,]  0.7672685136
#>  [756,]  0.2816475758
#>  [757,] -0.2592014816
#>  [758,] -0.3337014884
#>  [759,]  0.3641773979
#>  [760,]  0.9735354173
#>  [761,] -0.3217674343
#>  [762,] -0.7869705236
#>  [763,] -0.4184631778
#>  [764,] -0.4806315187
#>  [765,] -0.8255594923
#>  [766,] -0.5302905623
#>  [767,]  1.3100569299
#>  [768,] -0.9604082716
#>  [769,] -0.0885782501
#>  [770,]  0.2058064428
#>  [771,] -0.7842536888
#>  [772,]  0.0635042829
#>  [773,] -0.6950925494
#>  [774,] -0.1846474044
#>  [775,] -0.3363954059
#>  [776,] -0.1788591583
#>  [777,]  0.8910521535
#>  [778,] -0.1378227628
#>  [779,] -0.5356047624
#>  [780,]  0.3572859763
#>  [781,]  0.1585359589
#>  [782,]  1.0687386210
#>  [783,]  1.1066232209
#>  [784,] -0.2037525593
#>  [785,]  0.9385824119
#>  [786,]  0.6147245298
#>  [787,] -0.6015432004
#>  [788,]  0.1344282108
#>  [789,] -1.2289962722
#>  [790,] -0.2334211092
#>  [791,]  0.6008688337
#>  [792,] -0.3406579794
#>  [793,]  0.0389008284
#>  [794,]  0.6983065370
#>  [795,]  0.3481366840
#>  [796,] -1.3029117293
#>  [797,] -0.2717453710
#>  [798,] -0.2855158791
#>  [799,] -0.0936068076
#>  [800,] -0.3429542740
#>  [801,]  0.0655775755
#>  [802,]  0.4265481641
#>  [803,] -0.1910330382
#>  [804,] -0.9140217304
#>  [805,]  0.3288473978
#>  [806,] -0.8873835566
#>  [807,]  1.0038330933
#>  [808,]  0.9868409024
#>  [809,] -0.2429658122
#>  [810,] -0.9556264039
#>  [811,]  0.4624149222
#>  [812,]  0.3846087081
#>  [813,]  0.2159086782
#>  [814,]  0.6700748295
#>  [815,] -0.4725947523
#>  [816,] -0.4929570629
#>  [817,]  0.0322196526
#>  [818,]  0.3226433888
#>  [819,] -0.6958603876
#>  [820,]  0.9750553841
#>  [821,] -0.2351513849
#>  [822,]  0.0729400038
#>  [823,] -0.1591442294
#>  [824,] -0.5117365549
#>  [825,] -0.4795145227
#>  [826,]  0.4052723189
#>  [827,]  0.4204350231
#>  [828,]  0.2745340319
#>  [829,] -0.0205288446
#>  [830,] -0.4314104679
#>  [831,]  0.6094979258
#>  [832,]  1.0722988030
#>  [833,] -0.2645999857
#>  [834,] -0.3071173700
#>  [835,] -1.0247635403
#>  [836,] -0.0066208534
#>  [837,] -0.3715076949
#>  [838,] -0.9059752575
#>  [839,]  0.3446600453
#>  [840,] -0.4288191535
#>  [841,]  0.5646836894
#>  [842,] -0.3737715842
#>  [843,]  0.0442036396
#>  [844,] -0.3987125842
#>  [845,] -0.7878825453
#>  [846,]  0.2526586331
#>  [847,] -0.7940285746
#>  [848,]  0.9474003371
#>  [849,] -0.4694263366
#>  [850,]  0.3828985035
#>  [851,]  0.6224160814
#>  [852,] -0.5851068967
#>  [853,]  0.0889072141
#>  [854,] -0.7938838275
#>  [855,]  0.3155360575
#>  [856,] -0.1221251818
#>  [857,] -1.3300318663
#>  [858,] -0.5719262933
#>  [859,] -0.9125700934
#>  [860,] -0.2369379061
#>  [861,] -0.1245151235
#>  [862,]  0.1233020735
#>  [863,] -0.3210539672
#>  [864,]  0.2492034537
#>  [865,] -0.2825124975
#>  [866,] -0.5326973174
#>  [867,]  0.8425370589
#>  [868,] -0.1497229644
#>  [869,] -0.0707902538
#>  [870,]  1.4702428576
#>  [871,]  0.3788469193
#>  [872,]  0.1531686590
#>  [873,] -0.0760090370
#>  [874,]  0.4299919989
#>  [875,]  0.2210141261
#>  [876,]  0.2148670830
#>  [877,]  0.1065669273
#>  [878,]  0.3748454738
#>  [879,]  1.3183841846
#>  [880,]  0.1440800451
#>  [881,] -0.2546892976
#>  [882,]  0.3841600772
#>  [883,] -0.4562272063
#>  [884,]  0.0205490481
#>  [885,] -0.5283100660
#>  [886,] -0.0380574856
#>  [887,]  0.8189167854
#>  [888,] -0.3172761464
#>  [889,] -0.0775258939
#>  [890,]  0.6523293945
#>  [891,] -0.5292016138
#>  [892,] -0.2480728586
#>  [893,] -0.5409339254
#>  [894,]  1.6438803488
#>  [895,] -0.0285476169
#>  [896,]  0.5037556924
#>  [897,]  0.1677093800
#>  [898,]  0.4745106066
#>  [899,] -0.3244890126
#>  [900,]  0.2792941285
#>  [901,] -0.5484027083
#>  [902,] -0.0305746231
#>  [903,] -0.1514615113
#>  [904,]  0.2612436109
#>  [905,]  0.2779320184
#>  [906,] -0.5204612763
#>  [907,]  0.2379718307
#>  [908,] -0.0438157478
#>  [909,] -0.1526758471
#>  [910,] -0.0887342023
#>  [911,] -0.1575879331
#>  [912,]  0.1630577357
#>  [913,]  0.7016822801
#>  [914,] -0.8003373510
#>  [915,] -0.5517511555
#>  [916,]  0.5581564281
#>  [917,]  0.9733522042
#>  [918,]  0.4126658591
#>  [919,]  0.2656260533
#>  [920,]  0.9628163371
#>  [921,] -0.9312161253
#>  [922,] -0.4634518706
#>  [923,] -1.3692961321
#>  [924,]  0.2544493720
#>  [925,]  0.4049702770
#>  [926,] -0.8171378803
#>  [927,]  0.9692739744
#>  [928,]  0.2155657176
#>  [929,] -0.0075318714
#>  [930,] -0.3707496850
#>  [931,]  0.3387147326
#>  [932,]  0.6080738833
#>  [933,]  0.3844741556
#>  [934,] -0.4457022782
#>  [935,]  0.4762845993
#>  [936,]  0.8586670129
#>  [937,] -0.8078143291
#>  [938,]  1.1776193794
#>  [939,]  0.2418576279
#>  [940,]  0.3335424183
#>  [941,]  0.1285061590
#>  [942,]  1.3411691849
#>  [943,] -0.5559034024
#>  [944,]  0.1590959420
#>  [945,] -0.6951256650
#>  [946,] -0.2605413739
#>  [947,] -0.5783492469
#>  [948,]  0.3986691761
#>  [949,] -0.1698461841
#>  [950,]  0.3858948403
#>  [951,]  0.2774495262
#>  [952,]  0.9474472543
#>  [953,]  0.2077086640
#>  [954,]  0.2806206151
#>  [955,] -0.1963288230
#>  [956,] -0.8031052176
#>  [957,] -0.2807516689
#>  [958,] -0.3026777028
#>  [959,]  0.5083306761
#>  [960,] -0.2461716835
#>  [961,] -0.4289264617
#>  [962,]  0.7402283775
#>  [963,]  0.4969502832
#>  [964,] -0.2984628662
#>  [965,]  1.1478227038
#>  [966,]  0.4754709249
#>  [967,]  0.3991731054
#>  [968,] -0.9194334524
#>  [969,]  0.1032756551
#>  [970,]  0.8412846453
#>  [971,]  0.0857675933
#>  [972,] -0.0358834420
#>  [973,] -0.6591447420
#>  [974,] -0.0544416064
#>  [975,]  0.0107341666
#>  [976,] -1.0040242470
#>  [977,]  0.2431192539
#>  [978,] -0.2085788371
#>  [979,] -0.5573513809
#>  [980,] -0.0040653231
#>  [981,] -0.1772535627
#>  [982,] -0.1807462099
#>  [983,] -0.7999498494
#>  [984,]  0.6382018936
#>  [985,] -0.7711527131
#>  [986,] -0.5226049467
#>  [987,]  0.0978378885
#>  [988,] -0.2031567217
#>  [989,] -0.0101076878
#>  [990,] -0.1496906807
#>  [991,] -0.2085568221
#>  [992,]  0.6977879109
#>  [993,]  0.2173127207
#>  [994,]  0.4405165300
#>  [995,] -0.0108882877
#>  [996,]  0.1153967578
#>  [997,] -0.3793987382
#>  [998,] -0.6079901381
#>  [999,]  0.6614728418
#> [1000,] -0.0739633480