Euclidean Norm

\(\|x\|_2 = \left(\sum_{i=1}^n x_i^2\right)^{1/2}\).

norm2(x, axis = NA_real_, keepdims = FALSE)

Arguments

x
An Expression, vector, or matrix.

axis
(Optional) The dimension across which to apply the function: 1 indicates rows, 2 indicates columns, and NA indicates rows and columns. The default is NA.

keepdims
(Optional) Should dimensions be maintained when applying the atom along an axis? If FALSE, result will be collapsed into an \(n x 1\) column vector. The default is FALSE.

Value

An Expression representing the Euclidean norm of the input.

Examples

a <- Variable()
prob <- Problem(Minimize(norm2(a)), list(a <= -2))
result <- solve(prob)

#> Problem
#>   Name                   :                 
#>   Objective sense        : min             
#>   Type                   : CONIC (conic optimization problem)
#>   Constraints            : 3               
#>   Cones                  : 1               
#>   Scalar variables       : 4               
#>   Matrix variables       : 0               
#>   Integer variables      : 0               
#> 
#> Optimizer started.
#> Presolve started.
#> Linear dependency checker started.
#> Linear dependency checker terminated.
#> Eliminator started.
#> Freed constraints in eliminator : 1
#> Eliminator terminated.
#> Eliminator started.
#> Freed constraints in eliminator : 0
#> Eliminator terminated.
#> Eliminator - tries                  : 2                 time                   : 0.00            
#> Lin. dep.  - tries                  : 1                 time                   : 0.00            
#> Lin. dep.  - number                 : 0               
#> Presolve terminated. Time: 0.00    
#> Problem
#>   Name                   :                 
#>   Objective sense        : min             
#>   Type                   : CONIC (conic optimization problem)
#>   Constraints            : 3               
#>   Cones                  : 1               
#>   Scalar variables       : 4               
#>   Matrix variables       : 0               
#>   Integer variables      : 0               
#> 
#> Optimizer  - threads                : 8               
#> Optimizer  - solved problem         : the primal      
#> Optimizer  - Constraints            : 0
#> Optimizer  - Cones                  : 1
#> Optimizer  - Scalar variables       : 2                 conic                  : 2               
#> Optimizer  - Semi-definite variables: 0                 scalarized             : 0               
#> Factor     - setup time             : 0.00              dense det. time        : 0.00            
#> Factor     - ML order time          : 0.00              GP order time          : 0.00            
#> Factor     - nonzeros before factor : 0                 after factor           : 0               
#> Factor     - dense dim.             : 0                 flops                  : 0.00e+00        
#> ITE PFEAS    DFEAS    GFEAS    PRSTATUS   POBJ              DOBJ              MU       TIME  
#> 0   3.0e+00  0.0e+00  2.0e+00  0.00e+00   1.000000000e+00   0.000000000e+00   1.0e+00  0.00  
#> 1   6.4e-01  3.3e-16  4.0e-01  -5.00e-01  1.698109875e+00   1.822469528e+00   2.1e-01  0.00  
#> 2   4.8e-02  1.1e-16  5.0e-03  1.05e+00   1.991877841e+00   1.962906966e+00   1.6e-02  0.00  
#> 3   1.8e-05  1.2e-14  2.5e-08  1.00e+00   1.999998804e+00   1.999983751e+00   5.9e-06  0.00  
#> 4   6.1e-11  2.2e-11  8.6e-14  1.00e+00   2.000000000e+00   2.000000000e+00   2.2e-11  0.00  
#> Optimizer terminated. Time: 0.00    
#> 
#> 
#> Interior-point solution summary
#>   Problem status  : PRIMAL_AND_DUAL_FEASIBLE
#>   Solution status : OPTIMAL
#>   Primal.  obj: 2.0000000000e+00    nrm: 2e+00    Viol.  con: 0e+00    var: 0e+00    cones: 0e+00  
#>   Dual.    obj: 2.0000000000e+00    nrm: 1e+00    Viol.  con: 0e+00    var: 0e+00    cones: 0e+00  
#> Optimizer summary
#>   Optimizer                 -                        time: 0.00    
#>     Interior-point          - iterations : 4         time: 0.00    
#>       Basis identification  -                        time: 0.00    
#>         Primal              - iterations : 0         time: 0.00    
#>         Dual                - iterations : 0         time: 0.00    
#>         Clean primal        - iterations : 0         time: 0.00    
#>         Clean dual          - iterations : 0         time: 0.00    
#>     Simplex                 -                        time: 0.00    
#>       Primal simplex        - iterations : 0         time: 0.00    
#>       Dual simplex          - iterations : 0         time: 0.00    
#>     Mixed integer           - relaxations: 0         time: 0.00    
#> 

result$value

#> [1] 2

result$getValue(a)

#> [1] -2

prob <- Problem(Maximize(-norm2(a)), list(a <= -2))
result <- solve(prob)

#> Problem
#>   Name                   :                 
#>   Objective sense        : min             
#>   Type                   : CONIC (conic optimization problem)
#>   Constraints            : 3               
#>   Cones                  : 1               
#>   Scalar variables       : 4               
#>   Matrix variables       : 0               
#>   Integer variables      : 0               
#> 
#> Optimizer started.
#> Presolve started.
#> Linear dependency checker started.
#> Linear dependency checker terminated.
#> Eliminator started.
#> Freed constraints in eliminator : 1
#> Eliminator terminated.
#> Eliminator started.
#> Freed constraints in eliminator : 0
#> Eliminator terminated.
#> Eliminator - tries                  : 2                 time                   : 0.00            
#> Lin. dep.  - tries                  : 1                 time                   : 0.00            
#> Lin. dep.  - number                 : 0               
#> Presolve terminated. Time: 0.00    
#> Problem
#>   Name                   :                 
#>   Objective sense        : min             
#>   Type                   : CONIC (conic optimization problem)
#>   Constraints            : 3               
#>   Cones                  : 1               
#>   Scalar variables       : 4               
#>   Matrix variables       : 0               
#>   Integer variables      : 0               
#> 
#> Optimizer  - threads                : 8               
#> Optimizer  - solved problem         : the primal      
#> Optimizer  - Constraints            : 0
#> Optimizer  - Cones                  : 1
#> Optimizer  - Scalar variables       : 2                 conic                  : 2               
#> Optimizer  - Semi-definite variables: 0                 scalarized             : 0               
#> Factor     - setup time             : 0.00              dense det. time        : 0.00            
#> Factor     - ML order time          : 0.00              GP order time          : 0.00            
#> Factor     - nonzeros before factor : 0                 after factor           : 0               
#> Factor     - dense dim.             : 0                 flops                  : 0.00e+00        
#> ITE PFEAS    DFEAS    GFEAS    PRSTATUS   POBJ              DOBJ              MU       TIME  
#> 0   3.0e+00  0.0e+00  2.0e+00  0.00e+00   1.000000000e+00   0.000000000e+00   1.0e+00  0.00  
#> 1   6.4e-01  3.3e-16  4.0e-01  -5.00e-01  1.698109875e+00   1.822469528e+00   2.1e-01  0.00  
#> 2   4.8e-02  1.1e-16  5.0e-03  1.05e+00   1.991877841e+00   1.962906966e+00   1.6e-02  0.00  
#> 3   1.8e-05  1.2e-14  2.5e-08  1.00e+00   1.999998804e+00   1.999983751e+00   5.9e-06  0.00  
#> 4   6.1e-11  2.2e-11  8.6e-14  1.00e+00   2.000000000e+00   2.000000000e+00   2.2e-11  0.00  
#> Optimizer terminated. Time: 0.00    
#> 
#> 
#> Interior-point solution summary
#>   Problem status  : PRIMAL_AND_DUAL_FEASIBLE
#>   Solution status : OPTIMAL
#>   Primal.  obj: 2.0000000000e+00    nrm: 2e+00    Viol.  con: 0e+00    var: 0e+00    cones: 0e+00  
#>   Dual.    obj: 2.0000000000e+00    nrm: 1e+00    Viol.  con: 0e+00    var: 0e+00    cones: 0e+00  
#> Optimizer summary
#>   Optimizer                 -                        time: 0.00    
#>     Interior-point          - iterations : 4         time: 0.00    
#>       Basis identification  -                        time: 0.00    
#>         Primal              - iterations : 0         time: 0.00    
#>         Dual                - iterations : 0         time: 0.00    
#>         Clean primal        - iterations : 0         time: 0.00    
#>         Clean dual          - iterations : 0         time: 0.00    
#>     Simplex                 -                        time: 0.00    
#>       Primal simplex        - iterations : 0         time: 0.00    
#>       Dual simplex          - iterations : 0         time: 0.00    
#>     Mixed integer           - relaxations: 0         time: 0.00    
#> 

result$value

#> [1] -2

result$getValue(a)

#> [1] -2

x <- Variable(2)
z <- Variable(2)
prob <- Problem(Minimize(norm2(x - z) + 5), list(x >= c(2,3), z <= c(-1,-4)))
result <- solve(prob)

#> Problem
#>   Name                   :                 
#>   Objective sense        : min             
#>   Type                   : CONIC (conic optimization problem)
#>   Constraints            : 7               
#>   Cones                  : 1               
#>   Scalar variables       : 8               
#>   Matrix variables       : 0               
#>   Integer variables      : 0               
#> 
#> Optimizer started.
#> Presolve started.
#> Eliminator started.
#> Freed constraints in eliminator : 0
#> Eliminator terminated.
#> Eliminator - tries                  : 1                 time                   : 0.00            
#> Lin. dep.  - tries                  : 0                 time                   : 0.00            
#> Lin. dep.  - number                 : 0               
#> Presolve terminated. Time: 0.00    
#> Problem
#>   Name                   :                 
#>   Objective sense        : min             
#>   Type                   : CONIC (conic optimization problem)
#>   Constraints            : 7               
#>   Cones                  : 1               
#>   Scalar variables       : 8               
#>   Matrix variables       : 0               
#>   Integer variables      : 0               
#> 
#> Optimizer  - threads                : 8               
#> Optimizer  - solved problem         : the primal      
#> Optimizer  - Constraints            : 0
#> Optimizer  - Cones                  : 1
#> Optimizer  - Scalar variables       : 3                 conic                  : 3               
#> Optimizer  - Semi-definite variables: 0                 scalarized             : 0               
#> Factor     - setup time             : 0.00              dense det. time        : 0.00            
#> Factor     - ML order time          : 0.00              GP order time          : 0.00            
#> Factor     - nonzeros before factor : 0                 after factor           : 0               
#> Factor     - dense dim.             : 0                 flops                  : 0.00e+00        
#> ITE PFEAS    DFEAS    GFEAS    PRSTATUS   POBJ              DOBJ              MU       TIME  
#> 0   8.0e+00  0.0e+00  2.0e+00  0.00e+00   1.000000000e+00   0.000000000e+00   1.0e+00  0.00  
#> 1   2.3e+00  9.4e-16  1.1e+00  -1.19e+00  4.351249225e+00   6.224848237e+00   2.9e-01  0.00  
#> 2   1.1e+00  6.7e-16  1.0e-01  6.56e-01   6.275695273e+00   5.697578120e+00   1.3e-01  0.00  
#> 3   1.2e-01  7.8e-16  3.3e-03  7.96e-01   7.513204796e+00   7.439239084e+00   1.5e-02  0.00  
#> 4   7.5e-03  3.4e-15  6.0e-05  1.04e+00   7.612044255e+00   7.607975562e+00   9.3e-04  0.00  
#> 5   6.5e-04  7.8e-14  1.6e-06  1.00e+00   7.615085348e+00   7.614732520e+00   8.2e-05  0.00  
#> 6   5.0e-05  8.4e-13  3.5e-08  1.00e+00   7.615708313e+00   7.615682367e+00   6.2e-06  0.00  
#> 7   1.4e-06  1.7e-11  1.6e-10  1.00e+00   7.615771155e+00   7.615770451e+00   1.7e-07  0.00  
#> 8   9.7e-09  3.5e-10  9.8e-14  1.00e+00   7.615773091e+00   7.615773086e+00   1.2e-09  0.00  
#> Optimizer terminated. Time: 0.00    
#> 
#> 
#> Interior-point solution summary
#>   Problem status  : PRIMAL_AND_DUAL_FEASIBLE
#>   Solution status : OPTIMAL
#>   Primal.  obj: 7.6157730914e+00    nrm: 8e+00    Viol.  con: 2e-08    var: 0e+00    cones: 0e+00  
#>   Dual.    obj: 7.6157730864e+00    nrm: 1e+00    Viol.  con: 0e+00    var: 1e-09    cones: 0e+00  
#> Optimizer summary
#>   Optimizer                 -                        time: 0.00    
#>     Interior-point          - iterations : 8         time: 0.00    
#>       Basis identification  -                        time: 0.00    
#>         Primal              - iterations : 0         time: 0.00    
#>         Dual                - iterations : 0         time: 0.00    
#>         Clean primal        - iterations : 0         time: 0.00    
#>         Clean dual          - iterations : 0         time: 0.00    
#>     Simplex                 -                        time: 0.00    
#>       Primal simplex        - iterations : 0         time: 0.00    
#>       Dual simplex          - iterations : 0         time: 0.00    
#>     Mixed integer           - relaxations: 0         time: 0.00    
#> 

result$value

#> [1] 12.61577

result$getValue(x)

#>      [,1]
#> [1,]    2
#> [2,]    3

result$getValue(z)

#>      [,1]
#> [1,]   -1
#> [2,]   -4

prob <- Problem(Minimize(norm2(t(x - z)) + 5), list(x >= c(2,3), z <= c(-1,-4)))
result <- solve(prob)

#> Problem
#>   Name                   :                 
#>   Objective sense        : min             
#>   Type                   : CONIC (conic optimization problem)
#>   Constraints            : 7               
#>   Cones                  : 1               
#>   Scalar variables       : 8               
#>   Matrix variables       : 0               
#>   Integer variables      : 0               
#> 
#> Optimizer started.
#> Presolve started.
#> Eliminator started.
#> Freed constraints in eliminator : 0
#> Eliminator terminated.
#> Eliminator - tries                  : 1                 time                   : 0.00            
#> Lin. dep.  - tries                  : 0                 time                   : 0.00            
#> Lin. dep.  - number                 : 0               
#> Presolve terminated. Time: 0.00    
#> Problem
#>   Name                   :                 
#>   Objective sense        : min             
#>   Type                   : CONIC (conic optimization problem)
#>   Constraints            : 7               
#>   Cones                  : 1               
#>   Scalar variables       : 8               
#>   Matrix variables       : 0               
#>   Integer variables      : 0               
#> 
#> Optimizer  - threads                : 8               
#> Optimizer  - solved problem         : the primal      
#> Optimizer  - Constraints            : 0
#> Optimizer  - Cones                  : 1
#> Optimizer  - Scalar variables       : 3                 conic                  : 3               
#> Optimizer  - Semi-definite variables: 0                 scalarized             : 0               
#> Factor     - setup time             : 0.00              dense det. time        : 0.00            
#> Factor     - ML order time          : 0.00              GP order time          : 0.00            
#> Factor     - nonzeros before factor : 0                 after factor           : 0               
#> Factor     - dense dim.             : 0                 flops                  : 0.00e+00        
#> ITE PFEAS    DFEAS    GFEAS    PRSTATUS   POBJ              DOBJ              MU       TIME  
#> 0   8.0e+00  0.0e+00  2.0e+00  0.00e+00   1.000000000e+00   0.000000000e+00   1.0e+00  0.00  
#> 1   2.3e+00  9.4e-16  1.1e+00  -1.19e+00  4.351249225e+00   6.224848237e+00   2.9e-01  0.00  
#> 2   1.1e+00  6.7e-16  1.0e-01  6.56e-01   6.275695273e+00   5.697578120e+00   1.3e-01  0.00  
#> 3   1.2e-01  7.8e-16  3.3e-03  7.96e-01   7.513204796e+00   7.439239084e+00   1.5e-02  0.00  
#> 4   7.5e-03  3.4e-15  6.0e-05  1.04e+00   7.612044255e+00   7.607975562e+00   9.3e-04  0.00  
#> 5   6.5e-04  7.8e-14  1.6e-06  1.00e+00   7.615085348e+00   7.614732520e+00   8.2e-05  0.00  
#> 6   5.0e-05  8.4e-13  3.5e-08  1.00e+00   7.615708313e+00   7.615682367e+00   6.2e-06  0.00  
#> 7   1.4e-06  1.7e-11  1.6e-10  1.00e+00   7.615771155e+00   7.615770451e+00   1.7e-07  0.00  
#> 8   9.7e-09  3.5e-10  9.8e-14  1.00e+00   7.615773091e+00   7.615773086e+00   1.2e-09  0.00  
#> Optimizer terminated. Time: 0.00    
#> 
#> 
#> Interior-point solution summary
#>   Problem status  : PRIMAL_AND_DUAL_FEASIBLE
#>   Solution status : OPTIMAL
#>   Primal.  obj: 7.6157730914e+00    nrm: 8e+00    Viol.  con: 2e-08    var: 0e+00    cones: 0e+00  
#>   Dual.    obj: 7.6157730864e+00    nrm: 1e+00    Viol.  con: 0e+00    var: 1e-09    cones: 0e+00  
#> Optimizer summary
#>   Optimizer                 -                        time: 0.00    
#>     Interior-point          - iterations : 8         time: 0.00    
#>       Basis identification  -                        time: 0.00    
#>         Primal              - iterations : 0         time: 0.00    
#>         Dual                - iterations : 0         time: 0.00    
#>         Clean primal        - iterations : 0         time: 0.00    
#>         Clean dual          - iterations : 0         time: 0.00    
#>     Simplex                 -                        time: 0.00    
#>       Primal simplex        - iterations : 0         time: 0.00    
#>       Dual simplex          - iterations : 0         time: 0.00    
#>     Mixed integer           - relaxations: 0         time: 0.00    
#> 

result$value

#> [1] 12.61577

result$getValue(x)

#>      [,1]
#> [1,]    2
#> [2,]    3

result$getValue(z)

#>      [,1]
#> [1,]   -1
#> [2,]   -4