P-Norm

The vector p-norm. If given a matrix variable, p_norm will treat it as a vector and compute the p-norm of the concatenated columns.

p_norm(x, p = 2, axis = NA_real_, keepdims = FALSE, max_denom = 1024)

Arguments

x
An Expression, vector, or matrix.

p
A number greater than or equal to 1, or equal to positive infinity.

axis
(Optional) The dimension across which to apply the function: 1 indicates rows, 2 indicates columns, and NA indicates rows and columns. The default is NA.

keepdims
(Optional) Should dimensions be maintained when applying the atom along an axis? If FALSE, result will be collapsed into an \(n x 1\) column vector. The default is FALSE.

max_denom
(Optional) The maximum denominator considered in forming a rational approximation for \(p\). The default is 1024.

Value

An Expression representing the p-norm of the input.

Details

For \(p \geq 1\), the p-norm is given by $$\|x\|_p = \left(\sum_{i=1}^n |x_i|^p\right)^{1/p}$$ with domain \(x \in \mathbf{R}^n\). For \(p < 1, p \neq 0\), the p-norm is given by $$\|x\|_p = \left(\sum_{i=1}^n x_i^p\right)^{1/p}$$ with domain \(x \in \mathbf{R}^n_+\).

Note that the "p-norm" is actually a norm only when \(p \geq 1\) or \(p = +\infty\). For these cases, it is convex.

The expression is undefined when \(p = 0\).

Otherwise, when \(p < 1\), the expression is concave, but not a true norm.

Examples

x <- Variable(3)
prob <- Problem(Minimize(p_norm(x,2)))
result <- solve(prob)

#> Problem
#>   Name                   :                 
#>   Objective sense        : min             
#>   Type                   : CONIC (conic optimization problem)
#>   Constraints            : 4               
#>   Cones                  : 1               
#>   Scalar variables       : 8               
#>   Matrix variables       : 0               
#>   Integer variables      : 0               
#> 
#> Optimizer started.
#> Presolve started.
#> Eliminator started.
#> Freed constraints in eliminator : 0
#> Eliminator terminated.
#> Eliminator started.
#> Freed constraints in eliminator : 0
#> Eliminator terminated.
#> Eliminator - tries                  : 2                 time                   : 0.00            
#> Lin. dep.  - tries                  : 0                 time                   : 0.00            
#> Lin. dep.  - number                 : 0               
#> Presolve terminated. Time: 0.00    
#> Optimizer terminated. Time: 0.00    
#> 
#> 
#> Interior-point solution summary
#>   Problem status  : PRIMAL_AND_DUAL_FEASIBLE
#>   Solution status : OPTIMAL
#>   Primal.  obj: 0.0000000000e+00    nrm: 0e+00    Viol.  con: 0e+00    var: 0e+00    cones: 0e+00  
#>   Dual.    obj: 0.0000000000e+00    nrm: 1e+00    Viol.  con: 0e+00    var: 0e+00    cones: 0e+00  
#> Optimizer summary
#>   Optimizer                 -                        time: 0.00    
#>     Interior-point          - iterations : 0         time: 0.00    
#>       Basis identification  -                        time: 0.00    
#>         Primal              - iterations : 0         time: 0.00    
#>         Dual                - iterations : 0         time: 0.00    
#>         Clean primal        - iterations : 0         time: 0.00    
#>         Clean dual          - iterations : 0         time: 0.00    
#>     Simplex                 -                        time: 0.00    
#>       Primal simplex        - iterations : 0         time: 0.00    
#>       Dual simplex          - iterations : 0         time: 0.00    
#>     Mixed integer           - relaxations: 0         time: 0.00    
#> 

result$value

#> [1] 0

result$getValue(x)

#>      [,1]
#> [1,]    0
#> [2,]    0
#> [3,]    0

prob <- Problem(Minimize(p_norm(x,Inf)))
result <- solve(prob)
result$value

#> [1] -2.775514e-23

result$getValue(x)

#>               [,1]
#> [1,] -3.231174e-27
#> [2,]  0.000000e+00
#> [3,]  0.000000e+00

a <- c(1.0, 2, 3)
prob <- Problem(Minimize(p_norm(x,1.6)), list(t(x) %*% a >= 1))
result <- solve(prob)

#> Problem
#>   Name                   :                 
#>   Objective sense        : min             
#>   Type                   : CONIC (conic optimization problem)
#>   Constraints            : 476             
#>   Cones                  : 156             
#>   Scalar variables       : 631             
#>   Matrix variables       : 0               
#>   Integer variables      : 0               
#> 
#> Optimizer started.
#> Presolve started.
#> Linear dependency checker started.
#> Linear dependency checker terminated.
#> Eliminator started.
#> Freed constraints in eliminator : 6
#> Eliminator terminated.
#> Eliminator - tries                  : 1                 time                   : 0.00            
#> Lin. dep.  - tries                  : 1                 time                   : 0.00            
#> Lin. dep.  - number                 : 0               
#> Presolve terminated. Time: 0.00    
#> Problem
#>   Name                   :                 
#>   Objective sense        : min             
#>   Type                   : CONIC (conic optimization problem)
#>   Constraints            : 476             
#>   Cones                  : 156             
#>   Scalar variables       : 631             
#>   Matrix variables       : 0               
#>   Integer variables      : 0               
#> 
#> Optimizer  - threads                : 8               
#> Optimizer  - solved problem         : the primal      
#> Optimizer  - Constraints            : 311
#> Optimizer  - Cones                  : 157
#> Optimizer  - Scalar variables       : 471               conic                  : 470             
#> Optimizer  - Semi-definite variables: 0                 scalarized             : 0               
#> Factor     - setup time             : 0.00              dense det. time        : 0.00            
#> Factor     - ML order time          : 0.00              GP order time          : 0.00            
#> Factor     - nonzeros before factor : 1.73e+04          after factor           : 2.21e+04        
#> Factor     - dense dim.             : 0                 flops                  : 2.00e+06        
#> ITE PFEAS    DFEAS    GFEAS    PRSTATUS   POBJ              DOBJ              MU       TIME  
#> 0   3.0e+00  1.0e+00  1.0e+00  0.00e+00   0.000000000e+00   0.000000000e+00   1.0e+00  0.00  
#> 1   7.6e-01  2.5e-01  5.0e-02  1.72e+00   1.939900787e-01   1.168975544e-01   2.5e-01  0.00  
#> 2   8.0e-02  2.7e-02  3.6e-03  1.65e+00   3.765836668e-01   3.832238457e-01   2.7e-02  0.00  
#> 3   2.1e-02  7.1e-03  3.8e-04  1.58e+00   3.035311772e-01   3.039497835e-01   7.1e-03  0.01  
#> 4   3.4e-03  1.1e-03  1.5e-05  1.11e+00   2.956190469e-01   2.954341428e-01   1.1e-03  0.01  
#> 5   6.8e-04  2.3e-04  1.2e-06  9.99e-01   2.946629593e-01   2.946171318e-01   2.3e-04  0.01  
#> 6   8.8e-05  2.9e-05  5.3e-08  9.80e-01   2.944625759e-01   2.944561960e-01   2.9e-05  0.01  
#> 7   2.0e-05  6.8e-06  6.0e-09  9.75e-01   2.944194028e-01   2.944178989e-01   6.8e-06  0.01  
#> 8   1.7e-06  5.7e-07  1.5e-10  9.77e-01   2.944068359e-01   2.944067093e-01   5.7e-07  0.01  
#> 9   2.9e-07  9.6e-08  1.0e-11  9.60e-01   2.944050436e-01   2.944050216e-01   9.6e-08  0.01  
#> 10  2.3e-08  5.9e-09  1.4e-13  9.84e-01   2.944048264e-01   2.944048250e-01   5.8e-09  0.01  
#> Optimizer terminated. Time: 0.01    
#> 
#> 
#> Interior-point solution summary
#>   Problem status  : PRIMAL_AND_DUAL_FEASIBLE
#>   Solution status : OPTIMAL
#>   Primal.  obj: 2.9440482644e-01    nrm: 1e+00    Viol.  con: 8e-09    var: 0e+00    cones: 0e+00  
#>   Dual.    obj: 2.9440482499e-01    nrm: 9e-01    Viol.  con: 8e-10    var: 8e-10    cones: 0e+00  
#> Optimizer summary
#>   Optimizer                 -                        time: 0.01    
#>     Interior-point          - iterations : 10        time: 0.01    
#>       Basis identification  -                        time: 0.00    
#>         Primal              - iterations : 0         time: 0.00    
#>         Dual                - iterations : 0         time: 0.00    
#>         Clean primal        - iterations : 0         time: 0.00    
#>         Clean dual          - iterations : 0         time: 0.00    
#>     Simplex                 -                        time: 0.00    
#>       Primal simplex        - iterations : 0         time: 0.00    
#>       Dual simplex          - iterations : 0         time: 0.00    
#>     Mixed integer           - relaxations: 0         time: 0.00    
#> 

result$value

#> [1] 0.2944048

result$getValue(x)

#>            [,1]
#> [1,] 0.03835778
#> [2,] 0.12177801
#> [3,] 0.23936207

prob <- Problem(Minimize(sum(abs(x - a))), list(p_norm(x,-1) >= 0))
result <- solve(prob)

#> Problem
#>   Name                   :                 
#>   Objective sense        : min             
#>   Type                   : CONIC (conic optimization problem)
#>   Constraints            : 17              
#>   Cones                  : 3               
#>   Scalar variables       : 19              
#>   Matrix variables       : 0               
#>   Integer variables      : 0               
#> 
#> Optimizer started.
#> Presolve started.
#> Linear dependency checker started.
#> Linear dependency checker terminated.
#> Eliminator started.
#> Freed constraints in eliminator : 9
#> Eliminator terminated.
#> Eliminator started.
#> Freed constraints in eliminator : 0
#> Eliminator terminated.
#> Eliminator - tries                  : 2                 time                   : 0.00            
#> Lin. dep.  - tries                  : 1                 time                   : 0.00            
#> Lin. dep.  - number                 : 0               
#> Presolve terminated. Time: 0.00    
#> Problem
#>   Name                   :                 
#>   Objective sense        : min             
#>   Type                   : CONIC (conic optimization problem)
#>   Constraints            : 17              
#>   Cones                  : 3               
#>   Scalar variables       : 19              
#>   Matrix variables       : 0               
#>   Integer variables      : 0               
#> 
#> Optimizer  - threads                : 8               
#> Optimizer  - solved problem         : the primal      
#> Optimizer  - Constraints            : 6
#> Optimizer  - Cones                  : 3
#> Optimizer  - Scalar variables       : 15                conic                  : 9               
#> Optimizer  - Semi-definite variables: 0                 scalarized             : 0               
#> Factor     - setup time             : 0.00              dense det. time        : 0.00            
#> Factor     - ML order time          : 0.00              GP order time          : 0.00            
#> Factor     - nonzeros before factor : 16                after factor           : 16              
#> Factor     - dense dim.             : 0                 flops                  : 1.88e+02        
#> ITE PFEAS    DFEAS    GFEAS    PRSTATUS   POBJ              DOBJ              MU       TIME  
#> 0   3.0e+00  1.0e+00  2.5e+00  0.00e+00   1.500000000e+00   0.000000000e+00   1.0e+00  0.00  
#> 1   9.3e-01  3.1e-01  5.5e-01  -1.52e-01  2.177204806e+00   1.640277273e+00   3.1e-01  0.00  
#> 2   3.2e-01  1.1e-01  8.4e-02  1.22e+00   2.964547180e-01   1.022975930e-01   1.1e-01  0.00  
#> 3   2.3e-02  7.7e-03  1.2e-03  1.13e+00   8.830034873e-03   -6.678534467e-03  7.7e-03  0.00  
#> 4   4.7e-05  1.6e-05  9.3e-08  1.05e+00   -2.993614358e-07  -3.325307641e-05  1.6e-05  0.00  
#> 5   1.9e-13  6.6e-14  2.9e-20  1.00e+00   -7.086981226e-14  -1.996394403e-13  7.0e-14  0.00  
#> Optimizer terminated. Time: 0.00    
#> 
#> 
#> Interior-point solution summary
#>   Problem status  : PRIMAL_AND_DUAL_FEASIBLE
#>   Solution status : OPTIMAL
#>   Primal.  obj: -6.9833028249e-14   nrm: 3e+00    Viol.  con: 2e-13    var: 0e+00    cones: 0e+00  
#>   Dual.    obj: -1.9939605522e-13   nrm: 5e-01    Viol.  con: 8e-14    var: 3e-14    cones: 0e+00  
#> Optimizer summary
#>   Optimizer                 -                        time: 0.00    
#>     Interior-point          - iterations : 5         time: 0.00    
#>       Basis identification  -                        time: 0.00    
#>         Primal              - iterations : 0         time: 0.00    
#>         Dual                - iterations : 0         time: 0.00    
#>         Clean primal        - iterations : 0         time: 0.00    
#>         Clean dual          - iterations : 0         time: 0.00    
#>     Simplex                 -                        time: 0.00    
#>       Primal simplex        - iterations : 0         time: 0.00    
#>       Dual simplex          - iterations : 0         time: 0.00    
#>     Mixed integer           - relaxations: 0         time: 0.00    
#> 

result$value

#> [1] -6.983303e-14

result$getValue(x)

#>      [,1]
#> [1,]    1
#> [2,]    2
#> [3,]    3