The Atom class.

This virtual class represents atomic expressions in CVXR.

# S4 method for Atom
name(x)

# S4 method for Atom
validate_args(object)

# S4 method for Atom
dim(x)

# S4 method for Atom
nrow(x)

# S4 method for Atom
ncol(x)

# S4 method for Atom
allow_complex(object)

# S4 method for Atom
is_nonneg(object)

# S4 method for Atom
is_nonpos(object)

# S4 method for Atom
is_imag(object)

# S4 method for Atom
is_complex(object)

# S4 method for Atom
is_convex(object)

# S4 method for Atom
is_concave(object)

# S4 method for Atom
is_log_log_convex(object)

# S4 method for Atom
is_log_log_concave(object)

# S4 method for Atom
canonicalize(object)

# S4 method for Atom
graph_implementation(object, arg_objs, dim, data = NA_real_)

# S4 method for Atom
value_impl(object)

# S4 method for Atom
value(object)

# S4 method for Atom
grad(object)

# S4 method for Atom
domain(object)

# S4 method for Atom
atoms(object)

Arguments

x, object
An Atom object.

arg_objs
A list of linear expressions for each argument.

dim
A vector with two elements representing the dimensions of the resulting expression.

data
A list of additional data required by the atom.

Methods (by generic)

name: Returns the string representtation of the function call

validate_args: Raises an error if the arguments are invalid.

dim: The c(row, col) dimensions of the atom.

nrow: The number of rows in the atom.

ncol: The number of columns in the atom.

allow_complex: Does the atom handle complex numbers?

is_nonneg: A logical value indicating whether the atom is nonnegative.

is_nonpos: A logical value indicating whether the atom is nonpositive.

is_imag: A logical value indicating whether the atom is imaginary.

is_complex: A logical value indicating whether the atom is complex valued.

is_convex: A logical value indicating whether the atom is convex.

is_concave: A logical value indicating whether the atom is concave.

is_log_log_convex: A logical value indicating whether the atom is log-log convex.

is_log_log_concave: A logical value indicating whether the atom is log-log concave.

canonicalize: Represent the atom as an affine objective and conic constraints.

graph_implementation: The graph implementation of the atom.

value_impl: Returns the value of each of the componets in an Atom. Returns an empty matrix if it's an empty atom

value: Returns the value of the atom.

grad: The (sub/super)-gradient of the atom with respect to each variable.

domain: A list of constraints describing the closure of the region where the expression is finite.

atoms: Returns a list of the atom types present amongst this atom's arguments