The PSDConstraint class.

This class represents the positive semidefinite constraint, \(\frac{1}{2}(X + X^T) \succeq 0\), i.e. \(z^T(X + X^T)z \geq 0\) for all \(z\).

e1 %>>% e2

e1 %<<% e2

# S4 method for Expression,Expression
%>>%(e1, e2)

# S4 method for Expression,ConstVal
%>>%(e1, e2)

# S4 method for ConstVal,Expression
%>>%(e1, e2)

# S4 method for Expression,Expression
%<<%(e1, e2)

# S4 method for Expression,ConstVal
%<<%(e1, e2)

# S4 method for ConstVal,Expression
%<<%(e1, e2)

PSDConstraint(expr, id = NA_integer_)

# S4 method for PSDConstraint
name(x)

# S4 method for PSDConstraint
is_dcp(object)

# S4 method for PSDConstraint
is_dgp(object)

# S4 method for PSDConstraint
residual(object)

# S4 method for PSDConstraint
canonicalize(object)

Arguments

e1, e2
The Expression objects or numeric constants to compare.

expr
An Expression, numeric element, vector, or matrix representing \(X\).

id
(Optional) A numeric value representing the constraint ID.

x, object
A PSDConstraint object.

Methods (by generic)

name: The string representation of the constraint.

is_dcp: The constraint is DCP if the left-hand and right-hand expressions are affine.

is_dgp: Is the constraint DGP?

residual: A Expression representing the residual of the constraint.

canonicalize: The graph implementation of the object. Marks the top level constraint as the dual_holder so the dual value will be saved to the PSDConstraint.

Slots

expr

An Expression, numeric element, vector, or matrix representing \(X\).