The ProdEntries class.

The product of the entries in an expression.

ProdEntries(..., axis = NA_real_, keepdims = FALSE)

# S4 method for ProdEntries
to_numeric(object, values)

# S4 method for ProdEntries
sign_from_args(object)

# S4 method for ProdEntries
is_atom_convex(object)

# S4 method for ProdEntries
is_atom_concave(object)

# S4 method for ProdEntries
is_atom_log_log_convex(object)

# S4 method for ProdEntries
is_atom_log_log_concave(object)

# S4 method for ProdEntries
is_incr(object, idx)

# S4 method for ProdEntries
is_decr(object, idx)

# S4 method for ProdEntries
.column_grad(object, value)

# S4 method for ProdEntries
.grad(object, values)

Arguments

...

Expression objects, vectors, or matrices.

axis

(Optional) The dimension across which to apply the function: 1 indicates rows, 2 indicates columns, and NA indicates rows and columns. The default is NA.

keepdims

(Optional) Should dimensions be maintained when applying the atom along an axis? If FALSE, result will be collapsed into an \(n x 1\) column vector. The default is FALSE.

object

A ProdEntries object.

values

A list of numeric values for the arguments

idx

An index into the atom.

value

A numeric value.

Methods (by generic)

to_numeric(ProdEntries): The product of all the entries.

sign_from_args(ProdEntries): Returns the sign (is positive, is negative) of the atom.

is_atom_convex(ProdEntries): Is the atom convex?

is_atom_concave(ProdEntries): Is the atom concave?

is_atom_log_log_convex(ProdEntries): Is the atom log-log convex?

is_atom_log_log_concave(ProdEntries): is the atom log-log concave?

is_incr(ProdEntries): Is the atom weakly increasing in the argument idx?

is_decr(ProdEntries): Is the atom weakly decreasing in the argument idx?

.column_grad(ProdEntries): Gives the (sub/super)gradient of the atom w.r.t. each column variable

.grad(ProdEntries): Gives the (sub/super)gradient of the atom w.r.t. each variable

Slots

expr

An Expression representing a vector or matrix.

axis

(Optional) The dimension across which to apply the function: 1 indicates rows, 2 indicates columns, and NA indicates rows and columns. The default is NA.