The SOCAxis class.

This class represents a second-order cone constraint for each row/column. It Assumes \(t\) is a vector the same length as \(X\)'s rows (columns) for axis == 1 (2).

SOCAxis(t, X, axis, id = NA_integer_)

# S4 method for SOCAxis
as.character(x)

# S4 method for SOCAxis
format_constr(object, eq_constr, leq_constr, dims, solver)

# S4 method for SOCAxis
num_cones(object)

# S4 method for SOCAxis
cone_sizes(object)

# S4 method for SOCAxis
size(object)

Arguments

t
The scalar part of the second-order constraint.

X
A matrix whose rows/columns are each a cone.

axis
The dimension across which to take the slice: 1 indicates rows, and 2 indicates columns.

id
(Optional) A numeric value representing the constraint ID.

x, object
A SOCAxis object.

eq_constr
A list of the equality constraints in the canonical problem.

leq_constr
A list of the inequality constraints in the canonical problem.

dims
A list with the dimensions of the conic constraints.

solver
A string representing the solver to be called.

Methods (by generic)

format_constr: Format SOC constraints as inequalities for the solver.

num_cones: The number of elementwise cones.

cone_sizes: The dimensions of a single cone.

size: The dimensions of the (elementwise) second-order cones.

Slots

t

The scalar part of the second-order constraint.

x_elems

A list containing X, a matrix whose rows/columns are each a cone.

axis

The dimension across which to take the slice: 1 indicates rows, and 2 indicates columns.