Logistic Function

The elementwise logistic function, \(\log(1 + e^x)\). This is a special case of log(sum(exp)) that evaluates to a vector rather than to a scalar, which is useful for logistic regression.

logistic(x)

Arguments

x
An Expression, vector, or matrix.

Value

An Expression representing the logistic function evaluated at the input.

Examples

set.seed(92)
n <- 20
m <- 1000
sigma <- 45

beta_true <- stats::rnorm(n)
idxs <- sample(n, size = 0.8*n, replace = FALSE)
beta_true[idxs] <- 0
X <- matrix(stats::rnorm(m*n, 0, 5), nrow = m, ncol = n)
y <- sign(X %*% beta_true + stats::rnorm(m, 0, sigma))

beta <- Variable(n)
X_sign <- apply(X, 2, function(x) { ifelse(y <= 0, -1, 1) * x })
obj <- -sum(logistic(-X[y <= 0,] %*% beta)) - sum(logistic(X[y == 1,] %*% beta))
prob <- Problem(Maximize(obj))
result <- solve(prob)

log_odds <- result$getValue(X %*% beta)
beta_res <- result$getValue(beta)
y_probs <- 1/(1 + exp(-X %*% beta_res))
log(y_probs/(1 - y_probs))

#>                  [,1]
#>    [1,] -0.6704453347
#>    [2,] -0.5925033262
#>    [3,] -0.2870242361
#>    [4,] -0.3903343198
#>    [5,]  0.1512509246
#>    [6,] -0.3212271628
#>    [7,]  0.0710781834
#>    [8,] -0.0005963006
#>    [9,]  0.6231436185
#>   [10,] -0.3461505031
#>   [11,]  0.6165806599
#>   [12,]  0.2493753853
#>   [13,] -0.4229390081
#>   [14,]  1.3550387471
#>   [15,] -0.3265572499
#>   [16,] -0.5973987668
#>   [17,]  1.4250714788
#>   [18,] -0.8063993529
#>   [19,]  0.6354090777
#>   [20,] -0.5313213727
#>   [21,] -0.5836735104
#>   [22,] -0.2018595928
#>   [23,] -1.5740813377
#>   [24,] -0.2996789322
#>   [25,] -0.2075290025
#>   [26,]  0.4436891439
#>   [27,] -0.5653336504
#>   [28,] -1.2550371454
#>   [29,] -0.5346170032
#>   [30,]  0.4833947605
#>   [31,]  0.2310983335
#>   [32,]  0.2894445891
#>   [33,] -0.0125295287
#>   [34,]  0.7295611435
#>   [35,] -0.8746954703
#>   [36,] -0.3928855432
#>   [37,] -0.1155672306
#>   [38,]  1.1672248946
#>   [39,] -0.7217336659
#>   [40,]  0.0611244077
#>   [41,] -0.6167553932
#>   [42,] -0.0245106354
#>   [43,]  0.6908123405
#>   [44,] -0.8034296241
#>   [45,]  0.8836939284
#>   [46,] -1.0940449324
#>   [47,]  0.0151126356
#>   [48,] -0.0365083647
#>   [49,]  0.1208087986
#>   [50,]  0.0228312756
#>   [51,]  0.0203872140
#>   [52,] -0.4370818739
#>   [53,]  0.0381306851
#>   [54,] -0.0670260234
#>   [55,] -0.5450365469
#>   [56,] -0.1197476669
#>   [57,] -0.1350405683
#>   [58,]  0.1775439787
#>   [59,]  1.1397700810
#>   [60,] -0.4142503787
#>   [61,]  0.3992908663
#>   [62,] -0.4202041955
#>   [63,]  0.0406917986
#>   [64,] -0.6488958378
#>   [65,] -0.9702738094
#>   [66,]  0.1117432319
#>   [67,] -0.1453537443
#>   [68,]  0.0616547778
#>   [69,]  0.0779094456
#>   [70,]  0.4237326796
#>   [71,] -0.4578263388
#>   [72,] -0.2049677903
#>   [73,] -0.2303856203
#>   [74,]  0.1166916169
#>   [75,] -0.3651869512
#>   [76,] -0.2369442935
#>   [77,]  0.1457200085
#>   [78,] -1.0896236685
#>   [79,] -0.1308683889
#>   [80,]  0.2291634537
#>   [81,]  0.1373235012
#>   [82,] -0.0015650744
#>   [83,]  0.3467507623
#>   [84,] -0.1543849352
#>   [85,] -0.2308961305
#>   [86,] -0.5247030447
#>   [87,]  0.4603636944
#>   [88,] -0.3126440594
#>   [89,]  0.2120011598
#>   [90,]  0.3146879497
#>   [91,]  0.4192965891
#>   [92,]  0.6556651118
#>   [93,] -0.5111044185
#>   [94,]  1.3013045675
#>   [95,]  0.8029989074
#>   [96,]  0.2830294738
#>   [97,] -0.6703857621
#>   [98,]  0.6391940833
#>   [99,]  0.0848847220
#>  [100,] -0.0064341407
#>  [101,]  0.1647444783
#>  [102,]  0.2973095413
#>  [103,]  0.3564136244
#>  [104,] -0.8130308889
#>  [105,] -0.5596483897
#>  [106,]  0.2440581040
#>  [107,] -0.5330443838
#>  [108,]  0.4528347608
#>  [109,] -0.5708845586
#>  [110,] -0.4623901458
#>  [111,] -0.2336832233
#>  [112,]  0.1511174345
#>  [113,] -0.0619013253
#>  [114,]  0.5838843988
#>  [115,] -0.0005904018
#>  [116,] -0.7089594667
#>  [117,] -0.3172647492
#>  [118,]  0.3970538029
#>  [119,] -1.5819604763
#>  [120,]  0.5686175587
#>  [121,] -0.3273326450
#>  [122,] -0.8145817766
#>  [123,] -0.0947978576
#>  [124,]  0.3203397287
#>  [125,] -0.4429070831
#>  [126,]  0.1170315568
#>  [127,] -0.7652074680
#>  [128,] -0.0573319188
#>  [129,]  0.0095952395
#>  [130,] -0.1932122096
#>  [131,]  0.1964148442
#>  [132,] -0.3175564629
#>  [133,]  0.4246144515
#>  [134,]  0.1266876551
#>  [135,]  0.3305578083
#>  [136,] -0.0061843022
#>  [137,] -0.5537507083
#>  [138,] -0.1133649373
#>  [139,] -0.7994460084
#>  [140,]  0.8081443601
#>  [141,]  0.4152997083
#>  [142,]  0.8942076829
#>  [143,] -0.3139394385
#>  [144,]  0.1458332522
#>  [145,] -0.2050248457
#>  [146,] -0.0601878245
#>  [147,] -1.1747855199
#>  [148,]  0.5454614851
#>  [149,]  0.2075639180
#>  [150,] -0.2618695631
#>  [151,]  0.8038733334
#>  [152,] -0.0455072758
#>  [153,]  0.5741226696
#>  [154,]  0.6919727361
#>  [155,]  0.0716668903
#>  [156,] -0.4965672484
#>  [157,]  0.4423821040
#>  [158,] -0.3816781239
#>  [159,]  0.5205926880
#>  [160,] -0.1021905749
#>  [161,] -0.2226354032
#>  [162,]  0.7098704398
#>  [163,]  0.5765147698
#>  [164,]  0.2450268201
#>  [165,] -0.1737067622
#>  [166,]  0.3589608620
#>  [167,] -0.3289995213
#>  [168,] -0.8993194777
#>  [169,] -0.8217806990
#>  [170,] -0.7797630368
#>  [171,] -0.2137157866
#>  [172,]  0.5276187589
#>  [173,] -0.8487622034
#>  [174,]  0.1034534920
#>  [175,] -0.4103533107
#>  [176,] -0.7528612966
#>  [177,] -0.4402291430
#>  [178,] -1.0685425694
#>  [179,]  0.4078444526
#>  [180,]  0.6313454205
#>  [181,]  0.3750808043
#>  [182,] -0.0151773208
#>  [183,] -0.5310740716
#>  [184,]  0.4077445652
#>  [185,]  0.1571528762
#>  [186,] -0.6328460104
#>  [187,] -0.3822479607
#>  [188,]  0.5205044282
#>  [189,]  0.2562664685
#>  [190,]  0.8532419819
#>  [191,] -0.5649047567
#>  [192,]  0.0370582991
#>  [193,] -0.3390711752
#>  [194,] -0.0375189430
#>  [195,]  0.0474701828
#>  [196,] -0.0326512691
#>  [197,]  0.2159940396
#>  [198,] -1.0662311100
#>  [199,] -0.9549719251
#>  [200,] -0.6453697425
#>  [201,]  0.5183243342
#>  [202,] -0.6994842476
#>  [203,]  0.2180911231
#>  [204,]  0.1893071187
#>  [205,]  0.4929788960
#>  [206,]  0.1844609174
#>  [207,]  0.3573840719
#>  [208,] -0.2500474091
#>  [209,]  1.3916851618
#>  [210,] -0.7617650344
#>  [211,]  0.5279142933
#>  [212,] -1.0117523217
#>  [213,] -0.2423680151
#>  [214,] -0.6133807654
#>  [215,] -1.1655975634
#>  [216,]  0.3821228532
#>  [217,] -0.4501122375
#>  [218,] -1.0973090754
#>  [219,]  0.8937125027
#>  [220,]  0.3140605903
#>  [221,]  0.1340493006
#>  [222,] -1.2230623401
#>  [223,]  0.5678078653
#>  [224,]  0.8463424950
#>  [225,]  0.6349024362
#>  [226,]  0.7005746849
#>  [227,] -0.7991994979
#>  [228,]  0.5683273780
#>  [229,]  0.4112913272
#>  [230,]  0.0902408533
#>  [231,] -0.3665086703
#>  [232,]  0.0127776610
#>  [233,]  0.1629133709
#>  [234,] -0.3527103812
#>  [235,] -0.0758193874
#>  [236,]  0.0091694550
#>  [237,] -0.4611102201
#>  [238,]  0.0657050399
#>  [239,] -0.5298472866
#>  [240,] -0.5962056366
#>  [241,]  0.0637607910
#>  [242,] -0.3510055419
#>  [243,]  0.5475924703
#>  [244,]  0.5265491214
#>  [245,] -0.8597965122
#>  [246,]  0.6395161581
#>  [247,]  0.3215859416
#>  [248,] -0.4795797320
#>  [249,]  0.3975118712
#>  [250,]  0.1215612676
#>  [251,] -0.3140228777
#>  [252,] -0.3046456703
#>  [253,]  0.3923103128
#>  [254,] -0.1856414048
#>  [255,] -0.0851592610
#>  [256,] -0.0029471854
#>  [257,]  0.5734643658
#>  [258,] -0.0359961444
#>  [259,] -0.9567746238
#>  [260,]  0.6395524374
#>  [261,]  0.4795642258
#>  [262,]  0.1872567596
#>  [263,] -0.2498838735
#>  [264,]  0.3260895114
#>  [265,]  0.1986580349
#>  [266,] -0.6790478120
#>  [267,]  0.6809372096
#>  [268,] -0.2952483929
#>  [269,]  0.5265701185
#>  [270,]  0.6812761039
#>  [271,] -0.4787134298
#>  [272,]  0.6944539424
#>  [273,]  0.3116134751
#>  [274,]  0.6770255038
#>  [275,]  0.0024164923
#>  [276,]  0.1779713618
#>  [277,]  0.1332630698
#>  [278,]  0.2881547033
#>  [279,] -0.8398067868
#>  [280,]  0.8292850904
#>  [281,]  0.3237866299
#>  [282,]  0.3688617559
#>  [283,] -0.2838370573
#>  [284,]  0.1698026725
#>  [285,] -0.0972784562
#>  [286,]  0.1692273882
#>  [287,]  0.6139191106
#>  [288,]  0.1736792102
#>  [289,] -0.0806333149
#>  [290,]  0.0115378218
#>  [291,] -0.3183805271
#>  [292,]  0.0866691819
#>  [293,] -0.7586172170
#>  [294,]  0.0584511072
#>  [295,] -0.1537339461
#>  [296,]  0.8435127736
#>  [297,]  0.8815233190
#>  [298,]  0.1534426032
#>  [299,]  1.2270988519
#>  [300,]  0.3098845844
#>  [301,] -0.2555223621
#>  [302,] -0.2073252683
#>  [303,] -0.2298546000
#>  [304,] -0.5772026710
#>  [305,] -0.4885279158
#>  [306,]  0.5318205876
#>  [307,] -0.1440895864
#>  [308,]  0.6375758538
#>  [309,]  0.2563528651
#>  [310,] -0.0603814616
#>  [311,] -0.2223125884
#>  [312,] -0.4950958463
#>  [313,] -0.2533095210
#>  [314,]  0.3508712647
#>  [315,]  0.0868122401
#>  [316,] -0.3995683909
#>  [317,] -0.2455178403
#>  [318,] -0.0987995925
#>  [319,]  0.6998107989
#>  [320,] -0.4848105052
#>  [321,] -0.3735351934
#>  [322,] -0.0385923050
#>  [323,]  0.0463439328
#>  [324,] -0.8583527115
#>  [325,] -0.4583106421
#>  [326,]  0.6281303288
#>  [327,] -0.6350815766
#>  [328,] -0.9581641027
#>  [329,]  0.9368312179
#>  [330,]  0.1483808768
#>  [331,] -0.0538224424
#>  [332,] -1.2322811374
#>  [333,] -0.1706880584
#>  [334,]  0.1205463820
#>  [335,] -0.4870755913
#>  [336,]  0.0540016845
#>  [337,] -0.1344821276
#>  [338,]  1.1543367271
#>  [339,]  0.8618837713
#>  [340,]  0.0228898162
#>  [341,] -0.5238893804
#>  [342,]  0.3351152543
#>  [343,] -0.7517029800
#>  [344,]  0.6050191980
#>  [345,] -0.0211553385
#>  [346,]  0.4027099531
#>  [347,]  0.8252682680
#>  [348,] -0.3307009358
#>  [349,]  0.3177218738
#>  [350,] -0.0224511326
#>  [351,] -0.5546639556
#>  [352,] -0.2831516821
#>  [353,] -0.1794628448
#>  [354,] -0.5933377907
#>  [355,]  0.5243736983
#>  [356,]  0.2004083999
#>  [357,]  0.2904632325
#>  [358,] -0.4817000886
#>  [359,] -0.1010029642
#>  [360,]  0.3605441981
#>  [361,]  0.3838495823
#>  [362,] -0.6909134092
#>  [363,] -0.3578611094
#>  [364,]  0.7571538146
#>  [365,]  0.4263300833
#>  [366,] -0.7291872632
#>  [367,]  1.3050823544
#>  [368,]  0.3406635859
#>  [369,]  0.1624588446
#>  [370,] -0.4134209783
#>  [371,]  0.8469619231
#>  [372,]  0.1445684872
#>  [373,] -1.0050931581
#>  [374,]  0.1813224235
#>  [375,]  0.3127918662
#>  [376,] -0.2938397701
#>  [377,] -1.1220530024
#>  [378,]  0.0153629172
#>  [379,]  0.6470195736
#>  [380,] -1.1136056906
#>  [381,] -0.6871635608
#>  [382,]  0.5672943352
#>  [383,] -0.2756557981
#>  [384,] -0.8617931819
#>  [385,] -0.1516807841
#>  [386,] -0.2098155181
#>  [387,] -1.2010110327
#>  [388,]  1.2558242729
#>  [389,] -0.5077999668
#>  [390,] -0.8335684689
#>  [391,]  0.8759112462
#>  [392,]  0.5860331464
#>  [393,] -1.2937757618
#>  [394,] -0.7939739920
#>  [395,] -0.5236258195
#>  [396,]  0.1775947238
#>  [397,] -1.4013158551
#>  [398,]  0.7131393681
#>  [399,]  0.0354480102
#>  [400,]  0.1344407543
#>  [401,]  0.4879663116
#>  [402,]  0.3637997202
#>  [403,] -0.1982373490
#>  [404,]  0.1426994803
#>  [405,]  0.0306165989
#>  [406,]  0.6818060774
#>  [407,]  0.0828430761
#>  [408,]  0.1285016870
#>  [409,]  0.7544342374
#>  [410,]  0.2249868028
#>  [411,]  0.0320392441
#>  [412,]  0.4893623450
#>  [413,]  0.3546844203
#>  [414,] -0.3652869328
#>  [415,]  0.8342500520
#>  [416,] -0.2503009131
#>  [417,]  0.5247153428
#>  [418,]  0.9032874501
#>  [419,]  0.6160097873
#>  [420,]  0.1392498710
#>  [421,]  0.7268128063
#>  [422,] -0.5110286542
#>  [423,] -0.6768615117
#>  [424,]  0.8344103909
#>  [425,]  0.3100125157
#>  [426,]  0.4325796525
#>  [427,]  0.1332976300
#>  [428,]  0.5181857250
#>  [429,] -0.3189950503
#>  [430,]  0.4315530076
#>  [431,]  0.3930177512
#>  [432,] -0.4950311571
#>  [433,] -0.7391120592
#>  [434,] -0.3018808056
#>  [435,] -0.2371839322
#>  [436,]  0.9647747222
#>  [437,] -0.5663492238
#>  [438,] -0.2494091035
#>  [439,] -0.8171679329
#>  [440,] -0.7060429892
#>  [441,] -0.5057521976
#>  [442,]  0.3789690687
#>  [443,] -0.9727434095
#>  [444,]  0.4803623718
#>  [445,]  0.6154845316
#>  [446,] -0.2193745775
#>  [447,]  0.2770402644
#>  [448,]  0.1446113232
#>  [449,]  0.7294546275
#>  [450,]  0.1318361544
#>  [451,] -0.9285746070
#>  [452,]  0.5649806808
#>  [453,] -0.0734972867
#>  [454,]  0.6055621101
#>  [455,]  0.5110814887
#>  [456,]  0.3522342164
#>  [457,] -0.7569807629
#>  [458,] -0.1832483549
#>  [459,] -0.7545137128
#>  [460,] -0.0289813132
#>  [461,]  0.7737892086
#>  [462,]  0.4915856086
#>  [463,] -0.3123237227
#>  [464,] -1.0281488382
#>  [465,]  0.2270953888
#>  [466,]  0.7107637759
#>  [467,] -0.3046677070
#>  [468,] -0.5839001367
#>  [469,]  1.4978781904
#>  [470,] -0.2798494190
#>  [471,]  0.0598967088
#>  [472,] -0.1166410597
#>  [473,]  0.5759075420
#>  [474,]  0.0739393456
#>  [475,]  0.5817677102
#>  [476,]  0.2869903788
#>  [477,] -0.1296553772
#>  [478,]  0.4669442246
#>  [479,] -0.0833626521
#>  [480,]  0.0444081667
#>  [481,] -0.2424424749
#>  [482,]  0.4827907250
#>  [483,]  0.3103862607
#>  [484,]  0.0108827236
#>  [485,]  0.7864241538
#>  [486,] -0.9775717255
#>  [487,] -0.3104691518
#>  [488,] -0.3305362956
#>  [489,] -0.2215519861
#>  [490,]  0.0721115971
#>  [491,]  0.2177784197
#>  [492,]  0.1394645249
#>  [493,]  0.2520048072
#>  [494,]  0.5967186867
#>  [495,]  0.3058341929
#>  [496,]  0.5917215092
#>  [497,]  0.3640586456
#>  [498,]  0.1549052732
#>  [499,] -0.4251273010
#>  [500,] -0.5954529701
#>  [501,]  0.1971876202
#>  [502,]  0.3262305365
#>  [503,]  0.2089549141
#>  [504,] -0.3834597414
#>  [505,]  0.4277870201
#>  [506,]  0.0551658217
#>  [507,] -0.0386648456
#>  [508,]  0.2733740708
#>  [509,] -0.5725964158
#>  [510,] -0.2647254937
#>  [511,] -0.1636669286
#>  [512,]  0.2262513127
#>  [513,] -0.0280529750
#>  [514,] -0.2338495196
#>  [515,]  0.6861805944
#>  [516,] -0.5119036512
#>  [517,] -0.0523599818
#>  [518,]  0.1044167901
#>  [519,] -0.1237383020
#>  [520,]  0.0757063377
#>  [521,] -0.5462248701
#>  [522,]  0.0475506691
#>  [523,]  1.1280535495
#>  [524,]  2.1538163845
#>  [525,] -0.3064752785
#>  [526,] -0.0496078548
#>  [527,] -0.8555858119
#>  [528,]  0.0941729942
#>  [529,] -0.2142659155
#>  [530,]  0.4862829294
#>  [531,] -0.6682323049
#>  [532,] -0.1670030432
#>  [533,]  0.5639612985
#>  [534,]  0.0179409716
#>  [535,] -0.3645739127
#>  [536,]  0.0408205578
#>  [537,]  0.9396594391
#>  [538,] -0.1184783558
#>  [539,]  0.4494242094
#>  [540,] -0.2161860050
#>  [541,] -0.0674615778
#>  [542,]  0.2276741042
#>  [543,]  0.1874728224
#>  [544,] -0.2695303256
#>  [545,] -0.1355865426
#>  [546,]  0.8939255425
#>  [547,] -0.3868893319
#>  [548,]  0.0633856714
#>  [549,]  0.4120400293
#>  [550,]  0.2301468991
#>  [551,]  0.9149259708
#>  [552,]  1.1397376612
#>  [553,] -0.4758946337
#>  [554,] -0.2316751954
#>  [555,]  0.5579611805
#>  [556,] -0.4722311790
#>  [557,] -0.1854912788
#>  [558,] -1.1757858095
#>  [559,]  0.4799406479
#>  [560,]  0.2889562282
#>  [561,] -1.2229192272
#>  [562,] -0.6987098408
#>  [563,] -0.7161380972
#>  [564,] -0.6979051072
#>  [565,] -0.3020397058
#>  [566,] -0.4899860006
#>  [567,] -0.8275812528
#>  [568,] -0.4674889715
#>  [569,]  0.1342921939
#>  [570,]  0.4109774332
#>  [571,]  0.6809566638
#>  [572,]  0.0412533738
#>  [573,] -0.1366347001
#>  [574,] -1.1207120449
#>  [575,]  0.3470010672
#>  [576,] -0.0923410133
#>  [577,] -0.3819289088
#>  [578,]  0.8332908523
#>  [579,] -0.8232105044
#>  [580,]  0.4841817954
#>  [581,]  0.3710496055
#>  [582,] -0.1472127127
#>  [583,] -0.6323757400
#>  [584,] -0.0760900757
#>  [585,]  0.3433188031
#>  [586,]  0.2907384936
#>  [587,]  0.1804607748
#>  [588,]  0.0960759132
#>  [589,]  0.6891629172
#>  [590,] -0.5049501924
#>  [591,]  1.2848648214
#>  [592,] -0.3799215887
#>  [593,]  0.7477565518
#>  [594,] -0.1956145465
#>  [595,] -0.4801228225
#>  [596,] -1.1238499849
#>  [597,] -0.1742700192
#>  [598,]  0.7378822923
#>  [599,] -0.3434125379
#>  [600,] -0.1916625519
#>  [601,]  0.5453817042
#>  [602,]  0.5152003173
#>  [603,] -0.0849452916
#>  [604,] -0.9375529171
#>  [605,]  0.0342835238
#>  [606,] -0.6654814558
#>  [607,]  0.1873300464
#>  [608,] -0.0769402673
#>  [609,]  0.0459705256
#>  [610,] -1.1788484236
#>  [611,]  0.3286350893
#>  [612,] -0.0070420490
#>  [613,]  0.0901089713
#>  [614,] -0.6918417458
#>  [615,] -0.4205723266
#>  [616,]  0.4543193001
#>  [617,] -1.0384407390
#>  [618,]  0.1549844225
#>  [619,]  0.0778093405
#>  [620,]  0.7125842631
#>  [621,] -0.4782746693
#>  [622,]  0.0135683112
#>  [623,]  0.4106745399
#>  [624,] -1.0563061493
#>  [625,] -0.9110607773
#>  [626,]  0.4652517767
#>  [627,]  0.0051099251
#>  [628,] -0.7323503285
#>  [629,]  0.1945683311
#>  [630,]  0.1503659036
#>  [631,] -0.4189291406
#>  [632,] -0.1985242271
#>  [633,]  0.3485411921
#>  [634,] -0.2546291094
#>  [635,]  0.1227095439
#>  [636,] -0.1856085091
#>  [637,]  0.2738276935
#>  [638,]  1.0011421355
#>  [639,]  0.4347856874
#>  [640,] -1.6299348478
#>  [641,]  0.2096823217
#>  [642,]  0.0085592911
#>  [643,]  0.8348900942
#>  [644,] -0.3045576293
#>  [645,]  0.6219601186
#>  [646,] -0.2574879577
#>  [647,]  0.0345334908
#>  [648,]  0.0117637611
#>  [649,] -1.4764159674
#>  [650,] -0.1628203429
#>  [651,] -1.2739192277
#>  [652,]  0.8570660887
#>  [653,]  0.1669280652
#>  [654,] -0.5317774271
#>  [655,] -1.0839026209
#>  [656,] -0.3629678451
#>  [657,]  0.2942355157
#>  [658,] -0.7617421179
#>  [659,] -0.1690519620
#>  [660,]  0.4039901792
#>  [661,] -0.3634546193
#>  [662,] -0.8665464355
#>  [663,] -0.0941806774
#>  [664,]  0.5124614462
#>  [665,]  0.3797252344
#>  [666,]  0.4798017278
#>  [667,] -0.5581775235
#>  [668,]  0.3612841383
#>  [669,]  0.5622882008
#>  [670,]  0.1053955495
#>  [671,]  0.8528268700
#>  [672,]  0.3209294723
#>  [673,] -0.2066689528
#>  [674,] -0.4629094725
#>  [675,]  0.0776330813
#>  [676,]  0.0504105294
#>  [677,]  0.0518247852
#>  [678,]  0.2186326924
#>  [679,]  0.9749334285
#>  [680,] -1.1149300513
#>  [681,]  0.3516316231
#>  [682,]  0.3667412156
#>  [683,]  0.8789119858
#>  [684,] -0.2670321867
#>  [685,] -0.1829066756
#>  [686,]  0.1890757856
#>  [687,]  1.0046470615
#>  [688,]  0.3744898351
#>  [689,] -0.0766467366
#>  [690,]  0.6810402860
#>  [691,]  0.4759812419
#>  [692,]  0.6531340523
#>  [693,]  1.4328779844
#>  [694,] -0.7168405125
#>  [695,] -0.9394319744
#>  [696,]  0.2211740313
#>  [697,] -0.6708222751
#>  [698,] -0.6266545243
#>  [699,] -0.2085761345
#>  [700,] -0.5009710120
#>  [701,] -0.0736218067
#>  [702,]  0.0828168122
#>  [703,]  0.3307873932
#>  [704,] -0.7439959522
#>  [705,] -0.3728238902
#>  [706,]  0.2977398577
#>  [707,]  0.7548079992
#>  [708,]  0.3886465372
#>  [709,]  0.4132793359
#>  [710,]  0.1895206926
#>  [711,]  0.3993927746
#>  [712,]  0.9802145915
#>  [713,]  0.7551323066
#>  [714,] -0.2736229097
#>  [715,]  0.1938497691
#>  [716,]  0.2522080231
#>  [717,]  0.4456269069
#>  [718,] -0.1725763001
#>  [719,]  1.0796864291
#>  [720,]  0.3771343337
#>  [721,]  1.0646941348
#>  [722,]  0.5670733393
#>  [723,] -0.3559408053
#>  [724,] -0.7792547329
#>  [725,]  0.3513126354
#>  [726,] -0.2204865616
#>  [727,] -0.7115928475
#>  [728,] -0.1274099215
#>  [729,]  0.1965992107
#>  [730,] -0.3877820454
#>  [731,] -0.0199304098
#>  [732,]  0.1184656790
#>  [733,]  0.2743231664
#>  [734,]  0.5775043377
#>  [735,] -0.5056715572
#>  [736,] -0.2552136408
#>  [737,] -0.6560041220
#>  [738,]  0.0348106517
#>  [739,]  0.0902853628
#>  [740,] -0.6224927764
#>  [741,]  0.0797086856
#>  [742,] -0.1237983907
#>  [743,] -0.2795830616
#>  [744,]  0.4582542386
#>  [745,] -0.2354688880
#>  [746,]  0.2209488311
#>  [747,] -0.5618114993
#>  [748,]  0.2230977528
#>  [749,]  0.4956305404
#>  [750,] -0.5831842240
#>  [751,]  0.2152600274
#>  [752,]  0.0542749526
#>  [753,] -0.2871894307
#>  [754,]  0.8426608799
#>  [755,]  0.7672473362
#>  [756,]  0.2816404784
#>  [757,] -0.2591957050
#>  [758,] -0.3336942127
#>  [759,]  0.3641665889
#>  [760,]  0.9735082098
#>  [761,] -0.3217613371
#>  [762,] -0.7869473541
#>  [763,] -0.4184548076
#>  [764,] -0.4806189341
#>  [765,] -0.8255404247
#>  [766,] -0.5302742500
#>  [767,]  1.3100234120
#>  [768,] -0.9603843925
#>  [769,] -0.0885758472
#>  [770,]  0.2058009446
#>  [771,] -0.7842318910
#>  [772,]  0.0635001487
#>  [773,] -0.6950726912
#>  [774,] -0.1846424759
#>  [775,] -0.3363854002
#>  [776,] -0.1788552817
#>  [777,]  0.8910277049
#>  [778,] -0.1378184219
#>  [779,] -0.5355897766
#>  [780,]  0.3572767375
#>  [781,]  0.1585311640
#>  [782,]  1.0687100264
#>  [783,]  1.1065919606
#>  [784,] -0.2037459092
#>  [785,]  0.9385581262
#>  [786,]  0.6147091491
#>  [787,] -0.6015304298
#>  [788,]  0.1344253546
#>  [789,] -1.2289615184
#>  [790,] -0.2334132884
#>  [791,]  0.6008544272
#>  [792,] -0.3406511807
#>  [793,]  0.0389014523
#>  [794,]  0.6982869963
#>  [795,]  0.3481276196
#>  [796,] -1.3028777651
#>  [797,] -0.2717359116
#>  [798,] -0.2855081223
#>  [799,] -0.0936022934
#>  [800,] -0.3429438315
#>  [801,]  0.0655738344
#>  [802,]  0.4265372797
#>  [803,] -0.1910291208
#>  [804,] -0.9139994983
#>  [805,]  0.3288406400
#>  [806,] -0.8873585176
#>  [807,]  1.0038064778
#>  [808,]  0.9868168359
#>  [809,] -0.2429606318
#>  [810,] -0.9555984839
#>  [811,]  0.4623998917
#>  [812,]  0.3845989521
#>  [813,]  0.2159034786
#>  [814,]  0.6700576583
#>  [815,] -0.4725835818
#>  [816,] -0.4929440841
#>  [817,]  0.0322173427
#>  [818,]  0.3226373863
#>  [819,] -0.6958443341
#>  [820,]  0.9750287071
#>  [821,] -0.2351430206
#>  [822,]  0.0729387393
#>  [823,] -0.1591413479
#>  [824,] -0.5117258141
#>  [825,] -0.4795020602
#>  [826,]  0.4052633987
#>  [827,]  0.4204229114
#>  [828,]  0.2745261034
#>  [829,] -0.0205264551
#>  [830,] -0.4313962757
#>  [831,]  0.6094847593
#>  [832,]  1.0722682184
#>  [833,] -0.2645932011
#>  [834,] -0.3071082929
#>  [835,] -1.0247370843
#>  [836,] -0.0066212789
#>  [837,] -0.3714976419
#>  [838,] -0.9059503371
#>  [839,]  0.3446491465
#>  [840,] -0.4288086655
#>  [841,]  0.5646666572
#>  [842,] -0.3737606260
#>  [843,]  0.0441999138
#>  [844,] -0.3987035740
#>  [845,] -0.7878595116
#>  [846,]  0.2526524324
#>  [847,] -0.7940052286
#>  [848,]  0.9473779464
#>  [849,] -0.4694131845
#>  [850,]  0.3828915754
#>  [851,]  0.6224010941
#>  [852,] -0.5850916293
#>  [853,]  0.0889049578
#>  [854,] -0.7938630153
#>  [855,]  0.3155271067
#>  [856,] -0.1221225739
#>  [857,] -1.3299966172
#>  [858,] -0.5719104419
#>  [859,] -0.9125463945
#>  [860,] -0.2369340306
#>  [861,] -0.1245122382
#>  [862,]  0.1232996653
#>  [863,] -0.3210474198
#>  [864,]  0.2491969326
#>  [865,] -0.2825076817
#>  [866,] -0.5326847966
#>  [867,]  0.8425149446
#>  [868,] -0.1497184644
#>  [869,] -0.0707916058
#>  [870,]  1.4702040420
#>  [871,]  0.3788357935
#>  [872,]  0.1531630624
#>  [873,] -0.0760085783
#>  [874,]  0.4299814490
#>  [875,]  0.2210071373
#>  [876,]  0.2148609655
#>  [877,]  0.1065624019
#>  [878,]  0.3748336763
#>  [879,]  1.3183488157
#>  [880,]  0.1440760182
#>  [881,] -0.2546846663
#>  [882,]  0.3841516958
#>  [883,] -0.4562163052
#>  [884,]  0.0205488582
#>  [885,] -0.5282935657
#>  [886,] -0.0380554254
#>  [887,]  0.8188943401
#>  [888,] -0.3172667979
#>  [889,] -0.0775247766
#>  [890,]  0.6523131677
#>  [891,] -0.5291892562
#>  [892,] -0.2480661026
#>  [893,] -0.5409176114
#>  [894,]  1.6438376568
#>  [895,] -0.0285456035
#>  [896,]  0.5037423075
#>  [897,]  0.1677051070
#>  [898,]  0.4744941829
#>  [899,] -0.3244793425
#>  [900,]  0.2792857015
#>  [901,] -0.5483914330
#>  [902,] -0.0305736734
#>  [903,] -0.1514560108
#>  [904,]  0.2612388698
#>  [905,]  0.2779239571
#>  [906,] -0.5204464218
#>  [907,]  0.2379673728
#>  [908,] -0.0438145426
#>  [909,] -0.1526715836
#>  [910,] -0.0887326797
#>  [911,] -0.1575820644
#>  [912,]  0.1630536866
#>  [913,]  0.7016632587
#>  [914,] -0.8003156123
#>  [915,] -0.5517380024
#>  [916,]  0.5581397047
#>  [917,]  0.9733277968
#>  [918,]  0.4126554405
#>  [919,]  0.2656192917
#>  [920,]  0.9627910507
#>  [921,] -0.9311922756
#>  [922,] -0.4634413070
#>  [923,] -1.3692604538
#>  [924,]  0.2544433611
#>  [925,]  0.4049597731
#>  [926,] -0.8171165402
#>  [927,]  0.9692491441
#>  [928,]  0.2155599695
#>  [929,] -0.0075325145
#>  [930,] -0.3707412229
#>  [931,]  0.3387058619
#>  [932,]  0.6080580224
#>  [933,]  0.3844642886
#>  [934,] -0.4456922282
#>  [935,]  0.4762716555
#>  [936,]  0.8586442873
#>  [937,] -0.8077942673
#>  [938,]  1.1775882994
#>  [939,]  0.2418526966
#>  [940,]  0.3335351724
#>  [941,]  0.1285010705
#>  [942,]  1.3411337155
#>  [943,] -0.5558895146
#>  [944,]  0.1590928912
#>  [945,] -0.6951070206
#>  [946,] -0.2605365993
#>  [947,] -0.5783344632
#>  [948,]  0.3986588187
#>  [949,] -0.1698419638
#>  [950,]  0.3858857353
#>  [951,]  0.2774403061
#>  [952,]  0.9474207427
#>  [953,]  0.2077049221
#>  [954,]  0.2806160272
#>  [955,] -0.1963228133
#>  [956,] -0.8030850098
#>  [957,] -0.2807446681
#>  [958,] -0.3026694307
#>  [959,]  0.5083196478
#>  [960,] -0.2461653834
#>  [961,] -0.4289191253
#>  [962,]  0.7402087768
#>  [963,]  0.4969375776
#>  [964,] -0.2984568853
#>  [965,]  1.1477921192
#>  [966,]  0.4754584583
#>  [967,]  0.3991623271
#>  [968,] -0.9194068347
#>  [969,]  0.1032743784
#>  [970,]  0.8412642044
#>  [971,]  0.0857649533
#>  [972,] -0.0358843175
#>  [973,] -0.6591287615
#>  [974,] -0.0544429240
#>  [975,]  0.0107334267
#>  [976,] -1.0039959602
#>  [977,]  0.2431124707
#>  [978,] -0.2085744569
#>  [979,] -0.5573374263
#>  [980,] -0.0040653528
#>  [981,] -0.1772500967
#>  [982,] -0.1807420473
#>  [983,] -0.7999288658
#>  [984,]  0.6381859612
#>  [985,] -0.7711332290
#>  [986,] -0.5225925145
#>  [987,]  0.0978368183
#>  [988,] -0.2031494938
#>  [989,] -0.0101084787
#>  [990,] -0.1496896072
#>  [991,] -0.2085520348
#>  [992,]  0.6977698312
#>  [993,]  0.2173072834
#>  [994,]  0.4405061463
#>  [995,] -0.0108895078
#>  [996,]  0.1153938005
#>  [997,] -0.3793888995
#>  [998,] -0.6079741286
#>  [999,]  0.6614573931
#> [1000,] -0.0739580171