Logistic Function

The elementwise logistic function, \(\log(1 + e^x)\). This is a special case of log(sum(exp)) that evaluates to a vector rather than to a scalar, which is useful for logistic regression.

logistic(x)

Arguments

x

An Expression, vector, or matrix.

Value

An Expression representing the logistic function evaluated at the input.

Examples

set.seed(92)
n <- 20
m <- 1000
sigma <- 45

beta_true <- stats::rnorm(n)
idxs <- sample(n, size = 0.8*n, replace = FALSE)
beta_true[idxs] <- 0
X <- matrix(stats::rnorm(m*n, 0, 5), nrow = m, ncol = n)
y <- sign(X %*% beta_true + stats::rnorm(m, 0, sigma))

beta <- Variable(n)
X_sign <- apply(X, 2, function(x) { ifelse(y <= 0, -1, 1) * x })
obj <- -sum(logistic(-X[y <= 0,] %*% beta)) - sum(logistic(X[y == 1,] %*% beta))
prob <- Problem(Maximize(obj))
result <- solve(prob)

log_odds <- result$getValue(X %*% beta)
beta_res <- result$getValue(beta)
y_probs <- 1/(1 + exp(-X %*% beta_res))
log(y_probs/(1 - y_probs))
#>                  [,1]
#>    [1,] -0.6704566601
#>    [2,] -0.5925155914
#>    [3,] -0.2870297658
#>    [4,] -0.3903425916
#>    [5,]  0.1512533989
#>    [6,] -0.3212308442
#>    [7,]  0.0710815720
#>    [8,] -0.0005972517
#>    [9,]  0.6231543856
#>   [10,] -0.3461542128
#>   [11,]  0.6165907888
#>   [12,]  0.2493802797
#>   [13,] -0.4229470194
#>   [14,]  1.3550590126
#>   [15,] -0.3265597137
#>   [16,] -0.5974060590
#>   [17,]  1.4250937991
#>   [18,] -0.8064117254
#>   [19,]  0.6354245008
#>   [20,] -0.5313297468
#>   [21,] -0.5836826542
#>   [22,] -0.2018604692
#>   [23,] -1.5741079507
#>   [24,] -0.2996817503
#>   [25,] -0.2075326459
#>   [26,]  0.4436962967
#>   [27,] -0.5653426412
#>   [28,] -1.2550595374
#>   [29,] -0.5346239907
#>   [30,]  0.4834049559
#>   [31,]  0.2311018280
#>   [32,]  0.2894463877
#>   [33,] -0.0125268284
#>   [34,]  0.7295741839
#>   [35,] -0.8747089801
#>   [36,] -0.3928902742
#>   [37,] -0.1155668459
#>   [38,]  1.1672435760
#>   [39,] -0.7217417178
#>   [40,]  0.0611243125
#>   [41,] -0.6167645485
#>   [42,] -0.0245103839
#>   [43,]  0.6908204543
#>   [44,] -0.8034455356
#>   [45,]  0.8837094798
#>   [46,] -1.0940587657
#>   [47,]  0.0151093307
#>   [48,] -0.0365091373
#>   [49,]  0.1208098298
#>   [50,]  0.0228335289
#>   [51,]  0.0203841255
#>   [52,] -0.4370865446
#>   [53,]  0.0381300050
#>   [54,] -0.0670260414
#>   [55,] -0.5450479974
#>   [56,] -0.1197495178
#>   [57,] -0.1350421079
#>   [58,]  0.1775461764
#>   [59,]  1.1397934859
#>   [60,] -0.4142561404
#>   [61,]  0.3992984051
#>   [62,] -0.4202125247
#>   [63,]  0.0406907572
#>   [64,] -0.6489089858
#>   [65,] -0.9702903642
#>   [66,]  0.1117445633
#>   [67,] -0.1453568149
#>   [68,]  0.0616516205
#>   [69,]  0.0779096420
#>   [70,]  0.4237417182
#>   [71,] -0.4578309643
#>   [72,] -0.2049702780
#>   [73,] -0.2303885529
#>   [74,]  0.1166963282
#>   [75,] -0.3651955569
#>   [76,] -0.2369500750
#>   [77,]  0.1457228397
#>   [78,] -1.0896406573
#>   [79,] -0.1308710754
#>   [80,]  0.2291674619
#>   [81,]  0.1373238859
#>   [82,] -0.0015643758
#>   [83,]  0.3467565323
#>   [84,] -0.1543888962
#>   [85,] -0.2309014974
#>   [86,] -0.5247106391
#>   [87,]  0.4603747857
#>   [88,] -0.3126465765
#>   [89,]  0.2120047116
#>   [90,]  0.3146963593
#>   [91,]  0.4193015146
#>   [92,]  0.6556778133
#>   [93,] -0.5111104188
#>   [94,]  1.3013236680
#>   [95,]  0.8030124821
#>   [96,]  0.2830361899
#>   [97,] -0.6703954799
#>   [98,]  0.6392057972
#>   [99,]  0.0848836974
#>  [100,] -0.0064337716
#>  [101,]  0.1647458661
#>  [102,]  0.2973106680
#>  [103,]  0.3564182328
#>  [104,] -0.8130441051
#>  [105,] -0.5596560141
#>  [106,]  0.2440618391
#>  [107,] -0.5330512912
#>  [108,]  0.4528402173
#>  [109,] -0.5708931835
#>  [110,] -0.4623969739
#>  [111,] -0.2336890208
#>  [112,]  0.1511207911
#>  [113,] -0.0618991731
#>  [114,]  0.5838949464
#>  [115,] -0.0005892470
#>  [116,] -0.7089720216
#>  [117,] -0.3172697223
#>  [118,]  0.3970567787
#>  [119,] -1.5819880428
#>  [120,]  0.5686251944
#>  [121,] -0.3273432678
#>  [122,] -0.8145940559
#>  [123,] -0.0947979896
#>  [124,]  0.3203428780
#>  [125,] -0.4429133655
#>  [126,]  0.1170332022
#>  [127,] -0.7652234333
#>  [128,] -0.0573318811
#>  [129,]  0.0095928844
#>  [130,] -0.1932131076
#>  [131,]  0.1964228615
#>  [132,] -0.3175599181
#>  [133,]  0.4246197779
#>  [134,]  0.1266881215
#>  [135,]  0.3305602671
#>  [136,] -0.0061834939
#>  [137,] -0.5537621931
#>  [138,] -0.1133697510
#>  [139,] -0.7994572391
#>  [140,]  0.8081578245
#>  [141,]  0.4153083028
#>  [142,]  0.8942222160
#>  [143,] -0.3139444517
#>  [144,]  0.1458380321
#>  [145,] -0.2050259510
#>  [146,] -0.0601892036
#>  [147,] -1.1748051406
#>  [148,]  0.5454688323
#>  [149,]  0.2075669928
#>  [150,] -0.2618746549
#>  [151,]  0.8038868939
#>  [152,] -0.0455069481
#>  [153,]  0.5741311897
#>  [154,]  0.6919848048
#>  [155,]  0.0716646406
#>  [156,] -0.4965733924
#>  [157,]  0.4423892111
#>  [158,] -0.3816839141
#>  [159,]  0.5205994919
#>  [160,] -0.1021932394
#>  [161,] -0.2226407248
#>  [162,]  0.7098824842
#>  [163,]  0.5765239949
#>  [164,]  0.2450293204
#>  [165,] -0.1737074723
#>  [166,]  0.3589634540
#>  [167,] -0.3290028521
#>  [168,] -0.8993322171
#>  [169,] -0.8217942301
#>  [170,] -0.7797747655
#>  [171,] -0.2137166776
#>  [172,]  0.5276279942
#>  [173,] -0.8487795398
#>  [174,]  0.1034542895
#>  [175,] -0.4103611024
#>  [176,] -0.7528761525
#>  [177,] -0.4402384001
#>  [178,] -1.0685618040
#>  [179,]  0.4078520134
#>  [180,]  0.6313571340
#>  [181,]  0.3750879398
#>  [182,] -0.0151754002
#>  [183,] -0.5310802005
#>  [184,]  0.4077498342
#>  [185,]  0.1571536758
#>  [186,] -0.6328579544
#>  [187,] -0.3822521507
#>  [188,]  0.5205127750
#>  [189,]  0.2562676001
#>  [190,]  0.8532572121
#>  [191,] -0.5649117264
#>  [192,]  0.0370585167
#>  [193,] -0.3390790595
#>  [194,] -0.0375169486
#>  [195,]  0.0474707447
#>  [196,] -0.0326514547
#>  [197,]  0.2159972774
#>  [198,] -1.0662461286
#>  [199,] -0.9549865987
#>  [200,] -0.6453789741
#>  [201,]  0.5183307748
#>  [202,] -0.6994981183
#>  [203,]  0.2180974130
#>  [204,]  0.1893054841
#>  [205,]  0.4929863051
#>  [206,]  0.1844648176
#>  [207,]  0.3573890378
#>  [208,] -0.2500557609
#>  [209,]  1.3917106836
#>  [210,] -0.7617778892
#>  [211,]  0.5279206489
#>  [212,] -1.0117658159
#>  [213,] -0.2423709809
#>  [214,] -0.6133908459
#>  [215,] -1.1656157013
#>  [216,]  0.3821323274
#>  [217,] -0.4501177815
#>  [218,] -1.0973275111
#>  [219,]  0.8937308100
#>  [220,]  0.3140650093
#>  [221,]  0.1340498249
#>  [222,] -1.2230817731
#>  [223,]  0.5678172666
#>  [224,]  0.8463547383
#>  [225,]  0.6349126119
#>  [226,]  0.7005876882
#>  [227,] -0.7992087463
#>  [228,]  0.5683384775
#>  [229,]  0.4112971973
#>  [230,]  0.0902409614
#>  [231,] -0.3665163870
#>  [232,]  0.0127784696
#>  [233,]  0.1629153570
#>  [234,] -0.3527155111
#>  [235,] -0.0758198135
#>  [236,]  0.0091712200
#>  [237,] -0.4611139986
#>  [238,]  0.0657035000
#>  [239,] -0.5298528191
#>  [240,] -0.5962161970
#>  [241,]  0.0637669046
#>  [242,] -0.3510112099
#>  [243,]  0.5476027637
#>  [244,]  0.5265563483
#>  [245,] -0.8598107283
#>  [246,]  0.6395255501
#>  [247,]  0.3215942360
#>  [248,] -0.4795886634
#>  [249,]  0.3975190764
#>  [250,]  0.1215631023
#>  [251,] -0.3140252989
#>  [252,] -0.3046458001
#>  [253,]  0.3923162069
#>  [254,] -0.1856423411
#>  [255,] -0.0851603773
#>  [256,] -0.0029482127
#>  [257,]  0.5734735381
#>  [258,] -0.0359977194
#>  [259,] -0.9567905661
#>  [260,]  0.6395616014
#>  [261,]  0.4795687762
#>  [262,]  0.1872599648
#>  [263,] -0.2498881709
#>  [264,]  0.3260938617
#>  [265,]  0.1986580454
#>  [266,] -0.6790600913
#>  [267,]  0.6809504849
#>  [268,] -0.2952543618
#>  [269,]  0.5265755493
#>  [270,]  0.6812868456
#>  [271,] -0.4787183436
#>  [272,]  0.6944639345
#>  [273,]  0.3116161602
#>  [274,]  0.6770352241
#>  [275,]  0.0024198961
#>  [276,]  0.1779762817
#>  [277,]  0.1332637066
#>  [278,]  0.2881592259
#>  [279,] -0.8398208677
#>  [280,]  0.8293027203
#>  [281,]  0.3237905695
#>  [282,]  0.3688696112
#>  [283,] -0.2838397598
#>  [284,]  0.1698065742
#>  [285,] -0.0972781640
#>  [286,]  0.1692293332
#>  [287,]  0.6139295310
#>  [288,]  0.1736829439
#>  [289,] -0.0806338472
#>  [290,]  0.0115350265
#>  [291,] -0.3183864173
#>  [292,]  0.0866710268
#>  [293,] -0.7586296972
#>  [294,]  0.0584507549
#>  [295,] -0.1537362781
#>  [296,]  0.8435282624
#>  [297,]  0.8815392891
#>  [298,]  0.1534469763
#>  [299,]  1.2271212600
#>  [300,]  0.3098916000
#>  [301,] -0.2555255745
#>  [302,] -0.2073281110
#>  [303,] -0.2298554674
#>  [304,] -0.5772110245
#>  [305,] -0.4885371748
#>  [306,]  0.5318245990
#>  [307,] -0.1440915396
#>  [308,]  0.6375851656
#>  [309,]  0.2563565473
#>  [310,] -0.0603848083
#>  [311,] -0.2223194676
#>  [312,] -0.4951070092
#>  [313,] -0.2533144391
#>  [314,]  0.3508803854
#>  [315,]  0.0868143920
#>  [316,] -0.3995760620
#>  [317,] -0.2455234345
#>  [318,] -0.0988023306
#>  [319,]  0.6998229047
#>  [320,] -0.4848186944
#>  [321,] -0.3735362873
#>  [322,] -0.0385919797
#>  [323,]  0.0463459559
#>  [324,] -0.8583646944
#>  [325,] -0.4583177299
#>  [326,]  0.6281397044
#>  [327,] -0.6350930132
#>  [328,] -0.9581823784
#>  [329,]  0.9368446576
#>  [330,]  0.1483849866
#>  [331,] -0.0538264807
#>  [332,] -1.2323011986
#>  [333,] -0.1706885164
#>  [334,]  0.1205507360
#>  [335,] -0.4870842705
#>  [336,]  0.0540006968
#>  [337,] -0.1344842486
#>  [338,]  1.1543564231
#>  [339,]  0.8618954880
#>  [340,]  0.0228928269
#>  [341,] -0.5239006990
#>  [342,]  0.3351221664
#>  [343,] -0.7517139168
#>  [344,]  0.6050284431
#>  [345,] -0.0211552810
#>  [346,]  0.4027175331
#>  [347,]  0.8252852085
#>  [348,] -0.3307087458
#>  [349,]  0.3177258224
#>  [350,] -0.0224526341
#>  [351,] -0.5546741854
#>  [352,] -0.2831584120
#>  [353,] -0.1794649773
#>  [354,] -0.5933507121
#>  [355,]  0.5243789964
#>  [356,]  0.2004109640
#>  [357,]  0.2904665363
#>  [358,] -0.4817097810
#>  [359,] -0.1010052942
#>  [360,]  0.3605534751
#>  [361,]  0.3838585972
#>  [362,] -0.6909236928
#>  [363,] -0.3578673016
#>  [364,]  0.7571636427
#>  [365,]  0.4263330459
#>  [366,] -0.7291975361
#>  [367,]  1.3051018150
#>  [368,]  0.3406674952
#>  [369,]  0.1624612727
#>  [370,] -0.4134237907
#>  [371,]  0.8469756334
#>  [372,]  0.1445681797
#>  [373,] -1.0051098115
#>  [374,]  0.1813248620
#>  [375,]  0.3127957905
#>  [376,] -0.2938446482
#>  [377,] -1.1220696942
#>  [378,]  0.0153639539
#>  [379,]  0.6470270083
#>  [380,] -1.1136243303
#>  [381,] -0.6871705750
#>  [382,]  0.5673074646
#>  [383,] -0.2756604522
#>  [384,] -0.8618049910
#>  [385,] -0.1516823240
#>  [386,] -0.2098201412
#>  [387,] -1.2010298142
#>  [388,]  1.2558427314
#>  [389,] -0.5078108344
#>  [390,] -0.8335823356
#>  [391,]  0.8759229459
#>  [392,]  0.5860428533
#>  [393,] -1.2937980410
#>  [394,] -0.7939846252
#>  [395,] -0.5236335284
#>  [396,]  0.1775975846
#>  [397,] -1.4013400929
#>  [398,]  0.7131508152
#>  [399,]  0.0354491390
#>  [400,]  0.1344366250
#>  [401,]  0.4879745284
#>  [402,]  0.3638074820
#>  [403,] -0.1982415573
#>  [404,]  0.1427058308
#>  [405,]  0.0306148385
#>  [406,]  0.6818142230
#>  [407,]  0.0828439412
#>  [408,]  0.1285023444
#>  [409,]  0.7544467283
#>  [410,]  0.2249889766
#>  [411,]  0.0320368984
#>  [412,]  0.4893712975
#>  [413,]  0.3546914975
#>  [414,] -0.3652916745
#>  [415,]  0.8342618286
#>  [416,] -0.2503049305
#>  [417,]  0.5247218764
#>  [418,]  0.9033035542
#>  [419,]  0.6160200872
#>  [420,]  0.1392514959
#>  [421,]  0.7268226327
#>  [422,] -0.5110357658
#>  [423,] -0.6768711729
#>  [424,]  0.8344246950
#>  [425,]  0.3100200429
#>  [426,]  0.4325878598
#>  [427,]  0.1333008872
#>  [428,]  0.5181933545
#>  [429,] -0.3190005722
#>  [430,]  0.4315606205
#>  [431,]  0.3930258753
#>  [432,] -0.4950415623
#>  [433,] -0.7391246227
#>  [434,] -0.3018831186
#>  [435,] -0.2371877521
#>  [436,]  0.9647911369
#>  [437,] -0.5663594260
#>  [438,] -0.2494167440
#>  [439,] -0.8171821433
#>  [440,] -0.7060567490
#>  [441,] -0.5057583678
#>  [442,]  0.3789756520
#>  [443,] -0.9727593552
#>  [444,]  0.4803698804
#>  [445,]  0.6154932259
#>  [446,] -0.2193750953
#>  [447,]  0.2770446970
#>  [448,]  0.1446123066
#>  [449,]  0.7294647899
#>  [450,]  0.1318375506
#>  [451,] -0.9285885489
#>  [452,]  0.5649870685
#>  [453,] -0.0734971335
#>  [454,]  0.6055726790
#>  [455,]  0.5110901373
#>  [456,]  0.3522389149
#>  [457,] -0.7569925274
#>  [458,] -0.1832521125
#>  [459,] -0.7545274294
#>  [460,] -0.0289834723
#>  [461,]  0.7738049808
#>  [462,]  0.4915963396
#>  [463,] -0.3123273178
#>  [464,] -1.0281669120
#>  [465,]  0.2270999811
#>  [466,]  0.7107782565
#>  [467,] -0.3046705517
#>  [468,] -0.5839071481
#>  [469,]  1.4979018365
#>  [470,] -0.2798531408
#>  [471,]  0.0598986526
#>  [472,] -0.1166454607
#>  [473,]  0.5759189086
#>  [474,]  0.0739414249
#>  [475,]  0.5817785017
#>  [476,]  0.2869966634
#>  [477,] -0.1296562333
#>  [478,]  0.4669492271
#>  [479,] -0.0833634990
#>  [480,]  0.0444096705
#>  [481,] -0.2424477777
#>  [482,]  0.4827976350
#>  [483,]  0.3103920469
#>  [484,]  0.0108840206
#>  [485,]  0.7864389056
#>  [486,] -0.9775841591
#>  [487,] -0.3104754891
#>  [488,] -0.3305425914
#>  [489,] -0.2215509763
#>  [490,]  0.0721143735
#>  [491,]  0.2177853569
#>  [492,]  0.1394636725
#>  [493,]  0.2520085800
#>  [494,]  0.5967270610
#>  [495,]  0.3058376600
#>  [496,]  0.5917353513
#>  [497,]  0.3640662733
#>  [498,]  0.1549089403
#>  [499,] -0.4251336377
#>  [500,] -0.5954614705
#>  [501,]  0.1971901463
#>  [502,]  0.3262361037
#>  [503,]  0.2089600274
#>  [504,] -0.3834680497
#>  [505,]  0.4277940938
#>  [506,]  0.0551671256
#>  [507,] -0.0386640086
#>  [508,]  0.2733780431
#>  [509,] -0.5726047975
#>  [510,] -0.2647283714
#>  [511,] -0.1636688193
#>  [512,]  0.2262534070
#>  [513,] -0.0280563189
#>  [514,] -0.2338544077
#>  [515,]  0.6861872381
#>  [516,] -0.5119125214
#>  [517,] -0.0523609926
#>  [518,]  0.1044164641
#>  [519,] -0.1237390590
#>  [520,]  0.0757027240
#>  [521,] -0.5462324268
#>  [522,]  0.0475500898
#>  [523,]  1.1280721475
#>  [524,]  2.1538521816
#>  [525,] -0.3064773536
#>  [526,] -0.0496098560
#>  [527,] -0.8555980974
#>  [528,]  0.0941731448
#>  [529,] -0.2142726218
#>  [530,]  0.4862901334
#>  [531,] -0.6682417761
#>  [532,] -0.1670060019
#>  [533,]  0.5639675660
#>  [534,]  0.0179407790
#>  [535,] -0.3645781902
#>  [536,]  0.0408228260
#>  [537,]  0.9396738163
#>  [538,] -0.1184797861
#>  [539,]  0.4494313329
#>  [540,] -0.2161908811
#>  [541,] -0.0674621273
#>  [542,]  0.2276748621
#>  [543,]  0.1874745461
#>  [544,] -0.2695337123
#>  [545,] -0.1355847719
#>  [546,]  0.8939394488
#>  [547,] -0.3868942722
#>  [548,]  0.0633859755
#>  [549,]  0.4120442669
#>  [550,]  0.2301485812
#>  [551,]  0.9149448647
#>  [552,]  1.1397572405
#>  [553,] -0.4759021874
#>  [554,] -0.2316796589
#>  [555,]  0.5579682640
#>  [556,] -0.4722367623
#>  [557,] -0.1854934565
#>  [558,] -1.1758058576
#>  [559,]  0.4799492889
#>  [560,]  0.2889615432
#>  [561,] -1.2229386995
#>  [562,] -0.6987219359
#>  [563,] -0.7161481667
#>  [564,] -0.6979187374
#>  [565,] -0.3020466031
#>  [566,] -0.4899953489
#>  [567,] -0.8275962847
#>  [568,] -0.4674950196
#>  [569,]  0.1342948804
#>  [570,]  0.4109874859
#>  [571,]  0.6809664531
#>  [572,]  0.0412525653
#>  [573,] -0.1366350706
#>  [574,] -1.1207337079
#>  [575,]  0.3470067929
#>  [576,] -0.0923409971
#>  [577,] -0.3819336931
#>  [578,]  0.8333035847
#>  [579,] -0.8232234358
#>  [580,]  0.4841882070
#>  [581,]  0.3710572238
#>  [582,] -0.1472150650
#>  [583,] -0.6323914920
#>  [584,] -0.0760915545
#>  [585,]  0.3433258183
#>  [586,]  0.2907435295
#>  [587,]  0.1804655638
#>  [588,]  0.0960746038
#>  [589,]  0.6891781694
#>  [590,] -0.5049569493
#>  [591,]  1.2848835754
#>  [592,] -0.3799283129
#>  [593,]  0.7477702212
#>  [594,] -0.1956200888
#>  [595,] -0.4801320903
#>  [596,] -1.1238694316
#>  [597,] -0.1742729665
#>  [598,]  0.7378953287
#>  [599,] -0.3434189294
#>  [600,] -0.1916673964
#>  [601,]  0.5453891039
#>  [602,]  0.5152067136
#>  [603,] -0.0849465285
#>  [604,] -0.9375702176
#>  [605,]  0.0342861184
#>  [606,] -0.6654940081
#>  [607,]  0.1873345150
#>  [608,] -0.0769441016
#>  [609,]  0.0459690299
#>  [610,] -1.1788679169
#>  [611,]  0.3286415798
#>  [612,] -0.0070398665
#>  [613,]  0.0901108970
#>  [614,] -0.6918504354
#>  [615,] -0.4205800723
#>  [616,]  0.4543268904
#>  [617,] -1.0384554020
#>  [618,]  0.1549874586
#>  [619,]  0.0778153993
#>  [620,]  0.7125930810
#>  [621,] -0.4782826046
#>  [622,]  0.0135671419
#>  [623,]  0.4106809996
#>  [624,] -1.0563256277
#>  [625,] -0.9110756256
#>  [626,]  0.4652583868
#>  [627,]  0.0051119560
#>  [628,] -0.7323652751
#>  [629,]  0.1945666269
#>  [630,]  0.1503672563
#>  [631,] -0.4189320285
#>  [632,] -0.1985243872
#>  [633,]  0.3485440840
#>  [634,] -0.2546306539
#>  [635,]  0.1227093653
#>  [636,] -0.1856114989
#>  [637,]  0.2738320503
#>  [638,]  1.0011583972
#>  [639,]  0.4347908674
#>  [640,] -1.6299630581
#>  [641,]  0.2096862729
#>  [642,]  0.0085548911
#>  [643,]  0.8349020227
#>  [644,] -0.3045636391
#>  [645,]  0.6219698623
#>  [646,] -0.2574906120
#>  [647,]  0.0345353917
#>  [648,]  0.0117664619
#>  [649,] -1.4764422104
#>  [650,] -0.1628241897
#>  [651,] -1.2739376707
#>  [652,]  0.8570820034
#>  [653,]  0.1669292965
#>  [654,] -0.5317850287
#>  [655,] -1.0839194580
#>  [656,] -0.3629707878
#>  [657,]  0.2942403152
#>  [658,] -0.7617545907
#>  [659,] -0.1690545735
#>  [660,]  0.4039956100
#>  [661,] -0.3634582941
#>  [662,] -0.8665629700
#>  [663,] -0.0941815160
#>  [664,]  0.5124719800
#>  [665,]  0.3797310939
#>  [666,]  0.4798095771
#>  [667,] -0.5581899881
#>  [668,]  0.3612915464
#>  [669,]  0.5622974939
#>  [670,]  0.1053961693
#>  [671,]  0.8528432195
#>  [672,]  0.3209320303
#>  [673,] -0.2066703154
#>  [674,] -0.4629168486
#>  [675,]  0.0776318452
#>  [676,]  0.0504121140
#>  [677,]  0.0518265536
#>  [678,]  0.2186372967
#>  [679,]  0.9749493164
#>  [680,] -1.1149482951
#>  [681,]  0.3516361891
#>  [682,]  0.3667470411
#>  [683,]  0.8789232773
#>  [684,] -0.2670357825
#>  [685,] -0.1829090230
#>  [686,]  0.1890744270
#>  [687,]  1.0046635884
#>  [688,]  0.3744975371
#>  [689,] -0.0766459083
#>  [690,]  0.6810551492
#>  [691,]  0.4759922006
#>  [692,]  0.6531461508
#>  [693,]  1.4328980712
#>  [694,] -0.7168506910
#>  [695,] -0.9394506137
#>  [696,]  0.2211753382
#>  [697,] -0.6708343014
#>  [698,] -0.6266708148
#>  [699,] -0.2085801270
#>  [700,] -0.5009781300
#>  [701,] -0.0736209105
#>  [702,]  0.0828169897
#>  [703,]  0.3307929430
#>  [704,] -0.7440077387
#>  [705,] -0.3728347114
#>  [706,]  0.2977447079
#>  [707,]  0.7548204272
#>  [708,]  0.3886572805
#>  [709,]  0.4132856791
#>  [710,]  0.1895251830
#>  [711,]  0.3993992893
#>  [712,]  0.9802321367
#>  [713,]  0.7551475071
#>  [714,] -0.2736283043
#>  [715,]  0.1938512839
#>  [716,]  0.2522079318
#>  [717,]  0.4456332365
#>  [718,] -0.1725801231
#>  [719,]  1.0797039576
#>  [720,]  0.3771421018
#>  [721,]  1.0647146570
#>  [722,]  0.5670786800
#>  [723,] -0.3559502304
#>  [724,] -0.7792681821
#>  [725,]  0.3513208478
#>  [726,] -0.2204885378
#>  [727,] -0.7116021712
#>  [728,] -0.1274112014
#>  [729,]  0.1966007892
#>  [730,] -0.3877896070
#>  [731,] -0.0199313005
#>  [732,]  0.1184631125
#>  [733,]  0.2743273010
#>  [734,]  0.5775105425
#>  [735,] -0.5056774630
#>  [736,] -0.2552224425
#>  [737,] -0.6560176353
#>  [738,]  0.0348101351
#>  [739,]  0.0902873096
#>  [740,] -0.6225028210
#>  [741,]  0.0797128348
#>  [742,] -0.1238000548
#>  [743,] -0.2795868175
#>  [744,]  0.4582617314
#>  [745,] -0.2354719079
#>  [746,]  0.2209533894
#>  [747,] -0.5618179465
#>  [748,]  0.2231021093
#>  [749,]  0.4956384436
#>  [750,] -0.5831960283
#>  [751,]  0.2152613123
#>  [752,]  0.0542747032
#>  [753,] -0.2871950654
#>  [754,]  0.8426783292
#>  [755,]  0.7672623956
#>  [756,]  0.2816446739
#>  [757,] -0.2591984330
#>  [758,] -0.3336978253
#>  [759,]  0.3641738199
#>  [760,]  0.9735269661
#>  [761,] -0.3217634758
#>  [762,] -0.7869635400
#>  [763,] -0.4184585485
#>  [764,] -0.4806272557
#>  [765,] -0.8255505286
#>  [766,] -0.5302860747
#>  [767,]  1.3100438376
#>  [768,] -0.9603983599
#>  [769,] -0.0885776340
#>  [770,]  0.2058034252
#>  [771,] -0.7842468232
#>  [772,]  0.0635044349
#>  [773,] -0.6950875052
#>  [774,] -0.1846452173
#>  [775,] -0.3363920013
#>  [776,] -0.1788561470
#>  [777,]  0.8910436190
#>  [778,] -0.1378227135
#>  [779,] -0.5356006359
#>  [780,]  0.3572819579
#>  [781,]  0.1585347313
#>  [782,]  1.0687287745
#>  [783,]  1.1066133406
#>  [784,] -0.2037504826
#>  [785,]  0.9385727458
#>  [786,]  0.6147191507
#>  [787,] -0.6015363271
#>  [788,]  0.1344273146
#>  [789,] -1.2289859526
#>  [790,] -0.2334185391
#>  [791,]  0.6008626409
#>  [792,] -0.3406541746
#>  [793,]  0.0389000135
#>  [794,]  0.6983003709
#>  [795,]  0.3481336234
#>  [796,] -1.3028989554
#>  [797,] -0.2717435179
#>  [798,] -0.2855132968
#>  [799,] -0.0936070333
#>  [800,] -0.3429509998
#>  [801,]  0.0655776669
#>  [802,]  0.4265438505
#>  [803,] -0.1910306454
#>  [804,] -0.9140136202
#>  [805,]  0.3288425975
#>  [806,] -0.8873754493
#>  [807,]  1.0038235569
#>  [808,]  0.9868305517
#>  [809,] -0.2429636438
#>  [810,] -0.9556180721
#>  [811,]  0.4624106351
#>  [812,]  0.3846051532
#>  [813,]  0.2159064515
#>  [814,]  0.6700672870
#>  [815,] -0.4725897050
#>  [816,] -0.4929518906
#>  [817,]  0.0322191879
#>  [818,]  0.3226393176
#>  [819,] -0.6958534240
#>  [820,]  0.9750456384
#>  [821,] -0.2351493784
#>  [822,]  0.0729391805
#>  [823,] -0.1591423649
#>  [824,] -0.5117309414
#>  [825,] -0.4795095958
#>  [826,]  0.4052688752
#>  [827,]  0.4204315647
#>  [828,]  0.2745312392
#>  [829,] -0.0205289924
#>  [830,] -0.4314073568
#>  [831,]  0.6094918348
#>  [832,]  1.0722886897
#>  [833,] -0.2645977153
#>  [834,] -0.3071153523
#>  [835,] -1.0247537799
#>  [836,] -0.0066203198
#>  [837,] -0.3715043925
#>  [838,] -0.9059665162
#>  [839,]  0.3446574226
#>  [840,] -0.4288151695
#>  [841,]  0.5646785101
#>  [842,] -0.3737684092
#>  [843,]  0.0442044857
#>  [844,] -0.3987089975
#>  [845,] -0.7878751902
#>  [846,]  0.2526558932
#>  [847,] -0.7940216721
#>  [848,]  0.9473900905
#>  [849,] -0.4694223589
#>  [850,]  0.3828941189
#>  [851,]  0.6224105475
#>  [852,] -0.5851015896
#>  [853,]  0.0889057826
#>  [854,] -0.7938766278
#>  [855,]  0.3155334026
#>  [856,] -0.1221232804
#>  [857,] -1.3300195007
#>  [858,] -0.5719212348
#>  [859,] -0.9125609384
#>  [860,] -0.2369346297
#>  [861,] -0.1245148780
#>  [862,]  0.1233000140
#>  [863,] -0.3210504101
#>  [864,]  0.2492006992
#>  [865,] -0.2825086429
#>  [866,] -0.5326912742
#>  [867,]  0.8425287094
#>  [868,] -0.1497211783
#>  [869,] -0.0707894714
#>  [870,]  1.4702293019
#>  [871,]  0.3788441173
#>  [872,]  0.1531670624
#>  [873,] -0.0760079516
#>  [874,]  0.4299880763
#>  [875,]  0.2210118748
#>  [876,]  0.2148637327
#>  [877,]  0.1065667170
#>  [878,]  0.3748418794
#>  [879,]  1.3183723882
#>  [880,]  0.1440785278
#>  [881,] -0.2546857904
#>  [882,]  0.3841561647
#>  [883,] -0.4562233748
#>  [884,]  0.0205479356
#>  [885,] -0.5283052893
#>  [886,] -0.0380575119
#>  [887,]  0.8189095851
#>  [888,] -0.3172733240
#>  [889,] -0.0775247548
#>  [890,]  0.6523233123
#>  [891,] -0.5291967793
#>  [892,] -0.2480709824
#>  [893,] -0.5409294040
#>  [894,]  1.6438647924
#>  [895,] -0.0285473602
#>  [896,]  0.5037505007
#>  [897,]  0.1677080631
#>  [898,]  0.4745067744
#>  [899,] -0.3244861939
#>  [900,]  0.2792913243
#>  [901,] -0.5483960392
#>  [902,] -0.0305749669
#>  [903,] -0.1514603314
#>  [904,]  0.2612410255
#>  [905,]  0.2779296922
#>  [906,] -0.5204567438
#>  [907,]  0.2379697183
#>  [908,] -0.0438160738
#>  [909,] -0.1526747367
#>  [910,] -0.0887330336
#>  [911,] -0.1575867912
#>  [912,]  0.1630557708
#>  [913,]  0.7016753047
#>  [914,] -0.8003311557
#>  [915,] -0.5517455382
#>  [916,]  0.5581515129
#>  [917,]  0.9733421026
#>  [918,]  0.4126615822
#>  [919,]  0.2656229969
#>  [920,]  0.9628071016
#>  [921,] -0.9312067534
#>  [922,] -0.4634455202
#>  [923,] -1.3692829613
#>  [924,]  0.2544472271
#>  [925,]  0.4049660112
#>  [926,] -0.8171301071
#>  [927,]  0.9692648139
#>  [928,]  0.2155638727
#>  [929,] -0.0075314663
#>  [930,] -0.3707470096
#>  [931,]  0.3387124092
#>  [932,]  0.6080674524
#>  [933,]  0.3844711084
#>  [934,] -0.4456982781
#>  [935,]  0.4762807668
#>  [936,]  0.8586590271
#>  [937,] -0.8078057559
#>  [938,]  1.1776087750
#>  [939,]  0.2418545468
#>  [940,]  0.3335388192
#>  [941,]  0.1285051732
#>  [942,]  1.3411563387
#>  [943,] -0.5558981950
#>  [944,]  0.1590944842
#>  [945,] -0.6951189757
#>  [946,] -0.2605388590
#>  [947,] -0.5783436803
#>  [948,]  0.3986646311
#>  [949,] -0.1698432634
#>  [950,]  0.3858898592
#>  [951,]  0.2774471829
#>  [952,]  0.9474389006
#>  [953,]  0.2077068563
#>  [954,]  0.2806173791
#>  [955,] -0.1963264975
#>  [956,] -0.8030974166
#>  [957,] -0.2807482662
#>  [958,] -0.3026745665
#>  [959,]  0.5083256933
#>  [960,] -0.2461689204
#>  [961,] -0.4289212344
#>  [962,]  0.7402211651
#>  [963,]  0.4969454558
#>  [964,] -0.2984586746
#>  [965,]  1.1478120924
#>  [966,]  0.4754655764
#>  [967,]  0.3991696073
#>  [968,] -0.9194252349
#>  [969,]  0.1032745219
#>  [970,]  0.8412753667
#>  [971,]  0.0857662245
#>  [972,] -0.0358825081
#>  [973,] -0.6591380693
#>  [974,] -0.0544398507
#>  [975,]  0.0107345295
#>  [976,] -1.0040146620
#>  [977,]  0.2431172533
#>  [978,] -0.2085764904
#>  [979,] -0.5573452864
#>  [980,] -0.0040646845
#>  [981,] -0.1772517418
#>  [982,] -0.1807434915
#>  [983,] -0.7999411671
#>  [984,]  0.6381955520
#>  [985,] -0.7711446892
#>  [986,] -0.5225992541
#>  [987,]  0.0978367780
#>  [988,] -0.2031551713
#>  [989,] -0.0101076912
#>  [990,] -0.1496888327
#>  [991,] -0.2085557226
#>  [992,]  0.6977805953
#>  [993,]  0.2173102994
#>  [994,]  0.4405124281
#>  [995,] -0.0108878601
#>  [996,]  0.1153961412
#>  [997,] -0.3793953827
#>  [998,] -0.6079843124
#>  [999,]  0.6614656504
#> [1000,] -0.0739641974