Logistic Function

The elementwise logistic function, \(\log(1 + e^x)\). This is a special case of log(sum(exp)) that evaluates to a vector rather than to a scalar, which is useful for logistic regression.

logistic(x)

Arguments

x
An Expression, vector, or matrix.

Value

An Expression representing the logistic function evaluated at the input.

Examples

n <- 20
m <- 1000
sigma <- 45

beta_true <- stats::rnorm(n)
idxs <- sample(n, size = 0.8*n, replace = FALSE)
beta_true[idxs] <- 0
X <- matrix(stats::rnorm(m*n, 0, 5), nrow = m, ncol = n)
y <- sign(X %*% beta_true + stats::rnorm(m, 0, sigma))

beta <- Variable(n)
X_sign <- apply(X, 2, function(x) { ifelse(y <= 0, -1, 1) * x })
obj <- -sum(logistic(-X[y <= 0,] %*% beta)) - sum(logistic(X[y == 1,] %*% beta))
prob <- Problem(Maximize(obj))
result <- solve(prob)

log_odds <- result$getValue(X %*% beta)
beta_res <- result$getValue(beta)
y_probs <- 1/(1 + exp(-X %*% beta_res))
log(y_probs/(1 - y_probs))

#>                  [,1]
#>    [1,] -0.1512327392
#>    [2,] -0.5257626738
#>    [3,]  0.5114658491
#>    [4,]  0.7992249360
#>    [5,] -0.5802269621
#>    [6,] -0.1552604273
#>    [7,]  0.2828194377
#>    [8,] -0.6371598635
#>    [9,]  0.4036526449
#>   [10,] -0.1608277079
#>   [11,]  0.0209090567
#>   [12,]  0.0077610053
#>   [13,] -0.4483641503
#>   [14,]  0.5149202700
#>   [15,] -0.2185827677
#>   [16,] -0.4208400759
#>   [17,] -0.2326037819
#>   [18,] -0.3449665432
#>   [19,]  0.4473529430
#>   [20,] -0.5881919972
#>   [21,] -0.3457877893
#>   [22,] -0.0013056403
#>   [23,]  0.1566956965
#>   [24,]  0.3498795850
#>   [25,]  0.4246660270
#>   [26,]  0.0819393134
#>   [27,] -0.4932219918
#>   [28,]  0.8131683143
#>   [29,]  0.5770575478
#>   [30,] -0.3631207881
#>   [31,]  0.0742711970
#>   [32,] -0.4363212564
#>   [33,] -0.1007047491
#>   [34,]  0.4988482000
#>   [35,] -0.5652759781
#>   [36,]  0.3451929339
#>   [37,]  0.2341367863
#>   [38,]  0.0209769711
#>   [39,] -0.7063816008
#>   [40,]  0.4423177797
#>   [41,] -0.1981304608
#>   [42,]  0.6294170518
#>   [43,]  0.8931286115
#>   [44,] -0.5977770539
#>   [45,] -0.0966905117
#>   [46,]  0.3815851524
#>   [47,] -0.1236579418
#>   [48,]  0.0555233827
#>   [49,] -0.7291781455
#>   [50,] -0.0138016235
#>   [51,]  0.5780875106
#>   [52,]  0.2016736074
#>   [53,] -0.4312627052
#>   [54,]  0.0244209511
#>   [55,] -0.1569526214
#>   [56,] -0.5379048253
#>   [57,] -0.6205552077
#>   [58,]  0.4913096170
#>   [59,] -0.4294142293
#>   [60,] -0.6115368771
#>   [61,]  0.2087933562
#>   [62,] -0.7023107348
#>   [63,] -0.3245072443
#>   [64,]  0.1911830206
#>   [65,] -0.4865232323
#>   [66,] -0.1496075925
#>   [67,] -0.0475827373
#>   [68,]  0.6939027918
#>   [69,] -0.4410365202
#>   [70,]  0.5756137361
#>   [71,]  0.2092143836
#>   [72,] -0.3561114645
#>   [73,] -0.0550636394
#>   [74,] -0.3936877007
#>   [75,] -0.0786513859
#>   [76,]  0.4344101908
#>   [77,]  0.1406404866
#>   [78,] -0.3142141673
#>   [79,] -0.1173922862
#>   [80,]  0.1320157261
#>   [81,]  0.7411364093
#>   [82,]  0.1460754232
#>   [83,]  0.0580000903
#>   [84,]  0.0099304029
#>   [85,] -1.1582595427
#>   [86,]  0.4611187721
#>   [87,]  0.4765030593
#>   [88,]  0.0234602516
#>   [89,] -0.3157590950
#>   [90,] -0.2978387062
#>   [91,] -0.5408820203
#>   [92,] -0.3067093374
#>   [93,]  0.5229142941
#>   [94,] -0.2460522859
#>   [95,] -0.4605067627
#>   [96,] -0.3128487767
#>   [97,] -0.1644772338
#>   [98,] -0.0892934900
#>   [99,]  0.0220211370
#>  [100,]  0.0755417922
#>  [101,] -0.1987794096
#>  [102,] -0.0162244935
#>  [103,]  0.1938774562
#>  [104,] -0.2453292953
#>  [105,] -0.0098544540
#>  [106,]  0.1190140437
#>  [107,] -0.0247059788
#>  [108,] -0.2371163861
#>  [109,] -0.1987502153
#>  [110,]  0.1623694494
#>  [111,] -0.6547616723
#>  [112,]  0.0007253448
#>  [113,]  0.4077962218
#>  [114,] -0.3337025014
#>  [115,] -0.2374322349
#>  [116,] -0.2384502526
#>  [117,] -0.0776120074
#>  [118,]  0.2444223750
#>  [119,] -0.5175979168
#>  [120,] -0.0408714336
#>  [121,]  0.2594029894
#>  [122,]  0.1851113679
#>  [123,] -0.3057020810
#>  [124,]  0.1817301504
#>  [125,] -0.3924323343
#>  [126,] -0.3747397106
#>  [127,] -0.2468914002
#>  [128,] -0.2867853573
#>  [129,]  0.0043721859
#>  [130,]  0.3797538358
#>  [131,] -0.4516364348
#>  [132,] -0.4742894511
#>  [133,] -0.1420311215
#>  [134,] -0.5189809733
#>  [135,] -0.0330336678
#>  [136,] -0.5795920381
#>  [137,] -0.1851819425
#>  [138,]  0.7509844144
#>  [139,] -0.1012214267
#>  [140,] -0.2504542826
#>  [141,]  0.3107235940
#>  [142,] -0.4465599951
#>  [143,]  0.4025964213
#>  [144,] -0.1782021801
#>  [145,] -0.7555277733
#>  [146,] -0.0794990160
#>  [147,] -0.0930925546
#>  [148,]  0.1833886452
#>  [149,] -0.9422945336
#>  [150,]  0.2061585236
#>  [151,] -0.4108247403
#>  [152,] -0.0641600265
#>  [153,]  0.1808929984
#>  [154,]  0.2339412372
#>  [155,]  0.3484593795
#>  [156,]  0.3269124135
#>  [157,] -0.2281267664
#>  [158,] -0.6359532582
#>  [159,]  0.6101790542
#>  [160,]  0.3406240610
#>  [161,]  0.5207300853
#>  [162,]  0.1597021800
#>  [163,]  0.4417974065
#>  [164,]  0.0478402976
#>  [165,] -0.1129608373
#>  [166,] -0.0970093215
#>  [167,]  0.2432596912
#>  [168,]  0.1165162162
#>  [169,]  0.4979461761
#>  [170,] -0.1255523791
#>  [171,]  0.4542965136
#>  [172,]  0.0115033881
#>  [173,]  0.3172201929
#>  [174,]  0.3332538277
#>  [175,] -0.4621708091
#>  [176,]  0.1207209329
#>  [177,]  0.0786659656
#>  [178,]  0.6078686199
#>  [179,] -0.3149489222
#>  [180,] -0.2151985986
#>  [181,] -0.1165496153
#>  [182,]  0.8404707788
#>  [183,] -0.2839275719
#>  [184,] -0.3553634774
#>  [185,] -0.2568180883
#>  [186,]  0.0605815789
#>  [187,] -0.3163220675
#>  [188,] -0.3408829137
#>  [189,] -0.5710357157
#>  [190,]  0.2246683588
#>  [191,]  0.1269968972
#>  [192,]  0.5585459079
#>  [193,]  0.4199687805
#>  [194,] -0.0407259140
#>  [195,]  0.3529405251
#>  [196,]  0.0417058111
#>  [197,] -0.3836661851
#>  [198,]  0.3262889429
#>  [199,]  0.5971360931
#>  [200,] -0.5353773777
#>  [201,] -0.1457956269
#>  [202,]  0.0985177921
#>  [203,] -0.8165465995
#>  [204,]  0.2789236510
#>  [205,] -0.3319939921
#>  [206,] -0.0917915106
#>  [207,] -0.2183611688
#>  [208,]  0.0682147918
#>  [209,] -0.2069582481
#>  [210,] -0.0134737598
#>  [211,]  0.6800813497
#>  [212,] -0.2728916611
#>  [213,]  0.2116024198
#>  [214,]  0.1420494405
#>  [215,] -0.1101618990
#>  [216,] -0.0492065809
#>  [217,] -0.2711077476
#>  [218,]  0.0454623562
#>  [219,] -0.7244669395
#>  [220,] -0.0974968420
#>  [221,] -0.1423415926
#>  [222,] -0.5813039758
#>  [223,]  0.8599241321
#>  [224,]  0.1453288787
#>  [225,] -0.0730882566
#>  [226,] -0.0960220079
#>  [227,] -0.0848013874
#>  [228,]  0.1456756558
#>  [229,]  0.3931986332
#>  [230,]  0.2312130013
#>  [231,]  0.3543677059
#>  [232,] -0.4344691834
#>  [233,] -0.2300551876
#>  [234,]  0.0243872959
#>  [235,]  0.5308972231
#>  [236,] -0.9252644791
#>  [237,]  0.3825488379
#>  [238,] -1.1241089570
#>  [239,] -0.6147235580
#>  [240,]  0.2155596158
#>  [241,] -0.0049826181
#>  [242,]  0.4166832420
#>  [243,] -0.8694854246
#>  [244,]  0.1175735131
#>  [245,]  0.4106108197
#>  [246,]  0.1308458107
#>  [247,]  0.3449266701
#>  [248,] -0.3606074892
#>  [249,] -0.0886816926
#>  [250,] -0.5744214999
#>  [251,]  0.3213368633
#>  [252,]  0.6346095290
#>  [253,] -0.0169508339
#>  [254,]  0.3276597492
#>  [255,] -0.3376062727
#>  [256,] -0.4674020058
#>  [257,] -0.4962799189
#>  [258,] -0.3096343104
#>  [259,] -0.1751939614
#>  [260,]  0.2693289665
#>  [261,]  0.2218169265
#>  [262,] -0.0827697946
#>  [263,]  0.5255193124
#>  [264,] -0.7045465094
#>  [265,]  0.0124056659
#>  [266,]  0.2458982994
#>  [267,]  0.2556801332
#>  [268,] -0.1402177570
#>  [269,] -0.3157987369
#>  [270,] -0.7645780537
#>  [271,] -0.9177089202
#>  [272,] -0.3650930352
#>  [273,]  0.0863373286
#>  [274,]  0.2565549004
#>  [275,]  0.3373832652
#>  [276,] -0.2055857503
#>  [277,] -0.3232327013
#>  [278,] -0.0607034818
#>  [279,]  0.1709288093
#>  [280,]  0.5996912804
#>  [281,]  0.1790988492
#>  [282,]  0.1677673725
#>  [283,] -0.4601642699
#>  [284,] -0.1896764137
#>  [285,]  0.5162372097
#>  [286,]  0.2247403628
#>  [287,] -0.1423247171
#>  [288,]  0.5647541132
#>  [289,]  0.1484333219
#>  [290,] -0.1529546478
#>  [291,]  0.0811973455
#>  [292,] -0.5171762860
#>  [293,]  0.8562018220
#>  [294,]  0.2466086901
#>  [295,]  0.2297255273
#>  [296,]  0.0780138405
#>  [297,] -0.3740006197
#>  [298,] -0.6569671307
#>  [299,]  0.4466854530
#>  [300,]  0.3193612480
#>  [301,] -0.7532541141
#>  [302,] -0.1042899458
#>  [303,] -0.0947362239
#>  [304,]  0.0072593085
#>  [305,] -1.0924472206
#>  [306,]  0.3672115965
#>  [307,]  0.4680411312
#>  [308,]  0.0518949199
#>  [309,] -0.0488201538
#>  [310,]  0.3931419626
#>  [311,] -0.3356293077
#>  [312,]  0.6219700506
#>  [313,] -0.0434779538
#>  [314,] -0.2120933488
#>  [315,]  0.2695342980
#>  [316,] -0.1912399107
#>  [317,] -0.2433399285
#>  [318,] -0.1649345719
#>  [319,] -0.4254217077
#>  [320,] -0.1458856292
#>  [321,] -0.0812110144
#>  [322,] -0.0504176231
#>  [323,]  0.5114320405
#>  [324,]  0.3635596159
#>  [325,]  0.5747097970
#>  [326,]  0.9475661227
#>  [327,]  0.0549484192
#>  [328,] -0.2248451020
#>  [329,] -0.3178432049
#>  [330,] -0.1462510755
#>  [331,] -0.4880356499
#>  [332,]  0.5576292798
#>  [333,]  0.1552707062
#>  [334,] -0.3338679479
#>  [335,]  0.0897429269
#>  [336,] -0.4802568788
#>  [337,]  0.3663155698
#>  [338,]  0.4243055157
#>  [339,]  0.1710895784
#>  [340,] -0.0107701197
#>  [341,]  0.1607196878
#>  [342,] -0.3166296245
#>  [343,] -0.0660252241
#>  [344,]  0.3841921773
#>  [345,] -0.1283484838
#>  [346,]  0.0640515356
#>  [347,]  0.6768916078
#>  [348,]  0.4538410633
#>  [349,] -0.5264205042
#>  [350,] -0.6216847768
#>  [351,] -0.0271642545
#>  [352,]  0.5553809140
#>  [353,] -0.2826731110
#>  [354,] -0.1568139209
#>  [355,]  0.0418222535
#>  [356,]  0.1364822469
#>  [357,]  0.5334601386
#>  [358,]  0.2536976773
#>  [359,]  0.0762468336
#>  [360,] -0.1968529600
#>  [361,] -0.4376436378
#>  [362,] -0.5721600647
#>  [363,]  0.4255226020
#>  [364,] -0.0630639852
#>  [365,] -0.4574771443
#>  [366,]  0.0664196830
#>  [367,] -0.3574212484
#>  [368,]  0.4968476922
#>  [369,] -0.5546525620
#>  [370,] -0.1435537607
#>  [371,]  0.0657813675
#>  [372,]  0.2500360574
#>  [373,]  0.2286563039
#>  [374,]  0.4367255825
#>  [375,] -0.8660086502
#>  [376,] -0.0839822282
#>  [377,] -0.1506327389
#>  [378,] -0.2131731938
#>  [379,] -0.1619823212
#>  [380,]  0.1336727752
#>  [381,]  0.3598033806
#>  [382,]  0.2547212979
#>  [383,] -0.0871878614
#>  [384,]  0.1199043666
#>  [385,]  0.1458629999
#>  [386,]  0.4655681537
#>  [387,]  0.7991641800
#>  [388,] -0.1161096889
#>  [389,] -0.9672200646
#>  [390,]  0.4618914498
#>  [391,]  0.4060583949
#>  [392,]  0.8001029138
#>  [393,]  0.1904674706
#>  [394,] -0.8720488299
#>  [395,]  0.3741143110
#>  [396,] -0.0517867060
#>  [397,]  0.3852316641
#>  [398,]  0.0556012842
#>  [399,]  0.0306596876
#>  [400,]  0.2202137769
#>  [401,]  0.4874155194
#>  [402,]  0.1997851031
#>  [403,]  0.2896000561
#>  [404,]  0.6915338444
#>  [405,]  0.0872412664
#>  [406,]  0.0079243561
#>  [407,]  0.1832743571
#>  [408,]  0.7240597614
#>  [409,] -0.6543770425
#>  [410,]  0.3852143608
#>  [411,]  0.4000673496
#>  [412,]  0.6134864308
#>  [413,] -0.2209624209
#>  [414,] -0.0307809491
#>  [415,] -0.5392675738
#>  [416,]  0.4937839987
#>  [417,]  0.5349402972
#>  [418,] -0.3105960334
#>  [419,]  0.1268738265
#>  [420,]  0.1194944651
#>  [421,]  0.0543274763
#>  [422,] -0.2743475593
#>  [423,]  0.1199134921
#>  [424,]  0.0565480398
#>  [425,] -0.1435945864
#>  [426,] -0.3995648497
#>  [427,]  0.3351326181
#>  [428,] -0.5944576358
#>  [429,] -0.2996651976
#>  [430,] -0.3547132067
#>  [431,] -0.0138812142
#>  [432,]  0.2659117943
#>  [433,]  0.7052634164
#>  [434,] -0.2412270230
#>  [435,]  0.1844981356
#>  [436,]  0.1166556938
#>  [437,] -0.0762763220
#>  [438,] -0.2875729660
#>  [439,] -0.5310292362
#>  [440,] -0.8341017439
#>  [441,]  0.5036530544
#>  [442,]  0.2566139956
#>  [443,] -0.4000612962
#>  [444,] -0.5060102779
#>  [445,] -0.4590976427
#>  [446,]  0.0657697216
#>  [447,]  0.3699856895
#>  [448,] -0.3682613048
#>  [449,] -0.6567561721
#>  [450,]  1.2191295918
#>  [451,] -0.1814469344
#>  [452,] -0.1633953368
#>  [453,] -0.2302876189
#>  [454,] -0.3550101208
#>  [455,]  0.1857538681
#>  [456,]  0.0941599096
#>  [457,]  0.2865929831
#>  [458,]  0.0805475862
#>  [459,] -0.2800686691
#>  [460,]  0.1046573221
#>  [461,]  0.2222321032
#>  [462,] -0.1195807253
#>  [463,] -0.3570815628
#>  [464,]  0.4830064900
#>  [465,] -0.8001300944
#>  [466,] -0.0403673986
#>  [467,] -0.1318656310
#>  [468,] -0.1133053172
#>  [469,]  0.2767749172
#>  [470,]  0.2566467389
#>  [471,] -0.0792394852
#>  [472,]  0.0109142539
#>  [473,]  0.0295057746
#>  [474,]  0.3033314437
#>  [475,] -0.3438333371
#>  [476,]  0.2781499215
#>  [477,] -0.2885882911
#>  [478,] -0.1994364475
#>  [479,] -0.1473120854
#>  [480,]  0.7231077404
#>  [481,]  0.0309672947
#>  [482,]  0.7910784616
#>  [483,] -0.0247686686
#>  [484,]  0.3524580610
#>  [485,]  0.2933760353
#>  [486,]  0.3799934689
#>  [487,] -0.1590741355
#>  [488,]  0.1206895217
#>  [489,] -0.4095845289
#>  [490,] -0.4269125092
#>  [491,]  0.4859982519
#>  [492,] -0.5568444140
#>  [493,]  0.4099477260
#>  [494,] -0.0226588241
#>  [495,]  0.4670069118
#>  [496,]  0.0641337326
#>  [497,] -0.0273365211
#>  [498,]  0.3107904523
#>  [499,]  0.1707052390
#>  [500,] -0.2494203545
#>  [501,] -0.0194051458
#>  [502,]  0.0935750845
#>  [503,]  0.4189438216
#>  [504,] -0.1275423881
#>  [505,] -0.3486635563
#>  [506,]  0.0422836888
#>  [507,]  0.0725527661
#>  [508,] -0.8200094190
#>  [509,] -0.0383596171
#>  [510,]  0.4148687154
#>  [511,]  0.1801690189
#>  [512,] -0.4011275924
#>  [513,]  0.2477614986
#>  [514,] -0.2049428027
#>  [515,]  0.5713928387
#>  [516,] -0.1709063742
#>  [517,] -0.1626115226
#>  [518,]  0.7878591741
#>  [519,] -0.4355097673
#>  [520,]  0.0243322221
#>  [521,]  0.6267695731
#>  [522,]  0.4753869260
#>  [523,]  0.1488071150
#>  [524,] -0.0904985437
#>  [525,] -0.2610315918
#>  [526,] -0.1334299044
#>  [527,]  0.3753536969
#>  [528,]  0.2890546475
#>  [529,] -0.2249871548
#>  [530,]  0.2083647444
#>  [531,] -0.1841955406
#>  [532,]  0.0379072650
#>  [533,] -0.6507300104
#>  [534,]  0.0746936691
#>  [535,]  0.0114813441
#>  [536,] -0.2615438360
#>  [537,] -0.4194030528
#>  [538,]  0.2296788484
#>  [539,]  0.3716487151
#>  [540,]  0.0724437381
#>  [541,] -0.3402727190
#>  [542,]  0.2090159986
#>  [543,]  0.0892026478
#>  [544,]  0.3528706321
#>  [545,]  0.2859247262
#>  [546,] -0.0642838221
#>  [547,]  0.0411478685
#>  [548,] -0.1515879895
#>  [549,]  0.3579396797
#>  [550,] -0.3073099987
#>  [551,]  0.3895683462
#>  [552,] -0.2121238148
#>  [553,]  0.2313092764
#>  [554,] -0.0259338495
#>  [555,] -0.0128369708
#>  [556,] -0.1571113351
#>  [557,] -0.2900717126
#>  [558,] -0.2411077452
#>  [559,]  0.1212344487
#>  [560,]  0.8808583531
#>  [561,] -0.0604675963
#>  [562,]  0.2469495020
#>  [563,] -0.5141370910
#>  [564,]  0.2487159061
#>  [565,]  1.0412714340
#>  [566,] -0.3158877760
#>  [567,]  0.1023146500
#>  [568,] -0.0928703233
#>  [569,]  0.5444258602
#>  [570,]  0.1346230779
#>  [571,]  0.1502155601
#>  [572,]  0.6917386432
#>  [573,]  0.1813858664
#>  [574,] -0.2829543068
#>  [575,] -0.1978569570
#>  [576,]  0.4702876422
#>  [577,]  0.5988809306
#>  [578,]  0.7517843709
#>  [579,]  0.3224234854
#>  [580,] -0.3979238758
#>  [581,]  0.1834813980
#>  [582,]  0.1668504219
#>  [583,] -0.0319079340
#>  [584,]  0.0022896694
#>  [585,]  0.2647128573
#>  [586,]  0.9741279611
#>  [587,] -0.2951108935
#>  [588,]  0.0674837527
#>  [589,] -0.1821264063
#>  [590,]  0.2285819245
#>  [591,]  0.3311209346
#>  [592,] -0.0410970638
#>  [593,]  0.1125051964
#>  [594,]  0.3180572319
#>  [595,] -0.5653290606
#>  [596,] -0.4292524643
#>  [597,]  0.1244314729
#>  [598,]  0.0062606067
#>  [599,] -0.2656081580
#>  [600,]  0.1067541964
#>  [601,]  0.2427983803
#>  [602,] -0.2316803278
#>  [603,]  0.6842827957
#>  [604,]  0.1646486598
#>  [605,]  1.0846701240
#>  [606,]  0.0774839585
#>  [607,] -0.5472544990
#>  [608,]  0.5023345183
#>  [609,] -0.3287775883
#>  [610,]  0.3993536321
#>  [611,] -0.1052874200
#>  [612,] -0.5143202397
#>  [613,]  0.3160836719
#>  [614,] -0.3992477226
#>  [615,] -0.3602901398
#>  [616,]  0.2558363109
#>  [617,] -0.0088864545
#>  [618,]  0.2037507504
#>  [619,] -0.1264163712
#>  [620,] -0.0338936822
#>  [621,]  0.0710027562
#>  [622,] -0.7359476863
#>  [623,] -0.0948652533
#>  [624,]  0.4756797131
#>  [625,]  0.0569094244
#>  [626,]  0.6998052521
#>  [627,]  0.2043694380
#>  [628,]  0.3919539399
#>  [629,]  0.0929610033
#>  [630,]  0.3273847582
#>  [631,] -0.0089605584
#>  [632,] -0.5400101037
#>  [633,]  1.0120085015
#>  [634,] -0.1705995089
#>  [635,]  0.5003082845
#>  [636,] -0.3983039657
#>  [637,]  0.5637172085
#>  [638,]  0.0132493505
#>  [639,]  0.0579068490
#>  [640,] -0.1452491444
#>  [641,]  0.5306467795
#>  [642,] -0.2076738254
#>  [643,] -0.3766632304
#>  [644,]  0.8316500566
#>  [645,]  0.3990736381
#>  [646,] -0.1193010113
#>  [647,]  0.0587067961
#>  [648,] -0.8801068397
#>  [649,]  0.1952729404
#>  [650,] -0.1295752384
#>  [651,] -0.5343537165
#>  [652,]  0.0829575131
#>  [653,] -0.0821056129
#>  [654,] -0.3646425919
#>  [655,]  0.0090243955
#>  [656,] -0.3043874620
#>  [657,] -0.2307501220
#>  [658,] -0.2816903583
#>  [659,]  0.5453213594
#>  [660,]  0.4943172892
#>  [661,] -0.0694865898
#>  [662,]  0.2464759737
#>  [663,] -0.1905581086
#>  [664,] -0.1927264183
#>  [665,] -0.4826220036
#>  [666,]  0.5685778783
#>  [667,] -0.2023288041
#>  [668,] -0.2769702852
#>  [669,] -0.0286869614
#>  [670,] -0.1058886767
#>  [671,] -0.4670400481
#>  [672,] -0.6622505419
#>  [673,] -0.1840387313
#>  [674,] -0.5631154572
#>  [675,] -0.0005141562
#>  [676,] -0.0769402405
#>  [677,]  0.0609791086
#>  [678,] -0.6105571289
#>  [679,] -0.2705945124
#>  [680,] -0.0468237564
#>  [681,]  0.7030817515
#>  [682,]  0.7426399908
#>  [683,]  0.0412670541
#>  [684,] -0.6023112562
#>  [685,]  0.1477146833
#>  [686,] -0.3045082662
#>  [687,] -0.2393067981
#>  [688,]  0.3687457224
#>  [689,] -0.7544859308
#>  [690,] -0.3660157330
#>  [691,]  0.4334016991
#>  [692,] -0.4145319368
#>  [693,]  0.0037068458
#>  [694,]  0.3508758179
#>  [695,] -0.6769988996
#>  [696,] -0.1877597483
#>  [697,] -0.0771017475
#>  [698,]  0.0358811146
#>  [699,] -0.1175038918
#>  [700,] -0.2619243185
#>  [701,] -0.1617555450
#>  [702,] -0.3315283925
#>  [703,] -0.4106360623
#>  [704,] -0.2956781256
#>  [705,] -0.3646458693
#>  [706,]  0.3889440277
#>  [707,] -0.5480080310
#>  [708,]  0.1164978546
#>  [709,] -0.0845048363
#>  [710,] -0.1698285575
#>  [711,]  0.4466174077
#>  [712,] -0.4736937767
#>  [713,] -0.0727101373
#>  [714,]  0.0802735225
#>  [715,] -0.0390090671
#>  [716,] -0.3466932500
#>  [717,] -0.3059619621
#>  [718,] -0.3758883501
#>  [719,] -0.5217983817
#>  [720,]  0.4086460554
#>  [721,] -0.6621661914
#>  [722,]  0.4073547189
#>  [723,] -0.1183767884
#>  [724,] -0.0447327638
#>  [725,]  0.2275078787
#>  [726,]  0.3128122281
#>  [727,] -1.1760251719
#>  [728,]  0.7656412962
#>  [729,]  0.5864054785
#>  [730,]  0.1133193943
#>  [731,]  0.0273731186
#>  [732,] -0.4127006032
#>  [733,] -0.2945462930
#>  [734,] -0.2686248663
#>  [735,]  0.0539699004
#>  [736,] -0.0918296723
#>  [737,] -0.5451948122
#>  [738,]  0.2867385177
#>  [739,] -0.9231771522
#>  [740,]  0.0462587873
#>  [741,] -0.3634809934
#>  [742,]  0.4285296356
#>  [743,]  0.4889006623
#>  [744,] -0.2358217003
#>  [745,] -0.0828868645
#>  [746,] -0.0018938395
#>  [747,]  0.4813928165
#>  [748,] -0.2677799074
#>  [749,]  0.2133374392
#>  [750,]  0.0800877362
#>  [751,]  0.6680924383
#>  [752,]  0.4135037390
#>  [753,]  0.6302711041
#>  [754,] -0.1911166133
#>  [755,] -0.3223945984
#>  [756,] -0.2938076475
#>  [757,]  0.0305289568
#>  [758,] -0.1906945111
#>  [759,] -0.2832617030
#>  [760,]  0.0498834363
#>  [761,]  0.4865394651
#>  [762,] -0.3178243753
#>  [763,] -0.4219994909
#>  [764,] -0.0720574735
#>  [765,] -0.4554098753
#>  [766,]  0.4472334730
#>  [767,]  0.0468158432
#>  [768,]  0.8033247463
#>  [769,] -0.1944011253
#>  [770,]  0.0939207051
#>  [771,]  0.1909370246
#>  [772,]  0.1167339321
#>  [773,]  0.3448069499
#>  [774,] -0.1143293802
#>  [775,] -0.1680775655
#>  [776,]  0.1543573342
#>  [777,]  0.5046278594
#>  [778,] -0.1104754670
#>  [779,]  0.6644782366
#>  [780,]  0.0675414141
#>  [781,] -0.5461751251
#>  [782,] -0.3213605758
#>  [783,] -0.5353628857
#>  [784,]  0.0219458055
#>  [785,]  0.5276263467
#>  [786,] -0.0085726091
#>  [787,]  0.9556108028
#>  [788,] -0.6050090979
#>  [789,]  0.5521619414
#>  [790,]  0.4333067776
#>  [791,] -0.0418444185
#>  [792,]  0.1616317504
#>  [793,]  0.4490524472
#>  [794,] -0.1003317371
#>  [795,]  0.7805950956
#>  [796,]  0.2425791278
#>  [797,]  0.0682830968
#>  [798,]  0.2222677563
#>  [799,] -0.2425209148
#>  [800,]  0.5778148648
#>  [801,] -0.3755514191
#>  [802,] -0.3123205316
#>  [803,] -0.1606939217
#>  [804,] -0.3095177936
#>  [805,] -0.4050472944
#>  [806,]  0.5393111879
#>  [807,]  0.1696857893
#>  [808,] -0.6866034406
#>  [809,]  0.0139085034
#>  [810,] -0.0550586743
#>  [811,]  0.3572861872
#>  [812,] -0.3562064109
#>  [813,]  0.2838849033
#>  [814,]  0.1638442440
#>  [815,]  0.5267597522
#>  [816,]  0.0302838283
#>  [817,] -1.2676363148
#>  [818,]  0.2984944896
#>  [819,]  0.1576442026
#>  [820,]  0.3934661744
#>  [821,] -0.3546490154
#>  [822,]  0.0901837218
#>  [823,] -0.1469907671
#>  [824,]  0.2370816275
#>  [825,]  0.0953856723
#>  [826,]  0.0895762503
#>  [827,]  0.6243555223
#>  [828,] -0.4257684032
#>  [829,] -0.5573656354
#>  [830,]  0.0549955037
#>  [831,]  0.7993057730
#>  [832,]  0.1236822101
#>  [833,] -0.0736000938
#>  [834,]  1.1421211853
#>  [835,] -0.2104748184
#>  [836,] -0.3393851007
#>  [837,]  0.2374172969
#>  [838,]  0.7730762345
#>  [839,] -0.2007182488
#>  [840,] -0.0374364221
#>  [841,] -0.4616721209
#>  [842,] -0.2027344723
#>  [843,] -0.8492626436
#>  [844,] -0.2594919667
#>  [845,]  0.3838682199
#>  [846,]  0.6817554386
#>  [847,] -0.2540124331
#>  [848,]  0.3102389844
#>  [849,]  0.4341193620
#>  [850,] -0.1977225907
#>  [851,]  0.5785195696
#>  [852,]  0.2150550083
#>  [853,] -0.0546424750
#>  [854,]  0.8616608460
#>  [855,]  0.8304799473
#>  [856,]  0.0487553035
#>  [857,] -0.3080410747
#>  [858,]  0.3471052263
#>  [859,]  0.1700263730
#>  [860,]  0.1652062776
#>  [861,]  0.5943450320
#>  [862,] -0.1846325509
#>  [863,]  0.0467204639
#>  [864,]  0.0406973468
#>  [865,]  0.0833793486
#>  [866,] -0.5260301353
#>  [867,] -0.3966062636
#>  [868,] -0.0042364977
#>  [869,]  0.4474306059
#>  [870,] -0.1682518178
#>  [871,]  0.3535679123
#>  [872,]  0.0415488558
#>  [873,] -0.0543031017
#>  [874,] -0.4093497508
#>  [875,]  0.0203590600
#>  [876,] -0.3782580772
#>  [877,]  0.5008275060
#>  [878,]  0.4234509051
#>  [879,]  0.4721844282
#>  [880,]  0.0956667336
#>  [881,]  0.1657111024
#>  [882,] -0.3424603254
#>  [883,]  0.0989191118
#>  [884,] -0.0753767552
#>  [885,]  0.3403171562
#>  [886,]  0.3980726209
#>  [887,] -0.1850779059
#>  [888,]  0.3337983398
#>  [889,]  0.6750158550
#>  [890,] -0.2300274342
#>  [891,] -0.9985796187
#>  [892,] -0.0063929949
#>  [893,]  0.3754610122
#>  [894,] -0.4335720922
#>  [895,] -0.3638371847
#>  [896,] -0.3808043826
#>  [897,] -0.0550613935
#>  [898,] -0.8826760099
#>  [899,]  0.0041194114
#>  [900,] -0.0693857785
#>  [901,]  0.0120821010
#>  [902,] -0.6113648898
#>  [903,] -0.7929882002
#>  [904,] -0.2683569095
#>  [905,]  0.0179779087
#>  [906,]  0.4087320109
#>  [907,] -0.2229934134
#>  [908,]  0.3134038781
#>  [909,] -0.5257990006
#>  [910,] -0.2738958448
#>  [911,]  0.9405235196
#>  [912,]  0.4179851954
#>  [913,] -0.5561979137
#>  [914,] -0.4528138415
#>  [915,] -0.0269244154
#>  [916,] -0.8978362052
#>  [917,] -0.3030352334
#>  [918,]  0.0429375221
#>  [919,] -0.1193569305
#>  [920,]  0.0245582614
#>  [921,]  0.3928934879
#>  [922,] -0.5981068497
#>  [923,] -0.0311874443
#>  [924,]  0.2372125248
#>  [925,] -0.2186911948
#>  [926,]  0.0440884478
#>  [927,]  0.2835218277
#>  [928,] -0.1613355792
#>  [929,] -0.1070804338
#>  [930,]  0.1867195591
#>  [931,]  0.4172857322
#>  [932,]  0.2298914698
#>  [933,] -0.3478589490
#>  [934,] -0.1047737708
#>  [935,]  0.0787040523
#>  [936,] -0.2377760639
#>  [937,]  0.0695474304
#>  [938,]  0.7887968443
#>  [939,] -0.5002676008
#>  [940,] -0.1752422197
#>  [941,] -0.1969026174
#>  [942,] -0.1389305208
#>  [943,]  0.0031322868
#>  [944,]  0.0105267136
#>  [945,]  0.1800443711
#>  [946,]  0.3131741994
#>  [947,] -0.3855556116
#>  [948,]  0.5354398441
#>  [949,]  0.0202200603
#>  [950,] -0.0554930007
#>  [951,]  0.4689179272
#>  [952,]  0.1790549916
#>  [953,] -0.1517863102
#>  [954,]  0.2355545324
#>  [955,] -0.5895087604
#>  [956,]  0.0487135963
#>  [957,] -0.3165505095
#>  [958,]  0.4542946561
#>  [959,] -0.4058444966
#>  [960,] -0.4740818982
#>  [961,]  0.0739455071
#>  [962,] -0.2225245316
#>  [963,]  0.2841121225
#>  [964,] -0.2698232800
#>  [965,] -0.3768978708
#>  [966,] -0.5811487279
#>  [967,] -0.0428299367
#>  [968,]  0.6700306287
#>  [969,] -0.0655353931
#>  [970,]  0.5843378197
#>  [971,] -0.4959691181
#>  [972,] -0.3986484414
#>  [973,]  0.1181301994
#>  [974,]  0.1010923946
#>  [975,] -0.0384638623
#>  [976,] -0.0215045011
#>  [977,]  0.0040034890
#>  [978,]  0.2485319191
#>  [979,] -0.0551796206
#>  [980,] -0.5194714489
#>  [981,] -0.1164642651
#>  [982,]  0.4288024740
#>  [983,]  0.4932344021
#>  [984,]  0.4628645898
#>  [985,] -0.7829866671
#>  [986,] -0.3842724572
#>  [987,] -0.1489560247
#>  [988,]  0.2020028790
#>  [989,]  0.1483075438
#>  [990,]  0.2697074951
#>  [991,]  0.0715721602
#>  [992,]  0.1161721766
#>  [993,]  0.2602138111
#>  [994,] -0.0632549338
#>  [995,]  0.8572976997
#>  [996,]  0.0959098563
#>  [997,]  0.2650915025
#>  [998,] -0.2243655633
#>  [999,]  0.1599730034
#> [1000,] -0.3287497051