Logistic Function

The elementwise logistic function, \(\log(1 + e^x)\). This is a special case of log(sum(exp)) that evaluates to a vector rather than to a scalar, which is useful for logistic regression.

logistic(x)

Arguments

x
An Expression, vector, or matrix.

Value

An Expression representing the logistic function evaluated at the input.

Examples

n <- 20
m <- 1000
sigma <- 45

beta_true <- stats::rnorm(n)
idxs <- sample(n, size = 0.8*n, replace = FALSE)
beta_true[idxs] <- 0
X <- matrix(stats::rnorm(m*n, 0, 5), nrow = m, ncol = n)
y <- sign(X %*% beta_true + stats::rnorm(m, 0, sigma))

beta <- Variable(n)
X_sign <- apply(X, 2, function(x) { ifelse(y <= 0, -1, 1) * x })
obj <- -sum(logistic(-X[y <= 0,] %*% beta)) - sum(logistic(X[y == 1,] %*% beta))
prob <- Problem(Maximize(obj))
result <- solve(prob)

log_odds <- result$getValue(X %*% beta)
beta_res <- result$getValue(beta)
y_probs <- 1/(1 + exp(-X %*% beta_res))
log(y_probs/(1 - y_probs))

#>                  [,1]
#>    [1,]  0.1776907115
#>    [2,]  0.1721152925
#>    [3,]  1.1377964580
#>    [4,] -0.0505105419
#>    [5,] -0.0163218706
#>    [6,] -0.5614381826
#>    [7,] -0.3020215001
#>    [8,]  0.2674898176
#>    [9,] -0.0702584611
#>   [10,]  0.2545010657
#>   [11,] -0.3683195593
#>   [12,] -0.2560717230
#>   [13,] -0.7500115304
#>   [14,] -0.0963328104
#>   [15,]  0.1707131829
#>   [16,] -0.4276964582
#>   [17,] -1.0206022373
#>   [18,] -0.5187657660
#>   [19,] -1.6157586813
#>   [20,]  0.6537929058
#>   [21,]  0.1573090906
#>   [22,] -0.2774123442
#>   [23,] -0.2832150360
#>   [24,]  0.9793205782
#>   [25,]  0.1094721531
#>   [26,]  0.1328184018
#>   [27,] -0.3417093037
#>   [28,]  0.9179094818
#>   [29,]  0.1714278526
#>   [30,]  0.5562017371
#>   [31,]  0.6028983923
#>   [32,]  0.5253803704
#>   [33,]  0.2123573660
#>   [34,] -0.4862166017
#>   [35,] -0.3169873806
#>   [36,]  0.1681282922
#>   [37,]  0.2369406048
#>   [38,] -0.2627413355
#>   [39,]  0.1692382063
#>   [40,]  0.3957017266
#>   [41,] -0.2634506329
#>   [42,]  0.4578548708
#>   [43,] -0.0892857416
#>   [44,]  0.2324009211
#>   [45,] -0.4219129862
#>   [46,] -0.0047330625
#>   [47,]  0.0316958199
#>   [48,]  0.0272015841
#>   [49,] -0.1803892374
#>   [50,]  0.0028465492
#>   [51,]  0.0433399531
#>   [52,] -0.1151626669
#>   [53,]  0.8621552438
#>   [54,]  0.9560444429
#>   [55,] -0.2534996343
#>   [56,]  0.4152445648
#>   [57,]  0.1335290956
#>   [58,]  0.4772782311
#>   [59,]  1.2285439013
#>   [60,]  0.4200917460
#>   [61,]  0.1531666150
#>   [62,] -0.3644060027
#>   [63,]  1.0428400875
#>   [64,]  0.4571327750
#>   [65,]  0.3574276567
#>   [66,]  0.4391722649
#>   [67,]  0.3986042409
#>   [68,] -0.0688277974
#>   [69,]  0.6783599988
#>   [70,]  1.1679155150
#>   [71,] -0.6011281051
#>   [72,]  0.2973700586
#>   [73,] -0.7769708311
#>   [74,] -0.8947706558
#>   [75,]  1.0639101478
#>   [76,]  0.3209930772
#>   [77,]  0.5651706223
#>   [78,] -0.2949455058
#>   [79,]  0.1119202177
#>   [80,]  0.0390219153
#>   [81,]  0.1339415754
#>   [82,] -0.0246384469
#>   [83,]  0.0467438255
#>   [84,] -0.1850473852
#>   [85,] -0.0920323623
#>   [86,] -0.0805591974
#>   [87,]  0.1172102063
#>   [88,]  1.0575856669
#>   [89,] -0.1514444845
#>   [90,]  0.3152233591
#>   [91,]  0.2214181988
#>   [92,]  0.4218284159
#>   [93,] -0.0052967516
#>   [94,]  0.1378266828
#>   [95,] -0.2366678471
#>   [96,] -0.6606634614
#>   [97,]  0.3854960299
#>   [98,] -0.7398575690
#>   [99,] -0.0828850465
#>  [100,] -0.1687009610
#>  [101,]  0.4510994755
#>  [102,]  0.5020440306
#>  [103,]  0.6887763453
#>  [104,]  0.2127265289
#>  [105,]  0.0451642584
#>  [106,]  0.6001819569
#>  [107,]  0.3506252740
#>  [108,]  0.2196077836
#>  [109,] -0.5496393075
#>  [110,] -0.1607367051
#>  [111,]  0.1746315869
#>  [112,]  0.3579256007
#>  [113,]  0.1429728390
#>  [114,]  0.3785122165
#>  [115,]  0.3613745987
#>  [116,] -0.8665770608
#>  [117,] -0.2502884636
#>  [118,] -0.9264183526
#>  [119,] -0.0414606892
#>  [120,]  0.4070613360
#>  [121,]  0.0943136945
#>  [122,] -0.1966374977
#>  [123,]  0.1432217346
#>  [124,]  0.4671563330
#>  [125,] -0.2108343993
#>  [126,] -0.6195288792
#>  [127,]  0.1098914353
#>  [128,] -0.2215009964
#>  [129,] -0.6850344158
#>  [130,]  0.4299204543
#>  [131,]  0.5350842045
#>  [132,]  0.0446499902
#>  [133,] -0.3138743204
#>  [134,] -0.5695055813
#>  [135,]  0.3833270360
#>  [136,]  0.0063149656
#>  [137,]  0.2510409368
#>  [138,] -0.5842293918
#>  [139,] -1.1585912797
#>  [140,]  0.4132382254
#>  [141,]  0.7797662876
#>  [142,] -0.0276929826
#>  [143,] -0.0396475254
#>  [144,]  1.0901027107
#>  [145,] -0.2654809642
#>  [146,]  0.1020471085
#>  [147,]  0.0658007519
#>  [148,] -0.5623279681
#>  [149,]  0.2857876362
#>  [150,]  0.6520650495
#>  [151,] -0.7595282327
#>  [152,] -0.7362477234
#>  [153,]  0.2662795303
#>  [154,] -0.3372488500
#>  [155,] -0.4726556460
#>  [156,] -0.6537661710
#>  [157,] -0.6580087601
#>  [158,]  1.4067207140
#>  [159,]  0.3885098554
#>  [160,] -0.1819626223
#>  [161,] -0.0336060266
#>  [162,] -0.4546785003
#>  [163,] -0.7102743358
#>  [164,]  0.2868654298
#>  [165,]  0.3653155371
#>  [166,]  0.2673807554
#>  [167,]  0.7609098242
#>  [168,]  0.1290319504
#>  [169,]  0.0826247277
#>  [170,]  0.1707556903
#>  [171,] -0.3099082784
#>  [172,]  0.0426394386
#>  [173,] -1.0973042738
#>  [174,]  0.5300176300
#>  [175,]  0.7497431346
#>  [176,]  0.0418639220
#>  [177,]  0.4140081612
#>  [178,]  0.1780662151
#>  [179,]  0.3546889823
#>  [180,] -0.0860076205
#>  [181,]  0.0230660289
#>  [182,] -0.2813022948
#>  [183,]  0.6723863300
#>  [184,]  0.3312060933
#>  [185,] -0.4260109430
#>  [186,]  1.0886457297
#>  [187,]  0.0743683247
#>  [188,] -0.1926127072
#>  [189,] -0.6279346584
#>  [190,] -0.1605792799
#>  [191,] -0.2775197243
#>  [192,] -0.2674607006
#>  [193,] -0.2672951171
#>  [194,]  0.9970393691
#>  [195,]  0.1484181047
#>  [196,] -0.4774167794
#>  [197,] -0.9677630789
#>  [198,]  0.2042775687
#>  [199,]  1.0878920614
#>  [200,] -0.1953002333
#>  [201,]  0.2146812879
#>  [202,] -0.0682836579
#>  [203,]  0.5063258457
#>  [204,]  0.1069250157
#>  [205,]  0.1242716650
#>  [206,]  0.2076928608
#>  [207,] -0.0077973451
#>  [208,] -0.1508006590
#>  [209,] -0.4066165799
#>  [210,] -0.2997131054
#>  [211,]  0.5063917278
#>  [212,] -0.4201125168
#>  [213,]  0.0432398494
#>  [214,]  0.5797247743
#>  [215,] -0.7814608155
#>  [216,] -0.1511388117
#>  [217,] -0.1171619642
#>  [218,] -0.7490174907
#>  [219,]  0.0906234385
#>  [220,] -0.4667914103
#>  [221,]  1.2946845835
#>  [222,]  0.5124123410
#>  [223,]  0.0584864508
#>  [224,] -0.1109322014
#>  [225,]  0.5025104521
#>  [226,]  0.1977432416
#>  [227,] -0.2288044655
#>  [228,] -1.0110543218
#>  [229,] -0.0600447336
#>  [230,]  0.0305551204
#>  [231,] -0.1109136611
#>  [232,]  0.7772760616
#>  [233,]  0.2854256871
#>  [234,]  0.3735098224
#>  [235,]  0.2729023747
#>  [236,]  0.1909806418
#>  [237,] -0.3690688391
#>  [238,]  0.2379489910
#>  [239,] -0.5481357880
#>  [240,]  0.2229691668
#>  [241,] -0.1011083351
#>  [242,] -0.7439805569
#>  [243,] -0.0668737470
#>  [244,] -0.0326960177
#>  [245,]  0.0518840388
#>  [246,]  1.1398275521
#>  [247,]  0.6138973156
#>  [248,] -0.4696811944
#>  [249,] -0.2633505940
#>  [250,]  0.2304621304
#>  [251,]  0.5315984513
#>  [252,]  0.6519470441
#>  [253,] -0.3209931032
#>  [254,] -0.3898239183
#>  [255,]  0.0024494451
#>  [256,]  0.3816448788
#>  [257,] -0.5082717679
#>  [258,] -0.4132994660
#>  [259,]  0.1285439012
#>  [260,] -1.1988339942
#>  [261,] -0.3820850325
#>  [262,]  0.3775828861
#>  [263,]  0.1840796265
#>  [264,]  0.1216456811
#>  [265,]  0.1626190867
#>  [266,] -0.1257799026
#>  [267,]  0.1416376168
#>  [268,]  0.0697864631
#>  [269,]  0.1048416828
#>  [270,] -0.6800112584
#>  [271,] -0.1958425587
#>  [272,] -0.4499996358
#>  [273,]  0.4072209166
#>  [274,]  0.5281841106
#>  [275,] -0.5028569335
#>  [276,]  1.4791932375
#>  [277,] -0.5635515505
#>  [278,]  0.0662859666
#>  [279,]  0.1201416683
#>  [280,] -1.2659322792
#>  [281,]  0.2012763157
#>  [282,] -0.2324961216
#>  [283,] -0.3894950834
#>  [284,]  0.0265850693
#>  [285,]  0.6231820459
#>  [286,] -0.7996572602
#>  [287,]  0.0328773004
#>  [288,] -0.5215483562
#>  [289,] -0.8133066586
#>  [290,]  0.8730857313
#>  [291,]  0.0816145520
#>  [292,] -0.1943770638
#>  [293,]  0.0050323079
#>  [294,] -0.3238461833
#>  [295,]  0.9223565459
#>  [296,]  0.0721060131
#>  [297,] -0.2188530621
#>  [298,] -0.1151640656
#>  [299,]  0.6922103309
#>  [300,] -0.6917114108
#>  [301,] -0.0511631821
#>  [302,] -0.1268238505
#>  [303,] -0.9222147016
#>  [304,]  0.8401176380
#>  [305,]  0.5467099176
#>  [306,] -0.2234080000
#>  [307,]  0.6331918686
#>  [308,] -0.4564516191
#>  [309,]  0.1283192533
#>  [310,]  0.4771424552
#>  [311,] -0.4296886265
#>  [312,] -0.6726226217
#>  [313,]  0.3095440543
#>  [314,]  0.3277623742
#>  [315,] -0.3945954777
#>  [316,] -0.4991604695
#>  [317,] -0.8143978674
#>  [318,]  0.4367783468
#>  [319,] -0.7046898056
#>  [320,] -0.9667603550
#>  [321,] -0.4664628654
#>  [322,] -0.8598556324
#>  [323,]  0.3313352821
#>  [324,]  0.6695977155
#>  [325,]  0.0772705613
#>  [326,]  0.2681650422
#>  [327,]  0.2919246728
#>  [328,] -0.0873255779
#>  [329,]  0.2192785694
#>  [330,] -0.5567796785
#>  [331,] -0.2472875387
#>  [332,]  0.1914931521
#>  [333,] -0.5415030163
#>  [334,] -0.6576327823
#>  [335,] -0.0447206601
#>  [336,]  0.5443013349
#>  [337,]  0.5519578534
#>  [338,] -0.4002237647
#>  [339,]  0.0007258544
#>  [340,] -0.2317686041
#>  [341,] -0.1829072702
#>  [342,] -0.2935466864
#>  [343,]  0.6739011989
#>  [344,] -0.0247984403
#>  [345,]  0.5538155469
#>  [346,]  1.3397921171
#>  [347,]  0.5201641696
#>  [348,]  0.8275002256
#>  [349,]  0.9504390765
#>  [350,] -0.4982757436
#>  [351,] -0.2077971759
#>  [352,]  0.1307661534
#>  [353,]  0.3724847487
#>  [354,]  0.6117978647
#>  [355,] -0.2876523045
#>  [356,]  0.2332566460
#>  [357,]  0.0381363000
#>  [358,] -0.3740218516
#>  [359,] -0.5264739062
#>  [360,] -0.4115133054
#>  [361,] -0.1932375630
#>  [362,]  0.4093522611
#>  [363,]  0.4791920576
#>  [364,]  0.1768915196
#>  [365,]  0.2082779160
#>  [366,] -0.4706286241
#>  [367,]  0.6824742165
#>  [368,] -0.5592148203
#>  [369,]  0.6504626660
#>  [370,]  0.3235202872
#>  [371,]  0.0323302571
#>  [372,] -0.1930001647
#>  [373,] -0.2364273898
#>  [374,]  0.1617875167
#>  [375,] -0.7089564965
#>  [376,]  0.1239008453
#>  [377,]  0.4958056048
#>  [378,] -0.6874016253
#>  [379,] -0.5564183450
#>  [380,]  0.2048117084
#>  [381,]  0.8037387683
#>  [382,] -0.1400540270
#>  [383,] -0.5620573347
#>  [384,] -0.1501679148
#>  [385,]  0.6571963529
#>  [386,] -0.3086314532
#>  [387,] -0.8172099155
#>  [388,]  0.3582183466
#>  [389,] -0.9094794155
#>  [390,]  0.0230602201
#>  [391,]  0.0768746214
#>  [392,] -0.1216990527
#>  [393,] -0.3548894397
#>  [394,]  0.9930633511
#>  [395,]  0.4543745488
#>  [396,]  0.2910185456
#>  [397,] -0.3267726073
#>  [398,]  0.2010271567
#>  [399,]  0.6218372164
#>  [400,] -0.2876498251
#>  [401,]  0.4163143277
#>  [402,]  0.2246430882
#>  [403,]  1.4784975381
#>  [404,]  1.0056784429
#>  [405,] -0.7509954285
#>  [406,]  0.1794672380
#>  [407,]  0.7562027871
#>  [408,]  0.6165544356
#>  [409,] -0.2183994016
#>  [410,] -0.0527553830
#>  [411,] -0.7246912452
#>  [412,] -0.4566500898
#>  [413,] -0.0907251144
#>  [414,]  0.3339971688
#>  [415,]  0.1516460383
#>  [416,]  0.0384277755
#>  [417,]  0.8068633770
#>  [418,]  0.1876844748
#>  [419,]  0.0666845372
#>  [420,] -0.4173224702
#>  [421,] -0.4897861816
#>  [422,]  0.3933083019
#>  [423,] -0.0426002928
#>  [424,]  0.1951690908
#>  [425,] -0.0401149330
#>  [426,]  0.4345596239
#>  [427,] -0.5918893100
#>  [428,] -0.0728314779
#>  [429,] -0.0309474491
#>  [430,] -0.1230564706
#>  [431,] -0.4189946363
#>  [432,] -0.1347375261
#>  [433,]  0.9321161170
#>  [434,] -0.4185095272
#>  [435,] -0.0279187191
#>  [436,]  1.1128083686
#>  [437,]  0.0212827093
#>  [438,]  0.2639946470
#>  [439,] -0.7966938956
#>  [440,]  0.7550103566
#>  [441,]  0.3474904472
#>  [442,]  0.2441124427
#>  [443,] -0.6375231590
#>  [444,]  0.6014857532
#>  [445,] -0.4594758656
#>  [446,]  0.4297484072
#>  [447,] -0.9138942081
#>  [448,] -0.5525703406
#>  [449,] -0.1006351497
#>  [450,] -0.3186712938
#>  [451,] -0.4994961034
#>  [452,]  0.6721814246
#>  [453,]  0.4503671717
#>  [454,] -0.4065065163
#>  [455,] -0.6267926895
#>  [456,]  0.2471981464
#>  [457,]  0.0721763816
#>  [458,]  0.6667049486
#>  [459,]  0.1456541375
#>  [460,]  0.0287646286
#>  [461,]  0.5299143720
#>  [462,]  0.3200415152
#>  [463,] -0.3383515523
#>  [464,]  0.4879991389
#>  [465,] -0.5022946512
#>  [466,]  0.6297160566
#>  [467,] -0.1252733598
#>  [468,]  0.4257977864
#>  [469,] -0.0629095875
#>  [470,]  0.1894516702
#>  [471,]  0.1184705010
#>  [472,] -0.1447714853
#>  [473,] -0.6110765393
#>  [474,]  0.6786655965
#>  [475,]  0.9614812125
#>  [476,] -0.0568347389
#>  [477,] -0.2916240209
#>  [478,] -0.8175062099
#>  [479,] -0.1543216873
#>  [480,] -0.3401276187
#>  [481,]  1.3300784577
#>  [482,]  0.1893576634
#>  [483,] -0.3776601512
#>  [484,] -0.1966158200
#>  [485,] -0.2539349416
#>  [486,] -0.3989713575
#>  [487,]  0.2801572935
#>  [488,]  0.1772668757
#>  [489,]  0.0084104878
#>  [490,]  0.7782181654
#>  [491,]  0.3663389959
#>  [492,] -0.2723405706
#>  [493,]  0.3001270507
#>  [494,] -0.5927366596
#>  [495,] -0.6635321370
#>  [496,]  0.5666832162
#>  [497,]  0.6938432375
#>  [498,]  0.5305614451
#>  [499,] -0.4033557707
#>  [500,]  0.6425921523
#>  [501,] -0.9172517025
#>  [502,] -0.3970932000
#>  [503,] -0.4574944009
#>  [504,] -0.0240719445
#>  [505,] -0.7125111024
#>  [506,]  0.3472637398
#>  [507,] -0.7461263904
#>  [508,]  0.1454987608
#>  [509,] -0.8878603548
#>  [510,]  0.1328385096
#>  [511,]  0.2181327963
#>  [512,] -0.6625587211
#>  [513,] -0.4052296425
#>  [514,] -0.3967186098
#>  [515,] -0.8221928159
#>  [516,] -0.1283502066
#>  [517,]  0.4730531405
#>  [518,] -0.1963935928
#>  [519,] -0.6260064802
#>  [520,]  0.1296011589
#>  [521,]  0.3547349636
#>  [522,]  0.3801378712
#>  [523,] -0.6574171211
#>  [524,] -0.1478013977
#>  [525,] -0.9156233428
#>  [526,] -0.2035286396
#>  [527,] -0.6353314039
#>  [528,]  0.1563472921
#>  [529,]  0.0782209244
#>  [530,]  0.7057672805
#>  [531,]  0.3853389238
#>  [532,]  0.5317541870
#>  [533,]  0.4447710105
#>  [534,]  0.4459407557
#>  [535,] -0.5446620879
#>  [536,] -0.3940128075
#>  [537,]  0.1100505610
#>  [538,] -0.5846443697
#>  [539,]  0.2109707387
#>  [540,] -0.9638412278
#>  [541,] -0.2476191786
#>  [542,] -1.0695476172
#>  [543,]  0.8128839457
#>  [544,] -0.5259262027
#>  [545,] -0.3377341138
#>  [546,]  0.2222906089
#>  [547,]  0.8227905823
#>  [548,]  0.0010390471
#>  [549,]  0.4214453692
#>  [550,] -0.6509115672
#>  [551,] -0.5174005757
#>  [552,] -0.5436047043
#>  [553,]  0.9311095981
#>  [554,] -0.9777211879
#>  [555,]  0.4613614602
#>  [556,] -0.1324951584
#>  [557,]  0.4274742681
#>  [558,]  0.0408026945
#>  [559,]  0.6847750172
#>  [560,]  0.3879844389
#>  [561,]  0.0566043982
#>  [562,]  0.6207376984
#>  [563,] -0.4587979561
#>  [564,] -0.0915227172
#>  [565,] -0.4120802139
#>  [566,] -0.1440063882
#>  [567,]  0.5013391828
#>  [568,]  0.3981734345
#>  [569,] -0.2867979747
#>  [570,] -0.3096104332
#>  [571,] -0.3992848725
#>  [572,]  0.8171918789
#>  [573,]  0.0917680744
#>  [574,] -0.7169050223
#>  [575,]  0.4751827291
#>  [576,]  0.9444863905
#>  [577,]  0.4395678955
#>  [578,] -0.0051960047
#>  [579,] -0.4598181375
#>  [580,]  0.0718779340
#>  [581,]  0.6475022672
#>  [582,] -0.5091874726
#>  [583,]  0.4219118228
#>  [584,]  0.5971651096
#>  [585,] -0.3556012648
#>  [586,] -1.3453384829
#>  [587,]  1.1481721519
#>  [588,]  0.0652020824
#>  [589,]  0.4605862347
#>  [590,]  0.2565459364
#>  [591,]  0.3141802431
#>  [592,] -0.3438278655
#>  [593,]  0.6075429676
#>  [594,] -0.0284025509
#>  [595,] -0.2070293559
#>  [596,] -0.0604363095
#>  [597,]  0.8527724630
#>  [598,] -0.1380275156
#>  [599,] -0.0566275853
#>  [600,]  0.3142942216
#>  [601,]  0.5983512106
#>  [602,] -1.0338600084
#>  [603,]  0.3255109489
#>  [604,]  1.1641816207
#>  [605,]  0.2947549481
#>  [606,]  0.2337680661
#>  [607,]  1.0570027663
#>  [608,]  0.4668916983
#>  [609,]  0.6732787996
#>  [610,]  0.7988335513
#>  [611,] -0.5899948166
#>  [612,]  0.2898282606
#>  [613,] -0.3362645137
#>  [614,]  0.9084442462
#>  [615,]  0.3714154065
#>  [616,]  0.8654416450
#>  [617,] -0.0408162853
#>  [618,] -0.7428350216
#>  [619,] -0.4377607563
#>  [620,] -0.0019000371
#>  [621,]  0.0861871919
#>  [622,] -0.0223366853
#>  [623,]  0.3614228254
#>  [624,] -0.9077503400
#>  [625,]  0.1434581905
#>  [626,]  0.2308681266
#>  [627,]  0.2748790624
#>  [628,] -0.3900325637
#>  [629,] -0.4671769776
#>  [630,]  0.3954189415
#>  [631,] -0.0657404122
#>  [632,]  0.2057425142
#>  [633,] -0.0372858475
#>  [634,] -0.4199082735
#>  [635,]  0.3129053979
#>  [636,] -0.1452760195
#>  [637,] -0.1951476893
#>  [638,]  0.6359626476
#>  [639,] -1.6716764981
#>  [640,]  0.7910114956
#>  [641,]  0.7797164124
#>  [642,]  0.1527725500
#>  [643,] -0.3023830244
#>  [644,]  0.6563003394
#>  [645,] -0.1921618616
#>  [646,]  1.2316362522
#>  [647,] -0.7645861426
#>  [648,] -0.3636012573
#>  [649,] -0.8083665093
#>  [650,]  0.9252842003
#>  [651,]  0.8920906785
#>  [652,] -0.1022566154
#>  [653,] -0.1320845616
#>  [654,]  0.6807791790
#>  [655,]  0.5281108788
#>  [656,] -0.6877304818
#>  [657,]  0.1778924100
#>  [658,]  0.2056367211
#>  [659,] -0.5795616126
#>  [660,] -0.5029533801
#>  [661,] -0.2186162716
#>  [662,] -0.2630843616
#>  [663,]  0.4887429921
#>  [664,]  0.1591163287
#>  [665,]  0.9706392354
#>  [666,] -0.7181128610
#>  [667,]  0.2151829627
#>  [668,] -0.2857272303
#>  [669,] -0.2430649423
#>  [670,]  0.2806792835
#>  [671,]  0.1140931180
#>  [672,] -0.1864630753
#>  [673,]  0.6223659372
#>  [674,]  0.6474024595
#>  [675,] -0.0255297561
#>  [676,]  0.0656662380
#>  [677,]  0.4407097335
#>  [678,]  0.2836559557
#>  [679,]  0.8825028657
#>  [680,] -0.5679307691
#>  [681,]  0.1711122651
#>  [682,]  0.5892535200
#>  [683,] -0.7147612910
#>  [684,] -0.6692618383
#>  [685,] -0.2498981123
#>  [686,]  1.0398814690
#>  [687,]  0.6363214807
#>  [688,]  1.0208611406
#>  [689,] -0.7273044024
#>  [690,]  0.1379464320
#>  [691,] -0.2363306120
#>  [692,]  0.4013956242
#>  [693,] -0.1056145632
#>  [694,] -0.5205431286
#>  [695,]  0.6585077561
#>  [696,] -0.6959418551
#>  [697,] -0.0358490258
#>  [698,]  0.0020611603
#>  [699,]  0.5749064769
#>  [700,]  0.4975171128
#>  [701,]  0.0947953733
#>  [702,]  0.6721902607
#>  [703,]  0.7500905496
#>  [704,]  0.0070173997
#>  [705,]  0.1227375595
#>  [706,]  0.4314698685
#>  [707,]  0.2215772509
#>  [708,] -0.1994428429
#>  [709,]  0.2137147746
#>  [710,]  0.0379239408
#>  [711,] -0.6146250062
#>  [712,] -0.3175560459
#>  [713,] -0.6371884919
#>  [714,] -0.0141508370
#>  [715,]  0.1219672386
#>  [716,]  0.5865084265
#>  [717,] -0.0238901253
#>  [718,] -0.0049189277
#>  [719,] -0.3135952258
#>  [720,] -0.7022238256
#>  [721,]  0.1619676662
#>  [722,] -0.2268906387
#>  [723,] -0.8643965190
#>  [724,] -0.4861267350
#>  [725,]  0.7274159843
#>  [726,] -0.7508179005
#>  [727,]  0.6714586314
#>  [728,] -0.9619082837
#>  [729,] -0.2664711453
#>  [730,]  0.5117646854
#>  [731,]  0.1604751309
#>  [732,]  0.3905025657
#>  [733,] -0.4646762772
#>  [734,]  0.1800738865
#>  [735,] -0.7736296140
#>  [736,]  0.3208409129
#>  [737,] -0.5601436894
#>  [738,] -0.7337537476
#>  [739,]  0.5954565625
#>  [740,]  0.8386306914
#>  [741,]  0.2534849207
#>  [742,] -0.6716317472
#>  [743,]  0.3715155181
#>  [744,]  0.3607661908
#>  [745,] -0.5352675629
#>  [746,]  0.2058877940
#>  [747,]  0.2071328217
#>  [748,] -0.0518143038
#>  [749,] -0.2923104985
#>  [750,]  0.4433264188
#>  [751,]  0.2139148104
#>  [752,] -0.7399927456
#>  [753,]  0.5738113370
#>  [754,] -0.6510142771
#>  [755,] -0.7249443590
#>  [756,]  0.0416870697
#>  [757,] -0.2502427467
#>  [758,]  0.0988794208
#>  [759,] -0.6580414474
#>  [760,] -0.6925772164
#>  [761,] -0.1214502964
#>  [762,]  0.1474913437
#>  [763,]  0.3659193986
#>  [764,]  0.5587561936
#>  [765,]  0.6723309797
#>  [766,] -0.0677874502
#>  [767,]  0.2344021185
#>  [768,] -0.2433366073
#>  [769,] -0.1338745321
#>  [770,] -0.9397830625
#>  [771,] -0.5499498580
#>  [772,] -0.0508363042
#>  [773,] -0.9965985444
#>  [774,]  0.2057107267
#>  [775,] -0.5225173160
#>  [776,] -0.6613778111
#>  [777,]  0.9107599743
#>  [778,] -0.1498270663
#>  [779,]  1.2559352983
#>  [780,]  0.7155865632
#>  [781,] -0.7021132563
#>  [782,]  0.3475784884
#>  [783,] -0.0815499070
#>  [784,] -0.1589835428
#>  [785,]  0.2703057739
#>  [786,]  0.3361317601
#>  [787,]  0.3752877738
#>  [788,]  0.1027118463
#>  [789,]  0.0856404436
#>  [790,] -0.2614355851
#>  [791,] -0.2208425573
#>  [792,] -0.9378160033
#>  [793,] -0.4370834653
#>  [794,] -0.9098863912
#>  [795,]  0.5713158293
#>  [796,] -0.6674911330
#>  [797,]  0.4611780940
#>  [798,]  0.4626379509
#>  [799,] -0.0349044128
#>  [800,] -0.1177186814
#>  [801,] -0.5133867516
#>  [802,]  0.2342673394
#>  [803,] -0.4184077812
#>  [804,] -0.3548978522
#>  [805,] -1.5184098937
#>  [806,]  0.7804743072
#>  [807,] -0.3217609889
#>  [808,]  0.1642284778
#>  [809,]  1.0110942537
#>  [810,]  0.0365897583
#>  [811,]  0.6498829631
#>  [812,]  0.0851840009
#>  [813,]  0.1598821264
#>  [814,]  0.6470192001
#>  [815,]  0.7748830534
#>  [816,]  0.5734029734
#>  [817,] -0.2677175765
#>  [818,]  0.0178966660
#>  [819,] -0.2322458538
#>  [820,]  0.5621722502
#>  [821,]  0.1439272562
#>  [822,]  0.5511197201
#>  [823,]  0.4824810041
#>  [824,] -0.3756285895
#>  [825,]  0.1553806906
#>  [826,]  0.4608082208
#>  [827,]  0.9368098477
#>  [828,] -1.1322840420
#>  [829,] -0.2158895234
#>  [830,] -0.4512739876
#>  [831,]  0.0625531705
#>  [832,] -0.5661467455
#>  [833,]  0.6041429095
#>  [834,] -0.3996687145
#>  [835,]  0.0682118070
#>  [836,] -0.0064293469
#>  [837,]  0.0802372786
#>  [838,]  0.4535963854
#>  [839,] -0.3830870870
#>  [840,] -0.3691984346
#>  [841,]  0.7531610682
#>  [842,]  0.3640038945
#>  [843,]  0.3641461237
#>  [844,] -0.0964456873
#>  [845,]  0.8652494334
#>  [846,] -0.1026015586
#>  [847,] -0.2294880142
#>  [848,] -0.0261248341
#>  [849,]  0.4989506822
#>  [850,] -0.3706895700
#>  [851,] -0.0779749899
#>  [852,] -0.2053775700
#>  [853,] -0.4125335715
#>  [854,] -0.4654336259
#>  [855,] -0.3016832396
#>  [856,]  0.0611033718
#>  [857,]  1.5645897523
#>  [858,]  0.2861803607
#>  [859,]  0.1122706204
#>  [860,]  0.2867486791
#>  [861,]  0.0345295655
#>  [862,]  1.2125852099
#>  [863,]  0.2838497864
#>  [864,] -0.2230030180
#>  [865,]  0.0234530494
#>  [866,]  0.0404364429
#>  [867,]  0.9742373826
#>  [868,] -0.2175996853
#>  [869,]  0.1564465002
#>  [870,]  0.4092963842
#>  [871,] -0.0390081028
#>  [872,]  0.9610048519
#>  [873,] -0.0864774574
#>  [874,] -0.5426580322
#>  [875,] -1.1484439519
#>  [876,] -0.5791089158
#>  [877,]  0.3894882153
#>  [878,] -0.2808359967
#>  [879,]  0.3369371568
#>  [880,] -0.0014836576
#>  [881,]  0.3015108945
#>  [882,] -0.0400998720
#>  [883,]  0.0750985518
#>  [884,]  0.0902451860
#>  [885,] -0.2457585125
#>  [886,] -0.1224372653
#>  [887,] -0.2712512588
#>  [888,]  0.7679508207
#>  [889,] -0.2790233047
#>  [890,] -0.2350635385
#>  [891,] -0.0834078034
#>  [892,] -0.1478644242
#>  [893,]  0.1222262666
#>  [894,]  0.4755440036
#>  [895,]  0.0802760331
#>  [896,] -0.8502794374
#>  [897,] -0.5175282193
#>  [898,]  0.0900530153
#>  [899,] -0.8461887779
#>  [900,] -0.3029168395
#>  [901,] -0.5342718027
#>  [902,]  0.5020859722
#>  [903,] -0.5158811447
#>  [904,]  0.4157612192
#>  [905,]  0.5198172508
#>  [906,] -0.4760634226
#>  [907,]  0.4700041072
#>  [908,]  0.0405778851
#>  [909,] -0.4259855735
#>  [910,]  0.0218440878
#>  [911,] -0.9403763208
#>  [912,]  0.0062327972
#>  [913,] -0.2687505157
#>  [914,]  0.3163775548
#>  [915,]  0.4624797179
#>  [916,]  0.6800427362
#>  [917,]  0.9145308144
#>  [918,]  0.2144462956
#>  [919,]  1.0442390077
#>  [920,] -0.5574362923
#>  [921,]  0.6317146913
#>  [922,] -0.2935027524
#>  [923,]  1.1708029158
#>  [924,] -0.0745971445
#>  [925,]  0.0477142244
#>  [926,] -0.0881109312
#>  [927,] -1.7298261009
#>  [928,]  0.7435413736
#>  [929,] -0.8537077069
#>  [930,] -0.0991184553
#>  [931,]  0.9656648800
#>  [932,]  0.1500970381
#>  [933,]  0.2447495573
#>  [934,]  0.0607344594
#>  [935,]  0.6060306180
#>  [936,]  0.0204860380
#>  [937,]  0.6395497162
#>  [938,]  0.8015427193
#>  [939,] -0.5584415332
#>  [940,]  0.4995169666
#>  [941,] -0.4757305467
#>  [942,] -0.3367299215
#>  [943,] -0.1466713825
#>  [944,] -0.4355117741
#>  [945,]  0.3795105342
#>  [946,]  0.1555184708
#>  [947,]  1.3964583974
#>  [948,]  0.5865874498
#>  [949,]  0.2535063989
#>  [950,]  0.8644228316
#>  [951,] -0.2425276076
#>  [952,]  0.5226873233
#>  [953,] -1.5463309414
#>  [954,] -0.3259245855
#>  [955,]  0.8979142861
#>  [956,]  0.3719277533
#>  [957,]  0.5727959519
#>  [958,] -0.2964437231
#>  [959,]  0.8044727632
#>  [960,] -0.9285414791
#>  [961,] -0.0878503516
#>  [962,]  0.0175626878
#>  [963,] -0.0212480689
#>  [964,] -0.9524807778
#>  [965,] -0.4132595855
#>  [966,] -0.4881203100
#>  [967,] -0.4274400385
#>  [968,]  0.1290420148
#>  [969,] -0.7123710125
#>  [970,]  0.2838214820
#>  [971,] -0.2616943360
#>  [972,] -0.1750908740
#>  [973,] -0.0155265443
#>  [974,] -0.6000764466
#>  [975,] -0.0744696216
#>  [976,]  0.1913857990
#>  [977,]  0.0244553839
#>  [978,]  0.2853784576
#>  [979,]  0.2191583292
#>  [980,]  0.3958771533
#>  [981,]  0.7072533342
#>  [982,] -0.1371858064
#>  [983,] -0.7392101844
#>  [984,]  0.4900691658
#>  [985,]  0.0345386428
#>  [986,] -0.3167561360
#>  [987,] -0.4494276898
#>  [988,] -0.6417035219
#>  [989,] -0.1483817259
#>  [990,] -0.0303069040
#>  [991,] -0.5647913602
#>  [992,]  0.1401685720
#>  [993,]  0.3743517335
#>  [994,] -0.7008106366
#>  [995,]  0.5056967189
#>  [996,] -0.5213673440
#>  [997,] -0.1539288948
#>  [998,] -0.6439742564
#>  [999,]  1.0662671932
#> [1000,] -0.2273797543