Logistic Function

The elementwise logistic function, \(\log(1 + e^x)\). This is a special case of log(sum(exp)) that evaluates to a vector rather than to a scalar, which is useful for logistic regression.

logistic(x)

Arguments

x
An Expression, vector, or matrix.

Value

An Expression representing the logistic function evaluated at the input.

Examples

set.seed(92)
n <- 20
m <- 1000
sigma <- 45

beta_true <- stats::rnorm(n)
idxs <- sample(n, size = 0.8*n, replace = FALSE)
beta_true[idxs] <- 0
X <- matrix(stats::rnorm(m*n, 0, 5), nrow = m, ncol = n)
y <- sign(X %*% beta_true + stats::rnorm(m, 0, sigma))

beta <- Variable(n)
X_sign <- apply(X, 2, function(x) { ifelse(y <= 0, -1, 1) * x })
obj <- -sum(logistic(-X[y <= 0,] %*% beta)) - sum(logistic(X[y == 1,] %*% beta))
prob <- Problem(Maximize(obj))
result <- solve(prob)

log_odds <- result$getValue(X %*% beta)
beta_res <- result$getValue(beta)
y_probs <- 1/(1 + exp(-X %*% beta_res))
log(y_probs/(1 - y_probs))

#>                  [,1]
#>    [1,] -0.6704749080
#>    [2,] -0.5925331341
#>    [3,] -0.2870379486
#>    [4,] -0.3903535908
#>    [5,]  0.1512572334
#>    [6,] -0.3212395200
#>    [7,]  0.0710835074
#>    [8,] -0.0005975890
#>    [9,]  0.6231703747
#>   [10,] -0.3461630663
#>   [11,]  0.6166061043
#>   [12,]  0.2493867196
#>   [13,] -0.4229593467
#>   [14,]  1.3550953148
#>   [15,] -0.3265674645
#>   [16,] -0.5974215781
#>   [17,]  1.4251324876
#>   [18,] -0.8064351654
#>   [19,]  0.6354417869
#>   [20,] -0.5313441052
#>   [21,] -0.5836981490
#>   [22,] -0.2018654474
#>   [23,] -1.5741504172
#>   [24,] -0.2996895406
#>   [25,] -0.2075379799
#>   [26,]  0.4437078272
#>   [27,] -0.5653587234
#>   [28,] -1.2550934700
#>   [29,] -0.5346378553
#>   [30,]  0.4834189324
#>   [31,]  0.2311073821
#>   [32,]  0.2894532632
#>   [33,] -0.0125270197
#>   [34,]  0.7295949020
#>   [35,] -0.8747332582
#>   [36,] -0.3929008992
#>   [37,] -0.1155698716
#>   [38,]  1.1672748767
#>   [39,] -0.7217613491
#>   [40,]  0.0611259465
#>   [41,] -0.6167815218
#>   [42,] -0.0245111125
#>   [43,]  0.6908385245
#>   [44,] -0.8034687352
#>   [45,]  0.8837323738
#>   [46,] -1.0940884841
#>   [47,]  0.0151085668
#>   [48,] -0.0365100274
#>   [49,]  0.1208122531
#>   [50,]  0.0228343416
#>   [51,]  0.0203844024
#>   [52,] -0.4370979065
#>   [53,]  0.0381293694
#>   [54,] -0.0670272408
#>   [55,] -0.5450633650
#>   [56,] -0.1197517038
#>   [57,] -0.1350462690
#>   [58,]  0.1775515650
#>   [59,]  1.1398264705
#>   [60,] -0.4142678545
#>   [61,]  0.3993093738
#>   [62,] -0.4202251303
#>   [63,]  0.0406899394
#>   [64,] -0.6489268823
#>   [65,] -0.9703162177
#>   [66,]  0.1117470265
#>   [67,] -0.1453610013
#>   [68,]  0.0616520054
#>   [69,]  0.0779097545
#>   [70,]  0.4237531713
#>   [71,] -0.4578432764
#>   [72,] -0.2049761304
#>   [73,] -0.2303946464
#>   [74,]  0.1166999738
#>   [75,] -0.3652055706
#>   [76,] -0.2369569670
#>   [77,]  0.1457274668
#>   [78,] -1.0896689907
#>   [79,] -0.1308748666
#>   [80,]  0.2291738424
#>   [81,]  0.1373270584
#>   [82,] -0.0015644653
#>   [83,]  0.3467662364
#>   [84,] -0.1543936453
#>   [85,] -0.2309081994
#>   [86,] -0.5247240511
#>   [87,]  0.4603877327
#>   [88,] -0.3126546021
#>   [89,]  0.2120101791
#>   [90,]  0.3147060672
#>   [91,]  0.4193117610
#>   [92,]  0.6556947163
#>   [93,] -0.5111241598
#>   [94,]  1.3013591608
#>   [95,]  0.8030345897
#>   [96,]  0.2830437402
#>   [97,] -0.6704132272
#>   [98,]  0.6392225446
#>   [99,]  0.0848846769
#>  [100,] -0.0064339091
#>  [101,]  0.1647498426
#>  [102,]  0.2973181797
#>  [103,]  0.3564272058
#>  [104,] -0.8130653486
#>  [105,] -0.5596710965
#>  [106,]  0.2440682488
#>  [107,] -0.5330644211
#>  [108,]  0.4528532021
#>  [109,] -0.5709081945
#>  [110,] -0.4624089698
#>  [111,] -0.2336951363
#>  [112,]  0.1511247964
#>  [113,] -0.0619000972
#>  [114,]  0.5839109922
#>  [115,] -0.0005886994
#>  [116,] -0.7089907186
#>  [117,] -0.3172781965
#>  [118,]  0.3970662565
#>  [119,] -1.5820311393
#>  [120,]  0.5686397332
#>  [121,] -0.3273534434
#>  [122,] -0.8146154149
#>  [123,] -0.0948009080
#>  [124,]  0.3203506407
#>  [125,] -0.4429261187
#>  [126,]  0.1170360716
#>  [127,] -0.7652446527
#>  [128,] -0.0573330382
#>  [129,]  0.0095922820
#>  [130,] -0.1932187839
#>  [131,]  0.1964297859
#>  [132,] -0.3175673886
#>  [133,]  0.4246308091
#>  [134,]  0.1266910764
#>  [135,]  0.3305676037
#>  [136,] -0.0061833383
#>  [137,] -0.5537780293
#>  [138,] -0.1133732566
#>  [139,] -0.7994786961
#>  [140,]  0.8081793815
#>  [141,]  0.4153197841
#>  [142,]  0.8942467428
#>  [143,] -0.3139529798
#>  [144,]  0.1458418479
#>  [145,] -0.2050302852
#>  [146,] -0.0601911905
#>  [147,] -1.1748373963
#>  [148,]  0.5454839772
#>  [149,]  0.2075718798
#>  [150,] -0.2618829788
#>  [151,]  0.8039084833
#>  [152,] -0.0455087036
#>  [153,]  0.5741461566
#>  [154,]  0.6920031718
#>  [155,]  0.0716658645
#>  [156,] -0.4965852510
#>  [157,]  0.4424015303
#>  [158,] -0.3816941039
#>  [159,]  0.5206131122
#>  [160,] -0.1021959291
#>  [161,] -0.2226467856
#>  [162,]  0.7099018593
#>  [163,]  0.5765383577
#>  [164,]  0.2450354023
#>  [165,] -0.1737111611
#>  [166,]  0.3589719591
#>  [167,] -0.3290109439
#>  [168,] -0.8993574951
#>  [169,] -0.8218173429
#>  [170,] -0.7797950736
#>  [171,] -0.2137224894
#>  [172,]  0.5276415943
#>  [173,] -0.8488030517
#>  [174,]  0.1034571419
#>  [175,] -0.4103726123
#>  [176,] -0.7528966456
#>  [177,] -0.4402504707
#>  [178,] -1.0685900439
#>  [179,]  0.4078634139
#>  [180,]  0.6313746430
#>  [181,]  0.3750978853
#>  [182,] -0.0151762326
#>  [183,] -0.5310945538
#>  [184,]  0.4077602313
#>  [185,]  0.1571578343
#>  [186,] -0.6328759481
#>  [187,] -0.3822627604
#>  [188,]  0.5205265869
#>  [189,]  0.2562737039
#>  [190,]  0.8532806338
#>  [191,] -0.5649273584
#>  [192,]  0.0370598596
#>  [193,] -0.3390881102
#>  [194,] -0.0375165914
#>  [195,]  0.0474716909
#>  [196,] -0.0326523624
#>  [197,]  0.2160031986
#>  [198,] -1.0662729792
#>  [199,] -0.9550127251
#>  [200,] -0.6453964206
#>  [201,]  0.5183438838
#>  [202,] -0.6995174659
#>  [203,]  0.2181033722
#>  [204,]  0.1893102591
#>  [205,]  0.4929997740
#>  [206,]  0.1844708992
#>  [207,]  0.3573983780
#>  [208,] -0.2500619277
#>  [209,]  1.3917468410
#>  [210,] -0.7617974025
#>  [211,]  0.5279347959
#>  [212,] -1.0117928246
#>  [213,] -0.2423778513
#>  [214,] -0.6134070164
#>  [215,] -1.1656475779
#>  [216,]  0.3821435254
#>  [217,] -0.4501285515
#>  [218,] -1.0973567869
#>  [219,]  0.8937551928
#>  [220,]  0.3140737673
#>  [221,]  0.1340535775
#>  [222,] -1.2231146644
#>  [223,]  0.5678337944
#>  [224,]  0.8463777708
#>  [225,]  0.6349313876
#>  [226,]  0.7006066196
#>  [227,] -0.7992304584
#>  [228,]  0.5683536334
#>  [229,]  0.4113090746
#>  [230,]  0.0902430478
#>  [231,] -0.3665264216
#>  [232,]  0.0127791673
#>  [233,]  0.1629195238
#>  [234,] -0.3527239730
#>  [235,] -0.0758217146
#>  [236,]  0.0091714057
#>  [237,] -0.4611267271
#>  [238,]  0.0657051463
#>  [239,] -0.5298670352
#>  [240,] -0.5962326343
#>  [241,]  0.0637696596
#>  [242,] -0.3510198222
#>  [243,]  0.5476188975
#>  [244,]  0.5265704357
#>  [245,] -0.8598342445
#>  [246,]  0.6395418137
#>  [247,]  0.3216021845
#>  [248,] -0.4796018987
#>  [249,]  0.3975296808
#>  [250,]  0.1215659824
#>  [251,] -0.3140334370
#>  [252,] -0.3046536470
#>  [253,]  0.3923268208
#>  [254,] -0.1856474560
#>  [255,] -0.0851623853
#>  [256,] -0.0029482261
#>  [257,]  0.5734885337
#>  [258,] -0.0359995209
#>  [259,] -0.9568170963
#>  [260,]  0.6395788469
#>  [261,]  0.4795820527
#>  [262,]  0.1872639814
#>  [263,] -0.2498944304
#>  [264,]  0.3261025406
#>  [265,]  0.1986625794
#>  [266,] -0.6790778022
#>  [267,]  0.6809693467
#>  [268,] -0.2952619942
#>  [269,]  0.5265892875
#>  [270,]  0.6813045811
#>  [271,] -0.4787303953
#>  [272,]  0.6944830921
#>  [273,]  0.3116230428
#>  [274,]  0.6770528090
#>  [275,]  0.0024197436
#>  [276,]  0.1779809769
#>  [277,]  0.1332662709
#>  [278,]  0.2881667635
#>  [279,] -0.8398434637
#>  [280,]  0.8293259048
#>  [281,]  0.3237992873
#>  [282,]  0.3688802418
#>  [283,] -0.2838473109
#>  [284,]  0.1698121211
#>  [285,] -0.0972806327
#>  [286,]  0.1692343466
#>  [287,]  0.6139453272
#>  [288,]  0.1736876979
#>  [289,] -0.0806372917
#>  [290,]  0.0115349687
#>  [291,] -0.3183947100
#>  [292,]  0.0866726177
#>  [293,] -0.7586501948
#>  [294,]  0.0584528451
#>  [295,] -0.1537406155
#>  [296,]  0.8435506692
#>  [297,]  0.8815634589
#>  [298,]  0.1534506970
#>  [299,]  1.2271542730
#>  [300,]  0.3099007508
#>  [301,] -0.2555325163
#>  [302,] -0.2073333583
#>  [303,] -0.2298609385
#>  [304,] -0.5772270213
#>  [305,] -0.4885501760
#>  [306,]  0.5318394650
#>  [307,] -0.1440944374
#>  [308,]  0.6376020377
#>  [309,]  0.2563626717
#>  [310,] -0.0603869045
#>  [311,] -0.2223269285
#>  [312,] -0.4951221112
#>  [313,] -0.2533214108
#>  [314,]  0.3508914131
#>  [315,]  0.0868162197
#>  [316,] -0.3995866116
#>  [317,] -0.2455304598
#>  [318,] -0.0988051681
#>  [319,]  0.6998414069
#>  [320,] -0.4848309538
#>  [321,] -0.3735448316
#>  [322,] -0.0385930571
#>  [323,]  0.0463465747
#>  [324,] -0.8583873942
#>  [325,] -0.4583305124
#>  [326,]  0.6281566773
#>  [327,] -0.6351111078
#>  [328,] -0.9582080306
#>  [329,]  0.9368686813
#>  [330,]  0.1483898265
#>  [331,] -0.0538294760
#>  [332,] -1.2323331581
#>  [333,] -0.1706922414
#>  [334,]  0.1205541411
#>  [335,] -0.4870975188
#>  [336,]  0.0540018722
#>  [337,] -0.1344872641
#>  [338,]  1.1543885441
#>  [339,]  0.8619172539
#>  [340,]  0.0228941159
#>  [341,] -0.5239152625
#>  [342,]  0.3351315798
#>  [343,] -0.7517341995
#>  [344,]  0.6050445917
#>  [345,] -0.0211553401
#>  [346,]  0.4027288878
#>  [347,]  0.8253071946
#>  [348,] -0.3307175561
#>  [349,]  0.3177338715
#>  [350,] -0.0224535923
#>  [351,] -0.5546897702
#>  [352,] -0.2831651729
#>  [353,] -0.1794697806
#>  [354,] -0.5933677809
#>  [355,]  0.5243923255
#>  [356,]  0.2004167777
#>  [357,]  0.2904739869
#>  [358,] -0.4817230280
#>  [359,] -0.1010078513
#>  [360,]  0.3605645234
#>  [361,]  0.3838694248
#>  [362,] -0.6909420794
#>  [363,] -0.3578757714
#>  [364,]  0.7571835448
#>  [365,]  0.4263435565
#>  [366,] -0.7292168067
#>  [367,]  1.3051359147
#>  [368,]  0.3406762012
#>  [369,]  0.1624659079
#>  [370,] -0.4134349196
#>  [371,]  0.8469980103
#>  [372,]  0.1445715116
#>  [373,] -1.0051365477
#>  [374,]  0.1813294342
#>  [375,]  0.3128038964
#>  [376,] -0.2938527057
#>  [377,] -1.1220996016
#>  [378,]  0.0153652504
#>  [379,]  0.6470437635
#>  [380,] -1.1136542298
#>  [381,] -0.6871887435
#>  [382,]  0.5673227358
#>  [383,] -0.2756680175
#>  [384,] -0.8618279200
#>  [385,] -0.1516862929
#>  [386,] -0.2098263141
#>  [387,] -1.2010625277
#>  [388,]  1.2558770932
#>  [389,] -0.5078250351
#>  [390,] -0.8336059042
#>  [391,]  0.8759464726
#>  [392,]  0.5860589952
#>  [393,] -1.2938329454
#>  [394,] -0.7940054866
#>  [395,] -0.5236482114
#>  [396,]  0.1776026402
#>  [397,] -1.4013776825
#>  [398,]  0.7131706504
#>  [399,]  0.0354515464
#>  [400,]  0.1344391764
#>  [401,]  0.4879868739
#>  [402,]  0.3638171643
#>  [403,] -0.1982469232
#>  [404,]  0.1427100932
#>  [405,]  0.0306156058
#>  [406,]  0.6818311465
#>  [407,]  0.0828461875
#>  [408,]  0.1285058156
#>  [409,]  0.7544672914
#>  [410,]  0.2249944479
#>  [411,]  0.0320371401
#>  [412,]  0.4893857952
#>  [413,]  0.3547016264
#>  [414,] -0.3653020636
#>  [415,]  0.8342836851
#>  [416,] -0.2503121391
#>  [417,]  0.5247354409
#>  [418,]  0.9033273035
#>  [419,]  0.6160372927
#>  [420,]  0.1392565245
#>  [421,]  0.7268418475
#>  [422,] -0.5110498624
#>  [423,] -0.6768888222
#>  [424,]  0.8344473424
#>  [425,]  0.3100291226
#>  [426,]  0.4325995916
#>  [427,]  0.1333037726
#>  [428,]  0.5182077801
#>  [429,] -0.3190092003
#>  [430,]  0.4315729484
#>  [431,]  0.3930368444
#>  [432,] -0.4950546902
#>  [433,] -0.7391441524
#>  [434,] -0.3018907243
#>  [435,] -0.2371939154
#>  [436,]  0.9648176627
#>  [437,] -0.5663752920
#>  [438,] -0.2494238998
#>  [439,] -0.8172055952
#>  [440,] -0.7060760924
#>  [441,] -0.5057711547
#>  [442,]  0.3789856095
#>  [443,] -0.9727862792
#>  [444,]  0.4803828630
#>  [445,]  0.6155097529
#>  [446,] -0.2193798926
#>  [447,]  0.2770511982
#>  [448,]  0.1446154275
#>  [449,]  0.7294830437
#>  [450,]  0.1318408716
#>  [451,] -0.9286134971
#>  [452,]  0.5650023131
#>  [453,] -0.0734986241
#>  [454,]  0.6055895147
#>  [455,]  0.5111041487
#>  [456,]  0.3522485854
#>  [457,] -0.7570125215
#>  [458,] -0.1832567398
#>  [459,] -0.7545485930
#>  [460,] -0.0289848871
#>  [461,]  0.7738267180
#>  [462,]  0.4916108827
#>  [463,] -0.3123352624
#>  [464,] -1.0281943629
#>  [465,]  0.2271060962
#>  [466,]  0.7107972248
#>  [467,] -0.3046784495
#>  [468,] -0.5839219734
#>  [469,]  1.4979409905
#>  [470,] -0.2798596786
#>  [471,]  0.0599010367
#>  [472,] -0.1166487972
#>  [473,]  0.5759355986
#>  [474,]  0.0739432337
#>  [475,]  0.5817939577
#>  [476,]  0.2870045625
#>  [477,] -0.1296604338
#>  [478,]  0.4669615188
#>  [479,] -0.0833652537
#>  [480,]  0.0444112459
#>  [481,] -0.2424532945
#>  [482,]  0.4828098904
#>  [483,]  0.3103995774
#>  [484,]  0.0108854020
#>  [485,]  0.7864599756
#>  [486,] -0.9776112921
#>  [487,] -0.3104853783
#>  [488,] -0.3305507854
#>  [489,] -0.2215564713
#>  [490,]  0.0721162638
#>  [491,]  0.2177921308
#>  [492,]  0.1394672836
#>  [493,]  0.2520153188
#>  [494,]  0.5967426405
#>  [495,]  0.3058456899
#>  [496,]  0.5917512017
#>  [497,]  0.3640754839
#>  [498,]  0.1549136439
#>  [499,] -0.4251448004
#>  [500,] -0.5954778951
#>  [501,]  0.1971949328
#>  [502,]  0.3262441011
#>  [503,]  0.2089657531
#>  [504,] -0.3834786294
#>  [505,]  0.4278060779
#>  [506,]  0.0551679061
#>  [507,] -0.0386641262
#>  [508,]  0.2733846342
#>  [509,] -0.5726200932
#>  [510,] -0.2647347535
#>  [511,] -0.1636733083
#>  [512,]  0.2262601673
#>  [513,] -0.0280576745
#>  [514,] -0.2338600247
#>  [515,]  0.6862045593
#>  [516,] -0.5119259955
#>  [517,] -0.0523622086
#>  [518,]  0.1044192908
#>  [519,] -0.1237421612
#>  [520,]  0.0757043097
#>  [521,] -0.5462479680
#>  [522,]  0.0475516375
#>  [523,]  1.1281028814
#>  [524,]  2.1539096000
#>  [525,] -0.3064854352
#>  [526,] -0.0496119115
#>  [527,] -0.8556209075
#>  [528,]  0.0941749169
#>  [529,] -0.2142782269
#>  [530,]  0.4863034378
#>  [531,] -0.6682602579
#>  [532,] -0.1670118545
#>  [533,]  0.5639834187
#>  [534,]  0.0179411828
#>  [535,] -0.3645879087
#>  [536,]  0.0408239965
#>  [537,]  0.9396996941
#>  [538,] -0.1184823983
#>  [539,]  0.4494436508
#>  [540,] -0.2161967552
#>  [541,] -0.0674644549
#>  [542,]  0.2276800728
#>  [543,]  0.1874794258
#>  [544,] -0.2695410954
#>  [545,] -0.1355878959
#>  [546,]  0.8939643342
#>  [547,] -0.3869048085
#>  [548,]  0.0633883690
#>  [549,]  0.4120551411
#>  [550,]  0.2301550321
#>  [551,]  0.9149700761
#>  [552,]  1.1397875473
#>  [553,] -0.4759153077
#>  [554,] -0.2316858674
#>  [555,]  0.5579831924
#>  [556,] -0.4722502241
#>  [557,] -0.1854983874
#>  [558,] -1.1758371853
#>  [559,]  0.4799621134
#>  [560,]  0.2889692348
#>  [561,] -1.2229723045
#>  [562,] -0.6987407012
#>  [563,] -0.7161668963
#>  [564,] -0.6979377754
#>  [565,] -0.3020553503
#>  [566,] -0.4900090880
#>  [567,] -0.8276189539
#>  [568,] -0.4675069429
#>  [569,]  0.1342977925
#>  [570,]  0.4109998070
#>  [571,]  0.6809850377
#>  [572,]  0.0412522766
#>  [573,] -0.1366387808
#>  [574,] -1.1207645050
#>  [575,]  0.3470158424
#>  [576,] -0.0923440906
#>  [577,] -0.3819436967
#>  [578,]  0.8333262625
#>  [579,] -0.8232459822
#>  [580,]  0.4842012497
#>  [581,]  0.3710676165
#>  [582,] -0.1472190861
#>  [583,] -0.6324098168
#>  [584,] -0.0760948993
#>  [585,]  0.3433345426
#>  [586,]  0.2907507556
#>  [587,]  0.1804708865
#>  [588,]  0.0960759358
#>  [589,]  0.6891969426
#>  [590,] -0.5049705797
#>  [591,]  1.2849174011
#>  [592,] -0.3799390674
#>  [593,]  0.7477907541
#>  [594,] -0.1956257732
#>  [595,] -0.4801454402
#>  [596,] -1.1238988958
#>  [597,] -0.1742779554
#>  [598,]  0.7379146064
#>  [599,] -0.3434285814
#>  [600,] -0.1916725347
#>  [601,]  0.5454044032
#>  [602,]  0.5152192026
#>  [603,] -0.0849482176
#>  [604,] -0.9375950478
#>  [605,]  0.0342875591
#>  [606,] -0.6655125097
#>  [607,]  0.1873396958
#>  [608,] -0.0769470245
#>  [609,]  0.0459691613
#>  [610,] -1.1788991053
#>  [611,]  0.3286496116
#>  [612,] -0.0070398492
#>  [613,]  0.0901132805
#>  [614,] -0.6918681224
#>  [615,] -0.4205909636
#>  [616,]  0.4543387666
#>  [617,] -1.0384831978
#>  [618,]  0.1549915832
#>  [619,]  0.0778184000
#>  [620,]  0.7126113770
#>  [621,] -0.4782955043
#>  [622,]  0.0135666161
#>  [623,]  0.4106911388
#>  [624,] -1.0563542288
#>  [625,] -0.9111005827
#>  [626,]  0.4652712967
#>  [627,]  0.0051123720
#>  [628,] -0.7323852363
#>  [629,]  0.1945706468
#>  [630,]  0.1503703349
#>  [631,] -0.4189430719
#>  [632,] -0.1985284766
#>  [633,]  0.3485533384
#>  [634,] -0.2546380753
#>  [635,]  0.1227124351
#>  [636,] -0.1856168193
#>  [637,]  0.2738393333
#>  [638,]  1.0011848581
#>  [639,]  0.4348026799
#>  [640,] -1.6300056973
#>  [641,]  0.2096922034
#>  [642,]  0.0085548475
#>  [643,]  0.8349257869
#>  [644,] -0.3045721882
#>  [645,]  0.6219872397
#>  [646,] -0.2574982682
#>  [647,]  0.0345361199
#>  [648,]  0.0117661476
#>  [649,] -1.4764823335
#>  [650,] -0.1628282831
#>  [651,] -1.2739730001
#>  [652,]  0.8571055603
#>  [653,]  0.1669340215
#>  [654,] -0.5317999836
#>  [655,] -1.0839479952
#>  [656,] -0.3629808608
#>  [657,]  0.2942477765
#>  [658,] -0.7617743380
#>  [659,] -0.1690583988
#>  [660,]  0.4040060554
#>  [661,] -0.3634683065
#>  [662,] -0.8665870590
#>  [663,] -0.0941847771
#>  [664,]  0.5124860508
#>  [665,]  0.3797416682
#>  [666,]  0.4798222983
#>  [667,] -0.5582050576
#>  [668,]  0.3613028444
#>  [669,]  0.5623126445
#>  [670,]  0.1053986631
#>  [671,]  0.8528668074
#>  [672,]  0.3209402806
#>  [673,] -0.2066761832
#>  [674,] -0.4629294530
#>  [675,]  0.0776329432
#>  [676,]  0.0504137697
#>  [677,]  0.0518271005
#>  [678,]  0.2186431836
#>  [679,]  0.9749758250
#>  [680,] -1.1149772463
#>  [681,]  0.3516449595
#>  [682,]  0.3667569078
#>  [683,]  0.8789467148
#>  [684,] -0.2670431662
#>  [685,] -0.1829144445
#>  [686,]  0.1890791001
#>  [687,]  1.0046912292
#>  [688,]  0.3745077898
#>  [689,] -0.0766475447
#>  [690,]  0.6810748700
#>  [691,]  0.4760056047
#>  [692,]  0.6531642461
#>  [693,]  1.4329355331
#>  [694,] -0.7168701945
#>  [695,] -0.9394765938
#>  [696,]  0.2211814773
#>  [697,] -0.6708518355
#>  [698,] -0.6266897475
#>  [699,] -0.2085854632
#>  [700,] -0.5009924909
#>  [701,] -0.0736232607
#>  [702,]  0.0828180310
#>  [703,]  0.3308022332
#>  [704,] -0.7440298311
#>  [705,] -0.3728463038
#>  [706,]  0.2977529814
#>  [707,]  0.7548414515
#>  [708,]  0.3886690323
#>  [709,]  0.4132967336
#>  [710,]  0.1895304218
#>  [711,]  0.3994095848
#>  [712,]  0.9802587308
#>  [713,]  0.7551687353
#>  [714,] -0.2736358031
#>  [715,]  0.1938561144
#>  [716,]  0.2522143263
#>  [717,]  0.4456446789
#>  [718,] -0.1725845564
#>  [719,]  1.0797331538
#>  [720,]  0.3771528861
#>  [721,]  1.0647437602
#>  [722,]  0.5670940759
#>  [723,] -0.3559599552
#>  [724,] -0.7792901329
#>  [725,]  0.3513301038
#>  [726,] -0.2204944772
#>  [727,] -0.7116203135
#>  [728,] -0.1274146704
#>  [729,]  0.1966052636
#>  [730,] -0.3877991857
#>  [731,] -0.0199317788
#>  [732,]  0.1184646922
#>  [733,]  0.2743341647
#>  [734,]  0.5775246710
#>  [735,] -0.5056920769
#>  [736,] -0.2552297910
#>  [737,] -0.6560364012
#>  [738,]  0.0348107583
#>  [739,]  0.0902894719
#>  [740,] -0.6225195913
#>  [741,]  0.0797149985
#>  [742,] -0.1238041900
#>  [743,] -0.2795945644
#>  [744,]  0.4582743141
#>  [745,] -0.2354779038
#>  [746,]  0.2209599286
#>  [747,] -0.5618323632
#>  [748,]  0.2231077962
#>  [749,]  0.4956506719
#>  [750,] -0.5832128235
#>  [751,]  0.2152663676
#>  [752,]  0.0542766240
#>  [753,] -0.2872032534
#>  [754,]  0.8427008077
#>  [755,]  0.7672825479
#>  [756,]  0.2816525002
#>  [757,] -0.2592055026
#>  [758,] -0.3337061736
#>  [759,]  0.3641846172
#>  [760,]  0.9735540327
#>  [761,] -0.3217716521
#>  [762,] -0.7869857554
#>  [763,] -0.4184688458
#>  [764,] -0.4806402163
#>  [765,] -0.8255720922
#>  [766,] -0.5303011557
#>  [767,]  1.3100794128
#>  [768,] -0.9604241691
#>  [769,] -0.0885802867
#>  [770,]  0.2058099786
#>  [771,] -0.7842681524
#>  [772,]  0.0635072318
#>  [773,] -0.6951058370
#>  [774,] -0.1846507711
#>  [775,] -0.3364022157
#>  [776,] -0.1788614748
#>  [777,]  0.8910687580
#>  [778,] -0.1378250806
#>  [779,] -0.5356146508
#>  [780,]  0.3572923264
#>  [781,]  0.1585391124
#>  [782,]  1.0687576636
#>  [783,]  1.1066435905
#>  [784,] -0.2037567501
#>  [785,]  0.9385984393
#>  [786,]  0.6147344907
#>  [787,] -0.6015521836
#>  [788,]  0.1344301864
#>  [789,] -1.2290188757
#>  [790,] -0.2334263497
#>  [791,]  0.6008788554
#>  [792,] -0.3406631419
#>  [793,]  0.0389005053
#>  [794,]  0.6983191793
#>  [795,]  0.3481429745
#>  [796,] -1.3029340982
#>  [797,] -0.2717512733
#>  [798,] -0.2855208576
#>  [799,] -0.0936098014
#>  [800,] -0.3429607837
#>  [801,]  0.0655800287
#>  [802,]  0.4265554152
#>  [803,] -0.1910353739
#>  [804,] -0.9140369084
#>  [805,]  0.3288519712
#>  [806,] -0.8874000188
#>  [807,]  1.0038508070
#>  [808,]  0.9868572631
#>  [809,] -0.2429694653
#>  [810,] -0.9556449477
#>  [811,]  0.4624246672
#>  [812,]  0.3846155391
#>  [813,]  0.2159116896
#>  [814,]  0.6700865982
#>  [815,] -0.4726019862
#>  [816,] -0.4929655010
#>  [817,]  0.0322210994
#>  [818,]  0.3226477467
#>  [819,] -0.6958715163
#>  [820,]  0.9750733295
#>  [821,] -0.2351567267
#>  [822,]  0.0729408638
#>  [823,] -0.1591468495
#>  [824,] -0.5117436494
#>  [825,] -0.4795229582
#>  [826,]  0.4052784649
#>  [827,]  0.4204432364
#>  [828,]  0.2745396169
#>  [829,] -0.0205302894
#>  [830,] -0.4314197701
#>  [831,]  0.6095073690
#>  [832,]  1.0723192064
#>  [833,] -0.2646044805
#>  [834,] -0.3071234158
#>  [835,] -1.0247811959
#>  [836,] -0.0066204419
#>  [837,] -0.3715147429
#>  [838,] -0.9059918430
#>  [839,]  0.3446672003
#>  [840,] -0.4288265062
#>  [841,]  0.5646947333
#>  [842,] -0.3737787556
#>  [843,]  0.0442058676
#>  [844,] -0.3987186493
#>  [845,] -0.7878976970
#>  [846,]  0.2526627617
#>  [847,] -0.7940440196
#>  [848,]  0.9474152815
#>  [849,] -0.4694354494
#>  [850,]  0.3829033622
#>  [851,]  0.6224263818
#>  [852,] -0.5851170881
#>  [853,]  0.0889089500
#>  [854,] -0.7938978228
#>  [855,]  0.3155417062
#>  [856,] -0.1221265985
#>  [857,] -1.3300553534
#>  [858,] -0.5719367575
#>  [859,] -0.9125861776
#>  [860,] -0.2369406740
#>  [861,] -0.1245172019
#>  [862,]  0.1233036406
#>  [863,] -0.3210582754
#>  [864,]  0.2492075329
#>  [865,] -0.2825158318
#>  [866,] -0.5327060313
#>  [867,]  0.8425520219
#>  [868,] -0.1497259276
#>  [869,] -0.0707899546
#>  [870,]  1.4702687405
#>  [871,]  0.3788539905
#>  [872,]  0.1531725633
#>  [873,] -0.0760097700
#>  [874,]  0.4299995019
#>  [875,]  0.2210190437
#>  [876,]  0.2148711390
#>  [877,]  0.1065693828
#>  [878,]  0.3748532475
#>  [879,]  1.3184077140
#>  [880,]  0.1440822963
#>  [881,] -0.2546920734
#>  [882,]  0.3841653829
#>  [883,] -0.4562342999
#>  [884,]  0.0205493941
#>  [885,] -0.5283209976
#>  [886,] -0.0380588581
#>  [887,]  0.8189318260
#>  [888,] -0.3172822011
#>  [889,] -0.0775267260
#>  [890,]  0.6523401756
#>  [891,] -0.5292098340
#>  [892,] -0.2480769098
#>  [893,] -0.5409446824
#>  [894,]  1.6439092955
#>  [895,] -0.0285486973
#>  [896,]  0.5037648342
#>  [897,]  0.1677124029
#>  [898,]  0.4745211666
#>  [899,] -0.3244950228
#>  [900,]  0.2792998540
#>  [901,] -0.5484102552
#>  [902,] -0.0305755861
#>  [903,] -0.1514650483
#>  [904,]  0.2612468517
#>  [905,]  0.2779371610
#>  [906,] -0.5204708983
#>  [907,]  0.2379749856
#>  [908,] -0.0438161645
#>  [909,] -0.1526783441
#>  [910,] -0.0887353319
#>  [911,] -0.1575921135
#>  [912,]  0.1630603823
#>  [913,]  0.7016950098
#>  [914,] -0.8003518385
#>  [915,] -0.5517602678
#>  [916,]  0.5581672564
#>  [917,]  0.9733687124
#>  [918,]  0.4126733289
#>  [919,]  0.2656305485
#>  [920,]  0.9628332288
#>  [921,] -0.9312318277
#>  [922,] -0.4634587877
#>  [923,] -1.3693203227
#>  [924,]  0.2544529775
#>  [925,]  0.4049774826
#>  [926,] -0.8171521325
#>  [927,]  0.9692904038
#>  [928,]  0.2155700616
#>  [929,] -0.0075314952
#>  [930,] -0.3707556074
#>  [931,]  0.3387207358
#>  [932,]  0.6080845672
#>  [933,]  0.3844809400
#>  [934,] -0.4457096685
#>  [935,]  0.4762935244
#>  [936,]  0.8586820168
#>  [937,] -0.8078280025
#>  [938,]  1.1776397967
#>  [939,]  0.2418611798
#>  [940,]  0.3335470973
#>  [941,]  0.1285093873
#>  [942,]  1.3411928506
#>  [943,] -0.5559123921
#>  [944,]  0.1590980120
#>  [945,] -0.6951381306
#>  [946,] -0.2605448390
#>  [947,] -0.5783589164
#>  [948,]  0.3986759282
#>  [949,] -0.1698487597
#>  [950,]  0.3859009383
#>  [951,]  0.2774559641
#>  [952,]  0.9474647318
#>  [953,]  0.2077114388
#>  [954,]  0.2806240367
#>  [955,] -0.1963331160
#>  [956,] -0.8031189376
#>  [957,] -0.2807566360
#>  [958,] -0.3026835432
#>  [959,]  0.5083381835
#>  [960,] -0.2461756180
#>  [961,] -0.4289318337
#>  [962,]  0.7402416421
#>  [963,]  0.4969591552
#>  [964,] -0.2984666143
#>  [965,]  1.1478432144
#>  [966,]  0.4754795523
#>  [967,]  0.3991802841
#>  [968,] -0.9194511096
#>  [969,]  0.1032765788
#>  [970,]  0.8412985369
#>  [971,]  0.0857694650
#>  [972,] -0.0358832121
#>  [973,] -0.6591555601
#>  [974,] -0.0544410786
#>  [975,]  0.0107344674
#>  [976,] -1.0040427022
#>  [977,]  0.2431236020
#>  [978,] -0.2085816932
#>  [979,] -0.5573609989
#>  [980,] -0.0040655722
#>  [981,] -0.1772557633
#>  [982,] -0.1807491451
#>  [983,] -0.7999641592
#>  [984,]  0.6382121094
#>  [985,] -0.7711656240
#>  [986,] -0.5226134063
#>  [987,]  0.0978386534
#>  [988,] -0.2031618775
#>  [989,] -0.0101071753
#>  [990,] -0.1496922230
#>  [991,] -0.2085600287
#>  [992,]  0.6978002890
#>  [993,]  0.2173160390
#>  [994,]  0.4405237066
#>  [995,] -0.0108876672
#>  [996,]  0.1153987786
#>  [997,] -0.3794054386
#>  [998,] -0.6080010257
#>  [999,]  0.6614832218
#> [1000,] -0.0739665548