Logistic Function

The elementwise logistic function, \(\log(1 + e^x)\). This is a special case of log(sum(exp)) that evaluates to a vector rather than to a scalar, which is useful for logistic regression.

logistic(x)

Arguments

x

An Expression, vector, or matrix.

Value

An Expression representing the logistic function evaluated at the input.

Examples

set.seed(92)
n <- 20
m <- 1000
sigma <- 45

beta_true <- stats::rnorm(n)
idxs <- sample(n, size = 0.8*n, replace = FALSE)
beta_true[idxs] <- 0
X <- matrix(stats::rnorm(m*n, 0, 5), nrow = m, ncol = n)
y <- sign(X %*% beta_true + stats::rnorm(m, 0, sigma))

beta <- Variable(n)
X_sign <- apply(X, 2, function(x) { ifelse(y <= 0, -1, 1) * x })
obj <- -sum(logistic(-X[y <= 0,] %*% beta)) - sum(logistic(X[y == 1,] %*% beta))
prob <- Problem(Maximize(obj))
result <- solve(prob)

log_odds <- result$getValue(X %*% beta)
beta_res <- result$getValue(beta)
y_probs <- 1/(1 + exp(-X %*% beta_res))
log(y_probs/(1 - y_probs))
#>                  [,1]
#>    [1,] -0.6704371536
#>    [2,] -0.5924970549
#>    [3,] -0.2870207194
#>    [4,] -0.3903305446
#>    [5,]  0.1512494762
#>    [6,] -0.3212221378
#>    [7,]  0.0710794679
#>    [8,] -0.0005961887
#>    [9,]  0.6231365028
#>   [10,] -0.3461443684
#>   [11,]  0.6165738322
#>   [12,]  0.2493729000
#>   [13,] -0.4229340567
#>   [14,]  1.3550204908
#>   [15,] -0.3265509828
#>   [16,] -0.5973892358
#>   [17,]  1.4250527921
#>   [18,] -0.8063875606
#>   [19,]  0.6354045317
#>   [20,] -0.5313145903
#>   [21,] -0.5836658886
#>   [22,] -0.2018558285
#>   [23,] -1.5740622176
#>   [24,] -0.2996736249
#>   [25,] -0.2075267120
#>   [26,]  0.4436835108
#>   [27,] -0.5653261234
#>   [28,] -1.2550228377
#>   [29,] -0.5346095797
#>   [30,]  0.4833898723
#>   [31,]  0.2310958007
#>   [32,]  0.2894397218
#>   [33,] -0.0125271282
#>   [34,]  0.7295520624
#>   [35,] -0.8746838352
#>   [36,] -0.3928786677
#>   [37,] -0.1155639087
#>   [38,]  1.1672097900
#>   [39,] -0.7217220191
#>   [40,]  0.0611228864
#>   [41,] -0.6167469517
#>   [42,] -0.0245098673
#>   [43,]  0.6908016675
#>   [44,] -0.8034209675
#>   [45,]  0.8836838590
#>   [46,] -1.0940286700
#>   [47,]  0.0151098659
#>   [48,] -0.0365081455
#>   [49,]  0.1208070217
#>   [50,]  0.0228325315
#>   [51,]  0.0203843663
#>   [52,] -0.4370748638
#>   [53,]  0.0381300909
#>   [54,] -0.0670253113
#>   [55,] -0.5450317899
#>   [56,] -0.1197466571
#>   [57,] -0.1350384362
#>   [58,]  0.1775410381
#>   [59,]  1.1397581600
#>   [60,] -0.4142450475
#>   [61,]  0.3992870501
#>   [62,] -0.4201999349
#>   [63,]  0.0406908904
#>   [64,] -0.6488893071
#>   [65,] -0.9702623739
#>   [66,]  0.1117415997
#>   [67,] -0.1453520467
#>   [68,]  0.0616511915
#>   [69,]  0.0779089693
#>   [70,]  0.4237295739
#>   [71,] -0.4578180245
#>   [72,] -0.2049643257
#>   [73,] -0.2303823007
#>   [74,]  0.1166922211
#>   [75,] -0.3651839099
#>   [76,] -0.2369421613
#>   [77,]  0.1457184733
#>   [78,] -1.0896100534
#>   [79,] -0.1308667941
#>   [80,]  0.2291608162
#>   [81,]  0.1373206225
#>   [82,] -0.0015645590
#>   [83,]  0.3467469103
#>   [84,] -0.1543835498
#>   [85,] -0.2308938190
#>   [86,] -0.5246963632
#>   [87,]  0.4603606858
#>   [88,] -0.3126381627
#>   [89,]  0.2119984253
#>   [90,]  0.3146859952
#>   [91,]  0.4192903982
#>   [92,]  0.6556585814
#>   [93,] -0.5110961916
#>   [94,]  1.3012866817
#>   [95,]  0.8029894017
#>   [96,]  0.2830272518
#>   [97,] -0.6703772068
#>   [98,]  0.6391863669
#>   [99,]  0.0848818575
#>  [100,] -0.0064332923
#>  [101,]  0.1647413201
#>  [102,]  0.2973037010
#>  [103,]  0.3564079114
#>  [104,] -0.8130206469
#>  [105,] -0.5596398431
#>  [106,]  0.2440550486
#>  [107,] -0.5330370810
#>  [108,]  0.4528272853
#>  [109,] -0.5708769383
#>  [110,] -0.4623841324
#>  [111,] -0.2336821167
#>  [112,]  0.1511168180
#>  [113,] -0.0618989021
#>  [114,]  0.5838774850
#>  [115,] -0.0005897145
#>  [116,] -0.7089513619
#>  [117,] -0.3172602390
#>  [118,]  0.3970471261
#>  [119,] -1.5819421843
#>  [120,]  0.5686097850
#>  [121,] -0.3273318204
#>  [122,] -0.8145712120
#>  [123,] -0.0947952405
#>  [124,]  0.3203346211
#>  [125,] -0.4429004042
#>  [126,]  0.1170302080
#>  [127,] -0.7652000886
#>  [128,] -0.0573301844
#>  [129,]  0.0095936089
#>  [130,] -0.1932077962
#>  [131,]  0.1964147884
#>  [132,] -0.3175516562
#>  [133,]  0.4246079181
#>  [134,]  0.1266855672
#>  [135,]  0.3305523081
#>  [136,] -0.0061834345
#>  [137,] -0.5537455889
#>  [138,] -0.1133659428
#>  [139,] -0.7994346883
#>  [140,]  0.8081345370
#>  [141,]  0.4152951840
#>  [142,]  0.8941963356
#>  [143,] -0.3139351763
#>  [144,]  0.1458335813
#>  [145,] -0.2050217529
#>  [146,] -0.0601875776
#>  [147,] -1.1747708036
#>  [148,]  0.5454530712
#>  [149,]  0.2075616513
#>  [150,] -0.2618668554
#>  [151,]  0.8038631576
#>  [152,] -0.0455059981
#>  [153,]  0.5741153650
#>  [154,]  0.6919653211
#>  [155,]  0.0716635518
#>  [156,] -0.4965605133
#>  [157,]  0.4423766989
#>  [158,] -0.3816733357
#>  [159,]  0.5205855546
#>  [160,] -0.1021902110
#>  [161,] -0.2226333602
#>  [162,]  0.7098615761
#>  [163,]  0.5765079241
#>  [164,]  0.2450234845
#>  [165,] -0.1737037973
#>  [166,]  0.3589537727
#>  [167,] -0.3289941328
#>  [168,] -0.8993056792
#>  [169,] -0.8217702772
#>  [170,] -0.7797528629
#>  [171,] -0.2137113010
#>  [172,]  0.5276129313
#>  [173,] -0.8487540548
#>  [174,]  0.1034511834
#>  [175,] -0.4103495194
#>  [176,] -0.7528536716
#>  [177,] -0.4402248633
#>  [178,] -1.0685305792
#>  [179,]  0.4078400388
#>  [180,]  0.6313384527
#>  [181,]  0.3750767342
#>  [182,] -0.0151760713
#>  [183,] -0.5310658138
#>  [184,]  0.4077389424
#>  [185,]  0.1571498830
#>  [186,] -0.6328385157
#>  [187,] -0.3822421104
#>  [188,]  0.5204977222
#>  [189,]  0.2562612439
#>  [190,]  0.8532319796
#>  [191,] -0.5648959945
#>  [192,]  0.0370575608
#>  [193,] -0.3390685700
#>  [194,] -0.0375178322
#>  [195,]  0.0474691628
#>  [196,] -0.0326508613
#>  [197,]  0.2159912492
#>  [198,] -1.0662168263
#>  [199,] -0.9549591875
#>  [200,] -0.6453604072
#>  [201,]  0.5183165577
#>  [202,] -0.6994760335
#>  [203,]  0.2180900675
#>  [204,]  0.1893010887
#>  [205,]  0.4929717779
#>  [206,]  0.1844585902
#>  [207,]  0.3573788914
#>  [208,] -0.2500475107
#>  [209,]  1.3916700683
#>  [210,] -0.7617566067
#>  [211,]  0.5279063041
#>  [212,] -1.0117377264
#>  [213,] -0.2423643713
#>  [214,] -0.6133731964
#>  [215,] -1.1655821251
#>  [216,]  0.3821205590
#>  [217,] -0.4501059373
#>  [218,] -1.0972958499
#>  [219,]  0.8937038193
#>  [220,]  0.3140554393
#>  [221,]  0.1340460686
#>  [222,] -1.2230466787
#>  [223,]  0.5678001497
#>  [224,]  0.8463307286
#>  [225,]  0.6348930242
#>  [226,]  0.7005666356
#>  [227,] -0.7991866479
#>  [228,]  0.5683221791
#>  [229,]  0.4112845264
#>  [230,]  0.0902384990
#>  [231,] -0.3665052768
#>  [232,]  0.0127778808
#>  [233,]  0.1629114679
#>  [234,] -0.3527060141
#>  [235,] -0.0758178894
#>  [236,]  0.0091703090
#>  [237,] -0.4611016997
#>  [238,]  0.0657026892
#>  [239,] -0.5298373480
#>  [240,] -0.5961982846
#>  [241,]  0.0637630897
#>  [242,] -0.3510008511
#>  [243,]  0.5475868321
#>  [244,]  0.5265410802
#>  [245,] -0.8597857166
#>  [246,]  0.6395072030
#>  [247,]  0.3215850911
#>  [248,] -0.4795746135
#>  [249,]  0.3975075566
#>  [250,]  0.1215588503
#>  [251,] -0.3140178253
#>  [252,] -0.3046388565
#>  [253,]  0.3923042433
#>  [254,] -0.1856373031
#>  [255,] -0.0851580547
#>  [256,] -0.0029478231
#>  [257,]  0.5734570564
#>  [258,] -0.0359951466
#>  [259,] -0.9567630443
#>  [260,]  0.6395436832
#>  [261,]  0.4795560341
#>  [262,]  0.1872548182
#>  [263,] -0.2498804581
#>  [264,]  0.3260846519
#>  [265,]  0.1986533529
#>  [266,] -0.6790411675
#>  [267,]  0.6809304320
#>  [268,] -0.2952457251
#>  [269,]  0.5265617009
#>  [270,]  0.6812674037
#>  [271,] -0.4787051835
#>  [272,]  0.6944439888
#>  [273,]  0.3116084802
#>  [274,]  0.6770165536
#>  [275,]  0.0024186543
#>  [276,]  0.1779699225
#>  [277,]  0.1332607901
#>  [278,]  0.2881507054
#>  [279,] -0.8397962713
#>  [280,]  0.8292775991
#>  [281,]  0.3237814549
#>  [282,]  0.3688578998
#>  [283,] -0.2838318843
#>  [284,]  0.1698004876
#>  [285,] -0.0972757171
#>  [286,]  0.1692241261
#>  [287,]  0.6139121635
#>  [288,]  0.1736774048
#>  [289,] -0.0806310173
#>  [290,]  0.0115356410
#>  [291,] -0.3183776053
#>  [292,]  0.0866688610
#>  [293,] -0.7586085651
#>  [294,]  0.0584494009
#>  [295,] -0.1537315352
#>  [296,]  0.8435040022
#>  [297,]  0.8815133713
#>  [298,]  0.1534422891
#>  [299,]  1.2270852111
#>  [300,]  0.3098811958
#>  [301,] -0.2555183449
#>  [302,] -0.2073228094
#>  [303,] -0.2298493122
#>  [304,] -0.5771943799
#>  [305,] -0.4885229217
#>  [306,]  0.5318102353
#>  [307,] -0.1440876774
#>  [308,]  0.6375668399
#>  [309,]  0.2563495290
#>  [310,] -0.0603826125
#>  [311,] -0.2223115878
#>  [312,] -0.4950909080
#>  [313,] -0.2533065105
#>  [314,]  0.3508685320
#>  [315,]  0.0868118473
#>  [316,] -0.3995643750
#>  [317,] -0.2455160246
#>  [318,] -0.0987993348
#>  [319,]  0.6998024621
#>  [320,] -0.4848056892
#>  [321,] -0.3735274255
#>  [322,] -0.0385913222
#>  [323,]  0.0463446624
#>  [324,] -0.8583407399
#>  [325,] -0.4583042041
#>  [326,]  0.6281218504
#>  [327,] -0.6350738922
#>  [328,] -0.9581536906
#>  [329,]  0.9368189991
#>  [330,]  0.1483795514
#>  [331,] -0.0538229856
#>  [332,] -1.2322662603
#>  [333,] -0.1706842695
#>  [334,]  0.1205466826
#>  [335,] -0.4870705950
#>  [336,]  0.0539993440
#>  [337,] -0.1344803643
#>  [338,]  1.1543219382
#>  [339,]  0.8618718530
#>  [340,]  0.0228912762
#>  [341,] -0.5238850005
#>  [342,]  0.3351121133
#>  [343,] -0.7516925263
#>  [344,]  0.6050107669
#>  [345,] -0.0211546138
#>  [346,]  0.4027054784
#>  [347,]  0.8252612310
#>  [348,] -0.3306991151
#>  [349,]  0.3177167638
#>  [350,] -0.0224511937
#>  [351,] -0.5546576024
#>  [352,] -0.2831497250
#>  [353,] -0.1794602347
#>  [354,] -0.5933322824
#>  [355,]  0.5243653691
#>  [356,]  0.2004051609
#>  [357,]  0.2904586692
#>  [358,] -0.4816958366
#>  [359,] -0.1010020247
#>  [360,]  0.3605417321
#>  [361,]  0.3838467719
#>  [362,] -0.6909039469
#>  [363,] -0.3578574113
#>  [364,]  0.7571424542
#>  [365,]  0.4263218358
#>  [366,] -0.7291761536
#>  [367,]  1.3050651706
#>  [368,]  0.3406579942
#>  [369,]  0.1624561632
#>  [370,] -0.4134130589
#>  [371,]  0.8469514246
#>  [372,]  0.1445649511
#>  [373,] -1.0050810016
#>  [374,]  0.1813196002
#>  [375,]  0.3127873156
#>  [376,] -0.2938361177
#>  [377,] -1.1220384937
#>  [378,]  0.0153628659
#>  [379,]  0.6470094251
#>  [380,] -1.1135920633
#>  [381,] -0.6871516634
#>  [382,]  0.5672899540
#>  [383,] -0.2756522856
#>  [384,] -0.8617810850
#>  [385,] -0.1516781513
#>  [386,] -0.2098135313
#>  [387,] -1.2009950399
#>  [388,]  1.2558065399
#>  [389,] -0.5077951560
#>  [390,] -0.8335578399
#>  [391,]  0.8758981206
#>  [392,]  0.5860262464
#>  [393,] -1.2937601351
#>  [394,] -0.7939624455
#>  [395,] -0.5236187946
#>  [396,]  0.1775919943
#>  [397,] -1.4012999931
#>  [398,]  0.7131303228
#>  [399,]  0.0354470315
#>  [400,]  0.1344346042
#>  [401,]  0.4879601434
#>  [402,]  0.3637964696
#>  [403,] -0.1982356901
#>  [404,]  0.1427005201
#>  [405,]  0.0306143571
#>  [406,]  0.6817964609
#>  [407,]  0.0828413736
#>  [408,]  0.1284989089
#>  [409,]  0.7544252469
#>  [410,]  0.2249827926
#>  [411,]  0.0320374644
#>  [412,]  0.4893562385
#>  [413,]  0.3546810600
#>  [414,] -0.3652811586
#>  [415,]  0.8342382221
#>  [416,] -0.2502976035
#>  [417,]  0.5247080610
#>  [418,]  0.9032774598
#>  [419,]  0.6160013791
#>  [420,]  0.1392475009
#>  [421,]  0.7268018207
#>  [422,] -0.5110214346
#>  [423,] -0.6768522888
#>  [424,]  0.8344005978
#>  [425,]  0.3100101727
#>  [426,]  0.4325751680
#>  [427,]  0.1332970277
#>  [428,]  0.5181788324
#>  [429,] -0.3189916465
#>  [430,]  0.4315474267
#>  [431,]  0.3930133756
#>  [432,] -0.4950272369
#>  [433,] -0.7391026804
#>  [434,] -0.3018751084
#>  [435,] -0.2371812006
#>  [436,]  0.9647630791
#>  [437,] -0.5663435771
#>  [438,] -0.2494081269
#>  [439,] -0.8171576860
#>  [440,] -0.7060359456
#>  [441,] -0.5057450380
#>  [442,]  0.3789650418
#>  [443,] -0.9727302653
#>  [444,]  0.4803557203
#>  [445,]  0.6154759578
#>  [446,] -0.2193702957
#>  [447,]  0.2770369380
#>  [448,]  0.1446087161
#>  [449,]  0.7294448806
#>  [450,]  0.1318335772
#>  [451,] -0.9285619038
#>  [452,]  0.5649709765
#>  [453,] -0.0734955649
#>  [454,]  0.6055548961
#>  [455,]  0.5110751300
#>  [456,]  0.3522293641
#>  [457,] -0.7569706449
#>  [458,] -0.1832468898
#>  [459,] -0.7545049045
#>  [460,] -0.0289820837
#>  [461,]  0.7737808346
#>  [462,]  0.4915806942
#>  [463,] -0.3123188995
#>  [464,] -1.0281368072
#>  [465,]  0.2270935014
#>  [466,]  0.7107563475
#>  [467,] -0.3046626950
#>  [468,] -0.5838913933
#>  [469,]  1.4978597975
#>  [470,] -0.2798456702
#>  [471,]  0.0598968378
#>  [472,] -0.1166416643
#>  [473,]  0.5759006613
#>  [474,]  0.0739394932
#>  [475,]  0.5817607919
#>  [476,]  0.2869880844
#>  [477,] -0.1296524473
#>  [478,]  0.4669364460
#>  [479,] -0.0833612683
#>  [480,]  0.0444072863
#>  [481,] -0.2424410798
#>  [482,]  0.4827836556
#>  [483,]  0.3103832558
#>  [484,]  0.0108831268
#>  [485,]  0.7864166964
#>  [486,] -0.9775570658
#>  [487,] -0.3104653658
#>  [488,] -0.3305326402
#>  [489,] -0.2215454635
#>  [490,]  0.0721119666
#>  [491,]  0.2177769631
#>  [492,]  0.1394609002
#>  [493,]  0.2520017769
#>  [494,]  0.5967104024
#>  [495,]  0.3058290106
#>  [496,]  0.5917174095
#>  [497,]  0.3640554104
#>  [498,]  0.1549040815
#>  [499,] -0.4251214018
#>  [500,] -0.5954441973
#>  [501,]  0.1971851745
#>  [502,]  0.3262269633
#>  [503,]  0.2089535938
#>  [504,] -0.3834563407
#>  [505,]  0.4277818393
#>  [506,]  0.0551658550
#>  [507,] -0.0386635043
#>  [508,]  0.2733700808
#>  [509,] -0.5725888314
#>  [510,] -0.2647217788
#>  [511,] -0.1636635182
#>  [512,]  0.2262473011
#>  [513,] -0.0280542693
#>  [514,] -0.2338480245
#>  [515,]  0.6861692365
#>  [516,] -0.5118979748
#>  [517,] -0.0523596941
#>  [518,]  0.1044141927
#>  [519,] -0.1237360103
#>  [520,]  0.0757020992
#>  [521,] -0.5462163341
#>  [522,]  0.0475490115
#>  [523,]  1.1280395590
#>  [524,]  2.1537908187
#>  [525,] -0.3064695331
#>  [526,] -0.0496079312
#>  [527,] -0.8555735851
#>  [528,]  0.0941710872
#>  [529,] -0.2142656803
#>  [530,]  0.4862759981
#>  [531,] -0.6682230433
#>  [532,] -0.1670002253
#>  [533,]  0.5639511301
#>  [534,]  0.0179406785
#>  [535,] -0.3645673897
#>  [536,]  0.0408214074
#>  [537,]  0.9396472585
#>  [538,] -0.1184765291
#>  [539,]  0.4494183127
#>  [540,] -0.2161842346
#>  [541,] -0.0674599754
#>  [542,]  0.2276690892
#>  [543,]  0.1874693291
#>  [544,] -0.2695259607
#>  [545,] -0.1355822895
#>  [546,]  0.8939135418
#>  [547,] -0.3868838898
#>  [548,]  0.0633839832
#>  [549,]  0.4120333756
#>  [550,]  0.2301427253
#>  [551,]  0.9149168501
#>  [552,]  1.1397245688
#>  [553,] -0.4758874316
#>  [554,] -0.2316727118
#>  [555,]  0.5579533059
#>  [556,] -0.4722235629
#>  [557,] -0.1854874476
#>  [558,] -1.1757722104
#>  [559,]  0.4799355921
#>  [560,]  0.2889526594
#>  [561,] -1.2229036253
#>  [562,] -0.6987016036
#>  [563,] -0.7161284025
#>  [564,] -0.6978977999
#>  [565,] -0.3020373757
#>  [566,] -0.4899812295
#>  [567,] -0.8275713908
#>  [568,] -0.4674820092
#>  [569,]  0.1342915789
#>  [570,]  0.4109737195
#>  [571,]  0.6809472168
#>  [572,]  0.0412527630
#>  [573,] -0.1366310463
#>  [574,] -1.1206997740
#>  [575,]  0.3469964510
#>  [576,] -0.0923381986
#>  [577,] -0.3819228898
#>  [578,]  0.8332797200
#>  [579,] -0.8231996212
#>  [580,]  0.4841750320
#>  [581,]  0.3710459219
#>  [582,] -0.1472112292
#>  [583,] -0.6323713289
#>  [584,] -0.0760882618
#>  [585,]  0.3433160255
#>  [586,]  0.2907354369
#>  [587,]  0.1804593389
#>  [588,]  0.0960726924
#>  [589,]  0.6891577185
#>  [590,] -0.5049427939
#>  [591,]  1.2848475207
#>  [592,] -0.3799177276
#>  [593,]  0.7477480791
#>  [594,] -0.1956131097
#>  [595,] -0.4801180519
#>  [596,] -1.1238367472
#>  [597,] -0.1742673948
#>  [598,]  0.7378741080
#>  [599,] -0.3434082653
#>  [600,] -0.1916612128
#>  [601,]  0.5453737776
#>  [602,]  0.5151929960
#>  [603,] -0.0849446702
#>  [604,] -0.9375427053
#>  [605,]  0.0342838872
#>  [606,] -0.6654737580
#>  [607,]  0.1873288315
#>  [608,] -0.0769410194
#>  [609,]  0.0459691714
#>  [610,] -1.1788341018
#>  [611,]  0.3286316143
#>  [612,] -0.0070399490
#>  [613,]  0.0901083065
#>  [614,] -0.6918318235
#>  [615,] -0.4205678865
#>  [616,]  0.4543142198
#>  [617,] -1.0384260987
#>  [618,]  0.1549835014
#>  [619,]  0.0778117652
#>  [620,]  0.7125741976
#>  [621,] -0.4782691615
#>  [622,]  0.0135677547
#>  [623,]  0.4106698474
#>  [624,] -1.0562945351
#>  [625,] -0.9110488817
#>  [626,]  0.4652446431
#>  [627,]  0.0051111894
#>  [628,] -0.7323434187
#>  [629,]  0.1945627765
#>  [630,]  0.1503639205
#>  [631,] -0.4189210880
#>  [632,] -0.1985202306
#>  [633,]  0.3485346006
#>  [634,] -0.2546239065
#>  [635,]  0.1227065145
#>  [636,] -0.1856059292
#>  [637,]  0.2738245826
#>  [638,]  1.0011293900
#>  [639,]  0.4347783027
#>  [640,] -1.6299164906
#>  [641,]  0.2096804644
#>  [642,]  0.0085554771
#>  [643,]  0.8348778555
#>  [644,] -0.3045547260
#>  [645,]  0.6219522070
#>  [646,] -0.2574834042
#>  [647,]  0.0345338049
#>  [648,]  0.0117657026
#>  [649,] -1.4763989965
#>  [650,] -0.1628197087
#>  [651,] -1.2739006901
#>  [652,]  0.8570565365
#>  [653,]  0.1669252864
#>  [654,] -0.5317695971
#>  [655,] -1.0838890113
#>  [656,] -0.3629610924
#>  [657,]  0.2942318914
#>  [658,] -0.7617329976
#>  [659,] -0.1690507029
#>  [660,]  0.4039844421
#>  [661,] -0.3634482369
#>  [662,] -0.8665364350
#>  [663,] -0.0941790629
#>  [664,]  0.5124562764
#>  [665,]  0.3797202076
#>  [666,]  0.4797959767
#>  [667,] -0.5581728484
#>  [668,]  0.3612803581
#>  [669,]  0.5622815421
#>  [670,]  0.1053940567
#>  [671,]  0.8528174706
#>  [672,]  0.3209239401
#>  [673,] -0.2066647762
#>  [674,] -0.4629037239
#>  [675,]  0.0776301821
#>  [676,]  0.0504103802
#>  [677,]  0.0518251808
#>  [678,]  0.2186309562
#>  [679,]  0.9749210879
#>  [680,] -1.1149166299
#>  [681,]  0.3516259919
#>  [682,]  0.3667365382
#>  [683,]  0.8788987704
#>  [684,] -0.2670283150
#>  [685,] -0.1829033695
#>  [686,]  0.1890709841
#>  [687,]  1.0046336465
#>  [688,]  0.3744861564
#>  [689,] -0.0766435659
#>  [690,]  0.6810344262
#>  [691,]  0.4759775469
#>  [692,]  0.6531270721
#>  [693,]  1.4328584875
#>  [694,] -0.7168303718
#>  [695,] -0.9394229057
#>  [696,]  0.2211693267
#>  [697,] -0.6708147296
#>  [698,] -0.6266498849
#>  [699,] -0.2085742179
#>  [700,] -0.5009638785
#>  [701,] -0.0736191937
#>  [702,]  0.0828149280
#>  [703,]  0.3307833106
#>  [704,] -0.7439847991
#>  [705,] -0.3728216701
#>  [706,]  0.2977359381
#>  [707,]  0.7547982045
#>  [708,]  0.3886445430
#>  [709,]  0.4132736254
#>  [710,]  0.1895195157
#>  [711,]  0.3993880269
#>  [712,]  0.9802034278
#>  [713,]  0.7551247085
#>  [714,] -0.2736198953
#>  [715,]  0.1938462913
#>  [716,]  0.2522021664
#>  [717,]  0.4456210538
#>  [718,] -0.1725748975
#>  [719,]  1.0796724404
#>  [720,]  0.3771311537
#>  [721,]  1.0646832756
#>  [722,]  0.5670633867
#>  [723,] -0.3559389399
#>  [724,] -0.7792448504
#>  [725,]  0.3513107143
#>  [726,] -0.2204825461
#>  [727,] -0.7115830200
#>  [728,] -0.1274078617
#>  [729,]  0.1965955015
#>  [730,] -0.3877784328
#>  [731,] -0.0199314155
#>  [732,]  0.1184614808
#>  [733,]  0.2743204938
#>  [734,]  0.5774957181
#>  [735,] -0.5056630586
#>  [736,] -0.2552138070
#>  [737,] -0.6559975562
#>  [738,]  0.0348097420
#>  [739,]  0.0902843750
#>  [740,] -0.6224848658
#>  [741,]  0.0797100448
#>  [742,] -0.1237957604
#>  [743,] -0.2795788249
#>  [744,]  0.4582489032
#>  [745,] -0.2354659232
#>  [746,]  0.2209467376
#>  [747,] -0.5618031673
#>  [748,]  0.2230957584
#>  [749,]  0.4956249311
#>  [750,] -0.5831779870
#>  [751,]  0.2152564285
#>  [752,]  0.0542729034
#>  [753,] -0.2871861266
#>  [754,]  0.8426536721
#>  [755,]  0.7672403072
#>  [756,]  0.2816366370
#>  [757,] -0.2591920545
#>  [758,] -0.3336892143
#>  [759,]  0.3641623311
#>  [760,]  0.9734985416
#>  [761,] -0.3217553393
#>  [762,] -0.7869394343
#>  [763,] -0.4184477300
#>  [764,] -0.4806135252
#>  [765,] -0.8255281519
#>  [766,] -0.5302698128
#>  [767,]  1.3100060162
#>  [768,] -0.9603710998
#>  [769,] -0.0885757095
#>  [770,]  0.2057976200
#>  [771,] -0.7842226356
#>  [772,]  0.0635014179
#>  [773,] -0.6950667394
#>  [774,] -0.1846403957
#>  [775,] -0.3363813088
#>  [776,] -0.1788507353
#>  [777,]  0.8910171271
#>  [778,] -0.1378189439
#>  [779,] -0.5355852729
#>  [780,]  0.3572707139
#>  [781,]  0.1585296889
#>  [782,]  1.0686976361
#>  [783,]  1.1065801632
#>  [784,] -0.2037436928
#>  [785,]  0.9385459174
#>  [786,]  0.6147017846
#>  [787,] -0.6015202397
#>  [788,]  0.1344238655
#>  [789,] -1.2289485947
#>  [790,] -0.2334102667
#>  [791,]  0.6008456046
#>  [792,] -0.3406449647
#>  [793,]  0.0388999224
#>  [794,]  0.6982796634
#>  [795,]  0.3481236792
#>  [796,] -1.3028615822
#>  [797,] -0.2717341373
#>  [798,] -0.2855052943
#>  [799,] -0.0936034059
#>  [800,] -0.3429404588
#>  [801,]  0.0655747577
#>  [802,]  0.4265315904
#>  [803,] -0.1910251686
#>  [804,] -0.9139874111
#>  [805,]  0.3288341428
#>  [806,] -0.8873493970
#>  [807,]  1.0037938869
#>  [808,]  0.9868022556
#>  [809,] -0.2429568420
#>  [810,] -0.9555891094
#>  [811,]  0.4623956967
#>  [812,]  0.3845943807
#>  [813,]  0.2159000003
#>  [814,]  0.6700473531
#>  [815,] -0.4725773764
#>  [816,] -0.4929375156
#>  [817,]  0.0322172892
#>  [818,]  0.3226311055
#>  [819,] -0.6958346359
#>  [820,]  0.9750165658
#>  [821,] -0.2351415148
#>  [822,]  0.0729371318
#>  [823,] -0.1591380074
#>  [824,] -0.5117172701
#>  [825,] -0.4794958045
#>  [826,]  0.4052584912
#>  [827,]  0.4204194047
#>  [828,]  0.2745221730
#>  [829,] -0.0205273118
#>  [830,] -0.4313935626
#>  [831,]  0.6094751308
#>  [832,]  1.0722562499
#>  [833,] -0.2645906126
#>  [834,] -0.3071058500
#>  [835,] -1.0247243536
#>  [836,] -0.0066202888
#>  [837,] -0.3714932321
#>  [838,] -0.9059397683
#>  [839,]  0.3446469476
#>  [840,] -0.4288030195
#>  [841,]  0.5646611423
#>  [842,] -0.3737567336
#>  [843,]  0.0442022166
#>  [844,] -0.3986977469
#>  [845,] -0.7878512623
#>  [846,]  0.2526489031
#>  [847,] -0.7939966305
#>  [848,]  0.9473635039
#>  [849,] -0.4694084579
#>  [850,]  0.3828849604
#>  [851,]  0.6223936611
#>  [852,] -0.5850851118
#>  [853,]  0.0889029934
#>  [854,] -0.7938536687
#>  [855,]  0.3155238113
#>  [856,] -0.1221200251
#>  [857,] -1.3299805638
#>  [858,] -0.5719045011
#>  [859,] -0.9125344675
#>  [860,] -0.2369290342
#>  [861,] -0.1245118530
#>  [862,]  0.1232970400
#>  [863,] -0.3210417607
#>  [864,]  0.2491932528
#>  [865,] -0.2825020733
#>  [866,] -0.5326769112
#>  [867,]  0.8425040466
#>  [868,] -0.1497167423
#>  [869,] -0.0707890886
#>  [870,]  1.4701869812
#>  [871,]  0.3788327646
#>  [872,]  0.1531616726
#>  [873,] -0.0760065036
#>  [874,]  0.4299761551
#>  [875,]  0.2210052414
#>  [876,]  0.2148569902
#>  [877,]  0.1065630802
#>  [878,]  0.3748299869
#>  [879,]  1.3183336500
#>  [880,]  0.1440740556
#>  [881,] -0.2546796975
#>  [882,]  0.3841462697
#>  [883,] -0.4562112415
#>  [884,]  0.0205476811
#>  [885,] -0.5282887201
#>  [886,] -0.0380564559
#>  [887,]  0.8188855077
#>  [888,] -0.3172635106
#>  [889,] -0.0775231468
#>  [890,]  0.6523051264
#>  [891,] -0.5291824183
#>  [892,] -0.2480639855
#>  [893,] -0.5409122820
#>  [894,]  1.6438171323
#>  [895,] -0.0285455096
#>  [896,]  0.5037357682
#>  [897,]  0.1677028656
#>  [898,]  0.4744910107
#>  [899,] -0.3244755902
#>  [900,]  0.2792824991
#>  [901,] -0.5483815711
#>  [902,] -0.0305738705
#>  [903,] -0.1514546872
#>  [904,]  0.2612346130
#>  [905,]  0.2779208981
#>  [906,] -0.5204419343
#>  [907,]  0.2379637582
#>  [908,] -0.0438148167
#>  [909,] -0.1526698061
#>  [910,] -0.0887308975
#>  [911,] -0.1575809624
#>  [912,]  0.1630512943
#>  [913,]  0.7016546637
#>  [914,] -0.8003079386
#>  [915,] -0.5517306190
#>  [916,]  0.5581334416
#>  [917,]  0.9733141965
#>  [918,]  0.4126495220
#>  [919,]  0.2656149757
#>  [920,]  0.9627783592
#>  [921,] -0.9311797085
#>  [922,] -0.4634329810
#>  [923,] -1.3692431988
#>  [924,]  0.2544402163
#>  [925,]  0.4049540435
#>  [926,] -0.8171062158
#>  [927,]  0.9692371648
#>  [928,]  0.2155578850
#>  [929,] -0.0075319919
#>  [930,] -0.3707372457
#>  [931,]  0.3387029569
#>  [932,]  0.6080497308
#>  [933,]  0.3844598513
#>  [934,] -0.4456861777
#>  [935,]  0.4762665561
#>  [936,]  0.8586337111
#>  [937,] -0.8077824607
#>  [938,]  1.1775747224
#>  [939,]  0.2418477833
#>  [940,]  0.3335301005
#>  [941,]  0.1285003015
#>  [942,]  1.3411171383
#>  [943,] -0.5558827381
#>  [944,]  0.1590904246
#>  [945,] -0.6950985348
#>  [946,] -0.2605329228
#>  [947,] -0.5783275069
#>  [948,]  0.3986532474
#>  [949,] -0.1698383115
#>  [950,]  0.3858785617
#>  [951,]  0.2774383627
#>  [952,]  0.9474105143
#>  [953,]  0.2077017437
#>  [954,]  0.2806107137
#>  [955,] -0.1963208323
#>  [956,] -0.8030743922
#>  [957,] -0.2807400438
#>  [958,] -0.3026651383
#>  [959,]  0.5083128352
#>  [960,] -0.2461613303
#>  [961,] -0.4289107679
#>  [962,]  0.7401994321
#>  [963,]  0.4969320237
#>  [964,] -0.2984506724
#>  [965,]  1.1477782679
#>  [966,]  0.4754518386
#>  [967,]  0.3991576648
#>  [968,] -0.9193972102
#>  [969,]  0.1032722321
#>  [970,]  0.8412513882
#>  [971,]  0.0857632178
#>  [972,] -0.0358819416
#>  [973,] -0.6591196205
#>  [974,] -0.0544396306
#>  [975,]  0.0107342027
#>  [976,] -1.0039846599
#>  [977,]  0.2431098440
#>  [978,] -0.2085708407
#>  [979,] -0.5573298421
#>  [980,] -0.0040649299
#>  [981,] -0.1772469700
#>  [982,] -0.1807382833
#>  [983,] -0.7999187609
#>  [984,]  0.6381774395
#>  [985,] -0.7711219042
#>  [986,] -0.5225847604
#>  [987,]  0.0978345823
#>  [988,] -0.2031481975
#>  [989,] -0.0101080755
#>  [990,] -0.1496860642
#>  [991,] -0.2085505066
#>  [992,]  0.6977605795
#>  [993,]  0.2173046406
#>  [994,]  0.4405001890
#>  [995,] -0.0108882126
#>  [996,]  0.1153933185
#>  [997,] -0.3793848791
#>  [998,] -0.6079666500
#>  [999,]  0.6614474429
#> [1000,] -0.0739607664