Logistic Function

The elementwise logistic function, \(\log(1 + e^x)\). This is a special case of log(sum(exp)) that evaluates to a vector rather than to a scalar, which is useful for logistic regression.

logistic(x)

Arguments

x
An Expression, vector, or matrix.

Value

An Expression representing the logistic function evaluated at the input.

Examples

n <- 20
m <- 1000
sigma <- 45

beta_true <- stats::rnorm(n)
idxs <- sample(n, size = 0.8*n, replace = FALSE)
beta_true[idxs] <- 0
X <- matrix(stats::rnorm(m*n, 0, 5), nrow = m, ncol = n)
y <- sign(X %*% beta_true + stats::rnorm(m, 0, sigma))

beta <- Variable(n)
X_sign <- apply(X, 2, function(x) { ifelse(y <= 0, -1, 1) * x })
obj <- -sum(logistic(-X[y <= 0,] %*% beta)) - sum(logistic(X[y == 1,] %*% beta))
prob <- Problem(Maximize(obj))
result <- solve(prob)

log_odds <- result$getValue(X %*% beta)
beta_res <- result$getValue(beta)
y_probs <- 1/(1 + exp(-X %*% beta_res))
log(y_probs/(1 - y_probs))

#>                  [,1]
#>    [1,] -0.1247827797
#>    [2,]  0.1859585341
#>    [3,]  0.2934527476
#>    [4,] -0.4543832444
#>    [5,]  0.1066752546
#>    [6,] -0.3271244542
#>    [7,] -0.4170998280
#>    [8,]  0.9728848270
#>    [9,] -0.1776807250
#>   [10,]  0.1436550441
#>   [11,] -0.0549418865
#>   [12,] -0.2692309486
#>   [13,] -0.3053789946
#>   [14,]  0.3605003190
#>   [15,]  0.1219291475
#>   [16,]  0.0371039855
#>   [17,]  0.6286890908
#>   [18,]  0.5428560926
#>   [19,]  0.1926364206
#>   [20,] -1.0109885783
#>   [21,]  0.3620383311
#>   [22,] -0.1957165751
#>   [23,] -0.8572081614
#>   [24,] -0.0731686487
#>   [25,]  0.0294052920
#>   [26,] -0.0947085311
#>   [27,]  0.2851772339
#>   [28,] -0.0051224570
#>   [29,] -0.4919685433
#>   [30,] -0.0874609232
#>   [31,] -0.1682884313
#>   [32,]  0.2289977246
#>   [33,] -0.3772192220
#>   [34,] -0.4393554718
#>   [35,]  0.4075302455
#>   [36,]  0.0528801768
#>   [37,] -0.0242432523
#>   [38,] -0.0088881849
#>   [39,] -0.1495922334
#>   [40,] -0.0730820078
#>   [41,]  0.0707930774
#>   [42,] -0.0244825006
#>   [43,]  0.2394707544
#>   [44,]  0.3289927054
#>   [45,]  0.1144886304
#>   [46,]  0.5181707447
#>   [47,]  0.0896356523
#>   [48,]  0.2352031814
#>   [49,] -0.1983803659
#>   [50,]  0.2946291073
#>   [51,] -0.2397460595
#>   [52,]  0.3422657389
#>   [53,] -0.2381772861
#>   [54,]  0.2033116012
#>   [55,] -0.2300677549
#>   [56,] -0.4827541825
#>   [57,] -0.4378848230
#>   [58,] -0.0709108603
#>   [59,] -0.3823060905
#>   [60,]  0.3713380877
#>   [61,]  0.1087135162
#>   [62,] -0.0620274661
#>   [63,]  0.0271049469
#>   [64,] -0.3832746044
#>   [65,] -0.0222152945
#>   [66,]  0.3662865736
#>   [67,]  0.0692322112
#>   [68,]  1.1232860127
#>   [69,] -0.1009459511
#>   [70,]  0.1607562896
#>   [71,] -0.6249098020
#>   [72,] -0.0937169888
#>   [73,]  0.1097081288
#>   [74,] -0.4269507737
#>   [75,]  0.8080247517
#>   [76,]  0.2884019252
#>   [77,] -0.0475030346
#>   [78,] -0.0689719668
#>   [79,] -0.5504464518
#>   [80,]  0.1290017129
#>   [81,] -0.6122758292
#>   [82,]  0.3466752765
#>   [83,]  0.2800945747
#>   [84,] -0.3188255485
#>   [85,] -0.1561051110
#>   [86,]  0.1829894954
#>   [87,]  0.1852083654
#>   [88,]  0.5296341057
#>   [89,]  0.1040246728
#>   [90,] -0.3072681966
#>   [91,]  0.5386507518
#>   [92,] -0.2220046844
#>   [93,]  0.2480950764
#>   [94,] -0.3357081803
#>   [95,] -0.5785166723
#>   [96,] -0.4695519122
#>   [97,]  0.6116213299
#>   [98,] -0.1929490125
#>   [99,] -0.5272873545
#>  [100,]  0.5331197013
#>  [101,]  0.7360922730
#>  [102,] -0.1144690303
#>  [103,] -0.3140555953
#>  [104,] -0.3147212341
#>  [105,] -0.0834302181
#>  [106,]  0.1253871189
#>  [107,] -0.2613702379
#>  [108,]  0.3406113732
#>  [109,] -0.1893936969
#>  [110,]  0.4619607228
#>  [111,]  0.0554099538
#>  [112,] -0.1041431319
#>  [113,] -0.1837732283
#>  [114,]  0.2306608788
#>  [115,] -0.5559564107
#>  [116,]  0.2513702047
#>  [117,]  0.3620587704
#>  [118,] -0.4884142824
#>  [119,] -0.8753493053
#>  [120,] -0.0979406933
#>  [121,]  0.4574214533
#>  [122,]  0.0615222937
#>  [123,] -0.6604169398
#>  [124,]  0.1491773870
#>  [125,] -0.4124441479
#>  [126,] -0.0162907130
#>  [127,] -0.2946999896
#>  [128,] -0.2548855366
#>  [129,]  0.1636527418
#>  [130,]  0.2129284389
#>  [131,] -0.0046063440
#>  [132,]  0.3478307266
#>  [133,]  0.3622997298
#>  [134,]  0.0618772915
#>  [135,]  0.1904116759
#>  [136,]  0.0423616310
#>  [137,]  0.1580784740
#>  [138,]  0.1184997169
#>  [139,]  0.4233226527
#>  [140,]  0.0046128746
#>  [141,]  0.1398579222
#>  [142,] -0.5748309463
#>  [143,] -0.4782794034
#>  [144,] -0.5318798858
#>  [145,]  0.4663117036
#>  [146,]  0.2823847280
#>  [147,]  0.4430328090
#>  [148,]  0.0922219723
#>  [149,]  0.1937248365
#>  [150,] -0.3898188023
#>  [151,]  0.0900972659
#>  [152,] -0.0312402161
#>  [153,]  0.3677012162
#>  [154,]  0.1333331939
#>  [155,]  0.1828179332
#>  [156,] -0.2581006353
#>  [157,]  0.0481528914
#>  [158,]  0.5805537736
#>  [159,] -0.4012588599
#>  [160,]  0.0127451701
#>  [161,]  0.0541783696
#>  [162,]  0.0250511446
#>  [163,]  0.0360225896
#>  [164,]  0.5642285531
#>  [165,] -0.2218571383
#>  [166,]  0.0896519222
#>  [167,] -0.1195747436
#>  [168,]  0.6541343323
#>  [169,]  0.3847880775
#>  [170,] -0.0409223171
#>  [171,] -0.4319554361
#>  [172,] -0.0729777114
#>  [173,] -0.2641699735
#>  [174,]  0.1416173318
#>  [175,]  0.1587377997
#>  [176,]  0.0387577103
#>  [177,]  0.1340509763
#>  [178,] -0.4710414486
#>  [179,]  0.1897917705
#>  [180,]  0.0069972890
#>  [181,] -0.1908488511
#>  [182,]  0.0213767564
#>  [183,] -0.2383266934
#>  [184,] -0.1369210060
#>  [185,]  0.4244754048
#>  [186,] -0.0993262631
#>  [187,]  0.2658433543
#>  [188,]  0.0776092675
#>  [189,]  0.4551449134
#>  [190,]  0.5142705986
#>  [191,]  0.3629460955
#>  [192,]  0.4395322322
#>  [193,]  0.0961125276
#>  [194,]  0.1448485619
#>  [195,] -0.5747777436
#>  [196,]  0.0382977455
#>  [197,]  0.1688484073
#>  [198,] -0.3385206975
#>  [199,]  0.3746075751
#>  [200,]  0.3859384221
#>  [201,]  0.0425858639
#>  [202,] -0.0943794147
#>  [203,]  0.0603754180
#>  [204,]  0.0480546280
#>  [205,] -0.2001533466
#>  [206,]  0.2738802236
#>  [207,]  0.0835046237
#>  [208,] -0.2092004479
#>  [209,]  0.0782955815
#>  [210,]  0.5572564592
#>  [211,]  0.1183003043
#>  [212,] -0.2779499698
#>  [213,] -0.2109204821
#>  [214,]  0.4627575637
#>  [215,]  0.8609181746
#>  [216,] -0.2710966960
#>  [217,] -0.2184230769
#>  [218,]  0.6365336083
#>  [219,]  0.1986447440
#>  [220,] -0.0760110547
#>  [221,] -0.3005788793
#>  [222,]  0.2781998783
#>  [223,] -0.0368397793
#>  [224,] -0.0737137566
#>  [225,]  0.1807743979
#>  [226,] -0.0786305619
#>  [227,] -0.1343213458
#>  [228,]  0.0324828283
#>  [229,]  0.2190888270
#>  [230,] -0.3506397197
#>  [231,]  0.6042045406
#>  [232,] -0.2639122850
#>  [233,] -0.4447120830
#>  [234,] -0.0868911504
#>  [235,]  0.6862567438
#>  [236,]  0.3080154616
#>  [237,]  0.1844273749
#>  [238,] -0.0132818539
#>  [239,]  0.4523814595
#>  [240,] -0.2360716933
#>  [241,] -0.8410017897
#>  [242,] -0.2116943088
#>  [243,] -0.3872721194
#>  [244,] -0.1734927244
#>  [245,] -0.2352320343
#>  [246,]  0.1569660004
#>  [247,]  0.5446819334
#>  [248,]  0.4023934823
#>  [249,] -0.9246809927
#>  [250,] -0.5558538415
#>  [251,]  0.3287589924
#>  [252,] -0.3654865105
#>  [253,] -0.1763169047
#>  [254,] -0.4049533027
#>  [255,]  0.7081171084
#>  [256,]  0.0111063942
#>  [257,]  0.3394694989
#>  [258,] -0.5423888872
#>  [259,] -0.1647432156
#>  [260,] -0.2616365845
#>  [261,] -0.6002729929
#>  [262,]  0.0226421598
#>  [263,]  0.0593962803
#>  [264,]  0.2206155431
#>  [265,]  0.5604069688
#>  [266,] -0.4174288505
#>  [267,] -0.3880124116
#>  [268,]  0.1524967986
#>  [269,] -0.0807577795
#>  [270,]  0.1464280945
#>  [271,] -0.0120449927
#>  [272,] -0.6093237680
#>  [273,]  0.3525796506
#>  [274,] -0.0485854197
#>  [275,]  0.2271071934
#>  [276,] -0.3141715211
#>  [277,] -0.3057734301
#>  [278,] -0.2143686798
#>  [279,] -0.0361474758
#>  [280,] -0.0878718760
#>  [281,] -0.1849101711
#>  [282,] -0.4535910227
#>  [283,]  0.6939584000
#>  [284,]  0.0833171392
#>  [285,]  0.1224511181
#>  [286,]  0.4213024753
#>  [287,] -0.0666394874
#>  [288,]  0.3923593654
#>  [289,]  0.1086415368
#>  [290,] -0.4106417641
#>  [291,]  0.4734228487
#>  [292,]  0.0884659588
#>  [293,]  0.4547355195
#>  [294,]  0.7493063237
#>  [295,]  0.2730391874
#>  [296,] -0.1108986485
#>  [297,] -0.9479330793
#>  [298,]  0.0955911946
#>  [299,] -0.1812017612
#>  [300,] -0.5794246378
#>  [301,]  0.0027499678
#>  [302,] -0.3038638474
#>  [303,]  0.6575973439
#>  [304,]  0.1987480773
#>  [305,] -0.2531506762
#>  [306,] -0.8105595559
#>  [307,] -0.3415355547
#>  [308,]  0.3161707032
#>  [309,] -0.9379874834
#>  [310,]  0.2003508473
#>  [311,] -0.5175720071
#>  [312,] -0.3771902116
#>  [313,]  0.4072830226
#>  [314,]  0.0615926446
#>  [315,]  0.0630720262
#>  [316,] -0.7601235989
#>  [317,] -0.1620559273
#>  [318,] -0.2559585143
#>  [319,]  0.0713722434
#>  [320,]  0.1735304318
#>  [321,]  0.1296612111
#>  [322,] -0.4484766988
#>  [323,] -0.0237063205
#>  [324,]  0.4954932825
#>  [325,]  0.1344470870
#>  [326,] -0.2468405116
#>  [327,] -0.2260292616
#>  [328,] -0.3226365933
#>  [329,] -0.3081819372
#>  [330,] -0.1900753285
#>  [331,] -0.3799125754
#>  [332,]  0.1384650619
#>  [333,]  0.1915829916
#>  [334,]  0.4289845639
#>  [335,]  0.0491950453
#>  [336,]  0.2658167165
#>  [337,]  0.0635073365
#>  [338,]  0.2000257286
#>  [339,] -0.1540841651
#>  [340,]  0.4450770255
#>  [341,]  0.3991754538
#>  [342,]  0.3414347781
#>  [343,]  0.0018634572
#>  [344,] -0.4859689205
#>  [345,]  0.1039489378
#>  [346,]  0.2590177335
#>  [347,]  0.2619026497
#>  [348,]  0.0422554269
#>  [349,]  0.0552240628
#>  [350,]  0.1777700900
#>  [351,]  0.0270003048
#>  [352,] -0.4383651027
#>  [353,]  0.2652836870
#>  [354,]  0.1845852575
#>  [355,] -0.1244764604
#>  [356,]  0.2070422546
#>  [357,]  0.3803838286
#>  [358,] -0.0905758400
#>  [359,] -0.1608653108
#>  [360,]  0.0049708575
#>  [361,]  0.4580601899
#>  [362,] -0.0686122349
#>  [363,]  0.5710781136
#>  [364,]  0.0995399070
#>  [365,] -0.6072323641
#>  [366,]  0.7206951018
#>  [367,] -0.3495759743
#>  [368,]  0.9996833851
#>  [369,]  0.1811303230
#>  [370,] -0.0243998202
#>  [371,] -0.2924132334
#>  [372,] -0.5399238919
#>  [373,]  0.5575723936
#>  [374,] -0.0007124435
#>  [375,] -0.5529675585
#>  [376,]  0.1021587042
#>  [377,]  0.1493322448
#>  [378,] -0.3513944282
#>  [379,]  0.0206369015
#>  [380,]  0.1041201773
#>  [381,]  0.3658076331
#>  [382,]  0.0012375205
#>  [383,] -0.3756796457
#>  [384,]  0.2478254586
#>  [385,]  0.4986705512
#>  [386,]  0.3962671426
#>  [387,]  0.0529439660
#>  [388,] -0.4344315756
#>  [389,]  0.0313025266
#>  [390,] -0.2721494520
#>  [391,] -0.3583118865
#>  [392,] -0.6321597440
#>  [393,] -0.2993331957
#>  [394,] -0.2381657764
#>  [395,] -0.6926623087
#>  [396,] -0.1120481072
#>  [397,]  0.0223813795
#>  [398,]  0.2044852359
#>  [399,] -0.2041414158
#>  [400,]  0.5213473140
#>  [401,] -0.2644524161
#>  [402,] -0.2113434209
#>  [403,]  0.0859517178
#>  [404,] -0.0527985358
#>  [405,] -0.0215394198
#>  [406,] -0.0842642202
#>  [407,] -0.0573407150
#>  [408,] -0.2347653396
#>  [409,]  0.0318604924
#>  [410,] -0.2203259719
#>  [411,]  0.7879011029
#>  [412,]  0.0022759893
#>  [413,] -0.2839967876
#>  [414,] -0.2748047453
#>  [415,] -0.4406075485
#>  [416,] -0.8575800246
#>  [417,] -0.4408074975
#>  [418,]  0.4083194501
#>  [419,]  0.2713721219
#>  [420,]  0.4566515222
#>  [421,]  0.7165404433
#>  [422,]  0.5392538169
#>  [423,]  0.1393844202
#>  [424,]  0.4119182530
#>  [425,]  0.5007292430
#>  [426,]  0.3418196672
#>  [427,]  0.2046511989
#>  [428,] -0.1982708729
#>  [429,] -0.3617967733
#>  [430,] -0.0691845008
#>  [431,]  0.1011180993
#>  [432,]  0.1855763956
#>  [433,]  0.5335397757
#>  [434,] -0.4373886666
#>  [435,]  0.0330055104
#>  [436,] -0.0754783869
#>  [437,]  0.2656433258
#>  [438,]  0.1206709068
#>  [439,] -0.2027244193
#>  [440,] -0.0789518207
#>  [441,]  0.3159080215
#>  [442,] -0.8644653599
#>  [443,] -0.6628878068
#>  [444,]  0.3918609818
#>  [445,]  0.2116618895
#>  [446,] -0.3476800248
#>  [447,] -0.5811989190
#>  [448,] -0.2706334100
#>  [449,] -0.0172161523
#>  [450,]  0.5654268050
#>  [451,] -0.4775641780
#>  [452,] -0.7072760802
#>  [453,] -0.0530490856
#>  [454,]  0.1493340646
#>  [455,] -0.0251661129
#>  [456,] -0.0949812771
#>  [457,] -0.0710361048
#>  [458,]  0.1410860036
#>  [459,]  0.5658574178
#>  [460,]  0.4891438970
#>  [461,] -0.4937866115
#>  [462,]  0.3194016816
#>  [463,] -0.2669349751
#>  [464,] -0.3576184910
#>  [465,] -0.5405723215
#>  [466,]  0.4897407770
#>  [467,] -0.0475316900
#>  [468,]  0.0986384892
#>  [469,]  0.9586874474
#>  [470,] -0.5884326757
#>  [471,] -0.6279924255
#>  [472,]  0.2231455563
#>  [473,] -0.3262546239
#>  [474,] -0.0644010291
#>  [475,]  0.0663842571
#>  [476,]  0.2556924919
#>  [477,]  0.5833149174
#>  [478,]  0.2865180181
#>  [479,]  0.4426338783
#>  [480,] -0.0397855209
#>  [481,]  0.2172704064
#>  [482,] -0.4157866349
#>  [483,] -0.1527727934
#>  [484,] -0.4342039751
#>  [485,]  0.1525654357
#>  [486,]  0.1744813076
#>  [487,]  0.5924106727
#>  [488,]  0.0865424144
#>  [489,] -0.2380075038
#>  [490,]  0.1866326427
#>  [491,] -0.0965436817
#>  [492,]  0.0763023759
#>  [493,]  0.0355144749
#>  [494,]  0.0953295528
#>  [495,] -0.2170000302
#>  [496,] -0.5153255111
#>  [497,]  0.3178234615
#>  [498,] -0.0563932712
#>  [499,] -0.6321672753
#>  [500,]  0.3801004920
#>  [501,] -0.1644323560
#>  [502,] -0.1969349629
#>  [503,] -0.4464565013
#>  [504,]  0.4502703394
#>  [505,]  0.3557858756
#>  [506,]  0.3996610431
#>  [507,] -0.1312133553
#>  [508,]  0.0869539178
#>  [509,] -0.4859592558
#>  [510,]  0.0127049743
#>  [511,] -0.0391544763
#>  [512,] -0.3562682279
#>  [513,] -0.3977052628
#>  [514,] -0.0897218491
#>  [515,] -0.5483309663
#>  [516,] -0.0305356892
#>  [517,] -0.4247426452
#>  [518,] -0.2294222155
#>  [519,]  0.3165209661
#>  [520,]  0.4954903067
#>  [521,]  0.4140005121
#>  [522,] -0.4441484269
#>  [523,] -0.3314329804
#>  [524,]  0.5628589679
#>  [525,] -0.3205007483
#>  [526,]  0.5926507509
#>  [527,]  0.1355328933
#>  [528,] -0.2386397197
#>  [529,] -0.3522854317
#>  [530,] -0.3533261413
#>  [531,]  0.1889353135
#>  [532,]  0.3761590694
#>  [533,]  0.4672458580
#>  [534,]  0.3973814746
#>  [535,]  0.0687716750
#>  [536,] -0.1083247191
#>  [537,] -0.0933516243
#>  [538,] -0.0193925388
#>  [539,] -0.5734894191
#>  [540,] -0.6466967470
#>  [541,] -0.0196984690
#>  [542,]  0.2906306974
#>  [543,] -0.3731609131
#>  [544,]  0.2850785438
#>  [545,] -0.1922733245
#>  [546,]  0.0689415956
#>  [547,]  0.1635225249
#>  [548,]  0.2274616984
#>  [549,]  0.4971548392
#>  [550,]  0.0243326549
#>  [551,]  0.0617538273
#>  [552,]  0.4291923733
#>  [553,] -0.5384344813
#>  [554,]  0.0913655829
#>  [555,] -0.1845839070
#>  [556,] -0.1615256006
#>  [557,]  0.3057319816
#>  [558,] -0.0597370635
#>  [559,]  0.0102622856
#>  [560,] -0.0165925444
#>  [561,] -0.0744130243
#>  [562,] -0.1081030315
#>  [563,]  0.1289698584
#>  [564,]  0.2206441636
#>  [565,]  0.0043525289
#>  [566,] -0.0640802371
#>  [567,]  0.4296198825
#>  [568,]  0.2670359666
#>  [569,] -0.1286537631
#>  [570,] -0.3516768887
#>  [571,] -0.2840406743
#>  [572,]  0.1036883440
#>  [573,] -0.3433850096
#>  [574,] -0.1719692264
#>  [575,]  0.6350490915
#>  [576,] -0.6696113770
#>  [577,] -0.4194542774
#>  [578,]  0.0681754385
#>  [579,] -0.1662435976
#>  [580,] -0.1074813253
#>  [581,] -0.1031453941
#>  [582,] -0.0015722751
#>  [583,] -0.4624892381
#>  [584,]  0.6504833241
#>  [585,]  0.2080463497
#>  [586,]  0.2923037281
#>  [587,] -0.2383566326
#>  [588,] -0.0141739046
#>  [589,]  0.1479711529
#>  [590,]  0.4925948205
#>  [591,]  0.3409329442
#>  [592,] -0.1194838429
#>  [593,]  0.6290903568
#>  [594,] -0.0293046015
#>  [595,] -0.7703919162
#>  [596,] -0.2205265664
#>  [597,] -0.1886567239
#>  [598,] -0.0343684550
#>  [599,] -0.0512563038
#>  [600,]  0.1248159374
#>  [601,] -0.3092330549
#>  [602,] -0.3420661040
#>  [603,]  0.6708577868
#>  [604,]  0.2978608705
#>  [605,]  0.4021333255
#>  [606,]  0.2139951309
#>  [607,] -0.2705932858
#>  [608,] -0.0487649505
#>  [609,]  0.2081333262
#>  [610,]  0.3926603677
#>  [611,]  0.1251802335
#>  [612,] -0.0246212964
#>  [613,]  0.0721899139
#>  [614,] -0.2305361204
#>  [615,] -0.3605000527
#>  [616,]  0.3394538250
#>  [617,]  0.2516546700
#>  [618,]  0.0314847976
#>  [619,] -0.5203048659
#>  [620,] -0.5963212420
#>  [621,] -0.0404410860
#>  [622,] -0.1516187153
#>  [623,] -0.4426282660
#>  [624,]  0.0629959332
#>  [625,] -0.5151417404
#>  [626,]  0.3530017411
#>  [627,] -1.1467583682
#>  [628,]  0.0560486998
#>  [629,]  0.1697149690
#>  [630,] -0.0882178856
#>  [631,] -0.3261346885
#>  [632,] -0.2567548291
#>  [633,] -0.2513060581
#>  [634,] -0.6407496755
#>  [635,] -0.0671952075
#>  [636,]  0.3938539921
#>  [637,] -0.4881267029
#>  [638,] -0.1050855277
#>  [639,]  0.0130902014
#>  [640,]  0.1618365179
#>  [641,] -0.5095607424
#>  [642,]  0.5393575274
#>  [643,]  0.4837825523
#>  [644,]  0.2944133859
#>  [645,] -0.0555773951
#>  [646,]  0.4109381401
#>  [647,] -0.2316880426
#>  [648,] -0.1531729351
#>  [649,] -0.3944212364
#>  [650,]  0.1090134898
#>  [651,] -0.1059132321
#>  [652,]  0.0699814923
#>  [653,] -0.3631799728
#>  [654,] -0.3234323603
#>  [655,] -0.2084036493
#>  [656,]  0.1620082543
#>  [657,] -0.5579262056
#>  [658,] -0.0895773396
#>  [659,] -0.7293245571
#>  [660,]  0.3715510518
#>  [661,]  0.2782656648
#>  [662,] -0.0622552288
#>  [663,]  0.2984121768
#>  [664,]  0.1510140255
#>  [665,]  0.4262294760
#>  [666,]  0.0596985323
#>  [667,] -0.0059207065
#>  [668,]  0.2701489705
#>  [669,]  0.4218773757
#>  [670,]  0.1806377326
#>  [671,]  0.1828456228
#>  [672,] -0.0055322840
#>  [673,]  0.6350268830
#>  [674,] -0.0699025838
#>  [675,]  0.2436022482
#>  [676,] -0.0209803970
#>  [677,]  0.5513504086
#>  [678,] -0.4871976820
#>  [679,]  0.1988452687
#>  [680,] -0.1301970835
#>  [681,] -0.3792227548
#>  [682,]  0.1479503849
#>  [683,] -0.4214806303
#>  [684,]  0.4153826273
#>  [685,]  0.0785879331
#>  [686,]  0.0305566279
#>  [687,]  0.6420570415
#>  [688,] -0.3368148403
#>  [689,]  0.0074400968
#>  [690,] -0.1267684610
#>  [691,] -0.4034741385
#>  [692,] -0.1693226846
#>  [693,]  0.8380204002
#>  [694,] -0.1770109219
#>  [695,]  0.2510958377
#>  [696,]  0.1417574904
#>  [697,] -0.6012558534
#>  [698,] -0.4228307446
#>  [699,] -0.4976754451
#>  [700,] -0.5555965379
#>  [701,]  0.5060935644
#>  [702,] -0.4816901908
#>  [703,]  0.1261533764
#>  [704,]  0.2521607349
#>  [705,]  0.3916242136
#>  [706,]  0.5319918125
#>  [707,]  0.2244272022
#>  [708,]  0.2496769507
#>  [709,]  0.0791748071
#>  [710,] -0.1220230328
#>  [711,] -0.0192520850
#>  [712,] -0.3043339269
#>  [713,] -0.5058543731
#>  [714,]  0.5494115414
#>  [715,]  0.3294385922
#>  [716,]  0.4277470561
#>  [717,] -0.5689157186
#>  [718,] -0.0757544668
#>  [719,]  0.6037518814
#>  [720,]  0.0444897780
#>  [721,] -0.8300835901
#>  [722,] -0.4169407718
#>  [723,] -0.2476528216
#>  [724,] -0.1650809300
#>  [725,]  0.2219621992
#>  [726,]  0.3871119742
#>  [727,] -0.2401178812
#>  [728,]  0.0857554385
#>  [729,]  0.0290729074
#>  [730,] -0.1624557505
#>  [731,] -0.2303069619
#>  [732,] -0.4720540129
#>  [733,]  0.0221887295
#>  [734,]  0.6509882961
#>  [735,] -0.3713998141
#>  [736,]  0.0221792424
#>  [737,]  0.1487368905
#>  [738,]  0.1378165261
#>  [739,] -0.0409444590
#>  [740,] -0.2158311948
#>  [741,] -0.0173237453
#>  [742,] -0.2093368233
#>  [743,]  0.0916393373
#>  [744,]  0.4483235538
#>  [745,] -0.4823560413
#>  [746,] -0.4322351635
#>  [747,] -0.5346411304
#>  [748,] -0.0744454436
#>  [749,] -0.1244838304
#>  [750,] -0.2759559624
#>  [751,]  0.1798420959
#>  [752,]  0.0039565883
#>  [753,]  0.5550062873
#>  [754,]  0.4851459034
#>  [755,]  0.0643313085
#>  [756,]  0.0126495314
#>  [757,] -0.2548692046
#>  [758,]  0.1501797538
#>  [759,]  0.4194883323
#>  [760,]  0.2991568557
#>  [761,] -0.5472579042
#>  [762,] -0.2413597456
#>  [763,] -0.4673597612
#>  [764,] -0.0903004388
#>  [765,]  0.2163561191
#>  [766,] -0.0262373184
#>  [767,] -0.3817827848
#>  [768,] -0.0555084399
#>  [769,] -0.6147075991
#>  [770,]  0.3324088569
#>  [771,] -0.2385775541
#>  [772,]  0.2809861894
#>  [773,] -0.0754991746
#>  [774,] -0.3936344371
#>  [775,]  0.5360037868
#>  [776,]  0.2753143502
#>  [777,]  0.4762576074
#>  [778,] -0.3678688082
#>  [779,]  0.1524963711
#>  [780,] -0.7646332528
#>  [781,]  0.3451274757
#>  [782,]  0.1510672232
#>  [783,]  0.2762446565
#>  [784,]  0.2124645090
#>  [785,] -0.7553303210
#>  [786,]  0.6623675802
#>  [787,]  0.0915099573
#>  [788,]  0.0561408872
#>  [789,]  0.3908472667
#>  [790,] -0.7119934129
#>  [791,]  0.3519983787
#>  [792,] -0.5742776020
#>  [793,]  0.0113968586
#>  [794,]  0.1297896002
#>  [795,]  0.1260103058
#>  [796,] -0.4183267018
#>  [797,] -0.1249589768
#>  [798,] -0.0321596452
#>  [799,]  0.3020921357
#>  [800,] -0.0723542981
#>  [801,]  0.2363155628
#>  [802,] -0.3428174544
#>  [803,] -0.0181887442
#>  [804,] -0.0550165744
#>  [805,] -0.1285064182
#>  [806,] -0.1709458213
#>  [807,] -0.3164418167
#>  [808,] -0.0241863211
#>  [809,]  0.2070488147
#>  [810,] -0.0376321782
#>  [811,] -0.2753794426
#>  [812,] -0.7926904701
#>  [813,] -0.1439208882
#>  [814,]  0.5446603163
#>  [815,]  0.2740705644
#>  [816,]  0.0531017693
#>  [817,] -0.2911336905
#>  [818,]  0.0812750318
#>  [819,]  0.2182294259
#>  [820,] -0.0722091812
#>  [821,]  0.4369247137
#>  [822,]  0.5330458190
#>  [823,]  0.2342892403
#>  [824,]  0.0449455495
#>  [825,] -0.0793197630
#>  [826,]  0.4336114427
#>  [827,]  0.2620935782
#>  [828,] -0.3852840334
#>  [829,]  0.7108656350
#>  [830,]  0.0975113879
#>  [831,] -0.4212496267
#>  [832,]  0.3177679395
#>  [833,] -0.2908777951
#>  [834,]  0.3217279277
#>  [835,]  0.7221315500
#>  [836,]  0.5311227573
#>  [837,] -0.3159935647
#>  [838,]  0.3761050972
#>  [839,]  0.1944542561
#>  [840,]  0.0465616888
#>  [841,]  0.1512112809
#>  [842,]  0.4143104816
#>  [843,] -0.1579776288
#>  [844,] -0.0108949582
#>  [845,] -0.0327806865
#>  [846,] -0.1403897515
#>  [847,] -0.2263470678
#>  [848,] -0.9402802040
#>  [849,]  0.4070439913
#>  [850,]  0.1700888212
#>  [851,] -0.0999220249
#>  [852,]  0.2163044893
#>  [853,] -0.3794785078
#>  [854,]  0.3433613022
#>  [855,] -0.0608963120
#>  [856,]  0.5960226042
#>  [857,] -0.1725288354
#>  [858,]  0.3936073083
#>  [859,] -0.5109372224
#>  [860,]  0.3712730933
#>  [861,] -0.0452332677
#>  [862,]  0.5452347023
#>  [863,] -0.2791888522
#>  [864,]  0.3637556006
#>  [865,] -0.0859770327
#>  [866,] -0.0286928465
#>  [867,]  0.1829157637
#>  [868,] -0.4120835922
#>  [869,]  0.5203170394
#>  [870,] -0.2483559097
#>  [871,] -0.1886684163
#>  [872,] -0.1963299390
#>  [873,] -0.2274843033
#>  [874,]  0.2654271309
#>  [875,] -0.9671858351
#>  [876,]  0.2670604401
#>  [877,] -0.7033921219
#>  [878,]  0.0887371700
#>  [879,]  0.4241993142
#>  [880,] -0.0675198154
#>  [881,]  0.4976562152
#>  [882,]  0.7310256124
#>  [883,]  0.0826778481
#>  [884,]  0.1702171525
#>  [885,]  0.4567056327
#>  [886,] -0.1016224895
#>  [887,] -0.4864885060
#>  [888,]  0.4097898911
#>  [889,] -0.3320209095
#>  [890,]  0.2601581835
#>  [891,] -0.5200307014
#>  [892,]  0.0147046151
#>  [893,]  0.2353380349
#>  [894,] -0.1815600177
#>  [895,]  0.2542714923
#>  [896,]  0.1805389154
#>  [897,]  0.5760205156
#>  [898,]  0.4271148040
#>  [899,] -0.1470003419
#>  [900,] -0.0803736782
#>  [901,]  0.4220778572
#>  [902,] -0.0625351121
#>  [903,] -0.2112876902
#>  [904,] -0.3681294521
#>  [905,]  0.3027929519
#>  [906,]  0.3266191678
#>  [907,]  0.1275815002
#>  [908,]  0.0759227249
#>  [909,] -0.5688287140
#>  [910,]  0.7820587584
#>  [911,] -0.1673126338
#>  [912,]  0.1896248504
#>  [913,] -0.2567353560
#>  [914,]  0.3338806461
#>  [915,]  0.6228158491
#>  [916,] -0.2839102318
#>  [917,]  0.6397046541
#>  [918,]  0.0033884559
#>  [919,]  0.1834472627
#>  [920,] -0.7451649333
#>  [921,] -0.6043506727
#>  [922,] -0.2900488955
#>  [923,] -0.0744725339
#>  [924,] -0.0233093165
#>  [925,]  0.5104768907
#>  [926,]  0.1773872244
#>  [927,] -0.5029153088
#>  [928,]  0.0097134990
#>  [929,] -0.5979627024
#>  [930,] -0.1867461611
#>  [931,] -0.1997171811
#>  [932,]  0.2558872501
#>  [933,]  0.1681623435
#>  [934,] -0.1194956417
#>  [935,] -0.2809503943
#>  [936,] -0.3249071393
#>  [937,]  0.0276128294
#>  [938,]  0.2695596619
#>  [939,] -0.1363550103
#>  [940,]  0.3507158849
#>  [941,] -0.2372785612
#>  [942,] -0.2533311366
#>  [943,] -0.5199568790
#>  [944,] -0.4647813909
#>  [945,]  0.1356533581
#>  [946,] -0.4685073546
#>  [947,]  0.2921924439
#>  [948,] -0.5846393336
#>  [949,] -0.3490468047
#>  [950,] -0.0646555206
#>  [951,]  0.1511241323
#>  [952,]  0.2068271260
#>  [953,] -0.2217130997
#>  [954,] -0.5473370777
#>  [955,]  0.0919044228
#>  [956,] -0.4445274270
#>  [957,]  0.4631066685
#>  [958,]  0.2269931988
#>  [959,] -0.4369451911
#>  [960,] -0.6280680494
#>  [961,] -0.1642516812
#>  [962,] -0.0300862529
#>  [963,] -0.3249277283
#>  [964,] -0.0531678977
#>  [965,] -0.0298392588
#>  [966,] -0.1129208981
#>  [967,]  0.5482080978
#>  [968,]  0.1884419198
#>  [969,] -0.1865370262
#>  [970,] -0.5409784322
#>  [971,]  0.3893865652
#>  [972,] -0.2793440267
#>  [973,]  0.6515414799
#>  [974,]  0.4496001625
#>  [975,] -0.0736751364
#>  [976,]  0.3029993938
#>  [977,]  0.1990392911
#>  [978,]  0.2366455477
#>  [979,] -0.3223383589
#>  [980,]  0.2206396560
#>  [981,]  0.6431972580
#>  [982,] -0.0785946885
#>  [983,]  0.0234007895
#>  [984,] -0.4570646874
#>  [985,]  0.3133291074
#>  [986,]  0.0179074732
#>  [987,] -0.4607402837
#>  [988,]  0.0169748430
#>  [989,] -0.3013507937
#>  [990,]  0.0530491749
#>  [991,]  0.1174956629
#>  [992,] -0.3174131386
#>  [993,] -0.0021765329
#>  [994,]  0.0642416225
#>  [995,] -0.5312288341
#>  [996,]  0.1488616977
#>  [997,]  0.5868275724
#>  [998,] -0.1221531403
#>  [999,] -0.7002145053
#> [1000,] -0.2208789744