Logistic Function

The elementwise logistic function, \(\log(1 + e^x)\). This is a special case of log(sum(exp)) that evaluates to a vector rather than to a scalar, which is useful for logistic regression.

logistic(x)

Arguments

x

An Expression, vector, or matrix.

Value

An Expression representing the logistic function evaluated at the input.

Examples

set.seed(92)
n <- 20
m <- 1000
sigma <- 45

beta_true <- stats::rnorm(n)
idxs <- sample(n, size = 0.8*n, replace = FALSE)
beta_true[idxs] <- 0
X <- matrix(stats::rnorm(m*n, 0, 5), nrow = m, ncol = n)
y <- sign(X %*% beta_true + stats::rnorm(m, 0, sigma))

beta <- Variable(n)
X_sign <- apply(X, 2, function(x) { ifelse(y <= 0, -1, 1) * x })
obj <- -sum(logistic(-X[y <= 0,] %*% beta)) - sum(logistic(X[y == 1,] %*% beta))
prob <- Problem(Maximize(obj))
result <- solve(prob)

log_odds <- result$getValue(X %*% beta)
beta_res <- result$getValue(beta)
y_probs <- 1/(1 + exp(-X %*% beta_res))
log(y_probs/(1 - y_probs))
#>                  [,1]
#>    [1,] -0.6704747453
#>    [2,] -0.5925328212
#>    [3,] -0.2870377968
#>    [4,] -0.3903535369
#>    [5,]  0.1512570887
#>    [6,] -0.3212394186
#>    [7,]  0.0710834791
#>    [8,] -0.0005976271
#>    [9,]  0.6231703117
#>   [10,] -0.3461630253
#>   [11,]  0.6166059988
#>   [12,]  0.2493867405
#>   [13,] -0.4229590165
#>   [14,]  1.3550948223
#>   [15,] -0.3265674802
#>   [16,] -0.5974214291
#>   [17,]  1.4251319269
#>   [18,] -0.8064347421
#>   [19,]  0.6354415476
#>   [20,] -0.5313438902
#>   [21,] -0.5836979792
#>   [22,] -0.2018653764
#>   [23,] -1.5741498252
#>   [24,] -0.2996895657
#>   [25,] -0.2075379701
#>   [26,]  0.4437075976
#>   [27,] -0.5653584061
#>   [28,] -1.2550929261
#>   [29,] -0.5346377291
#>   [30,]  0.4834186388
#>   [31,]  0.2311073293
#>   [32,]  0.2894531798
#>   [33,] -0.0125271013
#>   [34,]  0.7295945149
#>   [35,] -0.8747328120
#>   [36,] -0.3929007100
#>   [37,] -0.1155697899
#>   [38,]  1.1672745037
#>   [39,] -0.7217610157
#>   [40,]  0.0611259460
#>   [41,] -0.6167812678
#>   [42,] -0.0245112095
#>   [43,]  0.6908383315
#>   [44,] -0.8034683028
#>   [45,]  0.8837321215
#>   [46,] -1.0940881168
#>   [47,]  0.0151087256
#>   [48,] -0.0365100536
#>   [49,]  0.1208122471
#>   [50,]  0.0228343539
#>   [51,]  0.0203843954
#>   [52,] -0.4370977832
#>   [53,]  0.0381296246
#>   [54,] -0.0670272720
#>   [55,] -0.5450630415
#>   [56,] -0.1197518068
#>   [57,] -0.1350460388
#>   [58,]  0.1775514401
#>   [59,]  1.1398260247
#>   [60,] -0.4142674984
#>   [61,]  0.3993092330
#>   [62,] -0.4202249437
#>   [63,]  0.0406900070
#>   [64,] -0.6489266545
#>   [65,] -0.9703157831
#>   [66,]  0.1117470893
#>   [67,] -0.1453609166
#>   [68,]  0.0616521968
#>   [69,]  0.0779098757
#>   [70,]  0.4237529659
#>   [71,] -0.4578431809
#>   [72,] -0.2049760543
#>   [73,] -0.2303945316
#>   [74,]  0.1166998759
#>   [75,] -0.3652053517
#>   [76,] -0.2369567192
#>   [77,]  0.1457272538
#>   [78,] -1.0896686258
#>   [79,] -0.1308746810
#>   [80,]  0.2291736953
#>   [81,]  0.1373271433
#>   [82,] -0.0015643759
#>   [83,]  0.3467661311
#>   [84,] -0.1543936115
#>   [85,] -0.2309079599
#>   [86,] -0.5247239950
#>   [87,]  0.4603874797
#>   [88,] -0.3126546162
#>   [89,]  0.2120101364
#>   [90,]  0.3147059544
#>   [91,]  0.4193116299
#>   [92,]  0.6556946429
#>   [93,] -0.5111240660
#>   [94,]  1.3013586628
#>   [95,]  0.8030343104
#>   [96,]  0.2830436450
#>   [97,] -0.6704129231
#>   [98,]  0.6392222750
#>   [99,]  0.0848846684
#>  [100,] -0.0064339648
#>  [101,]  0.1647498071
#>  [102,]  0.2973181065
#>  [103,]  0.3564271017
#>  [104,] -0.8130651251
#>  [105,] -0.5596709529
#>  [106,]  0.2440681383
#>  [107,] -0.5330643786
#>  [108,]  0.4528529740
#>  [109,] -0.5709081712
#>  [110,] -0.4624088876
#>  [111,] -0.2336950084
#>  [112,]  0.1511247448
#>  [113,] -0.0619001744
#>  [114,]  0.5839107079
#>  [115,] -0.0005887716
#>  [116,] -0.7089905383
#>  [117,] -0.3172780305
#>  [118,]  0.3970662275
#>  [119,] -1.5820303785
#>  [120,]  0.5686395206
#>  [121,] -0.3273532649
#>  [122,] -0.8146151770
#>  [123,] -0.0948010100
#>  [124,]  0.3203504916
#>  [125,] -0.4429258729
#>  [126,]  0.1170359830
#>  [127,] -0.7652443479
#>  [128,] -0.0573330291
#>  [129,]  0.0095923205
#>  [130,] -0.1932187132
#>  [131,]  0.1964295918
#>  [132,] -0.3175673688
#>  [133,]  0.4246307440
#>  [134,]  0.1266911131
#>  [135,]  0.3305676080
#>  [136,] -0.0061833793
#>  [137,] -0.5537778127
#>  [138,] -0.1133729598
#>  [139,] -0.7994784829
#>  [140,]  0.8081791592
#>  [141,]  0.4153194889
#>  [142,]  0.8942464415
#>  [143,] -0.3139527223
#>  [144,]  0.1458417356
#>  [145,] -0.2050302408
#>  [146,] -0.0601911319
#>  [147,] -1.1748369016
#>  [148,]  0.5454838585
#>  [149,]  0.2075718762
#>  [150,] -0.2618829301
#>  [151,]  0.8039082164
#>  [152,] -0.0455086160
#>  [153,]  0.5741459944
#>  [154,]  0.6920029468
#>  [155,]  0.0716658788
#>  [156,] -0.4965852192
#>  [157,]  0.4424012961
#>  [158,] -0.3816940147
#>  [159,]  0.5206130555
#>  [160,] -0.1021958137
#>  [161,] -0.2226467066
#>  [162,]  0.7099016303
#>  [163,]  0.5765382476
#>  [164,]  0.2450351330
#>  [165,] -0.1737110973
#>  [166,]  0.3589718972
#>  [167,] -0.3290109210
#>  [168,] -0.8993571740
#>  [169,] -0.8218169271
#>  [170,] -0.7797947156
#>  [171,] -0.2137223959
#>  [172,]  0.5276414892
#>  [173,] -0.8488026286
#>  [174,]  0.1034571399
#>  [175,] -0.4103722955
#>  [176,] -0.7528964425
#>  [177,] -0.4402503001
#>  [178,] -1.0685897810
#>  [179,]  0.4078631608
#>  [180,]  0.6313743728
#>  [181,]  0.3750977960
#>  [182,] -0.0151761111
#>  [183,] -0.5310943750
#>  [184,]  0.4077601021
#>  [185,]  0.1571578122
#>  [186,] -0.6328755783
#>  [187,] -0.3822625490
#>  [188,]  0.5205264608
#>  [189,]  0.2562736494
#>  [190,]  0.8532803165
#>  [191,] -0.5649272190
#>  [192,]  0.0370599525
#>  [193,] -0.3390879441
#>  [194,] -0.0375167676
#>  [195,]  0.0474717908
#>  [196,] -0.0326523828
#>  [197,]  0.2160030272
#>  [198,] -1.0662726662
#>  [199,] -0.9550123054
#>  [200,] -0.6453962911
#>  [201,]  0.5183437663
#>  [202,] -0.6995172274
#>  [203,]  0.2181032498
#>  [204,]  0.1893102213
#>  [205,]  0.4929995215
#>  [206,]  0.1844709863
#>  [207,]  0.3573983187
#>  [208,] -0.2500617709
#>  [209,]  1.3917463231
#>  [210,] -0.7617971564
#>  [211,]  0.5279346337
#>  [212,] -1.0117924511
#>  [213,] -0.2423777791
#>  [214,] -0.6134068322
#>  [215,] -1.1656471094
#>  [216,]  0.3821435128
#>  [217,] -0.4501285466
#>  [218,] -1.0973564770
#>  [219,]  0.8937549140
#>  [220,]  0.3140736705
#>  [221,]  0.1340535416
#>  [222,] -1.2231142137
#>  [223,]  0.5678335252
#>  [224,]  0.8463775306
#>  [225,]  0.6349311266
#>  [226,]  0.7006064459
#>  [227,] -0.7992302457
#>  [228,]  0.5683533931
#>  [229,]  0.4113089826
#>  [230,]  0.0902430997
#>  [231,] -0.3665262918
#>  [232,]  0.0127791973
#>  [233,]  0.1629194027
#>  [234,] -0.3527239106
#>  [235,] -0.0758217088
#>  [236,]  0.0091714254
#>  [237,] -0.4611265405
#>  [238,]  0.0657049711
#>  [239,] -0.5298668867
#>  [240,] -0.5962323527
#>  [241,]  0.0637695497
#>  [242,] -0.3510196810
#>  [243,]  0.5476184607
#>  [244,]  0.5265704966
#>  [245,] -0.8598340525
#>  [246,]  0.6395416471
#>  [247,]  0.3216020007
#>  [248,] -0.4796016137
#>  [249,]  0.3975296388
#>  [250,]  0.1215660180
#>  [251,] -0.3140333609
#>  [252,] -0.3046537339
#>  [253,]  0.3923267337
#>  [254,] -0.1856474517
#>  [255,] -0.0851623037
#>  [256,] -0.0029482542
#>  [257,]  0.5734883239
#>  [258,] -0.0359994195
#>  [259,] -0.9568166616
#>  [260,]  0.6395786993
#>  [261,]  0.4795818681
#>  [262,]  0.1872640071
#>  [263,] -0.2498943726
#>  [264,]  0.3261025397
#>  [265,]  0.1986625843
#>  [266,] -0.6790775642
#>  [267,]  0.6809690638
#>  [268,] -0.2952619239
#>  [269,]  0.5265890631
#>  [270,]  0.6813044039
#>  [271,] -0.4787301999
#>  [272,]  0.6944828574
#>  [273,]  0.3116230615
#>  [274,]  0.6770526435
#>  [275,]  0.0024198299
#>  [276,]  0.1779808418
#>  [277,]  0.1332661721
#>  [278,]  0.2881665987
#>  [279,] -0.8398431590
#>  [280,]  0.8293255980
#>  [281,]  0.3237992271
#>  [282,]  0.3688801756
#>  [283,] -0.2838471205
#>  [284,]  0.1698122005
#>  [285,] -0.0972806304
#>  [286,]  0.1692343277
#>  [287,]  0.6139450545
#>  [288,]  0.1736876288
#>  [289,] -0.0806372505
#>  [290,]  0.0115350238
#>  [291,] -0.3183946567
#>  [292,]  0.0866726899
#>  [293,] -0.7586498460
#>  [294,]  0.0584527216
#>  [295,] -0.1537405419
#>  [296,]  0.8435502793
#>  [297,]  0.8815631542
#>  [298,]  0.1534506381
#>  [299,]  1.2271539246
#>  [300,]  0.3099006614
#>  [301,] -0.2555325748
#>  [302,] -0.2073332472
#>  [303,] -0.2298608509
#>  [304,] -0.5772267215
#>  [305,] -0.4885499609
#>  [306,]  0.5318392173
#>  [307,] -0.1440944330
#>  [308,]  0.6376019511
#>  [309,]  0.2563626362
#>  [310,] -0.0603867354
#>  [311,] -0.2223266568
#>  [312,] -0.4951219409
#>  [313,] -0.2533212765
#>  [314,]  0.3508912176
#>  [315,]  0.0868161343
#>  [316,] -0.3995864739
#>  [317,] -0.2455304201
#>  [318,] -0.0988050511
#>  [319,]  0.6998411767
#>  [320,] -0.4848307427
#>  [321,] -0.3735448307
#>  [322,] -0.0385930033
#>  [323,]  0.0463465904
#>  [324,] -0.8583872350
#>  [325,] -0.4583302625
#>  [326,]  0.6281562759
#>  [327,] -0.6351108033
#>  [328,] -0.9582076940
#>  [329,]  0.9368683748
#>  [330,]  0.1483897578
#>  [331,] -0.0538293612
#>  [332,] -1.2323328505
#>  [333,] -0.1706923734
#>  [334,]  0.1205541752
#>  [335,] -0.4870973620
#>  [336,]  0.0540019005
#>  [337,] -0.1344871681
#>  [338,]  1.1543881301
#>  [339,]  0.8619170996
#>  [340,]  0.0228940601
#>  [341,] -0.5239151219
#>  [342,]  0.3351314933
#>  [343,] -0.7517339551
#>  [344,]  0.6050444076
#>  [345,] -0.0211553558
#>  [346,]  0.4027287306
#>  [347,]  0.8253067806
#>  [348,] -0.3307174435
#>  [349,]  0.3177337512
#>  [350,] -0.0224534039
#>  [351,] -0.5546894137
#>  [352,] -0.2831652441
#>  [353,] -0.1794698638
#>  [354,] -0.5933675715
#>  [355,]  0.5243922024
#>  [356,]  0.2004166414
#>  [357,]  0.2904739323
#>  [358,] -0.4817228183
#>  [359,] -0.1010077958
#>  [360,]  0.3605644153
#>  [361,]  0.3838692985
#>  [362,] -0.6909417859
#>  [363,] -0.3578756871
#>  [364,]  0.7571834531
#>  [365,]  0.4263435681
#>  [366,] -0.7292165902
#>  [367,]  1.3051354198
#>  [368,]  0.3406761981
#>  [369,]  0.1624659087
#>  [370,] -0.4134347323
#>  [371,]  0.8469977407
#>  [372,]  0.1445715350
#>  [373,] -1.0051362405
#>  [374,]  0.1813293979
#>  [375,]  0.3128037785
#>  [376,] -0.2938526668
#>  [377,] -1.1220992987
#>  [378,]  0.0153651685
#>  [379,]  0.6470436426
#>  [380,] -1.1136538960
#>  [381,] -0.6871886562
#>  [382,]  0.5673225676
#>  [383,] -0.2756679717
#>  [384,] -0.8618274857
#>  [385,] -0.1516861964
#>  [386,] -0.2098262882
#>  [387,] -1.2010622994
#>  [388,]  1.2558766881
#>  [389,] -0.5078247977
#>  [390,] -0.8336056154
#>  [391,]  0.8759462386
#>  [392,]  0.5860585689
#>  [393,] -1.2938325172
#>  [394,] -0.7940052589
#>  [395,] -0.5236478315
#>  [396,]  0.1776026459
#>  [397,] -1.4013771343
#>  [398,]  0.7131703354
#>  [399,]  0.0354513941
#>  [400,]  0.1344392169
#>  [401,]  0.4879866664
#>  [402,]  0.3638171292
#>  [403,] -0.1982469069
#>  [404,]  0.1427099257
#>  [405,]  0.0306156349
#>  [406,]  0.6818310155
#>  [407,]  0.0828461223
#>  [408,]  0.1285057261
#>  [409,]  0.7544668617
#>  [410,]  0.2249944604
#>  [411,]  0.0320371617
#>  [412,]  0.4893855472
#>  [413,]  0.3547013624
#>  [414,] -0.3653020142
#>  [415,]  0.8342834529
#>  [416,] -0.2503120423
#>  [417,]  0.5247353344
#>  [418,]  0.9033269903
#>  [419,]  0.6160371722
#>  [420,]  0.1392564258
#>  [421,]  0.7268415723
#>  [422,] -0.5110496849
#>  [423,] -0.6768887501
#>  [424,]  0.8344471002
#>  [425,]  0.3100289687
#>  [426,]  0.4325993985
#>  [427,]  0.1333037878
#>  [428,]  0.5182074205
#>  [429,] -0.3190090729
#>  [430,]  0.4315727731
#>  [431,]  0.3930367163
#>  [432,] -0.4950545403
#>  [433,] -0.7391438527
#>  [434,] -0.3018906083
#>  [435,] -0.2371938428
#>  [436,]  0.9648173805
#>  [437,] -0.5663750615
#>  [438,] -0.2494237461
#>  [439,] -0.8172050753
#>  [440,] -0.7060758808
#>  [441,] -0.5057710364
#>  [442,]  0.3789855200
#>  [443,] -0.9727858961
#>  [444,]  0.4803828289
#>  [445,]  0.6155096579
#>  [446,] -0.2193798156
#>  [447,]  0.2770512787
#>  [448,]  0.1446155292
#>  [449,]  0.7294828996
#>  [450,]  0.1318409037
#>  [451,] -0.9286130640
#>  [452,]  0.5650020528
#>  [453,] -0.0734986694
#>  [454,]  0.6055891661
#>  [455,]  0.5111038867
#>  [456,]  0.3522484553
#>  [457,] -0.7570123077
#>  [458,] -0.1832566529
#>  [459,] -0.7545482914
#>  [460,] -0.0289847891
#>  [461,]  0.7738263616
#>  [462,]  0.4916106414
#>  [463,] -0.3123352417
#>  [464,] -1.0281939101
#>  [465,]  0.2271059701
#>  [466,]  0.7107970846
#>  [467,] -0.3046782666
#>  [468,] -0.5839218929
#>  [469,]  1.4979405897
#>  [470,] -0.2798596271
#>  [471,]  0.0599008636
#>  [472,] -0.1166487011
#>  [473,]  0.5759354705
#>  [474,]  0.0739430541
#>  [475,]  0.5817937897
#>  [476,]  0.2870043563
#>  [477,] -0.1296602880
#>  [478,]  0.4669613857
#>  [479,] -0.0833653419
#>  [480,]  0.0444112344
#>  [481,] -0.2424533157
#>  [482,]  0.4828098326
#>  [483,]  0.3103995811
#>  [484,]  0.0108853685
#>  [485,]  0.7864597493
#>  [486,] -0.9776109757
#>  [487,] -0.3104852190
#>  [488,] -0.3305506836
#>  [489,] -0.2215565192
#>  [490,]  0.0721164318
#>  [491,]  0.2177919953
#>  [492,]  0.1394672788
#>  [493,]  0.2520151010
#>  [494,]  0.5967424555
#>  [495,]  0.3058456171
#>  [496,]  0.5917510440
#>  [497,]  0.3640752570
#>  [498,]  0.1549135530
#>  [499,] -0.4251447442
#>  [500,] -0.5954776967
#>  [501,]  0.1971947976
#>  [502,]  0.3262440537
#>  [503,]  0.2089656916
#>  [504,] -0.3834784387
#>  [505,]  0.4278057945
#>  [506,]  0.0551679071
#>  [507,] -0.0386641590
#>  [508,]  0.2733846180
#>  [509,] -0.5726199430
#>  [510,] -0.2647348435
#>  [511,] -0.1636731425
#>  [512,]  0.2262599970
#>  [513,] -0.0280575211
#>  [514,] -0.2338599267
#>  [515,]  0.6862043557
#>  [516,] -0.5119258140
#>  [517,] -0.0523622271
#>  [518,]  0.1044191346
#>  [519,] -0.1237421703
#>  [520,]  0.0757042560
#>  [521,] -0.5462476988
#>  [522,]  0.0475515224
#>  [523,]  1.1281024369
#>  [524,]  2.1539089500
#>  [525,] -0.3064852921
#>  [526,] -0.0496118139
#>  [527,] -0.8556205728
#>  [528,]  0.0941749894
#>  [529,] -0.2142782062
#>  [530,]  0.4863033008
#>  [531,] -0.6682599698
#>  [532,] -0.1670117661
#>  [533,]  0.5639834096
#>  [534,]  0.0179411966
#>  [535,] -0.3645877612
#>  [536,]  0.0408239665
#>  [537,]  0.9396993715
#>  [538,] -0.1184824885
#>  [539,]  0.4494434811
#>  [540,] -0.2161966446
#>  [541,] -0.0674644150
#>  [542,]  0.2276799998
#>  [543,]  0.1874793980
#>  [544,] -0.2695410576
#>  [545,] -0.1355879295
#>  [546,]  0.8939639729
#>  [547,] -0.3869045151
#>  [548,]  0.0633882753
#>  [549,]  0.4120550311
#>  [550,]  0.2301551764
#>  [551,]  0.9149697335
#>  [552,]  1.1397870966
#>  [553,] -0.4759152144
#>  [554,] -0.2316858624
#>  [555,]  0.5579830470
#>  [556,] -0.4722499795
#>  [557,] -0.1854982292
#>  [558,] -1.1758367263
#>  [559,]  0.4799620280
#>  [560,]  0.2889691993
#>  [561,] -1.2229720028
#>  [562,] -0.6987404448
#>  [563,] -0.7161666474
#>  [564,] -0.6979375453
#>  [565,] -0.3020552272
#>  [566,] -0.4900088603
#>  [567,] -0.8276186481
#>  [568,] -0.4675067376
#>  [569,]  0.1342977498
#>  [570,]  0.4109994207
#>  [571,]  0.6809848086
#>  [572,]  0.0412523455
#>  [573,] -0.1366388934
#>  [574,] -1.1207640744
#>  [575,]  0.3470157116
#>  [576,] -0.0923439571
#>  [577,] -0.3819436355
#>  [578,]  0.8333260513
#>  [579,] -0.8232455215
#>  [580,]  0.4842011563
#>  [581,]  0.3710673552
#>  [582,] -0.1472190217
#>  [583,] -0.6324095948
#>  [584,] -0.0760947748
#>  [585,]  0.3433343851
#>  [586,]  0.2907508525
#>  [587,]  0.1804708150
#>  [588,]  0.0960759319
#>  [589,]  0.6891966666
#>  [590,] -0.5049703707
#>  [591,]  1.2849170224
#>  [592,] -0.3799388947
#>  [593,]  0.7477904615
#>  [594,] -0.1956255577
#>  [595,] -0.4801452438
#>  [596,] -1.1238985849
#>  [597,] -0.1742780742
#>  [598,]  0.7379142724
#>  [599,] -0.3434283752
#>  [600,] -0.1916724538
#>  [601,]  0.5454042452
#>  [602,]  0.5152191528
#>  [603,] -0.0849481841
#>  [604,] -0.9375948660
#>  [605,]  0.0342874958
#>  [606,] -0.6655122644
#>  [607,]  0.1873396431
#>  [608,] -0.0769470189
#>  [609,]  0.0459691971
#>  [610,] -1.1788987151
#>  [611,]  0.3286496384
#>  [612,] -0.0070400147
#>  [613,]  0.0901132142
#>  [614,] -0.6918679066
#>  [615,] -0.4205908632
#>  [616,]  0.4543385287
#>  [617,] -1.0384827514
#>  [618,]  0.1549914894
#>  [619,]  0.0778184015
#>  [620,]  0.7126112773
#>  [621,] -0.4782953289
#>  [622,]  0.0135667163
#>  [623,]  0.4106910313
#>  [624,] -1.0563538650
#>  [625,] -0.9111001349
#>  [626,]  0.4652710955
#>  [627,]  0.0051122323
#>  [628,] -0.7323849608
#>  [629,]  0.1945705798
#>  [630,]  0.1503701506
#>  [631,] -0.4189429445
#>  [632,] -0.1985285669
#>  [633,]  0.3485532230
#>  [634,] -0.2546378116
#>  [635,]  0.1227124060
#>  [636,] -0.1856168231
#>  [637,]  0.2738391861
#>  [638,]  1.0011847180
#>  [639,]  0.4348025436
#>  [640,] -1.6300050791
#>  [641,]  0.2096920008
#>  [642,]  0.0085548659
#>  [643,]  0.8349255332
#>  [644,] -0.3045720720
#>  [645,]  0.6219870464
#>  [646,] -0.2574980996
#>  [647,]  0.0345360903
#>  [648,]  0.0117662361
#>  [649,] -1.4764816664
#>  [650,] -0.1628281251
#>  [651,] -1.2739725109
#>  [652,]  0.8571052232
#>  [653,]  0.1669338597
#>  [654,] -0.5317998231
#>  [655,] -1.0839476904
#>  [656,] -0.3629806704
#>  [657,]  0.2942477118
#>  [658,] -0.7617741397
#>  [659,] -0.1690582890
#>  [660,]  0.4040059630
#>  [661,] -0.3634682308
#>  [662,] -0.8665867243
#>  [663,] -0.0941847556
#>  [664,]  0.5124857752
#>  [665,]  0.3797413950
#>  [666,]  0.4798221425
#>  [667,] -0.5582048932
#>  [668,]  0.3613024842
#>  [669,]  0.5623123407
#>  [670,]  0.1053987295
#>  [671,]  0.8528664883
#>  [672,]  0.3209402321
#>  [673,] -0.2066760528
#>  [674,] -0.4629292996
#>  [675,]  0.0776331320
#>  [676,]  0.0504136953
#>  [677,]  0.0518272345
#>  [678,]  0.2186431744
#>  [679,]  0.9749752646
#>  [680,] -1.1149770130
#>  [681,]  0.3516448779
#>  [682,]  0.3667567430
#>  [683,]  0.8789464626
#>  [684,] -0.2670430477
#>  [685,] -0.1829143260
#>  [686,]  0.1890789497
#>  [687,]  1.0046909086
#>  [688,]  0.3745076220
#>  [689,] -0.0766475368
#>  [690,]  0.6810745650
#>  [691,]  0.4760055484
#>  [692,]  0.6531640580
#>  [693,]  1.4329352344
#>  [694,] -0.7168700347
#>  [695,] -0.9394761044
#>  [696,]  0.2211813993
#>  [697,] -0.6708515950
#>  [698,] -0.6266894849
#>  [699,] -0.2085854240
#>  [700,] -0.5009921796
#>  [701,] -0.0736230812
#>  [702,]  0.0828181296
#>  [703,]  0.3308020518
#>  [704,] -0.7440295308
#>  [705,] -0.3728460371
#>  [706,]  0.2977528453
#>  [707,]  0.7548409358
#>  [708,]  0.3886687772
#>  [709,]  0.4132967165
#>  [710,]  0.1895303077
#>  [711,]  0.3994095376
#>  [712,]  0.9802583573
#>  [713,]  0.7551683323
#>  [714,] -0.2736356264
#>  [715,]  0.1938560264
#>  [716,]  0.2522142274
#>  [717,]  0.4456445327
#>  [718,] -0.1725846586
#>  [719,]  1.0797327960
#>  [720,]  0.3771526153
#>  [721,]  1.0647433772
#>  [722,]  0.5670939083
#>  [723,] -0.3559597368
#>  [724,] -0.7792898224
#>  [725,]  0.3513300439
#>  [726,] -0.2204943660
#>  [727,] -0.7116200554
#>  [728,] -0.1274145254
#>  [729,]  0.1966053118
#>  [730,] -0.3877990615
#>  [731,] -0.0199316623
#>  [732,]  0.1184647714
#>  [733,]  0.2743340648
#>  [734,]  0.5775245168
#>  [735,] -0.5056918778
#>  [736,] -0.2552296520
#>  [737,] -0.6560361089
#>  [738,]  0.0348108554
#>  [739,]  0.0902894831
#>  [740,] -0.6225193926
#>  [741,]  0.0797149625
#>  [742,] -0.1238041063
#>  [743,] -0.2795944115
#>  [744,]  0.4582740839
#>  [745,] -0.2354778103
#>  [746,]  0.2209597599
#>  [747,] -0.5618322051
#>  [748,]  0.2231077549
#>  [749,]  0.4956505180
#>  [750,] -0.5832124874
#>  [751,]  0.2152663063
#>  [752,]  0.0542765705
#>  [753,] -0.2872030398
#>  [754,]  0.8427003840
#>  [755,]  0.7672822571
#>  [756,]  0.2816524067
#>  [757,] -0.2592054058
#>  [758,] -0.3337061531
#>  [759,]  0.3641843927
#>  [760,]  0.9735534858
#>  [761,] -0.3217715777
#>  [762,] -0.7869853510
#>  [763,] -0.4184688102
#>  [764,] -0.4806400169
#>  [765,] -0.8255718516
#>  [766,] -0.5303009645
#>  [767,]  1.3100789131
#>  [768,] -0.9604239522
#>  [769,] -0.0885800960
#>  [770,]  0.2058098768
#>  [771,] -0.7842676832
#>  [772,]  0.0635070675
#>  [773,] -0.6951055731
#>  [774,] -0.1846506440
#>  [775,] -0.3364020118
#>  [776,] -0.1788615765
#>  [777,]  0.8910682718
#>  [778,] -0.1378251475
#>  [779,] -0.5356144004
#>  [780,]  0.3572921198
#>  [781,]  0.1585389328
#>  [782,]  1.0687573128
#>  [783,]  1.1066432677
#>  [784,] -0.2037566827
#>  [785,]  0.9385980462
#>  [786,]  0.6147342806
#>  [787,] -0.6015520440
#>  [788,]  0.1344300897
#>  [789,] -1.2290185526
#>  [790,] -0.2334262359
#>  [791,]  0.6008786419
#>  [792,] -0.3406629922
#>  [793,]  0.0389005027
#>  [794,]  0.6983189943
#>  [795,]  0.3481428958
#>  [796,] -1.3029336624
#>  [797,] -0.2717512081
#>  [798,] -0.2855208050
#>  [799,] -0.0936096953
#>  [800,] -0.3429607269
#>  [801,]  0.0655799516
#>  [802,]  0.4265553240
#>  [803,] -0.1910354634
#>  [804,] -0.9140365385
#>  [805,]  0.3288517734
#>  [806,] -0.8873996926
#>  [807,]  1.0038504543
#>  [808,]  0.9868570307
#>  [809,] -0.2429694691
#>  [810,] -0.9556442907
#>  [811,]  0.4624245200
#>  [812,]  0.3846153743
#>  [813,]  0.2159116993
#>  [814,]  0.6700864072
#>  [815,] -0.4726018671
#>  [816,] -0.4929653767
#>  [817,]  0.0322210453
#>  [818,]  0.3226476893
#>  [819,] -0.6958711981
#>  [820,]  0.9750728462
#>  [821,] -0.2351565078
#>  [822,]  0.0729409505
#>  [823,] -0.1591467472
#>  [824,] -0.5117435269
#>  [825,] -0.4795227511
#>  [826,]  0.4052784243
#>  [827,]  0.4204429769
#>  [828,]  0.2745394668
#>  [829,] -0.0205303118
#>  [830,] -0.4314195352
#>  [831,]  0.6095072622
#>  [832,]  1.0723185263
#>  [833,] -0.2646044216
#>  [834,] -0.3071233167
#>  [835,] -1.0247807559
#>  [836,] -0.0066204918
#>  [837,] -0.3715145847
#>  [838,] -0.9059915186
#>  [839,]  0.3446669957
#>  [840,] -0.4288263609
#>  [841,]  0.5646944696
#>  [842,] -0.3737786325
#>  [843,]  0.0442057475
#>  [844,] -0.3987185617
#>  [845,] -0.7878972746
#>  [846,]  0.2526626105
#>  [847,] -0.7940436406
#>  [848,]  0.9474149951
#>  [849,] -0.4694352785
#>  [850,]  0.3829034298
#>  [851,]  0.6224261860
#>  [852,] -0.5851169554
#>  [853,]  0.0889088332
#>  [854,] -0.7938975780
#>  [855,]  0.3155416672
#>  [856,] -0.1221265788
#>  [857,] -1.3300548992
#>  [858,] -0.5719364882
#>  [859,] -0.9125857492
#>  [860,] -0.2369407266
#>  [861,] -0.1245172730
#>  [862,]  0.1233036288
#>  [863,] -0.3210583528
#>  [864,]  0.2492075228
#>  [865,] -0.2825158933
#>  [866,] -0.5327058882
#>  [867,]  0.8425517035
#>  [868,] -0.1497259153
#>  [869,] -0.0707899753
#>  [870,]  1.4702681642
#>  [871,]  0.3788538026
#>  [872,]  0.1531724426
#>  [873,] -0.0760096699
#>  [874,]  0.4299993469
#>  [875,]  0.2210187820
#>  [876,]  0.2148708876
#>  [877,]  0.1065695244
#>  [878,]  0.3748531349
#>  [879,]  1.3184073115
#>  [880,]  0.1440823004
#>  [881,] -0.2546922153
#>  [882,]  0.3841652278
#>  [883,] -0.4562342366
#>  [884,]  0.0205493219
#>  [885,] -0.5283206631
#>  [886,] -0.0380588914
#>  [887,]  0.8189316276
#>  [888,] -0.3172821376
#>  [889,] -0.0775267128
#>  [890,]  0.6523400957
#>  [891,] -0.5292097607
#>  [892,] -0.2480769909
#>  [893,] -0.5409444786
#>  [894,]  1.6439086647
#>  [895,] -0.0285487751
#>  [896,]  0.5037646911
#>  [897,]  0.1677123025
#>  [898,]  0.4745209905
#>  [899,] -0.3244949915
#>  [900,]  0.2792996126
#>  [901,] -0.5484100876
#>  [902,] -0.0305754905
#>  [903,] -0.1514649436
#>  [904,]  0.2612468742
#>  [905,]  0.2779371717
#>  [906,] -0.5204706826
#>  [907,]  0.2379749746
#>  [908,] -0.0438162304
#>  [909,] -0.1526783351
#>  [910,] -0.0887352467
#>  [911,] -0.1575918784
#>  [912,]  0.1630604585
#>  [913,]  0.7016946446
#>  [914,] -0.8003516511
#>  [915,] -0.5517600558
#>  [916,]  0.5581670457
#>  [917,]  0.9733683913
#>  [918,]  0.4126730853
#>  [919,]  0.2656304829
#>  [920,]  0.9628328353
#>  [921,] -0.9312316646
#>  [922,] -0.4634586015
#>  [923,] -1.3693199034
#>  [924,]  0.2544530232
#>  [925,]  0.4049772376
#>  [926,] -0.8171518112
#>  [927,]  0.9692900065
#>  [928,]  0.2155699646
#>  [929,] -0.0075315823
#>  [930,] -0.3707555828
#>  [931,]  0.3387206307
#>  [932,]  0.6080841703
#>  [933,]  0.3844807815
#>  [934,] -0.4457093847
#>  [935,]  0.4762933479
#>  [936,]  0.8586817570
#>  [937,] -0.8078277030
#>  [938,]  1.1776395219
#>  [939,]  0.2418611847
#>  [940,]  0.3335469611
#>  [941,]  0.1285094033
#>  [942,]  1.3411924674
#>  [943,] -0.5559121989
#>  [944,]  0.1590980075
#>  [945,] -0.6951377698
#>  [946,] -0.2605448495
#>  [947,] -0.5783587847
#>  [948,]  0.3986757797
#>  [949,] -0.1698486860
#>  [950,]  0.3859007651
#>  [951,]  0.2774557790
#>  [952,]  0.9474644343
#>  [953,]  0.2077114061
#>  [954,]  0.2806239981
#>  [955,] -0.1963328024
#>  [956,] -0.8031186123
#>  [957,] -0.2807564391
#>  [958,] -0.3026833064
#>  [959,]  0.5083380591
#>  [960,] -0.2461755267
#>  [961,] -0.4289317688
#>  [962,]  0.7402412460
#>  [963,]  0.4969588100
#>  [964,] -0.2984665628
#>  [965,]  1.1478428347
#>  [966,]  0.4754793196
#>  [967,]  0.3991800863
#>  [968,] -0.9194506870
#>  [969,]  0.1032765719
#>  [970,]  0.8412981682
#>  [971,]  0.0857694745
#>  [972,] -0.0358832540
#>  [973,] -0.6591553268
#>  [974,] -0.0544410040
#>  [975,]  0.0107344659
#>  [976,] -1.0040422513
#>  [977,]  0.2431234689
#>  [978,] -0.2085816925
#>  [979,] -0.5573606207
#>  [980,] -0.0040655126
#>  [981,] -0.1772558013
#>  [982,] -0.1807489925
#>  [983,] -0.7999637321
#>  [984,]  0.6382119649
#>  [985,] -0.7711653996
#>  [986,] -0.5226131846
#>  [987,]  0.0978385912
#>  [988,] -0.2031616225
#>  [989,] -0.0101070937
#>  [990,] -0.1496920926
#>  [991,] -0.2085600542
#>  [992,]  0.6977999929
#>  [993,]  0.2173160273
#>  [994,]  0.4405235825
#>  [995,] -0.0108877220
#>  [996,]  0.1153986710
#>  [997,] -0.3794053535
#>  [998,] -0.6080007944
#>  [999,]  0.6614830159
#> [1000,] -0.0739664324