Sum of Squares

The sum of the squared entries in a vector or matrix.

sum_squares(expr)

Arguments

expr
An Expression, vector, or matrix.

Value

An Expression representing the sum of squares of the input.

Examples

m <- 30
n <- 20
A <- matrix(stats::rnorm(m*n), nrow = m, ncol = n)
b <- matrix(stats::rnorm(m), nrow = m, ncol = 1)

x <- Variable(n)
obj <- Minimize(sum_squares(A %*% x - b))
constr <- list(0 <= x, x <= 1)
prob <- Problem(obj, constr)
result <- solve(prob)

result$value

#> [1] 9.813441

result$getValue(x)

#>                [,1]
#>  [1,]  3.075110e-01
#>  [2,]  4.746013e-01
#>  [3,]  1.719635e-01
#>  [4,]  3.546111e-08
#>  [5,]  4.013943e-02
#>  [6,]  8.963244e-10
#>  [7,]  1.984734e-09
#>  [8,]  8.050735e-02
#>  [9,]  2.689793e-01
#> [10,] -1.225041e-10
#> [11,]  2.921991e-09
#> [12,]  2.441208e-01
#> [13,]  2.974094e-02
#> [14,]  2.202891e-01
#> [15,]  1.954291e-01
#> [16,]  2.762106e-02
#> [17,]  1.751774e-01
#> [18,]  8.997645e-02
#> [19,]  1.760342e-01
#> [20,]  7.563559e-02

result$getDualValue(constr[[1]])

#>               [,1]
#>  [1,] 7.645705e-08
#>  [2,] 5.053300e-08
#>  [3,] 1.294602e-07
#>  [4,] 5.294848e-01
#>  [5,] 5.135023e-07
#>  [6,] 7.499881e+00
#>  [7,] 5.050196e+00
#>  [8,] 2.703664e-07
#>  [9,] 8.766303e-08
#> [10,] 9.733416e+00
#> [11,] 4.072360e+00
#> [12,] 9.280593e-08
#> [13,] 5.927891e-07
#> [14,] 1.021562e-07
#> [15,] 1.119543e-07
#> [16,] 6.391447e-07
#> [17,] 1.257439e-07
#> [18,] 2.092180e-07
#> [19,] 1.258850e-07
#> [20,] 2.885748e-07