Total Variation

The total variation of a vector, matrix, or list of matrices. Uses L1 norm of discrete gradients for vectors and L2 norm of discrete gradients for matrices.

tv(value, ...)

Arguments

value
An Expression, vector, or matrix.

...
(Optional) Expression objects or numeric constants that extend the third dimension of value.

Value

An Expression representing the total variation of the input.

Examples

rows <- 10
cols <- 10
Uorig <- matrix(sample(0:255, size = rows * cols, replace = TRUE), nrow = rows, ncol = cols)

# Known is 1 if the pixel is known, 0 if the pixel was corrupted
Known <- matrix(0, nrow = rows, ncol = cols)
for(i in 1:rows) {
   for(j in 1:cols) {
      if(stats::runif(1) > 0.7)
         Known[i,j] <- 1
   }
}
Ucorr <- Known %*% Uorig

# Recover the original image using total variation in-painting
U <- Variable(rows, cols)
obj <- Minimize(tv(U))
constraints <- list(Known * U == Known * Ucorr)
prob <- Problem(obj, constraints)
result <- solve(prob, solver = "SCS")
result$getValue(U)

#>           [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#>  [1,] 349.0000 349.0001 338.2889 307.6827 215.9995 244.6305 201.9996 309.9995
#>  [2,] 348.9998 358.4576 345.3494 310.0004 310.0004 310.0005 310.0003 310.0002
#>  [3,] 415.1481 430.0000 383.0000 310.0006 310.0007 310.0007 310.0004 328.2150
#>  [4,] 414.9997 520.0001 487.9997 289.9996 310.0007 310.0007 326.1265 388.8258
#>  [5,] 420.0440 420.0439 186.0003 304.0629 310.0006 324.4094 378.5166 388.8258
#>  [6,] 420.0440 403.9995 394.4084 387.9996 296.9999 369.7415 391.0167 406.9997
#>  [7,] 440.2605 427.4467 408.3662 360.5855 396.6152 446.0001 780.0002 407.0001
#>  [8,] 504.0005 504.0004 566.0004 265.9997 405.6461 411.7225 405.0002 428.4911
#>  [9,] 504.0004 504.0003 487.0001 404.9999 404.9999 477.0001 429.0475 429.0476
#> [10,] 504.0004 504.0003 487.0001 404.9999 238.0000 165.9998 429.0475 429.0476
#>           [,9]    [,10]
#>  [1,] 309.9991 277.0004
#>  [2,] 335.0775 450.9998
#>  [3,] 453.0001 453.0001
#>  [4,] 388.8258 388.8258
#>  [5,] 388.8261 365.9997
#>  [6,] 406.9996 406.9996
#>  [7,] 471.9997 471.9997
#>  [8,] 472.0000 532.9995
#>  [9,] 429.0478 429.0478
#> [10,] 429.0478 353.9999